### 2020-05-30 #2275 @Kevin-Mattheus-Moerman
@cybanical I hope you are doing well! I think this JOSS submission looks like your cup of tea. Would you be able to help review this? (perhaps you can also update your GitHub profile with your full name and a professional link e.g. https://www.media.mit.edu/people/mcarney/overview/)
Here is the short paper:
https://github.com/openjournals/joss-papers/blob/joss.02275/joss.02275/10.21105.joss.02275.pdf
and the software:
https://github.com/upb-lea/gym-electric-motor
