### 2020-05-31 #2275 @cybanical
Oh man, I wish I could. I'm really overwhelmed with work right now.  Trying
to get an open-hardware respirator shipped right now.  osrdesign.org.
<http://osrdesign.org>

*_____________________________________*
*MATTHEW CARNEY, PhD | Research Affiliate **| *Mechatronics Design Engineer
MIT Media Lab | Biomechatronics <http://biomech.media.mit.edu>
75 Amherst St. | E14-274G | Cambridge, MA 02142 USA
matthematic.com | @matt.thematic <https://www.instagram.com/matt.thematic/>
| LinkedIn <https://www.linkedin.com/in/matthematic/>

On Sat, May 30, 2020 at 4:24 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @cybanical <https://github.com/cybanical> I hope you are doing well! I
> think this JOSS submission looks like your cup of tea. Would you be able to
> help review this? (perhaps you can also update your GitHub profile with
> your full name and a professional link e.g.
> https://www.media.mit.edu/people/mcarney/overview/)
> Here is the short paper:
>
> https://github.com/openjournals/joss-papers/blob/joss.02275/joss.02275/10.21105.joss.02275.pdf
> and the software:
> https://github.com/upb-lea/gym-electric-motor
>
