+++ examples/agents/simple_controllers.py
@@ -10,8 +10,19 @@ import numpy as np
 
 class Controller:
 
+    """
+      The following is a base class for various controllers along with the motor environments
+
+    """
+
     @classmethod
     def make(cls, controller_type, environment, **controller_kwargs):
+        """
+        Args:
+            controller_type : Choose among the given set of controller at the end of this class
+            environment : Choose the corresponding motor environment from 'envs' class.
+
+        """
         assert controller_type in _controllers.keys(), f'Controller {controller_type} unknown'
         controller = _controllers[controller_type](environment, **controller_kwargs)
         return controller
@@ -24,6 +35,11 @@ class Controller:
 
 
 class OnOffController(Controller):
+
+    """
+        The following controller is a simple OnOffController which allows the state to choose high_action when referenced_state is below
+        ref_idx and low_action otherwise
+    """
 
     def __init__(self, environment, state_idx=None, reference_idx=0):
         action_space = environment.action_space
@@ -44,6 +60,13 @@ class OnOffController(Controller):
 
 
 class ThreePointController(Controller):
+
+    """
+    Below is an implementation of a 3 point controller: When state_idx is below the reference_idx, it choose high_action and when the
+    state_idx is above the reference_idx,it choose low_action. If it is between the reference_idx values, it will choose idle_action
+    A 'Hysteresis' value should be integrated(chosen as 0.01) in this controller because of the constant switching and high frequency around the reference_idx
+
+    """
 
     def __init__(self, environment, hysteresis=0.01, state_idx=None, reference_idx=0):
         action_space = environment.action_space
@@ -69,6 +92,13 @@ class ThreePointController(Controller):
 
 class PController(Controller):
 
+    """
+    In the below proportional control implementation, the controller output is proportional to the error signal,  which is the difference
+    between the reference_idx and the state_idx .i.e., the output of a proportional controller
+    is the multiplication product of the error signal and the proportional gain
+    Here kp =10 is assumed for proportionality gain.
+    """
+
     def __init__(self, environment, k_p=10, controller_no=0, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Box, 'No suitable action space for P Controller'
@@ -92,6 +122,11 @@ class PController(Controller):
 
 
 class PIController(PController):
+    """
+      The below Discrete PI Controller performs discrete-time PI controller computation using the error signal and proportional
+      and integral gain inputs. The error signal is the difference between the reference_idx and the referenced_state.
+      It outputs a weighted sum of the input error signal and the integral of the input error signal.
+    """
 
     def __init__(self, environment, k_p=10, k_i=0.01, controller_no=0, reference_idx=0):
         super().__init__(environment, k_p, controller_no, reference_idx)
@@ -118,6 +153,10 @@ class PIController(PController):
 
 
 class PmsmOnOffController(Controller):
+    """
+    The following controller is a simple OnOffController which allows the state to choose u_q values as '1'' when it is below
+        reference_state and '-1' otherwise
+    """
 
     def __init__(self, environment, state_idx=None, ref_idx=0):
         t32 = environment.physical_system.electrical_motor.t_32
@@ -150,6 +189,12 @@ class PmsmOnOffController(Controller):
 
 class SynRmOnOffController(PmsmOnOffController):
 
+    """
+     The following controller is a PmsmOnOffController which allows the state to choose u_d and u_q voltages values as '1'  when it is below
+        reference_state and '-1' otherwise
+
+    """
+
     def control(self, state, reference):
         if state[self._referenced_state] < reference[self._ref_idx]:
             u_q = 1
@@ -163,6 +208,20 @@ class SynRmOnOffController(PmsmOnOffController):
 
 
 class CascadedPIController(Controller):
+
+    """
[…]
