### 2020-06-01 09cb5dc
added a parameter to indicate update frequency. added commentary. fixed some minor visual errors
