### 2020-06-01 5a2f012
First draft of an RCVoltageSupply. Open for discussion
