### 2020-06-01 d5bf654
first changes in reset function
