### 2020-06-01 98f2e6f
added const/default/uniform init vlaue
