### 2020-06-02 fc10785
added const/default/uniform init vlaue update
