### 2020-06-02 e7ecade
added gaussian
