### 2020-06-02 4e23f1a
added userdefined init interval
