+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -2,6 +2,7 @@ import numpy as np
 import math
 from scipy.stats import truncnorm
 
+
 class ElectricMotor:
     """
         Base class for all technical electrical motor models.
@@ -138,29 +139,36 @@ class ElectricMotor:
         Returns:
             numpy.ndarray(float): The initial motors state.
         """
+        # todo better assertion messages
         assert isinstance(self._initializer, dict), \
-            'initializer have to be dict, but is %r' %type(self._initializer)
-        #using explicit dict keys?
-        initial_value = np.asarray(self._initializer['init_value'], dtype=float)
+            'initializer have to be dict, but is %r' % type(self._initializer)
+        # using explicit dict keys?
+        initial_value = np.atleast_1d(self._initializer['init_value'])
 
         if self._initializer['init_type'] == 'constant':
             # check if value is general constant or a constant for each current
             assert (initial_value.shape[0] == 1 or
                     initial_value.shape[0] == len(self.CURRENTS)), \
-                    'initial_value does not match right dimension'
+                'initial_value does not match right dimension'
+            #initial_value = initial_value or 0.0
             return np.ones(len(self.CURRENTS), dtype=float) * initial_value
 
         elif self._initializer['init_type'] in ['uniform', 'gaussian']:
             # todo catch scalar case with assert
-            # todo user definded interval
-            assert initial_value.shape[0] == 2, \
-                   'two limits needed as initial_value'
-            upper = np.asarray(
-                [self._nominal_values[i_x] for i_x in self.CURRENTS]
-                )
-            lower = np.zeros(len(self.CURRENTS), dtype=float)
-            # change lower respective to state space (could be negativ)
-            #lower = upper * self.
+            nominal_currents = np.asarray([self._nominal_values[ix]
+                                           for ix in self.CURRENTS])
+            upper = initial_value.T[-1] if initial_value.shape[0] == 2 \
+                else nominal_currents
+            lower = initial_value.T[0] if initial_value.shape[0] == 2 \
+                else np.zeros(len(self.CURRENTS), dtype=float)
+            # clip limits to nominal values, if to high
+            upper = np.clip(upper, a_min=None, a_max=nominal_currents)
+            # todo change lower nominal values could be negativ
+            lower = np.clip(lower,
+                            a_min=np.zeros(len(self.CURRENTS), dtype=float),
+                            a_max=None)
+            # todo change lower respective to state space (could be negativ)
+            # lower = upper * state space
             if self._initializer['init_type'] == 'uniform':
                 initial_value = (upper - lower) * \
                                 np.random.random_sample(len(self.CURRENTS)) + \
@@ -168,7 +176,7 @@ class ElectricMotor:
                 return np.ones(len(self.CURRENTS), dtype=float) * initial_value
 
[…]
