### 2020-06-02 7ebf0af
added userdefined init interval
