### 2020-06-02 0e5a42a
updated docstrings
