### 2020-06-02 fac6059
further experimentation on visualization
