### 2020-06-03 e941016
Revert "further experimentation on visualization"
This reverts commit fac60590ec8d545eff802eb3a6873d8fd7898c7d.
