### 2020-06-03 bf8b300
Revert "initial commit, plot reward in visualization"
This reverts commit ff3019058b0dcec85ab9040e066258171f690e3c.
