### 2020-06-03 915c78f
visulaization init with reward added to plotted variables
