+++ gym_electric_motor/visualization/motor_dashboard.py
@@ -6,7 +6,6 @@ import warnings
 
 
 class MotorDashboard(ElectricMotorVisualization):
-
     """
     Visualization of the variables of the motor in graphs. This can be the angular velocity omega, torque
     phase voltages, phase currents, supply voltage and the rotor angle at the PMSM.
@@ -139,7 +138,9 @@ class MotorDashboard(ElectricMotorVisualization):
             self.initialized = True
             self._update_physical_system_data()
             self._order_plotted_variables()
+
             self._set_up_plots()
+
             for var in self.dash_vars:
                 var.reset()
             plt.pause(0.05)
@@ -175,9 +176,10 @@ class MotorDashboard(ElectricMotorVisualization):
         """
         update the limits, nominal values and sampling time
         """
-        self._limits = self._physical_system.limits
-        self._nominal_state = self._physical_system.nominal_state
+        self._limits = self._physical_system.limits  # can add reward limit as 1
+        self._nominal_state = self._physical_system.nominal_state  # reward nominal state=??
         self._tau = self._physical_system.tau
+
         self._update_cycle = int(self._update_period / self._tau)
         try:
             self._labels = self._physical_system.labels
@@ -200,7 +202,8 @@ class MotorDashboard(ElectricMotorVisualization):
                 'u_sd': '$u_{sd}/V$',
                 'u_e': '$u_{e}/V$',
                 'u_sup': '$u_{sup}/V$',
-                'epsilon': r'$\epsilon/rad$'
+                'epsilon': r'$\epsilon/rad$',
+                'reward': r'$\r/units$'
             }
         self._state_space = self._physical_system.state_space
 
@@ -219,11 +222,12 @@ class MotorDashboard(ElectricMotorVisualization):
             fig_manager.window.showMaximized()
         else:
             self._update_figure = self._figure.canvas.draw
-        low = self._limits * self._state_space.low
-        high = self._limits * self._state_space.high
-
-        low_nominal = self._nominal_state * self._state_space.low
-        high_nominal = self._nominal_state * self._state_space.high
+        # state_low = self._limits * self._state_space.low
+        # state_high = self._limits * self._state_space.high
+        #
+        # state_low_nominal = self._nominal_state * self._state_space.low
+        # state_high_nominal = self._nominal_state * self._state_space.high
+        low, high, low_nominal, high_nominal, ref = self._get_variable_ranges()
 
         if len(self._plotted_state_index) > 1:
             self.dash_vars = np.array(
@@ -242,14 +246,13 @@ class MotorDashboard(ElectricMotorVisualization):
                         self._visu_period,
                         with_ref
                     )
-                    for var, ax, lo, hi, lo_nominal, hi_nominal, with_ref in zip(
-                        self._plotted_variables, axes,
-                        low[self._plotted_state_index],
-                        high[self._plotted_state_index],
-                        low_nominal[self._plotted_state_index],
-                        high_nominal[self._plotted_state_index],
-                        self._referenced_states[self._plotted_state_index]
-                    )
+                    for var, ax, lo, hi, lo_nominal, hi_nominal, with_ref in zip(self._plotted_variables, axes,
+                                                                                 low,
+                                                                                 high,
+                                                                                 low_nominal,
+                                                                                 high_nominal,
+                                                                                 ref
+                                                                                 )
                 ]
             )
         elif len(self._plotted_state_index) == 1:
@@ -285,26 +288,68 @@ class MotorDashboard(ElectricMotorVisualization):
         """
         self._plotted_state_index = []
[…]
