### 2020-06-03 466d43a
working visualization of instant rewards, subplot to be refined
