+++ gym_electric_motor/visualization/motor_dashboard.py
@@ -145,17 +145,20 @@ class MotorDashboard(ElectricMotorVisualization):
                 var.reset()
             plt.pause(0.05)
         k = self._k
-        state_denorm = state * self._limits
+        # state_denorm = state * self._limits
+
+        plot_vals_denorm, reference_denorm = self._calc_plotable_vals(state, reference, reward)
+
         if reference is not None:
-            reference_denorm = reference * self._limits
+            # reference_denorm = reference * self._limits
             for var, data, ref in zip(
                     self.dash_vars,
-                    state_denorm[self._plotted_state_index],
-                    reference_denorm[self._plotted_state_index]
+                    plot_vals_denorm,
+                    reference_denorm
             ):
                 var.step(data, k, ref)
         else:
-            for var, data in zip(self.dash_vars, state_denorm[self._plotted_state_index]):
+            for var, data in zip(self.dash_vars, plot_vals_denorm):
                 var.step(data, k)
 
         if (k + 1) % self._update_cycle == 0 or k == self._episode_length - 1 or done:
@@ -203,7 +206,7 @@ class MotorDashboard(ElectricMotorVisualization):
                 'u_e': '$u_{e}/V$',
                 'u_sup': '$u_{sup}/V$',
                 'epsilon': r'$\epsilon/rad$',
-                'reward': r'$\r/units$'
+                'reward': 'Reward'
             }
         self._state_space = self._physical_system.state_space
 
@@ -289,7 +292,7 @@ class MotorDashboard(ElectricMotorVisualization):
         self._plotted_state_index = []
         if self._plotted_variables == 'all':
             self._plotted_state_variables = list(self._physical_system.state_names)
-            self._plotted_state_index = list(range(len(self._plotted_variables)))
+            self._plotted_state_index = list(range(len(self._plotted_state_variables)))
         elif self._plotted_variables == 'none':
             self._plotted_state_variables = []
             warnings.warn("No valid variables for visualization", Warning, stacklevel=2)
@@ -322,27 +325,39 @@ class MotorDashboard(ElectricMotorVisualization):
     def _get_variable_ranges(self):
         rew_lo = -2000
         rew_hi = 0
-        rew_nom_lo = -1000
-        rew_nom_hi = -500
+        rew_nom_lo = 0
+        rew_nom_hi = 0
         state_low = self._limits * self._state_space.low
+        # state_low = list(state_low)
         state_high = self._limits * self._state_space.high
         state_low_nominal = self._nominal_state * self._state_space.low
         state_high_nominal = self._nominal_state * self._state_space.high
 
-        low = state_low[self._plotted_state_index],
-        high = state_high[self._plotted_state_index],
-        low_nominal = state_low_nominal[self._plotted_state_index],
-        high_nominal = state_high_nominal[self._plotted_state_index],
+        low = state_low[self._plotted_state_index]
+        high = state_high[self._plotted_state_index]
+        low_nominal = state_low_nominal[self._plotted_state_index]
+        high_nominal = state_high_nominal[self._plotted_state_index]
         is_referenced = self._referenced_states[self._plotted_state_index]
 
         if 'reward' in self._plotted_variables:
-            np.append(low, rew_lo)
-            np.append(high, rew_hi)
-            np.append(low_nominal, rew_nom_lo)
-            np.append(high_nominal, rew_nom_hi)
-            np.append(is_referenced, False)
+            low = np.append(low, rew_lo)
+            high = np.append(high, rew_hi)
+            low_nominal = np.append(low_nominal, rew_nom_lo)
+            high_nominal = np.append(high_nominal, rew_nom_hi)
[…]
