### 2020-06-04 474b0f4
cleaned reset fct
