### 2020-06-04 7219040
new reset for elec_motor
