### 2020-06-05 #2275 @moorepants
I just noticed this: https://arxiv.org/abs/1910.09434 It isn't clear if this JOSS paper would be a duplicate publication.
