### 2020-06-05 #2275 @Kevin-Mattheus-Moerman
Thanks @moorepants  for pointing this out:
> I just noticed this: https://arxiv.org/abs/1910.09434 It isn't clear if this JOSS paper would be a duplicate publication.
It does happen, and is allowed (and sometime even encouraged), for authors to publish a long traditional paper in a different journal which describes the background, theory, and typically some research findings/conclusions obtained with the software in full. Often authors publish the JOSS paper first and refer to the JOSS paper in this more traditional long article. The review of the traditional long paper should not focus on the software though so it is not a duplicate in that sense.

Perhaps @wkirgsn can clarify more on this arxiv preprint. What is the focus of that preprint and how does it differ from this JOSS submission? Has another paper on this been published anywhere? Can you please cite any such papers in the JOSS paper as well?
