### 2020-06-05 a962829
synch motor reset updated
