+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -35,10 +35,7 @@ class ElectricMotor:
     #: _default_limits(dict(float)): Default motor limits (0 for unbounded limits)
     _default_limits = {}
     #: _default_initial_state(dict): Default initial motor-state values
-    _default_initializer = {'init_value': 0.0,
-                            'random_init': None,
-                            'random_params': [None, None],
-                            'interval': None}
+    _default_initializer = {}
 
     @property
     def nominal_values(self):
@@ -135,80 +132,87 @@ class ElectricMotor:
         """
         pass
 
-    def reset(self, upper_bound, lower_bound):
+    def reset(self, initial_states=None):
         """
         Reset the motors state to a new initial state. (Default 0)
 
+        Args:
+            initial_states(dict): states needed to be initialized,
+                                  with initialization value
         Returns:
             numpy.ndarray(float): The initial motors state.
         """
-        # use default or use userdefined interval
+        initial_states = initial_states or self._initializer['init_value']
+
+        # use default or use user-defined interval
         if (self._initializer['interval'] is not None
-                and isinstance(self._initializer, (list, np.ndarray, tuple))):
-            upper_bound = self._initializer['interval'][0]
-            lower_bound = self._initializer['interval'][1]
-
+                and isinstance(self._initializer['interval'], (list, np.ndarray, tuple))):
+            lower_bound = np.asarray(self._initializer['interval']).T[0]
+            upper_bound = np.asarray(self._initializer['interval']).T[1]
+            # clip bounds to nominal values
+            nom_currents = np.asarray([self._nominal_values[s]
+                                       for s in initial_states.keys()])
+            upper_bound = np.clip(upper_bound, a_min=None, a_max=nom_currents)
+            lower_bound = np.clip(lower_bound,
+                                  a_min=np.zeros(len(initial_states.keys()), dtype=float),
+                                  a_max=None)
+        # use nominal limits as interval
         else:
-            # maybe redundant or bad style or good to make things clear?
-            # upper_bound = upper_bound
-            # lower_bound = lower_bound
-            pass
+            # todo when no parameters passed, when state space irrelevant, 0 correct?
+            upper_bound = np.asarray([self._nominal_values[s]
+                                      for s in initial_states.keys()])
+            lower_bound = np.zeros(len(initial_states.keys()), dtype=float)
 
-        # assert isinstance(self._initializer, dict), \
-        #     'initializer have to be dict, but is %r' % type(self._initializer)
-        #initial_value = np.atleast_1d(self._initializer['init_value'])
         # todo check ob error raises sinnvoll sind
-        # here the nominal values are used as limits/ boundaries
         # todo richtige begrenzungen berechnen/ übergeben bekommen
-        #upper_bound
-        #lower_bound
-        # nominal_currents = np.asarray([self._nominal_values[ix]
-                                       # for ix in self.CURRENTS])
+        # eingabe anpassen an restlichen init motor states (eps, ...) auch als dict?
+        # todo change lower nominal values could be negativ
+        # todo change lower respective to state space (could be negativ)
+        # lower = upper * state space
 
-        # upper = initial_value.T[-1] if initial_value.shape[0] == 2 \
-        #     else nominal_currents
-        # lower = initial_value.T[0] if initial_value.shape[0] == 2 \
-        #     else np.zeros(len(self.CURRENTS), dtype=float)
-        # # clip limits to nominal values, if to high
-        # upper = np.clip(upper, a_min=None, a_max=nominal_currents)
[…]
