### 2020-06-05 #2275 @wkirgsn
Dear @moorepants and @Kevin-Mattheus-Moerman,

as already pointed out by Mr. Moerman, the arxiv paper represents a traditional long paper that describes the background and theory behind electric motor control, and emphasizes the potential in data-driven learning of a control scheme through reinforcement learning, for which gym-electric-motor is a new platform.
Conversely, the JOSS paper focuses on the toolbox itself and the software used.

The order of submission is due to us learning about JOSS just recently.
In contrast to arxiv.org, having our toolbox included in a peer-reviewed journal that JOSS is increases the accountability and credibility of this open software.

In case of further concerns regarding overlaps, I would like to point out [JOSS' preprint policy](https://joss.readthedocs.io/en/latest/submitting.html#preprint-policy).
