### 2020-06-05 45296ed
synch motor reset updated
