### 2020-06-05 #2275 @Kevin-Mattheus-Moerman
@bmagyar We understand if you cannot review at this time. However, we do not expect the same "timely review" as usual given the COVID 19 crisis. We can be much more flexible with time. If given this you can help here let me know otherwise no worries and thanks for your response.
