### 2020-06-05 d4affde
synch motor reset updated
