+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -142,17 +142,24 @@ class ElectricMotor:
         Returns:
             numpy.ndarray(float): The initial motors state.
         """
-        initial_states = initial_states or self._initializer['init_value']
+        initial_states = initial_states or self._initializer['em_init']['states']
+
+        # for order and organization purposes
+        interval = self._initializer['em_init']['interval']
+        random_type = self._initializer['random_init']['type']
+        random_params = self._initializer['random_init']['params']
 
         # use default or use user-defined interval
-        if (self._initializer['interval'] is not None
-                and isinstance(self._initializer['interval'], (list, np.ndarray, tuple))):
-            lower_bound = np.asarray(self._initializer['interval']).T[0]
-            upper_bound = np.asarray(self._initializer['interval']).T[1]
+        if (interval is not None
+                and isinstance(interval, (list, np.ndarray, tuple))):
+            lower_bound = np.asarray(interval).T[0]
+            upper_bound = np.asarray(interval).T[1]
             # clip bounds to nominal values
             nom_currents = np.asarray([self._nominal_values[s]
                                        for s in initial_states.keys()])
-            upper_bound = np.clip(upper_bound, a_min=None, a_max=nom_currents)
+            upper_bound = np.clip(upper_bound,
+                                  a_min=None,
+                                  a_max=nom_currents)
             lower_bound = np.clip(lower_bound,
                                   a_min=np.zeros(len(initial_states.keys()), dtype=float),
                                   a_max=None)
@@ -171,19 +178,18 @@ class ElectricMotor:
         # lower = upper * state space
 
         # random initialization for each motor state (current, epsilon)
-        if (self._initializer['random_init'] is not None
-                and isinstance(self._initializer['random_init'], str)):
-            initial_states = initial_states.fromkeys(initial_states.keys(), None)
-            if self._initializer['random_init'] == 'uniform':
+        if random_type is not None and isinstance(random_type, str):
+            initial_states = dict.fromkeys(initial_states.keys(), None)
+            if random_type == 'uniform':
                 initial_value = (upper_bound - lower_bound) * \
                                 np.random.random_sample(len(initial_states.keys())) + \
                                 lower_bound
                 return np.ones(len(initial_states.keys()), dtype=float) * initial_value
 
-            elif self._initializer['random_init'] == 'gaussian':
+            elif random_type == 'gaussian':
                 # todo can mean be outside of intervall (ok with truncnorm)
-                mue = self._initializer['random_params']['mue'] or upper_bound / 2
-                sigma = self._initializer['random_params']['sigma'] or 1
+                mue = random_params['mue'] or upper_bound / 2
+                sigma = random_params['sigma'] or 1
                 a, b = (lower_bound - mue) / sigma, (upper_bound - mue) / sigma
                 initial_value = truncnorm.rvs(a, b,
                                               loc=mue,
@@ -200,8 +206,7 @@ class ElectricMotor:
         # constant initialization for each motor state (current, epsilon)
         elif (initial_states is not None
                 and len(initial_states.keys()) == len(initial_states.keys())):
-            initial_value = np.atleast_1d(
-                list(initial_states.values()))
+            initial_value = np.atleast_1d(list(initial_states.values()))
             # check init_value meets interval boundaries
             if ((lower_bound <= initial_value).all()
                     and (initial_value <= upper_bound).all()):
@@ -303,10 +308,13 @@ class DcMotor(ElectricMotor):
                                'i_e': 1.2, 'u': 420}
     _default_limits = {'omega': 500, 'torque': 0.0, 'i_a': 75, 'i_e': 2,
                        'u': 420}
-    _default_initializer = {'init_value': {'i_a': 0.0, 'i_e': 0.0},
[…]
