### 2020-06-05 8d6e02c
started to update load reset
