+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -318,6 +318,50 @@ class ConstantSpeedLoad(MechanicalLoad):
         return np.array([0]), np.array([0])
 
 
+class ExternalSpeedLoad(MechanicalLoad):
+    """
+       Constant speed mechanical load system which will always set the speed
+       to a predefined value.
+    """
+
+    HAS_JACOBIAN = True
+    # _default_initializer = {'ml_init': {'states': {'omega': 0.0},
+    #                                     'interval': None},
+    #                         'random_init': {'type': None,
+    #                                         'params': {'mue': None, 'sigma': None}}
+    #                         }
+
+    @property
+    def omega(self):
+        """
+        Returns:
+            float: Constant value for omega in rad/s.
+        """
+        return self._omega
+
+    def reset(self, *_, **__):
+        # Docstring from superclass
+        return np.array([self._omega])
+
+    def __init__(self, omega=0, **__):
+        """
+        Args:
+            omega_fixed(float)): Fix value for the speed in rad/s.
+        """
+        super().__init__()
+        self._omega = omega
+
+    def mechanical_ode(self, *_, **__):
+        # Docstring of superclass
+        return np.array([0])
+
+    def mechanical_jacobian(self, t, mechanical_state, torque):
+        # Docstring of superclass
+        return np.array([0]), np.array([0])
+
+
+
+
 class PositionalPolyStaticLoad(PolynomialStaticLoad):
 
     def __init__(self, load_parameter=None, limits=None):
