### 2020-06-06 1b0ddbd
motor init fixed, load init not yet
