+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -80,89 +80,8 @@ class MechanicalLoad:
         Returns:
             ndarray(float): Initial state of the mechanical-ODE.
         """
-        initial_states = self._initializer['ml_init']['states']
-        print(initial_states.keys())
-        # for order and organization purposes
-        interval = self._initializer['ml_init']['interval']
-        random_type = self._initializer['random_init']['type']
-        random_params = self._initializer['random_init']['params']
 
-        # use default or use user-defined interval
-        if (interval is not None
-                and isinstance(interval, (list, np.ndarray, tuple))):
-            lower_bound = np.asarray(interval).T[0]
-            upper_bound = np.asarray(interval).T[1]
-            # clip bounds to nominal values
-            nom_currents = np.asarray([self._nominal_values[s]
-                                       for s in initial_states.keys()])
-            upper_bound = np.clip(upper_bound,
-                                  a_min=None,
-                                  a_max=nom_currents)
-            lower_bound = np.clip(lower_bound,
-                                  a_min=np.zeros(len(initial_states.keys()),
-                                                 dtype=float),
-                                  a_max=None)
-        # use nominal limits as interval
-        else:
-            # todo when no parameters passed, when state space irrelevant, 0 correct?
-            upper_bound = np.asarray([self._nominal_values[s]
-                                      for s in initial_states.keys()])
-            lower_bound = np.zeros(len(initial_states.keys()), dtype=float)
-
-        # todo check ob error raises sinnvoll sind
-        # todo richtige begrenzungen berechnen/ übergeben bekommen
-        # eingabe anpassen an restlichen init motor states (eps, ...) auch als dict?
-        # todo change lower nominal values could be negativ
-        # todo change lower respective to state space (could be negativ)
-        # lower = upper * state space
-
-        # random initialization for each motor state (current, epsilon)
-        if random_type is not None and isinstance(random_type, str):
-            initial_states = dict.fromkeys(initial_states.keys(), None)
-            if random_type == 'uniform':
-                initial_value = (upper_bound - lower_bound) * \
-                                np.random.random_sample(
-                                    len(initial_states.keys())) + \
-                                lower_bound
-                return np.ones(len(initial_states.keys()),
-                               dtype=float) * initial_value
-
-            elif random_type == 'gaussian':
-                # todo can mean be outside of intervall (ok with truncnorm)
-                mue = random_params['mue'] or upper_bound / 2
-                sigma = random_params['sigma'] or 1
-                a, b = (lower_bound - mue) / sigma, (upper_bound - mue) / sigma
-                initial_value = truncnorm.rvs(a, b,
-                                              loc=mue,
-                                              scale=sigma,
-                                              size=(
-                                                  len(initial_states.keys())))
-                return np.ones(len(initial_states.keys()),
-                               dtype=float) * initial_value
-
-            else:
-                """
-                possible to implement other Distributions
-                """
-                raise NotImplementedError
-
-        # constant initialization for each motor state (current, epsilon)
-        elif (initial_states is not None
-              and len(initial_states.keys()) == len(initial_states.keys())):
-            initial_value = np.atleast_1d(list(initial_states.values()))
-            # check init_value meets interval boundaries
-            if ((lower_bound <= initial_value).all()
-                    and (initial_value <= upper_bound).all()):
-                return np.ones(len(initial_states.keys()),
-                               dtype=float) * initial_value
-
-            else:
-                raise Exception('Initialization Value have to be in nominal '
[…]
