+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -132,21 +132,27 @@ class ElectricMotor:
         """
         pass
 
-    def reset(self, initial_states=None):
+    def initialize(self, state_space=None,
+                   state_positions=None, initial_states=None):
         """
-        Reset the motors state to a new initial state. (Default 0)
+        Initializes give state values. Values can be given as a constant or
+        sampled random out of a statistical distribution. Initial value is in
+        range of the nominal values or a given interval.
 
         Args:
-            initial_states(dict): states needed to be initialized,
-                                  with initialization value
+            state_space(gym.Box): normalized state space boundaries
+            state_positions(dict): indexes of system states
+            initial_states(list): system states to be initilized (when not
+                                    explicitly given by 'initializer')
         Returns:
-            numpy.ndarray(float): The initial motors state.
+            ndarray(float): initial value for each state
+
         """
         initial_states = initial_states or self._initializer['em_init']['states']
 
         # for order and organization purposes
         interval = self._initializer['em_init']['interval']
-        random_type = self._initializer['random_init']['type']
+        random_dist = self._initializer['random_init']['distribution']
         random_params = self._initializer['random_init']['params']
 
         # use default or use user-defined interval
@@ -161,56 +167,54 @@ class ElectricMotor:
                                   a_min=None,
                                   a_max=nom_currents)
             lower_bound = np.clip(lower_bound,
-                                  a_min=np.zeros(len(initial_states.keys()), dtype=float),
+                                  a_min=np.zeros(len(initial_states.keys()),
+                                                 dtype=float),
                                   a_max=None)
-        # use nominal limits as interval
+        # use nominal limits as interval in state space
         else:
-            # todo when no parameters passed, when state space irrelevant, 0 correct?
+            state_idx = [state_positions[state] for state in initial_states]
             upper_bound = np.asarray([self._nominal_values[s]
                                       for s in initial_states.keys()])
-            lower_bound = np.zeros(len(initial_states.keys()), dtype=float)
-
-        #todo check ob error raises sinnvoll sind
-        #todo richtige begrenzungen berechnen/ übergeben bekommen
-        #eingabe anpassen an restlichen init motor states (eps, ...) auch als dict?
-        #todo change lower nominal values could be negativ
-        #todo change lower respective to state space (could be negativ)
-        #lower = upper * state space
+            upper_bound *= state_space.high[state_idx]
+            lower_bound = upper_bound * state_space.low[state_idx]
 
         # random initialization for each motor state (current, epsilon)
-        if random_type is not None and isinstance(random_type, str):
+        if random_dist is not None and isinstance(random_dist, str):
             initial_states = dict.fromkeys(initial_states.keys(), None)
-            if random_type == 'uniform':
+            if random_dist == 'uniform':
                 initial_value = (upper_bound - lower_bound) * \
-                                np.random.random_sample(len(initial_states.keys())) + \
+                                np.random.random_sample(
+                                    len(initial_states.keys())) + \
                                 lower_bound
-                return np.ones(len(initial_states.keys()), dtype=float) * initial_value
+                return np.ones(len(initial_states.keys()),
+                               dtype=float) * initial_value
 
-            elif random_type == 'gaussian':
+            elif random_dist == 'gaussian' or random_dist == 'normal':
                 # todo can mean be outside of intervall (ok with truncnorm)
-                mue = random_params['mue'] or upper_bound / 2
-                sigma = random_params['sigma'] or 1
+                mue = random_params[0] or upper_bound / 2
+                sigma = random_params[1] or 1
                 a, b = (lower_bound - mue) / sigma, (upper_bound - mue) / sigma
                 initial_value = truncnorm.rvs(a, b,
                                               loc=mue,
[…]

[…]
