### 2020-06-06 82d00e7
legend added to limits and nominal values
