### 2020-06-06 1618636
fixed load state init, user input still  to complicated
