### 2020-06-07 6fbc5ee
new class externalspeedload
