+++ gym_electric_motor/physical_systems/electric_motors_LOCAL_23360.py
@@ -1,1961 +1,0 @@
-import numpy as np
-import math
-from scipy.stats import truncnorm
-
-
-class ElectricMotor:
-    """
-        Base class for all technical electrical motor models.
-
-        A motor consists of the ode-state. These are the dynamic quantities of its ODE.
-        For example:
-            ODE-State of a DC-shunt motor: `` [i_a, i_e ] ``
-                * i_a: Anchor circuit current
-                * i_e: Exciting circuit current
-
-        Each electric motor can be parametrized by a dictionary of motor parameters, the nominal state dictionary
-        and the limit dictionary.
-    """
-
-    #: Parameter indicating if the class is implementing the optional jacobian function
-    HAS_JACOBIAN = False
-
-    #: CURRENTS_IDX(list(int)): Indices for accessing all motor currents.
-    CURRENTS_IDX = []
-
-    #: CURRENTS(list(str)): List of the motor currents names
-    CURRENTS = []
-    #: VOLTAGES(list(str)): List of the motor input voltages names
-    VOLTAGES = []
-
-    #: _default_motor_parameter(dict): Default parameter dictionary for the motor
-    _default_motor_parameter = {}
-    #: _default_nominal_values(dict(float)): Default nominal motor state array
-    _default_nominal_values = {}
-    #: _default_limits(dict(float)): Default motor limits (0 for unbounded limits)
-    _default_limits = {}
-    #: _default_initial_state(dict): Default initial motor-state values
-    _default_initializer = {}
-
-    @property
-    def nominal_values(self):
-        """
-        Readonly motors nominal values.
-
-        Returns:
-            dict(float): Current nominal values of the motor.
-        """
-        return self._nominal_values
-
-    @property
-    def limits(self):
-        """
-        Readonly motors limit state array. Entries are set to the maximum physical possible values
-        in case of unspecified limits.
-
-        Returns:
-            dict(float): Limits of the motor.
-        """
-        return self._limits
-
-    @property
-    def motor_parameter(self):
-        """
-        Returns:
-             dict(float): The motors parameter dictionary
-        """
-        return self._motor_parameter
-
-    @property
-    def initializer(self):
-        """
-        Returns:
-            dict: Motor initial state and additional initializer parameter
-        """
-        return self._initializer
-
-    def __init__(self, motor_parameter=None, nominal_values=None,
-                 limit_values=None, initializer=None, **__):
-        """
-        :param  motor_parameter: Motor parameter dictionary. Contents specified
-                for each motor.
-        :param  nominal_values: Nominal values for the motor quantities.
-        :param  limit_values: Limits for the motor quantities.
-        :param  initializer: Initial motor states (currents)
-                            ('constant', 'uniform', 'gaussian' sampled from
-                             given interval or out of nominal motor values)
-        """
-        motor_parameter = motor_parameter or {}
-        self._motor_parameter = self._default_motor_parameter.copy()
-        self._motor_parameter.update(motor_parameter)
-        limit_values = limit_values or {}
-        self._limits = self._default_limits.copy()
-        self._limits.update(limit_values)
-        nominal_values = nominal_values or {}
-        self._nominal_values = self._default_nominal_values.copy()
-        self._nominal_values.update(nominal_values)
-        initializer = initializer or {}
-        self._initializer = self._default_initializer.copy()
-        self._initializer.update(initializer)
-
-    def electrical_ode(self, state, u_in, omega, *_):
-        """
-        Calculation of the derivatives of each motor state variable for the given inputs / The motors ODE-System.
-
-        Args:
-            state(ndarray(float)): The motors state.
-            u_in(list(float)): The motors input voltages.
-            omega(float): Angular velocity of the motor
-
[…]

[…]
