### 2020-06-08 5e13354
complicated input fixed, seperate inputs for load, motor
