+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -75,7 +75,7 @@ class ElectricMotor:
         return self._initializer
 
     def __init__(self, motor_parameter=None, nominal_values=None,
-                 limit_values=None, initializer=None, **__):
+                 limit_values=None, motor_initializer=None, **__):
         """
         :param  motor_parameter: Motor parameter dictionary. Contents specified
                 for each motor.
@@ -94,9 +94,9 @@ class ElectricMotor:
         nominal_values = nominal_values or {}
         self._nominal_values = self._default_nominal_values.copy()
         self._nominal_values.update(nominal_values)
-        initializer = initializer or {}
+        motor_initializer = motor_initializer or {}
         self._initializer = self._default_initializer.copy()
-        self._initializer.update(initializer)
+        self._initializer.update(motor_initializer)
 
     def electrical_ode(self, state, u_in, omega, *_):
         """
@@ -149,13 +149,12 @@ class ElectricMotor:
             ndarray(float): initial value for each state
 
         """
-        initial_states = initial_states or self._initializer['em_init']['states']
-
+        initial_states = initial_states or self._initializer['states']
 
         # for order and organization purposes
-        interval = self._initializer['em_init']['interval']
-        random_dist = self._initializer['random_init']['distribution']
-        random_params = self._initializer['random_init']['params']
+        interval = self._initializer['interval']
+        random_dist = self._initializer['random_init']
+        random_params = self._initializer['random_params']
 
         # use default or use user-defined interval
         if (interval is not None
@@ -201,8 +200,7 @@ class ElectricMotor:
                 initial_value = truncnorm.rvs(a, b,
                                               loc=mue,
                                               scale=sigma,
-                                              size=(
-                                                  len(initial_states.keys())))
+                                              size=len(initial_states.keys()))
                 return np.ones(len(initial_states.keys()),
                                dtype=float) * initial_value
 
@@ -339,13 +337,10 @@ class DcMotor(ElectricMotor):
                                'i_e': 1.2, 'u': 420}
     _default_limits = {'omega': 500, 'torque': 0.0, 'i_a': 75, 'i_e': 2,
                        'u': 420}
-    _default_initializer = {'em_init': {'states': {'i_a': 0.0, 'i_e': 0.0},
-                                        'interval': None},
-                            'ml_init': {'states': {'omega': None},
-                                        'interval': None},
-                            'random_init': {'distribution': None,
-                                            'params': {'mue': None, 'sigma': None}}
-                            }
+    _default_initializer = {'states': {'i_a': 0.0, 'i_e': 0.0},
+                            'interval': None,
+                            'random_init': None,
+                            'random_params': (None, None)}
 
     def __init__(self,
                  motor_parameter=None, nominal_values=None,
@@ -481,13 +476,10 @@ class DcShuntMotor(DcMotor):
                                'i_e': 1.2, 'u': 420}
     _default_limits = {'omega': 500, 'torque': 0.0, 'i_a': 75, 'i_e': 2,
                        'u': 420}
-    _default_initializer = {'em_init': {'states': {'i_a': 0.0, 'i_e': 0.0},
-                                        'interval': None},
-                            'ml_init': {'states': {'omega': None},
-                                        'interval': None},
-                            'random_init': {'distribution': None,
-                                            'params': {'mue': None, 'sigma': None}}
-                            }
+    _default_initializer = {'states': {'i_a': 0.0, 'i_e': 0.0},
+                            'interval': None,
+                            'random_init': None,
+                            'random_params': (None, None)}
 
 
     def i_in(self, state):
[…]
