### 2020-06-08 8fec43b
cleanup and refactoring of the dashboard
