+++ gym_electric_motor/visualization/motor_dashboard.py
@@ -1,524 +1,54 @@
-from ..core import ElectricMotorVisualization
-from matplotlib import pyplot as plt
-import numpy as np
-import matplotlib
-import warnings
+from gym_electric_motor.core import ElectricMotorVisualization
+import gym_electric_motor.visualization.motor_dashboard_plots as mdp
+import matplotlib.pyplot as plt
 
 
 class MotorDashboard(ElectricMotorVisualization):
-    """
-    Visualization of the variables of the motor in graphs. This can be the angular velocity omega, torque
-    phase voltages, phase currents, supply voltage and the rotor angle at the PMSM.
 
-    In the table are all variables given that can be shown for a specific motor.
-
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |variable| extex | shunt | permex |series | PMSM, | Description                                |
-    |        |       |       |        |       | SynRM |                                            |
-    +========+=======+=======+========+=======+=======+============================================+
-    |omega   | x     | x     | x      | x     | x     | mechanical angular velocity                |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |torque  | x     | x     | x      | x     | x     | Motor generated torque                     |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_sup   | x     | x     | x      | x     | x     | supply voltage of the converter            |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |i       |       |       | x      | x     |       | current in permex and series motors        |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |i_e     | x     | x     |        |       |       | current in excitation circuit              |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u       |       | x     | x      | x     |       | voltage at shunt, permex and series motors |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_e     | x     |       |        |       |       | voltage at the excitation circuit          |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_a     | x     |       |        |       | x     | armature voltage or voltage of phase a     |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_b     |       |       |        |       | x     | voltage of phase b                         |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_c     |       |       |        |       | x     | voltage of phase c                         |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_d     |       |       |        |       | x     | d-axis voltage                             |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |u_q     |       |       |        |       | x     | q-axis voltage                             |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |i_a     | x     | x     |        |       | x     | armature current or current in phase a     |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |i_b     |       |       |        |       | x     | current in phase a                         |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |i_c     |       |       |        |       | x     | current in phase b                         |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
-    |i_d     |       |       |        |       | x     | d-axis current                             |
-    +--------+-------+-------+--------+-------+-------+--------------------------------------------+
[…]
