+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -1,0 +1,232 @@
+import numpy as np
+import matplotlib.lines as lin
+
+
+class MotorDashboardPlot:
+
+    def __init__(self):
+        self._axis = None
+
+    def initialize(self, axis):
+        self._axis = axis
+
+    def set_modules(self, ps, rg, rf):
+        pass
+
+    def step(self, k, state, reference, action, reward, done):
+        raise NotImplementedError
+
+    def update(self):
+        pass
+
+    def reset(self):
+        pass
+
+
+class StatePlot(MotorDashboardPlot):
+
+    x_width = 3
+    mode = 'continuous'
+    state_line_cfg = {
+        'color': 'blue',
+        'linestyle': '',
+        'linewidth': 0.75,
+        'marker': '.',
+        'markersize': .5
+    }
+    reference_line_cfg = {
+        'color': 'green',
+        'linewidth': 0.75,
+        'linestyle': '',
+        'marker': '.',
+        'markersize': .5
+    }
+    limit_line_cfg = {
+        'color': 'red',
+        'linestyle': '--',
+        'linewidth': 1
+    }
+
+    state_labels = {
+        'omega': r'$\omega/(rad/s)$',
+        'torque': '$T/Nm$',
+        'i': '$i/A$',
+        'i_a': '$i_{a}/A$',
+        'i_e': '$i_{e}/A$',
+        'i_b': '$i_{b}/A$',
+        'i_c': '$i_{c}/A$',
+        'i_sq': '$i_{sq}/A$',
+        'i_sd': '$i_{sd}/A$',
+        'u': '$u/V$',
+        'u_a': '$u_{a}/V$',
+        'u_b': '$u_{b}/V$',
+        'u_c': '$u_{c}/V$',
+        'u_sq': '$u_{sq}/V$',
+        'u_sd': '$u_{sd}/V$',
+        'u_e': '$u_{e}/V$',
+        'u_sup': '$u_{sup}/V$',
+        'epsilon': r'$\epsilon/rad$'
+    }
+
+    def __init__(self, state):
+        self._state_space = None
+        self._state = state
+        self._state_idx = None
+        self._limits = None
+        self._referenced = None
+        self._state_line = None
+        self._reference_line = None
+        self._state_data = []
+        self._ref_data = []
+        self._t_data = []
+        self._tau = None
+        self._normalized = True
+        self._x_points = None
+        self._t = 0
+        super().__init__()
+
+    def set_modules(self, ps, rg, rf):
+        self._state_idx = ps.state_positions[self._state]
+        self._limits = ps.limits[self._state_idx]
+        self._state_space = ps.state_space.low[self._state_idx], ps.state_space.high[self._state_idx]
+        self._referenced = rg.referenced_states[self._state_idx]
+        self._tau = ps.tau
+        self._x_points = int(self.x_width / self._tau)
+        if self._limits == self._state_space[1]:
+            self._normalized = False
+
+    def initialize(self, axis):
+        super().initialize(axis)
+        axis.grid()
+        if self._referenced:
+            self._reference_line, = self._axis.plot(self._t_data, self._ref_data, **self.reference_line_cfg)
+        self._state_line, = self._axis.plot(self._t_data, self._state_data, **self.state_line_cfg)
+        min_limit = self._limits * self._state_space[0] if self._normalized else self._state_space[0]
+        max_limit = self._limits * self._state_space[1] if self._normalized else self._state_space[1]
+
+        if self._state_space[0] < 0:
+            self._axis.axhline(min_limit, **self.limit_line_cfg)
+        lim = self._axis.axhline(max_limit, **self.limit_line_cfg)
+
[…]
