### 2020-06-08 c114669
added comments to dashboard, commented out old tests
