+++ gym_electric_motor/physical_systems/voltage_supplies.py
@@ -37,7 +37,7 @@ class VoltageSupply:
         """
         return self.get_voltage(0, 0)
 
-    def get_voltage(self, i_sup, t, *args, **kwargs):
+    def get_voltage(self, t, i_sup, *args, **kwargs):
         """
         Get the supply voltage based on the floating supply current i_sup, the time t and optional further arguments.
 
@@ -70,15 +70,17 @@ class IdealVoltageSupply(VoltageSupply):
 class RCVoltageSupply(VoltageSupply):
     """Voltage supply moduled as RC element"""
     
-    def __init__(self, u_nominal, R, C, **__):
+    def __init__(self, u_nominal, supply_parameter = {'R':1,'C':1}, **__):
         super().__init__(u_nominal)
         # Supply range is between 0 - capacitor completely unloaded - and u_nominal - capacitor is completely loaded
+        
         self.supply_range = (0,u_nominal) 
-        self._R = R
-        self._C = C
+        self._R = supply_parameter['R']
+        self._C = supply_parameter['C']
         self._u_sup = 0.0
         self._u_0 = u_nominal
         self._solver = EulerSolver()
+        print(self._u_0)
         def system_equation(t, u_sup, u_0, i_sup, R, C):
             # calculates ODE for derivate of u_sup
             return (u_0 - u_sup - R*i_sup)/(R*C)
@@ -87,9 +89,11 @@ class RCVoltageSupply(VoltageSupply):
     def reset(self):
         # On reset the capacitor is unloaded again
         self._solver.set_initial_value(0.0)
+        self._u_sup = 0.0
+        return self._u_sup
                 
     
-    def get_voltage(self, i_sup, t):
-        self._solver.set_f_params(self._u_sup, self._u_0, i_sup, self._R, self._C)
+    def get_voltage(self, t,i_sup):
+        self._solver.set_f_params(self._u_0, i_sup, self._R, self._C)
         self._u_sup = self._solver.integrate(t)
         return self._u_sup
