### 2020-06-08 5b2402a
Initialize the capicitance at u_0. Added warning for very small values of R*C
