### 2020-06-08 e539468
Small style fixes
