### 2020-06-08 ff03569
Bugfix in MotorDashboard
