### 2020-06-08 867ac27
added support for the discrete switching case
