### 2020-06-09 4043400
started external speed load
