+++ gym_electric_motor/physical_systems/voltage_supplies.py
@@ -75,8 +75,7 @@ class RCVoltageSupply(VoltageSupply):
         """
         RC circuit takes additional values for it's electrical elements.
         Args:
-            R(float): Reluctance in Ohm
-            C(float): Capacitance in Farad
+            supply_parameter(dict): Consists or Reluctance R in Ohm and Capacitance C in Farad
         Additional notes:
             If the product of R and C get too small the numerical stability of the ODE is not given anymore
             since our time difference tau is only in the range of 10e-3. You might want to consider R as a
@@ -85,17 +84,17 @@ class RCVoltageSupply(VoltageSupply):
         super().__init__(u_nominal)
         # Supply range is between 0 - capacitor completely unloaded - and u_nominal - capacitor is completely loaded
         self.supply_range = (0,u_nominal) 
-        self._R = supply_parameter['R']
-        self._C = supply_parameter['C']
-        if self._R*self._C < 1e-4:
+        self._r = supply_parameter['R']
+        self._c = supply_parameter['C']
+        if self._r*self._c < 1e-4:
             warnings.warn("The product of R and C might be too small for the correct calculation of the supply voltage. You might want to consider R as a time constant.")
 
         self._u_sup = u_nominal
         self._u_0 = u_nominal
         self._solver = EulerSolver()
-        def system_equation(t, u_sup, u_0, i_sup, R, C):
+        def system_equation(t, u_sup, u_0, i_sup, r, c):
             # ODE for derivate of u_sup
-            return np.array([(u_0 - u_sup[0] - R*i_sup)/(R*C)])
+            return np.array([(u_0 - u_sup[0] - r*i_sup)/(r*c)])
         self._solver.set_system_equation(system_equation)
         
     def reset(self):
@@ -106,7 +105,7 @@ class RCVoltageSupply(VoltageSupply):
                 
     
     def get_voltage(self, t,i_sup):
-        self._solver.set_f_params(self._u_0, i_sup, self._R, self._C)
+        self._solver.set_f_params(self._u_0, i_sup, self._r, self._c)
         self._u_sup = self._solver.integrate(t)[0]
         return self._u_sup
     
