### 2020-06-09 989dad5
Some commentary fixes
