### 2020-06-09 e2a3e41
uncommented failing assertion until proper understanding is available
