### 2020-06-09 fd3be0e
modified examples to work with the new visualization
