### 2020-06-09 8a4ecc4
added ExternalSpeedLoad for adding a userdefinded speedprofile
