### 2020-06-10 d779042
updated docstring for ExternalSpeedLoad
