### 2020-06-10 2dc8f1b
actions added to visualization. bugfix req for None-type action at the start
