+++ examples/agents/simple_controllers.py
@@ -1,5 +1,7 @@
 from gym.spaces import Discrete, Box
-import sys, os
+import sys
+import os
+
 sys.path.append(os.path.abspath(os.path.join('..')))
 from gym_electric_motor.physical_systems.electric_motors import DcShuntMotor, DcExternallyExcitedMotor, \
     DcPermanentlyExcitedMotor, DcSeriesMotor
@@ -9,7 +11,6 @@ import numpy as np
 
 
 class Controller:
-
     """
       The following is a base class for various controllers along with the motor environments
 
@@ -35,12 +36,10 @@ class Controller:
 
 
 class OnOffController(Controller):
-
     """
         The following controller is a simple OnOffController which allows the state to choose high_action when referenced_state is below
         ref_idx and low_action otherwise
     """
-
     def __init__(self, environment, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Discrete, 'Not suitable action space for On off controller'
@@ -60,14 +59,13 @@ class OnOffController(Controller):
 
 
 class ThreePointController(Controller):
-
     """
     Below is an implementation of a 3 point controller: When state_idx is below the reference_idx, it choose high_action and when the
     state_idx is above the reference_idx,it choose low_action. If it is between the reference_idx values, it will choose idle_action
-    A 'Hysteresis' value should be integrated(chosen as 0.01) in this controller because of the constant switching and high frequency around the reference_idx
+    A 'Hysteresis' value should be integrated(chosen as 0.01) in this controller because of the constant switching and
+    high frequency around the reference_idx
 
     """
-
     def __init__(self, environment, hysteresis=0.01, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Discrete, 'Not suitable action space for three point controller'
@@ -91,13 +89,11 @@ class ThreePointController(Controller):
 
 
 class PController(Controller):
-
     """
     In the below proportional control implementation, the controller output is proportional to the error signal,  which is the difference
     between the reference_idx and the state_idx .i.e., the output of a proportional controller
     is the multiplication product of the error signal and the proportional gain.Here kp =10 is assumed for proportionality gain.
     """
-
     def __init__(self, environment, k_p=10, controller_no=0, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Box, 'No suitable action space for P Controller'
@@ -126,15 +122,17 @@ class PIController(PController):
       and integral gain inputs. The error signal is the difference between the reference_idx and the referenced_state.
       It outputs a weighted sum of the input error signal and the integral of the input error signal.
     """
-
     def __init__(self, environment, k_p=10, k_i=0.01, controller_no=0, reference_idx=0):
         super().__init__(environment, k_p, controller_no, reference_idx)
+        assert type(environment.physical_system) is DcSeriesMotor  #
         self._k_i = k_i
         self._tau = environment.physical_system.tau
         self._limits = environment.physical_system.limits
         self._integrated_value = 0
-        self._referenced_state_max = self._limits[self._referenced_state] * environment.physical_system.state_space.high[self._referenced_state]
-        self._referenced_state_min = self._limits[self._referenced_state] * environment.physical_system.state_space.low[self._referenced_state]
+        self._referenced_state_max = self._limits[self._referenced_state] * \
+                                     environment.physical_system.state_space.high[self._referenced_state]
[…]
