### 2020-06-11 9ad4499
action visualization bugfix done
