### 2020-06-11 c3dd149
comments added. examples modified
