### 2020-06-12 61c8dd7
cleaned up intialization, Inductionmotors not quite working yet
