+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -97,6 +97,7 @@ class ElectricMotor:
         motor_initializer = motor_initializer or {}
         self._initializer = self._default_initializer.copy()
         self._initializer.update(motor_initializer)
+        self._initial_states = self._initializer['states']
 
     def electrical_ode(self, state, u_in, omega, *_):
         """
@@ -132,103 +133,123 @@ class ElectricMotor:
         """
         pass
 
-    def initialize(self, state_space=None,
-                   state_positions=None, initial_states=None,
+    def initialize(self, state_space,
+                   state_positions,
                    **__):
         """
-        Initializes give state values. Values can be given as a constant or
+        Initializes given state values. Values can be given as a constant or
         sampled random out of a statistical distribution. Initial value is in
         range of the nominal values or a given interval.
 
         Args:
-            state_space(gym.Box): normalized state space boundaries
-            state_positions(dict): indexes of system states
-            initial_states(list): system states to be initilized (when not
-                                    explicitly given by 'initializer')
+            state_space(gym.Box): normalized state space boundaries (given by
+                                  physical system)
+            state_positions(dict): indexes of system states (given by physical
+                                   system)
         Returns:
             ndarray(float): initial value for each state
-
         """
-        initial_states = initial_states or self._initializer['states']
-
-        # for order and organization purposes
+        # for organization purposes
         interval = self._initializer['interval']
         random_dist = self._initializer['random_init']
         random_params = self._initializer['random_params']
+        if isinstance(self._nominal_values, dict):
+            nominal_values_ = np.asarray(list(self._nominal_values.values()))
+        else:
+            nominal_values_ = self._nominal_values
 
-        # use default or use user-defined interval
-        if (interval is not None
-                and isinstance(interval, (list, np.ndarray, tuple))):
-            lower_bound = np.asarray(interval, dtype=float).T[0]
-            upper_bound = np.asarray(interval, dtype=float).T[1]
-            # clip bounds to nominal values
-            nom_currents = np.asarray([self._nominal_values[s]
-                                       for s in initial_states.keys()],
-                                      dtype=float)
+        # setting nominal values as interval limits
+        state_idx = [state_positions[state] for state in self._initial_states.keys()]
+        upper_bound = np.asarray(nominal_values_[state_idx], dtype=float)
+        lower_bound = upper_bound * np.asarray(state_space.low, dtype=float)[state_idx]
+        # clip nominal boundaries to user defined
+        if interval is not None:
+            lower_bound = np.clip(lower_bound,
+                                  a_min=
+                                  np.asarray(self.interval, dtype=float).T[0],
+                                  a_max=None)
             upper_bound = np.clip(upper_bound,
                                   a_min=None,
-                                  a_max=nom_currents)
-            lower_bound = np.clip(lower_bound,
-                                  a_min=np.zeros(len(initial_states.keys()),
-                                                 dtype=float),
-                                  a_max=None)
-        # use nominal limits as interval in state space
+                                  a_max=
+                                  np.asarray(self.interval, dtype=float).T[1])
         else:
-            state_idx = [state_positions[state] for state in initial_states]
-            upper_bound = np.asarray([self._nominal_values[s]
-                                for s in initial_states.keys()], dtype=float)
-            upper_bound *= np.asarray(state_space.high[state_idx], dtype=float)
-            lower_bound = upper_bound * \
[…]
