+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -72,93 +72,131 @@ class MechanicalLoad:
         load_initializer = load_initializer or {}
         self._initializer = self._default_initializer.copy()
         self._initializer.update(load_initializer)
+        self._initial_states = self._initializer['states']
 
-    def initialize(self, nominal_values, **__):
+    def initialize(self,
+                   nominal_state,
+                   state_space=None,
+                   state_positions=None,
+                   **__):
         """
         Initializes give state values. Values can be given as a constant or
         sampled random out of a statistical distribution. Initial value is in
         range of the nominal values or a given interval.
 
         Args:
-            nominal_values(dict): nominal values for states
+            nominal_state(list): nominal values fpr each state given from
+                                  physical system
+            state_space(gym.Box): normalized state space boundaries
+            state_positions(dict): indexes of system states
         Returns:
             ndarray(float): initial value for each state
 
         """
         # for order and organization purposes
-        initial_states = self._initializer['states']
         interval = self._initializer['interval']
         random_dist = self._initializer['random_init']
         random_params = self._initializer['random_params']
 
-        # use default or use user-defined interval
-        if (interval is not None
-                and isinstance(interval, (list, np.ndarray, tuple))):
-            lower_bound = np.asarray(interval).T[0]
-            upper_bound = np.asarray(interval).T[1]
-            # clip bounds to nominal values
-            nom_currents = np.asarray([nominal_values[s]
-                                       for s in initial_states.keys()])
+        # setting nominal values as interval limits
+        state_idx = [state_positions[state] for state in self._initial_states.keys()]
+        upper_bound = np.asarray(nominal_state[state_idx], dtype=float)
+        lower_bound = upper_bound * np.asarray(state_space.low, dtype=float)
+        # clip nominal boundaries to user defined
+        if interval is not None:
+            lower_bound = np.clip(lower_bound,
+                                  a_min=
+                                  np.asarray(self.interval, dtype=float).T[0],
+                                  a_max=None)
             upper_bound = np.clip(upper_bound,
                                   a_min=None,
-                                  a_max=nom_currents)
-            lower_bound = np.clip(lower_bound,
-                                  a_min=np.zeros(len(initial_states.keys()),
-                                                 dtype=float),
-                                  a_max=None)
-        # use nominal limits as interval in state space
+                                  a_max=
+                                  np.asarray(self.interval, dtype=float).T[1])
         else:
-            upper_bound = np.asarray([nominal_values[s]
-                                      for s in initial_states.keys()])
-            lower_bound = np.zeros_like(upper_bound)
+            pass
+        # if (interval is not None
+        #         and isinstance(interval, (list, np.ndarray, tuple))):
+        #     lower_bound = np.asarray(interval).T[0]
+        #     upper_bound = np.asarray(interval).T[1]
+        #     # clip bounds to nominal values
+        #     nom_currents = np.asarray([nominal_values[s]
+        #                                for s in initial_states.keys()])
+        #     upper_bound = np.clip(upper_bound,
+        #                           a_min=None,
+        #                           a_max=nom_currents)
+        #     lower_bound = np.clip(lower_bound,
+        #                           a_min=np.zeros(len(initial_states.keys()),
+        #                                          dtype=float),
+        #                           a_max=None)
+        # # use nominal limits as interval in state space
+        # else:
+        #     upper_bound = np.asarray([nominal_values[s]
+        #                               for s in initial_states.keys()])
+        #     lower_bound = np.zeros_like(upper_bound)
 
         # random initialization for each motor state (current, epsilon)
-        if random_dist is not None and isinstance(random_dist, str):
-            initial_states = dict.fromkeys(initial_states.keys(), None)
-            if random_dist == 'uniform':
[…]
