### 2020-06-12 53839c0
working DC, Synch Motors, Inductionmotors had to be fixed
