### 2020-06-12 5f0f3ab
added comments and docstrings
