### 2020-06-12 4d40464
added comments and docstrings
