+++ examples/agents/simple_controllers.py
@@ -1,11 +1,12 @@
 from gym.spaces import Discrete, Box
 import sys
 import os
+import statistics as s
 
 sys.path.append(os.path.abspath(os.path.join('..')))
 from gym_electric_motor.physical_systems.electric_motors import DcShuntMotor, DcExternallyExcitedMotor, \
     DcPermanentlyExcitedMotor, DcSeriesMotor
-from gym_electric_motor.physical_systems import SynchronousMotorSystem
+from gym_electric_motor.physical_systems import SynchronousMotorSystem, DcMotorSystem
 import math
 import numpy as np
 
@@ -38,8 +39,11 @@ class Controller:
 class OnOffController(Controller):
     """
         The following controller is a simple OnOffController which allows the state to choose high_action when referenced_state is below
-        ref_idx and low_action otherwise
+        ref_idx and low_action otherwise.
+
+        Valid for DcPermExDisc-v1, DcSeriesDisc-v1, PMSMDisc-v1, SynRMDisc-v1 ,SCIMDisc-v1
     """
+
     def __init__(self, environment, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Discrete, 'Not suitable action space for On off controller'
@@ -63,9 +67,12 @@ class ThreePointController(Controller):
     Below is an implementation of a 3 point controller: When state_idx is below the reference_idx, it choose high_action and when the
     state_idx is above the reference_idx,it choose low_action. If it is between the reference_idx values, it will choose idle_action
     A 'Hysteresis' value should be integrated(chosen as 0.01) in this controller because of the constant switching and
-    high frequency around the reference_idx
+    high frequency around the reference_idx.
+
+    Valid for DcPermExDisc-v1, DcSeriesDisc-v1,DcShuntDisc-v1, PMSMDisc-v1, SynRMDisc-v1 ,SCIMDisc-v1
 
     """
+
     def __init__(self, environment, hysteresis=0.01, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Discrete, 'Not suitable action space for three point controller'
@@ -88,15 +95,19 @@ class ThreePointController(Controller):
             return self._idle_action
 
 
-class PController(Controller):
+class DCPController(Controller):
     """
     In the below proportional control implementation, the controller output is proportional to the error signal,  which is the difference
     between the reference_idx and the state_idx .i.e., the output of a proportional controller
     is the multiplication product of the error signal and the proportional gain.Here kp =10 is assumed for proportionality gain.
+
+    Valid for DC Motor System
     """
+
     def __init__(self, environment, k_p=10, controller_no=0, state_idx=None, reference_idx=0):
         action_space = environment.action_space
-        assert type(action_space) is Box, 'No suitable action space for P Controller'
+        assert type(action_space) is Box and type(
+            environment.physical_system) is DcMotorSystem, 'No suitable action space for P Controller'
         self._k_p = k_p
         self._controller_no = controller_no
         self._action_min = action_space.low[controller_no]
@@ -116,15 +127,20 @@ class PController(Controller):
         ])
 
 
-class PIController(PController):
+class DCPIController(DCPController):
     """
       The below Discrete PI Controller performs discrete-time PI controller computation using the error signal and proportional
       and integral gain inputs. The error signal is the difference between the reference_idx and the referenced_state.
       It outputs a weighted sum of the input error signal and the integral of the input error signal.
+
+      Valid for DC Motor System
     """
+
[…]
