### 2020-06-15 9053fdf
Added assertion for missing keys
