### 2020-06-16 e9dc2c4
in progess update flux init
