### 2020-06-16 d805902
fixed init. for Induction motors
