+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -87,8 +87,17 @@ class ElectricMotor:
         """
         return self._initializer
 
+    @property
+    def initial_limits(self):
+        """
+        Returns:
+            dict: nominal motor limits for choosing initial values
+        """
+        return self._initial_limits
+
     def __init__(self, motor_parameter=None, nominal_values=None,
-                 limit_values=None, motor_initializer=None, **__):
+                 limit_values=None, motor_initializer=None, initial_limits=None,
+                 **__):
         """
         :param  motor_parameter: Motor parameter dictionary. Contents specified
                 for each motor.
@@ -113,7 +122,9 @@ class ElectricMotor:
         self._initial_states = self._initializer['states']
         # intialize limits, because in general they're not needed to be changed
         # during during training or episodes
+        initial_limits = initial_limits or {}
         self._initial_limits = self._nominal_values.copy()
+        self._initial_limits.update(initial_limits)
 
     def electrical_ode(self, state, u_in, omega, *_):
         """
@@ -149,50 +160,6 @@ class ElectricMotor:
         """
         pass
 
-    # def update_initial_limits(self, new_limits=None
-    #                        state_space,
-    #                        state_positions,):
-    #     """
-    #
-    #     Args:
-    #         new_limits(dict): new limitations for initial_limits
-    #         new_limits(dict): new limitations for initial_limits
-    #     """
-    #     interval = self._initializer['interval']
-    #     if isinstance(self._nominal_values, dict):
-    #         nominal_values_ = np.asarray(list(self._nominal_values.values()))
-    #     else:
-    #         nominal_values_ = self._nominal_values
-    #
-    #     #for Induction motors new limitations
-    #     if any(map(lambda state: state in self._initial_states.keys(),
-    #               ['psi_ralpha', 'psi_rbeta'])):
-    #        # draw a sample magnetic field angle from [-pi,pi]
-    #        eps_mag = 2 * np.pi * np.random.random_sample() - np.pi
-    #        # calc maximal alphabeta-currents, given dq-currents
-    #         i_alphabeta_max = self.
-    #
-    #
-    #
-    #
-    #     # setting upper and lower limits regarding state space
-    #     state_idx = [state_positions[state] for state in self._initial_states.keys()]
-    #     upper_bound = np.asarray(nominal_values_[state_idx], dtype=float)
-    #     lower_bound = upper_bound * np.asarray(state_space.low, dtype=float)[state_idx]
-    #     # clip nominal boundaries to user defined limits
-    #     if interval is not None:
-    #         lower_bound = np.clip(lower_bound,
-    #                               a_min=
-    #                               np.asarray(interval, dtype=float).T[0],
-    #                               a_max=None)
-    #         upper_bound = np.clip(upper_bound,
-    #                               a_min=None,
-    #                               a_max=
-    #                               np.asarray(interval, dtype=float).T[1])
-    #
-    #     return upper_bound, lower_bound
-
-
     def initialize(self,
                    state_space,
                    state_positions,
@@ -219,26 +186,22 @@ class ElectricMotor:
         else:
             nominal_values_ = self._nominal_values
 
-
-        # for Induction motors seperate limitations
-        #if any(map(lambda state: state in self._initial_states.keys(),
-         #          ['psi_ralpha', 'psi_rbeta'])):
-            # draw a sample magnetic field angle from [-pi,pi]
-          #  eps_mag = 2 * np.pi * np.random.random_sample() - np.pi
[…]
