### 2020-06-16 da8f8d3
updated init.py in physical_systems, cleaned init() in mechanical loads
