### 2020-06-17 658ee65
added comments, merged issue 9, 30 - integrated both functionalitis as far as possible
