### 2020-06-17 28da6bd
fixed initialization, merged issue 9 and 30 together
