### 2020-06-18 78e41d5
updated valid controllers for DC-series
