+++ examples/pi_series_omega_control.py
@@ -13,7 +13,7 @@ from gym_electric_motor.visualization import MotorDashboard
 
 if __name__ == '__main__':
     env = gem.make(
-        'emotor-dc-series-cont-v1',
+        'DcSeriesCont-v1',   # replace with 'DcSeriesDisc-v1' for discrete controllers
         # Pass an instance
         visualization=MotorDashboard(plotted_variables='all', visu_period=1),
         motor_parameter=dict(r_a=15e-3, r_e=15e-3, l_a=1e-3, l_e=1e-3),
@@ -29,7 +29,7 @@ if __name__ == '__main__':
             ], p=[0.1, 0.8, 0.1], super_episode_length=(1000, 10000)
         )
     )
-    controller = Controller.make('cascaded_pi', env)
+    controller = Controller.make('pi_controller', env)  # works for 'on_off'(disc),three_point(disc),'p_controller', 'cascaded pi' too
     state, reference = env.reset()
     start = time.time()
     cum_rew = 0
@@ -37,5 +37,7 @@ if __name__ == '__main__':
         env.render()
         action = controller.control(state, reference)
         (state, reference), reward, done, _ = env.step(action)
+        if done:
+            env.reset()
         cum_rew += reward
     print(cum_rew)
