### 2020-06-18 ba78773
updated valid controllers for DC-Perm
