+++ examples/perm_dc_omega.py
@@ -8,7 +8,7 @@ from gym_electric_motor.visualization import MotorDashboard
 
 if __name__ == '__main__':
     env = gem.make(
-        'DcPermExDisc-v1',
+        'DcPermExDisc-v1',  # replace with 'DcPermExCont-v1' for continuous mode
         # Pass an instance
         visualization=MotorDashboard(plotted_variables=['omega', 'torque', 'i', 'u', 'u_sup'], visu_period=1),
         # motor_parameter=dict(r_a=15e-3, r_e=15e-3, l_a=1e-3, l_e=1e-3),
@@ -23,7 +23,7 @@ if __name__ == '__main__':
             ], p=[0.1, 0.8, 0.1], super_episode_length=(1000, 10000)
         )
     )
-    controller = Controller.make('three_point', env) #'on_off'  works for on_off too
+    controller = Controller.make('three_point', env)   # works for 'on_off','pi_controller'(cont),'p_controller'(cont), 'cascaded_pi'(cont) too
     state, reference = env.reset()
     start = time.time()
     cum_rew = 0
