### 2020-06-20 a3f5fbf
bugfix in motor_dashboard + common X-axis + Dark mode added
