+++ examples/pi_series_omega_control.py
@@ -16,7 +16,7 @@ if __name__ == '__main__':
         'emotor-dc-series-cont-v1',
         # Pass an instance
         #visualization=MotorDashboard(plotted_variables='all', visu_period=1),
-        visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u', 'action_0', 'reward']),
+        visualization=MotorDashboard(plots=['omega','reward', 'i'], dark_mode=True),
         motor_parameter=dict(r_a=15e-3, r_e=15e-3, l_a=1e-3, l_e=1e-3),
         # Take standard class and pass parameters (Load)
         load_parameter=dict(a=0.01, b=.1, c=0.1, j_load=.06),
