+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -144,8 +144,7 @@ class StatePlot(MotorDashboardPlot):
         if self._limits == self._state_space[1]:
             self._normalized = False
 
-        # self._action_space = ps.action_space
-        # a=5
+
 
     def initialize(self, axis):
         super().initialize(axis)
@@ -162,7 +161,7 @@ class StatePlot(MotorDashboardPlot):
 
         self._axis.set_ylim(min_limit - 0.1 * (max_limit - min_limit), max_limit + 0.1*(max_limit - min_limit))
         self._axis.set_xlim(0, self.x_width)
-        self._axis.set_xlabel('t/s')
+        #self._axis.set_xlabel('t/s')
         y_label = self.state_labels.get(self._state)  # fix
         self._axis.set_ylabel(y_label)
 
@@ -246,7 +245,7 @@ class RewardPlot(MotorDashboardPlot):
         spacing = 0.1 * (max_limit - min_limit)
         self._axis.set_ylim(min_limit - spacing, max_limit + spacing)
         self._axis.set_xlim(0, self.x_width)
-        self._axis.set_xlabel('t/s')
+        #self._axis.set_xlabel('t/s')
         y_label = 'reward'
         self._axis.set_ylabel(y_label)
         # adds a constant line at 0 which is eventually updated by the plot variable values. legend can be set here.
@@ -329,7 +328,7 @@ class ActionPlot(MotorDashboardPlot):
         spacing = (act_min - act_max) * 0.1
         self._axis.set_ylim(act_min - spacing, act_max + spacing)
         self._axis.set_xlim(0, self.x_width)
-        self._axis.set_xlabel('t/s')
+        #self._axis.set_xlabel('t/s')
         self._axis.set_ylabel(self._action)
         base_action_line = lin.Line2D([], [], color=self.action_line_cfg['color'])
         self._axis.legend((base_action_line,), (self._action,), loc='upper left')
