### 2020-06-21 daa1900
New Visu interface for Console Printer
