+++ examples/ddpg_pmsm_dq_current_control.py
@@ -74,14 +74,13 @@ if __name__ == '__main__':
     env = gem.make(
         'emotor-pmsm-cont-v1',
         # Pass a class with extra parameters
-        visualization=MotorDashboard(['i_sq', 'i_sd', 'action_1', 'action_0']), visu_period=1,
+        visualization=MotorDashboard(plots=['i_sq', 'i_sd', 'action_1', 'action_0']), visu_period=1,
         load=ConstantSpeedLoad(omega_fixed=1000 * np.pi / 30),
         control_space='dq',
         # Pass a string (with extra parameters)
         ode_solver='scipy.solve_ivp', solver_kwargs={},
         # Pass an instance
         reference_generator=rg,
-        #plotted_variables=['i_sq', 'i_sd', 'u_sq', 'u_sd', 'reward'],
         reward_weights={'i_sq': 1000, 'i_sd': 1000},
         reward_power=0.5,
         observed_states=['i_sq', 'i_sd'],
