### 2020-06-22 cd4fcff
Fixed a bug that resulted from a wrong execution order. Also fixed misusage of integers as booleans
