### 2020-06-22 8149d24
Test fixes for real testing of system equation
