### 2020-06-22 3edf8c8
remove qt5 dependency from requirements after vis refactoring
