+++ requirements.txt
@@ -4,4 +4,4 @@ scipy>=1.4.1
 gym>=0.15.4
 pytest>=5.2.2
 pytest-cov
-PyQT5==5.9.2
+
