### 2020-06-22 666baeb
restored changes for only issue 9 and 30
