+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -1,4 +1,5 @@
 import numpy as np
+from scipy.stats import truncnorm
 
 
 class MechanicalLoad:
@@ -8,6 +9,22 @@ class MechanicalLoad:
     It contains an mechanical ode system as well as the state names, limits and nominal values
     of the mechanical quantities. The only required state is 'omega' as the rotational speed of the motor shaft
     in rad/s.
+
+    ExternalSpeedLoad and ConstantSpeedLoad should be initialized with the own
+    initial value in class initialisation, when definded external
+    # todo updateting intialization with motor_initializer
+
+    Initialization is given by initializer(dict). Can be constant state value
+    or random value in given or nominal interval.
+    dict should be like:
+        { 'states': <dict with state names and initital values>,
+          'interval': < boundaries for each state (only for random init>,
+          'random_init': <str: 'uniform' or 'normal'>,
+          'random_params: {'mue': <const>, 'sigma': <const>}
+    Example initializer(dict) for constant initialization:
+        { 'states': {'omega': 15.5}}
+    Example  initializer(dict) for random initialization:
+        { 'random_init': 'normal'}
     """
 
     @property
@@ -43,12 +60,22 @@ class MechanicalLoad:
         """
         return self._nominal_values
 
+    @property
+    def initializer(self):
+        """
+        Returns:
+            dict: The motors initial state and additional initializer parameters
+        """
+        return self._initializer
+
     OMEGA_IDX = 0
 
     #: Parameter indicating if the class is implementing the optional jacobian function
     HAS_JACOBIAN = False
+    #: _default_initial_state(dict): Default initial motor-state values
+    _default_initializer = {}
 
-    def __init__(self, state_names=None, j_load=0.0, **__):
+    def __init__(self, state_names=None, j_load=0.0, load_initializer=None, **__):
         """
         Args:
             state_names(list(str)): List of the names of the states in the mechanical-ODE.
@@ -58,15 +85,124 @@ class MechanicalLoad:
         self._state_names = list(state_names or ['omega'])
         self._limits = {}
         self._nominal_values = {}
+        load_initializer = load_initializer or {}
+        self._initializer = self._default_initializer.copy()
+        self._initializer.update(load_initializer)
+        self._initial_states = self._initializer['states']
 
-    def reset(self, *_, **__):
+    def initialize(self,
+                   nominal_state,
+                   state_space=None,
+                   state_positions=None,
+                   **__):
         """
-        Reset the motors mechanical-ODE to an initial state (default: all zeros).
+        Initializes give state values. Values can be given as a constant or
+        sampled random out of a statistical distribution. Initial value is in
+        range of the nominal values or a given interval.
 
+        Args:
+            nominal_state(list): nominal values for each state given from
+                                  physical system
+            state_space(gym.Box): normalized state space boundaries
+            state_positions(dict): indexes of system states
         Returns:
-            ndarray(float): Initial state of the mechanical-ODE.
+            ndarray(float): initial value for each state
+
         """
-        return np.zeros_like(self._state_names, dtype=float)
+        # for order and organization purposes
+        interval = self._initializer['interval']
+        random_dist = self._initializer['random_init']
+        random_params = self._initializer['random_params']
+
+        # setting nominal values as interval limits
+        state_idx = [state_positions[state] for state in self._initial_states.keys()]
+        upper_bound = np.asarray(nominal_state[state_idx], dtype=float)
+        lower_bound = upper_bound * np.asarray(state_space.low, dtype=float)[state_idx]
+        # clip nominal boundaries to user defined
+        if interval is not None:
+            lower_bound = np.clip(lower_bound,
+                                  a_min=
+                                  np.asarray(interval, dtype=float).T[0],
+                                  a_max=None)
[…]

[…]
