### 2020-06-23 fcdbfab
Pull Request fixes
