### 2020-06-25 6c42ba9
added first tests for ExternalSpeedload and updated tests
