+++ tests/test_physical_systems/test_mechanical_loads.py
@@ -1,12 +1,30 @@
 import pytest
 import gym_electric_motor as gem
-from gym_electric_motor.physical_systems import PolynomialStaticLoad, MechanicalLoad, ConstantSpeedLoad
+from gym_electric_motor.physical_systems import PolynomialStaticLoad, MechanicalLoad, ConstantSpeedLoad, ExternalSpeedLoad
+from gym.spaces import Box
 import numpy as np
+from scipy import signal
 import math
 
 # The load parameter values used in the test
 load_parameter1 = {'j_load': 0.2, 'state_names': ['omega'], 'j_rot_load': 0.25, 'omega_range': (0, 1),
                    'parameter': dict(a=0.12, b=0.13, c=0.4, j_load=0.2)}
+# The different initializers used in test
+test_const_initializer = {'states': {'omega': 15.0},
+                          'interval': None,
+                          'random_init': None,
+                          'random_params': (None, None)}
+# todo rand init
+test_rand_initializer = { 'interval': None,
+                          'random_init': None,
+                          'random_params': (None, None)}
+# profile_parameter
+test_amp = 20
+test_bias = 10
+test_freq = 2
+# simple triangular profile as example
+test_speed_profile = (lambda t, amp, freq, bias:
+                      amp*signal.sawtooth(2*np.pi*freq*t, width=0.5)+bias)
 
 
 @pytest.fixture
@@ -27,7 +45,10 @@ def concreteMechanicalLoad():
     # Parameters picked from the parameter dict above
     state_names = load_parameter1['state_names']
     j_load = load_parameter1['j_load']
-    return MechanicalLoad(state_names, j_load, a=2, b=6)
+    test_initializer = test_const_initializer
+    # todo random init return
+    return MechanicalLoad(state_names, j_load, a=2, b=6,
+                          load_initializer=test_initializer)
 
 
 @pytest.fixture
@@ -47,7 +68,11 @@ def concretePolynomialLoad():
     """
     addnl_params = dict(p=0.5, q=0.98, r=2678.88)
     test_load_params = dict(a=0.01, b=0.05, c=0.1, j_load=0.1, p=0.5, q = 0.98, r= 2678.88)
-    return PolynomialStaticLoad(load_parameter=test_load_params, x=3, y=76)  # x, y are random kwargs
+    test_initializer = test_const_initializer
+    # x, y are random kwargs
+    # todo random init
+    return PolynomialStaticLoad(load_parameter=test_load_params, x=3, y=76.,
+                                load_initializer=test_initializer)
 
 
 def test_InitMechanicalLoad(defaultMechanicalLoad, concreteMechanicalLoad):
@@ -60,9 +85,11 @@ def test_InitMechanicalLoad(defaultMechanicalLoad, concreteMechanicalLoad):
     assert defaultMechanicalLoad.state_names == load_parameter1['state_names']
     assert defaultMechanicalLoad.j_total == 0
     assert defaultMechanicalLoad.limits == {}
+    assert defaultMechanicalLoad.initializer == {}
     assert concreteMechanicalLoad.state_names == load_parameter1['state_names']
     assert concreteMechanicalLoad.j_total == load_parameter1['j_load']
     assert concreteMechanicalLoad.limits == {}
+    assert concreteMechanicalLoad.initializer == test_const_initializer
 
 
 def test_MechanicalLoad_set_j_rotor(concreteMechanicalLoad):
@@ -86,15 +113,25 @@ def test_MechanicalLoad_MechanicalOde(concreteMechanicalLoad):
     with pytest.raises(NotImplementedError):
         concreteMechanicalLoad.mechanical_ode(*test_args)
 
-
[…]
