### 2020-06-25 704f3bd
added first tests for ExternalSpeedload and updated tests
