### 2020-06-25 ae71bfe
basic tests for mechanical loads added
