### 2020-06-25 5b0a585
basic tests for mechanical loads added
