### 2020-06-25 6a0c8d4
revert allowance of higher pyqt5 versions
