### 2020-06-25 0931e8b
revert allowance of higher pyqt5 versions again
