### 2020-06-27 01a38e6
added some test cases and fixed some problems with new implemented functions, added necessary updates in conf.py. Only 2 errors in test_e_m and 38 in test_env pending
