+++ tests/test_physical_systems/test_electric_motors.py
@@ -12,14 +12,22 @@ g_i_e = [0, 0.2, 0.5, -1, 1]
 g_u_in = [0, 400, -400, 325.2, -205.4]
 g_t_load = [0, 1, 5, -5, 4.35]
 
+# default/ global initializer
+test_initializer = {'states': {},
+                    'interval': None,
+                    'random_init': None,
+                    'random_params': None}
 
 # region general test functions
 
 
-def motor_testing(motor_1_default, motor_2_default, motor_1, motor_2, motor_state, limit_values, nominal_values,
-                  motor_parameter):
-    motor_reset_testing(motor_1_default, motor_2_default, motor_state)
-    motor_reset_testing(motor_1, motor_2, motor_state)
+def motor_testing(motor_1_default, motor_2_default, motor_1, motor_2,
+                  motor_state, limit_values, nominal_values, motor_parameter,
+                  state_positions, state_space):
+    motor_reset_testing(motor_1_default, motor_2_default, motor_state,
+                        state_positions, state_space, nominal_values)
+    motor_reset_testing(motor_1, motor_2, motor_state, state_positions,
+                        state_space, nominal_values)
 
     limit_nominal_testing(motor_1_default, motor_2_default, motor_1, motor_2, limit_values, nominal_values)
     motor_parameter_testing(motor_1_default, motor_2_default, motor_1, motor_2, motor_parameter)
@@ -45,31 +53,60 @@ def motor_parameter_testing(motor_1_default, motor_2_default, motor_1, motor_2, 
     assert motor_1_default.motor_parameter is not motor_1.motor_parameter, "Failed motor parameter test"
 
 
-def motor_reset_testing(motor_1, motor_2, motor_state):
+def motor_reset_testing(motor_1, motor_2, motor_state, state_positions,
+                        state_space, nominal_values):
     # tests if the reset function works correctly
-    assert (motor_1.reset() == motor_2.reset()).all()
-    assert all(motor_1.reset() == np.zeros(len(motor_state))), "Reset is not zero"
-
+    assert motor_1.initializer == motor_2.initializer
+    # test random initialization
+    if motor_1.initializer['states'] is None:
+        for idx, state in enumerate(motor_state):
+            assert np.abs(motor_1.reset(state_space, state_positions)[idx] \
+                     <= nominal_values[state])
+            assert np.abs(motor_2.reset(state_space, state_positions)[idx] \
+                     <= nominal_values[state])
+    # test constant initialization
+    else:
+        assert all(motor_1.reset(state_space, state_positions) ==
+                motor_2.reset(state_space, state_positions))
+        init_values = list(motor_1.initializer['states'].values())
+        assert all(motor_1.reset(state_space, state_positions) == init_values)
 
 # endregion
 
 # region DcSeriesMotor
-
-
-def test_dc_series_motor():
-    motor_state = ['i', 'i']  # list of state names
+@pytest.mark.parametrize(
+    'states, interval, random_init, random_params',
+    [({'i': 16}, None, None, (None, None)),
+     (None, None, 'uniform', (None, None)),
+     (None, [[10, 20]], 'uniform', (None, None)),
+     (None, None, 'gaussian', (None, None)),
+     (None, None, 'gaussian', (16, 2)),
+     (None, [[10, 20]], 'gaussian', (None, None))])
+def test_dc_series_motor(states, interval, random_init, random_params):
+    #motor_state = ['i', 'i']  # list of state names
+    motor_state = ['i']  # list of state names
     motor_parameter = test_motor_parameter['DcSeries']['motor_parameter']
     nominal_values = test_motor_parameter['DcSeries']['nominal_values']  # dict
     limit_values = test_motor_parameter['DcSeries']['limit_values']  # dict
[…]
