### 2020-06-29 6df7918
Added 1phase and 3phase AC voltage supply and made all voltage supply returns to lists
