+++ gym_electric_motor/physical_systems/voltage_supplies.py
@@ -48,7 +48,7 @@ class VoltageSupply:
             t(float): Current time of the system.
 
         Returns:
-             float: Supply Voltage at time t.
+             list(float): Supply Voltage at time t.
         """
         raise NotImplementedError
 
@@ -65,7 +65,7 @@ class IdealVoltageSupply(VoltageSupply):
 
     def get_voltage(self, *_, **__):
         # Docstring of superclass
-        return self._u_nominal
+        return [self._u_nominal]
 
 
 class RCVoltageSupply(VoltageSupply):
@@ -75,7 +75,7 @@ class RCVoltageSupply(VoltageSupply):
         """
         This Voltage Supply is a model of a non ideal Voltage supply where the ideal voltage source U_0 is part of an RC element
         Args:
-            supply_parameter(dict): Consists or Resistance R in Ohm and Capacitance C in Farad
+            supply_parameter(dict): Consists of Resistance R in Ohm and Capacitance C in Farad
         Additional notes:
             If the product of R and C get too small the numerical stability of the ODE is not given anymore
             typical time differences tau are only in the range of 10e-3. One might want to consider R*C as a
@@ -91,7 +91,7 @@ class RCVoltageSupply(VoltageSupply):
         self._c = supply_parameter['C']
         if self._r*self._c < 1e-4:
             warnings.warn("The product of R and C might be too small for the correct calculation of the supply voltage. You might want to consider R*C as a time constant.")
-        self._u_sup = u_nominal
+        self._u_sup = [u_nominal]
         self._u_0 = u_nominal
         self._solver = EulerSolver()
         self._solver.set_system_equation(self.system_equation)
@@ -104,12 +104,60 @@ class RCVoltageSupply(VoltageSupply):
         # Docstring of superclass
         # On reset the capacitor is loaded again
         self._solver.set_initial_value(np.array([self._u_0]))
-        self._u_sup = self._u_0
+        self._u_sup = [self._u_0]
         return self._u_sup
     
     def get_voltage(self, t,i_sup):
         # Docstring of superclass
         self._solver.set_f_params(self._u_0, i_sup, self._r, self._c)
-        self._u_sup = self._solver.integrate(t)[0]
+        self._u_sup = self._solver.integrate(t)
         return self._u_sup
+    
+class AC1PhaseSupply(VoltageSupply):
+    
+    def __init__(self, u_nominal=230, supply_parameter=None, **__):
+        super().__init__(u_nominal)
+        supply_parameter = supply_parameter or {'frequency': 50, 'phase': np.random.rand()*2*np.pi, 'fixed_phase': False}
+        assert 'frequency' in supply_parameter.keys(), "Pass key 'frequency' for frequency f in Hz in your dict"
+        assert 'phase' in supply_parameter.keys(), "Pass key 'phase' for the starting phase in in your dict"
+        assert 'fixed_phase' in supply_parameter.keys(), "Pass key 'fixed_phase' as Boolean to indicate whether the phase given by the user should be used for every episode"
+        assert type(supply_parameter['fixed_phase']) == type(True), "Value for key 'fixed_phase' should be Boolean"
+        self._fixed_phi = supply_parameter['fixed_phase']
+        self._f = supply_parameter['frequency']
+        self._phi = supply_parameter['phase']
+        
+    def reset(self):
+        if not self._fixed_phi:
+            self._phi = np.random.rand()*2*np.pi
+        return self.get_voltage(0)
+    
+    def get_voltage(self, t, **__):
+        # Docstring of superclass
+        self._u_sup = [self._u_nominal*np.sin(self._f*t + self._phi)]
+        return self._u_sup
+
[…]
