### 2020-06-29 32a72c4
Made SCML systems compatible to 1-dimensional voltage arrays
