### 2020-06-29 f8ab515
Added keys for the newly introduced voltage supplies. Fixed some bugs
