### 2020-06-29 a82515f
qd reordering ( visualisation not working )
