+++ gym_electric_motor/physical_systems/physical_systems.py
@@ -359,12 +359,12 @@ class ThreePhaseMotorSystem(SCMLSystem):
             normed_epsilon(bool): True, if epsilon is normed to [-1,1] else in [-pi, pi] (default)
 
         Returns:
-            (quantity_q, quantity_d): The quantities in the dq-space
+            (quantity_d, quantity_q): The quantities in the dq-space
         """
         if normed_epsilon:
             epsilon_el *= np.pi
         dq_quantity = self._electrical_motor.q_inv(self._electrical_motor.t_23(abc_quantities), epsilon_el)
-        return dq_quantity[::-1]
+        return dq_quantity
 
     def dq_to_abc_space(self, dq_quantities, epsilon_el, normed_epsilon=False):
         """
@@ -444,8 +444,8 @@ class SynchronousMotorSystem(ThreePhaseMotorSystem):
         # Docstring of superclass
         return (
             self._mechanical_load.state_names +['torque',
-                                                'i_a', 'i_b', 'i_c', 'i_sq', 'i_sd',
-                                                'u_a', 'u_b', 'u_c', 'u_sq', 'u_sd',
+                                                'i_a', 'i_b', 'i_c', 'i_sd', 'i_sq',
+                                                'u_a', 'u_b', 'u_c', 'u_sd', 'u_sq',
                                                 'epsilon', 'u_sup',
                                                 ]
         )
@@ -486,8 +486,8 @@ class SynchronousMotorSystem(ThreePhaseMotorSystem):
             u_sup = self._supply.get_voltage(self._t, i_sup)
             u_in = self._converter.convert(i_in, self._ode_solver.t)
             u_in = [u * u_sup for u in u_in]
-            u_qd = self.abc_to_dq_space(u_in, eps)
-            self._ode_solver.set_f_params(u_qd)
+            u_dq = self.abc_to_dq_space(u_in, eps)
+            self._ode_solver.set_f_params(u_dq)
             ode_state = self._ode_solver.integrate(t)
             eps = ode_state[self._ode_epsilon_idx]
             i_in = self.dq_to_abc_space(self._electrical_motor.i_in(ode_state[self._ode_currents_idx]), eps)
@@ -496,17 +496,17 @@ class SynchronousMotorSystem(ThreePhaseMotorSystem):
         u_sup = self._supply.get_voltage(self._t, i_sup)
         u_in = self._converter.convert(i_in, self._ode_solver.t)
         u_in = [u * u_sup for u in u_in]
-        u_qd = self.abc_to_dq_space(u_in, eps)
-        self._ode_solver.set_f_params(u_qd)
+        u_dq = self.abc_to_dq_space(u_in, eps)
+        self._ode_solver.set_f_params(u_dq)
         ode_state = self._ode_solver.integrate(self._t + self._tau)
         self._t = self._ode_solver.t
         self._k += 1
         torque = self._electrical_motor.torque(ode_state[self._motor_ode_idx])
         noise = self._noise_generator.noise()
         mechanical_state = ode_state[self._load_ode_idx]
-        i_qd = ode_state[self._ode_currents_idx]
+        i_dq = ode_state[self._ode_currents_idx]
         i_abc = list(
-            self.dq_to_abc_space(i_qd, eps)
+            self.dq_to_abc_space(i_dq, eps)
         )
         eps = ode_state[self._ode_epsilon_idx] % (2 * np.pi)
         if eps > np.pi:
[…]
