### 2020-06-29 b1e55a4
Added NoConverter class which throughputs the voltage supply
