### 2020-06-29 00d1fc7
Added conversion from effective value to sinus amplitude
