### 2020-06-29 d1e1dbe
Fixed conversion for the three phase supply
