+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib.lines as lin
+from gym.spaces import Box
 
 
 class MotorDashboardPlot:
@@ -15,6 +16,7 @@ class MotorDashboardPlot:
             axis(matplotlib.pyplot.axis): Axis to plot in
         """
         self._axis = axis
+        self._axis.grid(True)
 
     def set_modules(self, ps, rg, rf):
         """Interconnection of the environments modules.
@@ -52,7 +54,7 @@ class StatePlot(MotorDashboardPlot):
     """Class to plot any motor state and its reference."""
 
     # Width of the Plot in seconds
-    x_width = 3
+    x_width = 1
 
     # Either "continuous" or "repeating"
     mode = 'continuous'
@@ -144,11 +146,8 @@ class StatePlot(MotorDashboardPlot):
         if self._limits == self._state_space[1]:
             self._normalized = False
 
-
-
     def initialize(self, axis):
         super().initialize(axis)
-        axis.grid()
         if self._referenced:
             self._reference_line, = self._axis.plot(self._t_data, self._ref_data, **self.reference_line_cfg)
         self._state_line, = self._axis.plot(self._t_data, self._state_data, **self.state_line_cfg)
@@ -160,8 +159,8 @@ class StatePlot(MotorDashboardPlot):
         lim = self._axis.axhline(max_limit, **self.limit_line_cfg)
 
         self._axis.set_ylim(min_limit - 0.1 * (max_limit - min_limit), max_limit + 0.1*(max_limit - min_limit))
-        self._axis.set_xlim(0, self.x_width)
-        y_label = self.state_labels.get(self._state)  # fix
+        self._axis.set_xlim(-self.x_width, 0)
+        y_label = self.state_labels.get(self._state, self._state)
         self._axis.set_ylabel(y_label)
 
         self._t_data = np.linspace(0, self.x_width, self._x_points, endpoint=False).tolist()
@@ -187,12 +186,12 @@ class StatePlot(MotorDashboardPlot):
         if self._referenced:
             self._ref_data[idx] = ref
         if done:
-            self.update()
             self._axis.axvline(self._t, color='red', linewidth=1)
 
     def update(self):
         state_data = self._state_data
         ref_data = self._ref_data
+
         if self._normalized:
             state_data = state_data * self._limits
             if self._referenced:
@@ -211,7 +210,7 @@ class RewardPlot(MotorDashboardPlot):
     """ Class used to plot the instantaneous reward during the episode
     """
 
-    x_width = 3
+    x_width = 1
     mode = 'continuous'
     reward_line_cfg = {
         'color': 'gray',
@@ -222,7 +221,6 @@ class RewardPlot(MotorDashboardPlot):
     }
 
     def __init__(self):
-
         self._reward_range = None
         self._reward_line = None
         self._reward_data = None
@@ -233,9 +231,7 @@ class RewardPlot(MotorDashboardPlot):
         super().__init__()
 
     def initialize(self, axis):
-
         super().initialize(axis)
-        axis.grid()
         self._t_data = np.linspace(0, self.x_width, self._x_points, endpoint=False)
         self._reward_data = np.zeros_like(self._t_data, dtype=float)
         self._reward_line, = self._axis.plot(self._t_data, self._reward_data, **self.reward_line_cfg)
@@ -243,7 +239,7 @@ class RewardPlot(MotorDashboardPlot):
         max_limit = self._reward_range[1]
         spacing = 0.1 * (max_limit - min_limit)
         self._axis.set_ylim(min_limit - spacing, max_limit + spacing)
-        self._axis.set_xlim(0, self.x_width)
+        self._axis.set_xlim(-self.x_width, 0)
[…]
