### 2020-06-29 6f8428b
bugfix after speedup
