### 2020-06-29 fb06ba8
added docstrings and documentation file for extspeedload
