### 2020-06-29 c45d136
added docstrings and documentation file for extspeedload
