### 2020-06-30 7c162c8
Added comments for the new voltage supplies
