### 2020-06-30 98252cc
Added few comments to new NoConverter class
