### 2020-06-30 11d92bd
Added supplies to the docs
