+++ docs/parts/physical_systems/voltage_supplies/voltage_supply.rst
@@ -7,6 +7,8 @@ Voltage Supplies
 
     ideal_voltage_supply
     rc_voltage_supply
+    ac_1_phase_supply
+    ac_3_phase_supply
 
 
 Voltage Supply Base Class
