+++ gym_electric_motor/reward_functions/__init__.py
@@ -1,6 +1,8 @@
 from .weighted_sum_of_errors import WeightedSumOfErrors, ShiftedWeightedSumOfErrors
+from .constraint_monitor import ConstraintMonitor
 from ..utils import register_class
 from .. import RewardFunction
 
 register_class(WeightedSumOfErrors, RewardFunction, 'WSE')
 register_class(ShiftedWeightedSumOfErrors, RewardFunction, 'SWSE')
+register_class(ConstraintMonitor, RewardFunction, 'ConMon')
