### 2020-07-06 30f36ac
Added test for the 1 phase AC supply
