+++ tests/test_physical_systems/test_voltage_supplies.py
@@ -137,7 +137,7 @@ class TestAC1PhaseSupply(TestVoltageSupply):
     def test_get_voltage(self):
         """Test the get voltage function for different times t."""
         supply_parameter = {'frequency': 1, 'phase': 0, 'fixed_phase': True}
-        supply = vs.AC1PhaseSupply(supply_parameter = supply_parameter)
+        supply = vs.AC1PhaseSupply(supply_parameter=supply_parameter)
         
         times = [0, 2*np.pi, 4*np.pi]
         for time in times:
@@ -150,7 +150,75 @@ class TestAC1PhaseSupply(TestVoltageSupply):
         times = [3*np.pi/2, 7*np.pi/2, 11*np.pi/2]
         for time in times:
             assert supply.get_voltage(time) == pytest.approx([-230.0 * np.sqrt(2)])
+          
+        #"Hand" calculated
+        supply_parameter = {'frequency': 36, 'phase': 0.5, 'fixed_phase': True}
+        supply = vs.AC1PhaseSupply(supply_parameter=supply_parameter)
+        assert supply.get_voltage(1) == pytest.approx([-303.058731])
+        assert supply.get_voltage(2) == pytest.approx([-78.381295])
+        assert supply.get_voltage(3) == pytest.approx([323.118651])
 
 
+class TestAC3PhaseSupply(TestVoltageSupply):
+    key = 'AC3PhaseSupply'
+    class_to_test = vs.AC3PhaseSupply
     
+    def test_default_initialization(self):
+        """Test for default initialization values"""
+        voltage_supply = vs.AC3PhaseSupply()
+        assert voltage_supply.u_nominal == 400.0
+        assert voltage_supply._max_amp == 400.0 / np.sqrt(3) * np.sqrt(2) 
+        assert voltage_supply.supply_range == [-400.0 / np.sqrt(3) * np.sqrt(2), 400.0 / np.sqrt(3) * np.sqrt(2)]
+        assert voltage_supply._fixed_phi == False
+        assert voltage_supply._f == 50
+        
+    def test_reset(self, u_nominal=400.0):
+        """Test the reset function for correct behavior on fixed phase""" 
+        supply_parameter = {'frequency': 50, 'phase': 0, 'fixed_phase': False}
+        voltage_supply = vs.AC3PhaseSupply(u_nominal, supply_parameter)
+        assert voltage_supply._phi == 0
+        _ = voltage_supply.reset()
+        assert voltage_supply._phi != 0, "Test this again and if this error doesn't appear next time you should consider playing lotto"
+
+        supply_parameter = {'frequency': 50, 'phase': 0, 'fixed_phase': True}
+        voltage_supply = vs.AC3PhaseSupply(u_nominal, supply_parameter)
+        assert voltage_supply._phi == 0
+        _ = voltage_supply.reset()
+        assert voltage_supply._phi == 0
+
+        
+    def test_initialization(self, u_nominal = 300.0):
+        supply_parameter = {'frequency': 35, 'phase': 0, 'fixed_phase': True}
+        voltage_supply = vs.AC3PhaseSupply(u_nominal, supply_parameter)
+        assert voltage_supply.u_nominal == 300.0
+        assert voltage_supply._max_amp == 300.0 / np.sqrt(3) * np.sqrt(2)
+        assert voltage_supply.supply_range == [-300.0 / np.sqrt(3) * np.sqrt(2), 300.0 / np.sqrt(3) * np.sqrt(2)]
+        assert voltage_supply._fixed_phi == True
+        assert voltage_supply._phi == 0
+        assert voltage_supply._f == 35
+        
+    def test_get_voltage(self):
[…]
