### 2020-07-06 21c4874
Changed DummyVoltageSupply to fit new voltage returns
