### 2020-07-06 dd07b4d
reduntant code removed, interfaces cleaned
