+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -8,6 +8,10 @@ class MotorDashboardPlot:
 
     def __init__(self):
         self._axis = None
+        self._tau = None
+        self._x_points = None
+        self._t = 0
+        self._t_data = []
 
     def initialize(self, axis):
         """Initialization of the plot. Set labels, legends,... here.
@@ -17,6 +21,7 @@ class MotorDashboardPlot:
         """
         self._axis = axis
         self._axis.grid(True)
+        self._axis.set_xlim(-self.x_width, 0)
 
     def set_modules(self, ps, rg, rf):
         """Interconnection of the environments modules.
@@ -26,7 +31,9 @@ class MotorDashboardPlot:
             rg(ReferenceGenerator): The ReferenceGenerator of the environment.
             rf(RewardFunction): The RewardFunction of the environment
         """
-        pass
+        self._tau = ps.tau
+        # Number of points on the x-axis in a plot (= x_width / tau)
+        self._x_points = int(self.x_width / self._tau)
 
     def step(self, k, state, reference, action, reward, done):
         """Passing of current environmental information..
@@ -43,7 +50,13 @@ class MotorDashboardPlot:
 
     def update(self):
         """Called by the MotorDashboard each time before the figure is updated."""
-        pass
+        # configure x-axis properties
+        if self.mode == 'continuous':
+            x_lim = self._axis.get_xlim()
+            upper_lim = max(self._t, x_lim[1])
+            lower_lim = upper_lim - self.x_width
+            self._axis.set_xlim(lower_lim, upper_lim)
+
 
     def reset(self):
         """Called by the MotorDashboard each time the environment is reset."""
@@ -125,24 +138,18 @@ class StatePlot(MotorDashboardPlot):
         # Data containers
         self._state_data = []
         self._ref_data = []
-        self._t_data = []
 
-        self._tau = None
         # Flag, if the passed data is normalized
         self._normalized = True
-        # Number of points on the x-axis in a plot (= x_width / tau)
-        self._x_points = None
-        self._t = 0
         super().__init__()
 
     def set_modules(self, ps, rg, rf):
         # Docstring of superclass
+        super(StatePlot, self).set_modules(ps, rg, rf)
         self._state_idx = ps.state_positions[self._state]
         self._limits = ps.limits[self._state_idx]
         self._state_space = ps.state_space.low[self._state_idx], ps.state_space.high[self._state_idx]
         self._referenced = rg.referenced_states[self._state_idx]
-        self._tau = ps.tau
-        self._x_points = int(self.x_width / self._tau)
         if self._limits == self._state_space[1]:
             self._normalized = False
 
@@ -159,7 +166,6 @@ class StatePlot(MotorDashboardPlot):
         lim = self._axis.axhline(max_limit, **self.limit_line_cfg)
 
         self._axis.set_ylim(min_limit - 0.1 * (max_limit - min_limit), max_limit + 0.1*(max_limit - min_limit))
-        self._axis.set_xlim(-self.x_width, 0)
         y_label = self.state_labels.get(self._state, self._state)
         self._axis.set_ylabel(y_label)
 
@@ -199,11 +205,8 @@ class StatePlot(MotorDashboardPlot):
         if self._referenced:
             self._reference_line.set_data(self._t_data, ref_data)
         self._state_line.set_data(self._t_data, state_data)
-        if self.mode == 'continuous':
-            x_lim = self._axis.get_xlim()
-            upper_lim = max(self._t, x_lim[1])
-            lower_lim = upper_lim - self.x_width
-            self._axis.set_xlim(lower_lim, upper_lim)
+
+        super(StatePlot, self).update()
 
 
[…]
