### 2020-07-06 f05b040
Added initial Callback class and modified the environment to support callbacks
