### 2020-07-06 ce9154e
Added AdaptiveLimitMargin callback
