### 2020-07-06 9fac60c
added P controller as separate class
