+++ examples/pi_series_omega_control.py
@@ -1,4 +1,3 @@
-
 import os
 import sys
 import time
@@ -14,36 +13,39 @@ from gym_electric_motor.visualization import MotorDashboard
 >> cd examples
 >> python pi_series_omega_control.py
 Description:
-        Environment to control a continuously controlled DC Series Motor.
+        Environment to control a Continuously controlled DC Series Motor.
 
         Controlled Quantity: 'omega'
 
-        Limitations: Physical limitations of the motor like Current,Speed
+        Limitations: Physical limitation of the  motor Current
 
         Converter : OneQuadrantConverter from converters.py
 """
 
 if __name__ == '__main__':
-    """Discrete mode: It is the mode where the input command causes the controlled variable to change
-     its value in discrete steps,with a brief static condition occurring between each step.
-    
-    Continuous mode is the one where the input command signal causes 
-    the controlled variable to remain constant,or to change in a linear(rather than step) mode. 
     """
-    env = gem.make(
-        'DcSeriesCont-v1',  # replace with 'DcSeriesDisc-v1' for discrete controllers
-        visualization=MotorDashboard(plots=['omega' , 'torque', 'i', 'u', 'u_sup'], dark_mode=True),
-        motor_parameter=dict(r_a=15e-3, r_e=15e-3, l_a=5e-3, l_e=5e-3),
-        load_parameter=dict(a=0.01, b=.1, c=0.1, j_load=.06),
-        ode_solver='scipy.solve_ivp', solver_kwargs=dict(),
-        # Pass a Class with extra parameters
-        reference_generator=rg.SwitchedReferenceGenerator(
-            sub_generators=[
-                rg.SinusoidalReferenceGenerator, rg.WienerProcessReferenceGenerator(), rg.StepReferenceGenerator()
-            ], p=[0.1, 0.8, 0.1], super_episode_length=(1000, 10000)
-        )
-    )
-    controller = Controller.make('pi_controller', env)
+           Continuous mode: The action is the average (normalized) voltage per time step which is assumed to be transferred to a PWM 
+                            converter to generate the switching sequence.
+
+           Discrete mode: The action is the switching state of the power converter i.e., a quantity from a discrete set of
+                          switching states which can be generated by the converter.   
+
+
+    """
+
+    env = gem.make('DcSeriesCont-v1',  # replace with 'DcSeriesDisc-v1' for discrete controllers
+                   state_filter=['omega', 'i'],
+                   # Pass an instance
+                   visualization=MotorDashboard(plots=['i', 'omega']),
+                   # Take standard class and pass parameters (Load)
+                   motor_parameter=dict(r_a=15e-3, r_e=15e-3, l_a=1e-3, l_e=1e-3),
+                   load_parameter=dict(a=0, b=.1, c=.1, j_load=0.04),
+                   # Pass a string (with extra parameters)
+                   ode_solver='scipy.solve_ivp', solver_kwargs={},
+                   # Pass a Class with extra parameters
+                   reference_generator=rg.WienerProcessReferenceGenerator())
+
+    controller = Controller.make('pi_controller',env)
     # The above controller can be replaced with 'on_off'(disc),three_point(disc),'p_controller', 'cascaded pi' from simple_controllers class
     state, reference = env.reset()
     start = time.time()
