+++ examples/perm_DC_omega.py
@@ -10,29 +10,26 @@ from gym_electric_motor.visualization import MotorDashboard
 """
 Run this file from within the 'examples' folder:
 >> cd examples
->> python perm_dc_omega.py
+>> python perm_DC_omega.py
 
 Description:
-        Environment to control a Continuously controlled DC Permanently Excited Motor.
-
+        Environment to control a Continuously Controlled DC Permanently Excited Motor.
         Controlled Quantity: 'omega'
-
-        Limitations: Physical limitations of the motor like Current,Speed
-
+        Limitations: Physical limitations of the motor will be Current.
         Converter : FourQuadrantConverter from converters.py
-
 """
 
 if __name__ == '__main__':
 
-    """Discrete mode: It is the mode where the input command causes the controlled variable to change its 
-    value in discrete steps,with a brief static condition occurring between each step.
-    
-    Continuous mode is the one where the input command signal causes  the controlled variable to remain constant,
-    or to change in a linear(rather than step) mode. 
+    """
+       Continuous mode: The action is the average (normalized) voltage per time step which is assumed to be transferred to a PWM/SVM 
+                        converter to generate the switching sequence.
+
+       Discrete mode: The action is the switching state of the power converter i.e., a quantity from a discrete set of
+                      switching states which can be generated by the converter.     
     """
     env = gem.make(
-        'DcPermExCont-v1',  # replace with 'DcPermExDisc-v1' for continuous mode
+        'DcPermExDisc-v1',  # replace with 'DcPermExCont-v1'  for continuous mode
         visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u', 'u_sup'], visu_period=1),
         ode_solver='scipy.solve_ivp', solver_kwargs=dict(),
         reference_generator=rg.SwitchedReferenceGenerator(
