+++ gym_electric_motor/core.py
@@ -724,3 +724,4 @@ class Callback:
         pass
     def on_close(self):
         pass
+    
