### 2020-07-07 59aa5f5
added separate voltage injection class
