+++ tests/test_physical_systems/test_electric_motors.py
@@ -347,28 +347,28 @@ def extex_motor_electrical_ode_testing(motor_1):
 def torque_testing(state, mp):
     """
     use this function as benchmark for the torque
-    :param state: [i_q, i_d, epsilon]
+    :param state: [i_d, i_q, epsilon]
     :param mp: motor parameter (dict)
     :return: generated torque
     """
-    return 1.5 * mp['p'] * (mp['psi_p'] + (mp['l_d'] - mp['l_q']) * state[1]) * state[0]
+    return 1.5 * mp['p'] * (mp['psi_p'] + (mp['l_d'] - mp['l_q']) * state[0]) * state[1]
 
 
 def synchronous_motor_ode_testing(state, voltage, omega, mp):
     """
     use this function as benchmark for the ode
-    :param state: [i_q, i_d, epsilon]
-    :param voltage: [u_q, u_d]
+    :param state: [i_d, i_q, epsilon]
+    :param voltage: [u_d, u_q]
     :param omega: angular velocity
     :param mp: motor parameter (dict)
     :return: ode system
     """
-    u_d = voltage[1]
-    u_q = voltage[0]
-    i_d = state[1]
-    i_q = state[0]
-    return np.array([(u_q - mp['r_s'] * i_q - omega * mp['p'] * (mp['l_d'] * i_d + mp['psi_p'])) / mp['l_q'],
-                     (u_d - mp['r_s'] * i_d + mp['l_q'] * omega * mp['p'] * i_q) / mp['l_d'],
+    u_d = voltage[0]
+    u_q = voltage[1]
+    i_d = state[0]
+    i_q = state[1]
+    return np.array([(u_d - mp['r_s'] * i_d + mp['l_q'] * omega * mp['p'] * i_q) / mp['l_d'],
+                     (u_q - mp['r_s'] * i_q - omega * mp['p'] * (mp['l_d'] * i_d + mp['psi_p'])) / mp['l_q'],
                      omega * mp['p']])
 
 
@@ -377,14 +377,13 @@ def synchronous_motor_ode_testing(state, voltage, omega, mp):
 def test_synchronous_motor_testing(motor_type, motor_class):
     """
     testing the synrm and pmsm
-    consider that it uses dq coordinates and the state is [i_q, i_d, epsilon]!!.
-    The same goes with the voltages u_qd and not as usual dq!!
+    consider that it uses dq coordinates and the state is [i_d, i_q, epsilon]!!.
+    voltage u_dq
     :return:
     """
     parameter = test_motor_parameter[motor_type]
-    # use this values for testing
-    state = np.array([0.5, 0.3, 0.68])  # i_q, i_d, epsilon
-    u_qd = np.array([200, 50])
+    state = np.array([0.3, 0.5, 0.68])  # i_d, i_q, epsilon
+    u_dq = np.array([50, 200])
     omega = 25
     # test default initialization first
     default_init_1 = make_module(ElectricMotor, motor_type)
@@ -398,10 +397,10 @@ def test_synchronous_motor_testing(motor_type, motor_class):
     if motor_type == 'SynRM':
         mp.update({'psi_p': 0})
     for motor in [default_init_1, default_init_2]:
-        assert motor.torque(state) == torque_testing(state, mp)
+        assert abs(motor.torque(state) - torque_testing(state, mp)) < 1E-8
         assert all(motor.i_in(state) == state[0:2])
[…]
