### 2020-07-07 #2275 @JFDuval
Hey Kevin,

Thank you for thinking about me but I'm not sure I'm the best reviewer.
However, I can recommend someone: Alejandro Azocar (afazocar@umich.edu)

On Mon, Jul 6, 2020 at 4:23 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @YashBansod <https://github.com/YashBansod> @podhrmic
> <https://github.com/podhrmic> @horychen <https://github.com/horychen>
> @epezent <https://github.com/epezent> @JFDuval
> <https://github.com/JFDuval> @sichkar-valentyn
> <https://github.com/sichkar-valentyn> @kimukook
> <https://github.com/kimukook> would you be interested in reviewing this
> submission for JOSS (The Journal of Open Source Software)
> <https://joss.theoj.org>:
> *gym-electric-motor (GEM): A Python toolbox for the simulation of electric
> drive systems*
>
> The review takes place on GitHub and focuses on this short paper:
>
> https://github.com/openjournals/joss-papers/blob/joss.02275/joss.02275/10.21105.joss.02275.pdf
> and the software:
> https://github.com/upb-lea/gym-electric-motor
>
> You can tell me if you are interested by replying here. Once I have
> sufficient reviewers I'll open a dedicated review issue.
>
