### 2020-07-08 5e5fa0a
corrected the order of dq also in the docs, renamed the stray inductances to be consistent in docs and code
