+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -1237,8 +1237,8 @@ class InductionMotor(ThreePhaseMotor):
         =============== ====== =============================================
         Motor Currents  Unit   Description
         =============== ====== =============================================
-        i_sq            A      Quadrature axis current
         i_sd            A      Direct axis current
+        i_sq            A      Quadrature axis current
         i_sa            A      Current through branch a
         i_sb            A      Current through branch b
         i_sc            A      Current through branch c
@@ -1248,8 +1248,8 @@ class InductionMotor(ThreePhaseMotor):
         =============== ====== =============================================
         Motor Voltages  Unit   Description
         =============== ====== =============================================
-        u_sq            V      Quadrature axis voltage
         u_sd            V      Direct axis voltage
+        u_sq            V      Quadrature axis voltage
         u_sa            V      Voltage through branch a
         u_sb            V      Voltage through branch b
         u_sc            V      Voltage through branch c
@@ -1313,8 +1313,8 @@ class InductionMotor(ThreePhaseMotor):
     _default_motor_parameter = {
         'p': 2,
         'l_m': 143.75e-3,
-        'l_ssig': 5.87e-3,
-        'l_rsig': 5.87e-3,
+        'l_sigs': 5.87e-3,
+        'l_sigr': 5.87e-3,
         'j_rotor': 1.1e-3,
         'r_s': 2.9338,
         'r_r': 1.355,
@@ -1387,18 +1387,18 @@ class InductionMotor(ThreePhaseMotor):
     def _torque_limit(self):
         # Docstring of superclass
         mp = self._motor_parameter
-        return 1.5 * mp['p'] * mp['l_m'] ** 2/(mp['l_m']+mp['l_rsig']) * self._limits['i_sd'] * self._limits['i_sq'] / 2
+        return 1.5 * mp['p'] * mp['l_m'] ** 2/(mp['l_m']+mp['l_sigr']) * self._limits['i_sd'] * self._limits['i_sq'] / 2
 
     def torque(self, states):
         # Docstring of superclass
         mp = self._motor_parameter
-        return 1.5 * mp['p'] * mp['l_m']/(mp['l_m'] + mp['l_rsig']) * (states[self.PSI_RALPHA_IDX] * states[self.I_SBETA_IDX] - states[self.PSI_RBETA_IDX] * states[self.I_SALPHA_IDX])
+        return 1.5 * mp['p'] * mp['l_m']/(mp['l_m'] + mp['l_sigr']) * (states[self.PSI_RALPHA_IDX] * states[self.I_SBETA_IDX] - states[self.PSI_RBETA_IDX] * states[self.I_SALPHA_IDX])
 
     def _update_model(self):
         # Docstring of superclass
         mp = self._motor_parameter
-        l_s = mp['l_m']+mp['l_ssig']
-        l_r = mp['l_m']+mp['l_rsig']
+        l_s = mp['l_m']+mp['l_sigs']
+        l_r = mp['l_m']+mp['l_sigr']
         sigma = (l_s*l_r-mp['l_m']**2) /(l_s*l_r)
         tau_r = l_r / mp['r_r']
         tau_sig = sigma * l_s / (mp['r_s'] + mp['r_r'] * (mp['l_m']**2) / (l_r**2))
@@ -1414,8 +1414,8 @@ class InductionMotor(ThreePhaseMotor):
 
     def electrical_jacobian(self, state, u_in, omega, *args):
         mp = self._motor_parameter
-        l_s = mp['l_m'] + mp['l_ssig']
-        l_r = mp['l_m'] + mp['l_rsig']
+        l_s = mp['l_m'] + mp['l_sigs']
[…]
