### 2020-07-08 3d7c50c
Basic fixes of FOC Controller
