+++ examples/agents/simple_controllers.py
@@ -51,7 +51,12 @@ class OnOffController(Controller):
             self._low_action = 2
         else:
             self._low_action = 0
-        self._referenced_state = state_idx or np.argmax(environment.reference_generator.referenced_states)
+        if state_idx is None:
+            self._referenced_state = np.argmax(
+                environment.reference_generator.referenced_states[environment.state_filter]
+            )
+        else:
+            self._referenced_state = state_idx
         self._ref_idx = reference_idx
 
     def control(self, state, reference):
@@ -83,7 +88,12 @@ class ThreePointController(Controller):
             self._low_action = 2
         else:
             self._low_action = 0
-        self._referenced_state = state_idx or np.argmax(environment.reference_generator.referenced_states)
+        if state_idx is None:
+            self._referenced_state = np.argmax(
+                environment.reference_generator.referenced_states[environment.state_filter]
+            )
+        else:
+            self._referenced_state = state_idx
 
     def control(self, state, reference):
         if state[self._referenced_state] < reference[self._ref_idx] - self._hysteresis:
@@ -112,7 +122,12 @@ class DCPController(Controller):
         self._action_min = action_space.low[controller_no]
         self._action_max = action_space.high[controller_no]
         self._ref_idx = reference_idx
-        self._referenced_state = state_idx or np.argmax(environment.reference_generator.referenced_states)
+        if state_idx is None:
+            self._referenced_state = np.argmax(
+                environment.reference_generator.referenced_states[environment.state_filter]
+            )
+        else:
+            self._referenced_state = state_idx
 
     def control(self, state, reference):
         return np.array([
@@ -141,10 +156,10 @@ class IController(Controller):
         self._action_max = action_space.low[controller_no]
         self._ref_idx = reference_idx
         self._referenced_state = state_idx or np.argmax(environment.reference_generator.referenced_states)
-        self._referenced_state_max = self._limits[self._referenced_state] * \
-                                     environment.physical_system.state_space.high[self._referenced_state]
-        self._referenced_state_min = self._limits[self._referenced_state] * environment.physical_system.state_space.low[
-            self._referenced_state]
+        self._referenced_state_max = self._limits[self._referenced_state] \
+            * environment.physical_system.state_space.high[self._referenced_state]
+        self._referenced_state_min = self._limits[self._referenced_state] \
+            * environment.physical_system.state_space.low[self._referenced_state]
 
     def control(self, state, reference):
         diff = reference[self._ref_idx] - state[self._referenced_state]
@@ -187,10 +202,10 @@ class DCPIController(DCPController):
         self._tau = environment.physical_system.tau
         self._limits = environment.physical_system.limits
         self._integrated_value = 0
-        self._referenced_state_max = self._limits[self._referenced_state] * \
-                                     environment.physical_system.state_space.high[self._referenced_state]
-        self._referenced_state_min = self._limits[self._referenced_state] * environment.physical_system.state_space.low[
-            self._referenced_state]
+        self._referenced_state_max = self._limits[self._referenced_state] \
+            * environment.physical_system.state_space.high[self._referenced_state]
+        self._referenced_state_min = self._limits[self._referenced_state] \
+            * environment.physical_system.state_space.low[self._referenced_state]
 
     def control(self, state, reference):
[…]
