### 2020-07-09 e8ca83a
new file with constraint_monitor class, added functionallity to rest of the code, added unit tests
