+++ gym_electric_motor/reward_functions/__init__.py
@@ -5,4 +5,3 @@ from .. import RewardFunction
 
 register_class(WeightedSumOfErrors, RewardFunction, 'WSE')
 register_class(ShiftedWeightedSumOfErrors, RewardFunction, 'SWSE')
-register_class(ConstraintMonitor, RewardFunction, 'ConMon')
