### 2020-07-09 813269a
fixed merge conflicts, corrected formula for psi_max
