+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -11,26 +11,6 @@ class MechanicalLoad:
     It contains an mechanical ode system as well as the state names, limits and nominal values
     of the mechanical quantities. The only required state is 'omega' as the rotational speed of the motor shaft
     in rad/s.
-<<<<<<< HEAD
-
-    ExternalSpeedLoad and ConstantSpeedLoad should be initialized with the own
-    initial value in class initialisation, when definded external
-    # todo updateting intialization with motor_initializer
-
-    Initialization is given by initializer(dict). Can be constant state value
-    or random value in given or nominal interval.
-    dict should be like:
-        { 'states': <dict with state names and initital values>,
-          'interval': < boundaries for each state (only for random init>,
-          'random_init': <str: 'uniform' or 'normal'>,
-          'random_params: {'mue': <const>, 'sigma': <const>}
-    Example initializer(dict) for constant initialization:
-        { 'states': {'omega': 15.5}}
-    Example  initializer(dict) for random initialization:
-        { 'random_init': 'normal'}
-||||||| merged common ancestors
-=======
-
     ExternalSpeedLoad and ConstantSpeedLoad should be initialized with the own
     initial value in class initialisation, when definded external
 
@@ -46,7 +26,6 @@ class MechanicalLoad:
         { 'states': {'omega': 16.0}}
     Example  initializer(dict) for random initialization:
         { 'random_init': 'normal'}
->>>>>>> new_issue9_30
     """
 
     @property
