### 2020-07-09 9880a3f
Added comments for Callback class
