### 2020-07-09 d78aeb6
examples for issue 9 and 30 added
