### 2020-07-09 f16341a
Added comments to callbacks.py
