+++ gym_electric_motor/callbacks.py
@@ -1,3 +1,5 @@
+"""This module introduces predefined callbacks for the GEM environment."""
+
 from .core import Callback
 from gym_electric_motor.reference_generators import SubepisodedReferenceGenerator, SwitchedReferenceGenerator
 
@@ -14,7 +16,24 @@ class DummyCallback(Callback):
         print('on_close')
         
 class AdaptiveLimitMargin(Callback):
+    """
+    Callback used to adapt the limit margin of a reference generator during runtime. Supports all SubepisodedReferenceGenerators 
+    and SwitchedReferenceGenerator with only SubepisodedReferenceGenerators as sub generators.
+    """
+    
     def __init__(self, initial_limit_margin=(-0.1,0.1), maximum_limit_margin=(-1,1), step_size=0.1, update_time='episode', update_freq=10):
+        """
+        Args:
+            initial_limit_margin(tuple(floats)): The initial limit margin which gets updated by AdaptiveLimitMargin until it reaches maximum_limit_margin
+            maximum_limit_margin(tuple(floats)): The maximum limit margin. This will be the limit margin after AdaptiveLimitMargin's last update
+            step_size(float): The value by which each limit gets updated at each step
+            update_time(string): When the update happens. "step" for the end of a step, "episode" for the end of an episode
+            update_freq(int): After how many cumulative units of update_time an update occurs
+        
+        Additional Notes:
+            All limit_margins should be between -1 and 1
+
+        """
         assert update_time in ['step', 'episode'], "Chose an option of either 'step' or 'episode' for updating on cumultative steps or episodes"
         assert initial_limit_margin[1] > initial_limit_margin[0], "First element of limit margin has to be smaller than second"
         assert maximum_limit_margin[1] > maximum_limit_margin[0], "First element of limit margin has to be smaller than second"
@@ -34,6 +53,8 @@ class AdaptiveLimitMargin(Callback):
         self._update_freq = update_freq
         
     def set_env(self, env):
+        #see docstring of superclass
+        #assertions added to check for the right reference generator 
         if env.reference_generator.__class__ == SwitchedReferenceGenerator:
             for sub_generator in env.reference_generator._sub_generators:
                 assert issubclass(sub_generator.__class__, SubepisodedReferenceGenerator), ("The AdaptiveLimitMargin does only support the SubepisodedReferenceGenerator as reference generator or "
@@ -42,6 +63,7 @@ class AdaptiveLimitMargin(Callback):
             assert issubclass(env.reference_generator.__class__, SubepisodedReferenceGenerator), ("The AdaptiveLimitMargin does only support the SubepisodedReferenceGenerator as reference generator or" 
                                                                                                  "SwitchedReferenceGenerator with SubepisodedReferenceGenerator as all sub reference generators")
         self._env = env
+        #initial image margin added to the reference generator
         if env.reference_generator.__class__ == SwitchedReferenceGenerator:
             for sub_generator in self._env.reference_generator._sub_generators:
                 sub_generator._limit_margin = self._limit_margin
@@ -49,6 +71,7 @@ class AdaptiveLimitMargin(Callback):
             self._env.reference_generator._limit_margin = self._limit_margin
         
     def on_step_end(self):
+        #see docstring of superclass
         if self._update_time == 0:
             self._step += 1
             if self._step % self._update_freq == 0:
@@ -56,6 +79,7 @@ class AdaptiveLimitMargin(Callback):
                 self._update_limit_margin()     
                 
     def on_reset_end(self):
+        #see docstring of superclass
         if self._update_time == 1:
             self._episode += 1
             if self._episode % self._update_freq == 0:
@@ -63,6 +87,7 @@ class AdaptiveLimitMargin(Callback):
[…]
