### 2020-07-09 866834d
Changed some comments and added file for sphinx documentation
