### 2020-07-09 9fd454a
episode based mean reward plot first implementaiton - error in updating x-lim
