+++ tests/test_physical_systems/test_voltage_supplies.py
@@ -2,7 +2,6 @@ import gym_electric_motor as gem
 import gym_electric_motor.physical_systems.voltage_supplies as vs
 from ..testing_utils import DummyOdeSolver
 import numpy as np
-import pytest
 
 class TestVoltageSupply:
 
@@ -111,13 +110,13 @@ class TestAC1PhaseSupply(TestVoltageSupply):
         
     def test_reset(self, u_nominal=230.0):
         """Test the reset function for correct behavior on fixed phase""" 
-        supply_parameter = {'frequency': 50, 'phase': 0, 'fixed_phase': False}
+        supply_parameter = {'frequency': 50}
         voltage_supply = vs.AC1PhaseSupply(u_nominal, supply_parameter)
-        assert voltage_supply._phi == 0
+        first_phi = voltage_supply._phi
         _ = voltage_supply.reset()
-        assert voltage_supply._phi != 0, "Test this again and if this error doesn't appear next time you should consider playing lotto"
+        assert voltage_supply._phi != first_phi, "Test this again and if this error doesn't appear next time you should consider playing lotto"
 
-        supply_parameter = {'frequency': 50, 'phase': 0, 'fixed_phase': True}
+        supply_parameter = {'frequency': 50, 'phase': 0}
         voltage_supply = vs.AC1PhaseSupply(u_nominal, supply_parameter)
         assert voltage_supply._phi == 0
         assert voltage_supply.reset() == [0.0]
@@ -125,7 +124,7 @@ class TestAC1PhaseSupply(TestVoltageSupply):
 
         
     def test_initialization(self, u_nominal = 300.0):
-        supply_parameter = {'frequency': 35, 'phase': 0, 'fixed_phase': True}
+        supply_parameter = {'frequency': 35, 'phase': 0}
         voltage_supply = vs.AC1PhaseSupply(u_nominal, supply_parameter)
         assert voltage_supply.u_nominal == 300.0
         assert voltage_supply._max_amp == 300.0 * np.sqrt(2)
@@ -136,29 +135,28 @@ class TestAC1PhaseSupply(TestVoltageSupply):
         
     def test_get_voltage(self):
         """Test the get voltage function for different times t."""
-        supply_parameter = {'frequency': 1, 'phase': 0, 'fixed_phase': True}
+        supply_parameter = {'frequency': 1, 'phase': 0}
         supply = vs.AC1PhaseSupply(supply_parameter=supply_parameter)
         
         #Test for default sinus values
-        times = [0, 2*np.pi, 4*np.pi]
+        times = [0, 1, 2]
         for time in times:
-            assert supply.get_voltage(time) == pytest.approx([0.0])
+            assert np.allclose(supply.get_voltage(time), [0.0])
             
-        times = [np.pi/2, 5*np.pi/2, 9*np.pi/2]
+        times = [1/4, 5/4, 9/4]
         for time in times:
-            assert supply.get_voltage(time) == pytest.approx([230.0 * np.sqrt(2)])
+            assert np.allclose(supply.get_voltage(time),[230.0 * np.sqrt(2)])
             
-        times = [3*np.pi/2, 7*np.pi/2, 11*np.pi/2]
+        times = [3/4, 7/4, 11/4]
         for time in times:
-            assert supply.get_voltage(time) == pytest.approx([-230.0 * np.sqrt(2)])
+            assert np.allclose(supply.get_voltage(time),[-230.0 * np.sqrt(2)])
           
         #"Hand" calculated
-        supply_parameter = {'frequency': 36, 'phase': 0.5, 'fixed_phase': True}
[…]
