### 2020-07-10 3c445ab
example added, docs updatet
