+++ docs/parts/reward_functions/constraint_monitor.rst
@@ -1,0 +1,7 @@
+Constraint Monitor
+#####################
+
+
+.. autoclass:: gym_electric_motor.reward_functions.constraint_monitor.ConstraintMonitor
+    :members:
+    :inherited-members:
