### 2020-07-10 1c1aa03
removed pyqt dependency in toolbox. tested examples
