### 2020-07-10 b3d744d
small final  changes
