### 2020-07-12 #2275 @Kevin-Mattheus-Moerman
@christiankral @dineshresearch @madcowswe would you be interested in reviewing this submission for [JOSS (The Journal of Open Source Software)](https://joss.theoj.org):
_gym-electric-motor (GEM): A Python toolbox for the simulation of electric drive systems_

The review takes place on GitHub and focuses on this short paper:
https://github.com/openjournals/joss-papers/blob/joss.02275/joss.02275/10.21105.joss.02275.pdf
and the software:
https://github.com/upb-lea/gym-electric-motor

You can tell me if you are interested by replying here. Once I have sufficient reviewers I'll open a dedicated review issue.
