### 2020-07-12 d04175e
added Symmetric Optimum tuning
