+++ examples/agents/simple_controllers.py
@@ -104,7 +104,7 @@ class ThreePointController(Controller):
             return self._idle_action
 
 
-class DCPController(Controller):
+class PController(Controller):
     """
     In the below proportional control implementation, the controller output is proportional to the error signal,  which is the difference
     between the reference_idx and the state_idx .i.e., the output of a proportional controller
@@ -157,25 +157,27 @@ class IController(Controller):
         self._ref_idx = reference_idx
         self._referenced_state = state_idx or np.argmax(environment.reference_generator.referenced_states)
         self._referenced_state_max = self._limits[self._referenced_state] \
-            * environment.physical_system.state_space.high[self._referenced_state]
+                                     * environment.physical_system.state_space.high[self._referenced_state]
         self._referenced_state_min = self._limits[self._referenced_state] \
-            * environment.physical_system.state_space.low[self._referenced_state]
+                                     * environment.physical_system.state_space.low[self._referenced_state]
 
     def control(self, state, reference):
         diff = reference[self._ref_idx] - state[self._referenced_state]
         self._integrated_value += diff * self._tau
-        if state[self._referenced_state] > self._referenced_state_max:  # check upper limit
-            state[self._referenced_state] = self._referenced_state_max
-            self._integrated_value = self._integrated_value - diff * self._tau  # anti-reset windup
-        if state[self._referenced_state] < self._referenced_state_min:  # check upper limit
-            state[self._referenced_state] = self._referenced_state_min
-            self._integrated_value = self._integrated_value - diff * self._tau  # anti-reset windup
+        if self._integrated_value > self._referenced_state_max:  # check upper limit
+            self._integrated_value = self._referenced_state_max
+        else:
+            self._integrated_value = self._integrated_value - diff * self._k_i  # anti-reset windup
+        if self._integrated_value < self._referenced_state_min:  # check upper limit
+            self._integrated_value = self._referenced_state_min
+        else:
+            self._integrated_value = self._integrated_value - diff * self._k_i  # anti-reset windup
 
         return np.array([
             max(
                 self._action_min,
                 min(
-                    self._action_max, self._k_i / self._tau * self._integrated_value
+                    self._action_max, self._tau / self._k_i * self._integrated_value
                 )
             )
         ])
@@ -184,7 +186,7 @@ class IController(Controller):
         self._integrated_value = 0
 
 
-class DCPIController(DCPController):
+class PIController(PController):
     """
       The below Discrete PI Controller performs discrete-time PI controller computation using the error signal and proportional
       and integral gain inputs. The error signal is the difference between the reference_idx and the referenced_state.
@@ -203,18 +205,27 @@ class DCPIController(DCPController):
         self._limits = environment.physical_system.limits
         self._integrated_value = 0
         self._referenced_state_max = self._limits[self._referenced_state] \
-            * environment.physical_system.state_space.high[self._referenced_state]
+                                     * environment.physical_system.state_space.high[self._referenced_state]
         self._referenced_state_min = self._limits[self._referenced_state] \
-            * environment.physical_system.state_space.low[self._referenced_state]
+                                     * environment.physical_system.state_space.low[self._referenced_state]
+        self._motor_parameter = environment.physical_system.electrical_motor.motor_parameter
 
     def control(self, state, reference):
+        mp = self._motor_parameter
+        t_motor = mp['l_a'] / mp['r_a']
+        t_t = 3 / 2 * self._tau
[…]
