### 2020-07-13 f685d75
Added tests for AdaptiveLimitMargin class
