### 2020-07-13 9baf780
error fixed, episode based plot visualization to be edited
