### 2020-07-13 fce1256
Added tests of callback calls into the environment tests
