+++ tests/test_core.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
-from tests.testing_utils import DummyPhysicalSystem, DummyReferenceGenerator, DummyRewardFunction, DummyVisualization, \
-    mock_instantiate, instantiate_dict
+from tests.testing_utils import DummyPhysicalSystem, DummyReferenceGenerator, DummyRewardFunction, DummyVisualization, DummyCallback, \
+    mock_instantiate, instantiate_dict 
 from gym.spaces import Tuple, Box
 import gym_electric_motor
 from gym_electric_motor.visualization import ConsolePrinter
@@ -21,11 +21,13 @@ class TestElectricMotorEnvironment:
         rg = DummyReferenceGenerator()
         rf = DummyRewardFunction()
         vs = DummyVisualization()
+        cb = DummyCallback()
         env = self.test_class(
             physical_system=ps,
             reference_generator=rg,
             reward_function=rf,
-            visualization=vs
+            visualization=vs,
+            callbacks = [cb]
         )
         return env
 
@@ -35,21 +37,21 @@ class TestElectricMotorEnvironment:
             assert type(env) == self.test_class
 
     @pytest.mark.parametrize(
-        "physical_system, reference_generator, reward_function, state_filter, visualization, kwargs",
+        "physical_system, reference_generator, reward_function, state_filter, visualization, callbacks, kwargs",
         [
-            (DummyPhysicalSystem, DummyReferenceGenerator, DummyRewardFunction, None, None, {}),
+            (DummyPhysicalSystem, DummyReferenceGenerator, DummyRewardFunction, None, None, [], {}),
             (
                     DummyPhysicalSystem(2), DummyReferenceGenerator, DummyRewardFunction(), ['dummy_state_0'],
-                    ConsolePrinter, {'a': 1, 'b': 2}
+                    ConsolePrinter, [DummyCallback()], {'a': 1, 'b': 2}
             ),
             (
                     DummyPhysicalSystem(10), DummyReferenceGenerator(),
-                    DummyRewardFunction(observed_states=['dummy_state_0']), ['dummy_state_0', 'dummy_state_2'], None, {}
+                    DummyRewardFunction(observed_states=['dummy_state_0']), ['dummy_state_0', 'dummy_state_2'], None, [DummyCallback(), DummyCallback()],{}
             ),
         ]
     )
     def test_initialization(self, monkeypatch, physical_system, reference_generator, reward_function,
-                            state_filter, visualization, kwargs):
+                            state_filter, visualization, callbacks, kwargs):
         with monkeypatch.context() as m:
             instantiate_dict.clear()
             m.setattr(gym_electric_motor.core, "instantiate", mock_instantiate)
@@ -59,6 +61,7 @@ class TestElectricMotorEnvironment:
                 reward_function=reward_function,
                 visualization=visualization,
                 state_filter=state_filter,
+                callbacks = callbacks,
                 **kwargs
             )
 
@@ -95,15 +98,25 @@ class TestElectricMotorEnvironment:
         assert reference_generator, DummyReferenceGenerator or env.reference_generator.kwargs == kwargs
         assert reward_function != DummyRewardFunction or env.reward_function.kwargs == kwargs
         assert visualization != DummyVisualization or env._visualization.kwargs == kwargs
+        for callback in callbacks:
+            assert callback._env == env
 
     def test_reset(self, env):
         ps = env.physical_system
         rg = env.reference_generator
         rf = env.reward_function
         vs = env._visualization
+        cbs = env._callbacks
         rf.last_state = rf.last_reference = ps.state = rg.get_reference_state = vs.reference_trajectory = None
-
+        # Initial amount of resets
+        for callback in cbs:
[…]
