### 2020-07-13 b1e93f6
Added support vor vectorial supply voltages
