+++ gym_electric_motor/physical_systems/physical_systems.py
@@ -168,7 +168,7 @@ class SCMLSystem(PhysicalSystem):
         voltages_lower = currents_upper
         voltages_upper = voltages_lower + len(self._electrical_motor.VOLTAGES)
         self.VOLTAGES_IDX = list(range(voltages_lower, voltages_upper))
-        self.U_SUP_IDX = voltages_upper
+        self.U_SUP_IDX = list(range(voltages_upper, voltages_upper + self._supply.voltage_len))
 
     def simulate(self, action, *_, **__):
         # Docstring of superclass
@@ -178,17 +178,17 @@ class SCMLSystem(PhysicalSystem):
 
         for t in switching_times[:-1]:
             i_sup = self._converter.i_sup(i_in)
-            u_sup = self._supply.get_voltage(self._t, i_sup)[0]
+            u_sup = self._supply.get_voltage(self._t, i_sup)
             u_in = self._converter.convert(i_in, self._ode_solver.t)
-            u_in = [u * u_sup for u in u_in]
+            u_in = [u * u_s for u in u_in for u_s in u_sup]
             self._ode_solver.set_f_params(u_in)
             ode_state = self._ode_solver.integrate(t)
             i_in = self._electrical_motor.i_in(ode_state[self._ode_currents_idx])
         
         i_sup = self._converter.i_sup(i_in)
-        u_sup = self._supply.get_voltage(self._t, i_sup)[0]
+        u_sup = self._supply.get_voltage(self._t, i_sup)
         u_in = self._converter.convert(i_in, self._ode_solver.t)
-        u_in = [u * u_sup for u in u_in]
+        u_in = [u * u_s for u in u_in for u_s in u_sup]
         self._ode_solver.set_f_params(u_in)
         ode_state = self._ode_solver.integrate(self._t + self._tau)
         self._t = self._ode_solver.t
@@ -272,9 +272,9 @@ class SCMLSystem(PhysicalSystem):
         motor_state = self._electrical_motor.reset()
         mechanical_state = self._mechanical_load.reset()
         ode_state = np.concatenate((mechanical_state, motor_state))
-        u_sup = self.supply.reset()[0]
+        u_sup = self.supply.reset()
         u_in = self.converter.reset()
-        u_in = [u * u_sup for u in u_in]
+        u_in = [u * u_s for u in u_in for u_s in u_sup]
         torque = self.electrical_motor.torque(motor_state)
         noise = self._noise_generator.reset()
         self._t = 0
@@ -285,7 +285,7 @@ class SCMLSystem(PhysicalSystem):
             [torque],
             motor_state[self._electrical_motor.CURRENTS_IDX],
             u_in,
-            [u_sup]
+            u_sup
         ))
         return (system_state + noise) / self._limits
 
@@ -469,7 +469,7 @@ class SynchronousMotorSystem(ThreePhaseMotorSystem):
         voltages_upper = voltages_lower + 5
         self.VOLTAGES_IDX = list(range(voltages_lower, voltages_upper))
         self.EPSILON_IDX = voltages_upper
-        self.U_SUP_IDX = self.EPSILON_IDX + 1
+        self.U_SUP_IDX = list(range(self.EPSILON_IDX + 1, self.EPSILON_IDX + 1 + self._supply.voltage_len))
         self._ode_epsilon_idx = self._motor_ode_idx[-1]
 
     def simulate(self, action, *_, **__):
@@ -483,9 +483,9 @@ class SynchronousMotorSystem(ThreePhaseMotorSystem):
 
         for t in switching_times[:-1]:
[…]
