+++ tests/test_callbacks.py
@@ -1,10 +1,10 @@
 from gym_electric_motor.reference_generators import SubepisodedReferenceGenerator, SwitchedReferenceGenerator
-from gym_electric_motor.callbacks import AdaptiveLimitMargin
+from gym_electric_motor.callbacks import RampingLimitMargin
 from tests.testing_utils import DummyElectricMotorEnvironment, DummyReferenceGenerator
 import pytest
 
-class TestAdaptiveLimitMargin:
-    test_class = AdaptiveLimitMargin
+class TestRampingLimitMargin:
+    test_class = RampingLimitMargin
     key = ''
 
     def test_update(self):
@@ -55,7 +55,7 @@ class TestAdaptiveLimitMargin:
         assert callback._limit_margin == (-0.1,0.1)
         assert callback._maximum_limit_margin == (-1,1)
         assert callback._step_size == 0.1
-        assert callback._update_time == 1
+        assert callback._update_time == 'episode'
         assert callback._update_freq == 10
 
     def test_update_switched(self):
@@ -80,14 +80,14 @@ class TestAdaptiveLimitMargin:
         #reference generator has to be a subclass of SubepisodedReferenceGenerator
         with pytest.raises(AssertionError) as excinfo:
             env = DummyElectricMotorEnvironment(reference_generator=DummyReferenceGenerator(), callbacks=callbacks)
-        assert "The AdaptiveLimitMargin does only support" in str(excinfo.value)
+        assert "The RampingLimitMargin does only support" in str(excinfo.value)
         
         # all sub generators have to subclasses of SubepisodedReferenceGenerator
         sub_generators = [SubepisodedReferenceGenerator(),SubepisodedReferenceGenerator(),DummyReferenceGenerator()]
         switched = SwitchedReferenceGenerator(sub_generators)
         with pytest.raises(AssertionError) as excinfo:
             env = DummyElectricMotorEnvironment(reference_generator=switched, callbacks=callbacks)
-        assert "The AdaptiveLimitMargin does only support" in str(excinfo.value)
+        assert "The RampingLimitMargin does only support" in str(excinfo.value)
 
         
 
