+++ gym_electric_motor/core.py
@@ -38,7 +38,7 @@ class ElectricMotorEnvironment(gym.core.Env):
             Generation of the reference for the motor to follow. Needs to be a subclass of *ReferenceGenerator*
 
         Reward Function:
-            Calculation of the reward  based on the state of the physical system and the generated reference
+            Calculation of the reward based on the state of the physical system and the generated reference
             and observation if the motor state is within the limits. Needs to be a subclass of *RewardFunction*.
 
         Visualization:
@@ -194,11 +194,11 @@ class ElectricMotorEnvironment(gym.core.Env):
         self._action = None
         
         self._callbacks = callbacks
-        self._call_callbacks(self._callbacks, 'set_env', self)
+        self._call_callbacks('set_env', self)
         
-    def _call_callbacks(self, callbacks, func_name, *args):
+    def _call_callbacks(self, func_name, *args):
         """Calls each callback's func_name function with *args"""
-        for callback in callbacks:
+        for callback in self._callbacks:
             func = getattr(callback, func_name)
             func(*args)
             
@@ -209,7 +209,7 @@ class ElectricMotorEnvironment(gym.core.Env):
         Returns:
              The initial observation consisting of the initial state and initial reference.
         """
-        self._call_callbacks(self._callbacks, 'on_reset_begin')
+        self._call_callbacks('on_reset_begin')
         self._reset_required = False
         self._state = self._physical_system.reset()
         self._reference, next_ref, trajectories = self.reference_generator.reset(self._state)
@@ -217,7 +217,7 @@ class ElectricMotorEnvironment(gym.core.Env):
         self._reward_function.reset(self._state, self._reference)
         self._reward = 0.0
         self._action = None
-        self._call_callbacks(self._callbacks, 'on_reset_end')
+        self._call_callbacks('on_reset_end')
         return self._state[self.state_filter], next_ref
 
     def render(self, *_, **__):
@@ -243,21 +243,21 @@ class ElectricMotorEnvironment(gym.core.Env):
         """
 
         assert not self._reset_required, 'A reset is required before the environment can perform further steps'
-        self._call_callbacks(self._callbacks, 'on_step_begin')
+        self._call_callbacks('on_step_begin')
         self._action = action
         self._state = self._physical_system.simulate(action)
         self._reference = self.reference_generator.get_reference(self._state)
         self._reward, self._done = self._reward_function.reward(self._state, self._reference, action)
         self._reset_required = self._done
         ref_next = self.reference_generator.get_reference_observation(self._state)
-        self._call_callbacks(self._callbacks, 'on_step_end')
+        self._call_callbacks('on_step_end')
         return (self._state[self.state_filter], ref_next), self._reward, self._reset_required, {}
 
     def close(self):
         """
         Called when the environment is deleted. Closes all its modules.
         """
-        self._call_callbacks(self._callbacks, 'on_close')
+        self._call_callbacks('on_close')
         self._reward_function.close()
         self._physical_system.close()
         self._reference_generator.close()
