### 2020-07-14 69c3c5a
Small addition to NoConverter
