### 2020-07-14 797aac9
Revert "Merge branch 'nightly' into issue_#48"
This reverts commit a2756d8260de3bafae454669771114e25330a7a0, reversing
changes made to 69c3c5a36088f8b8a09050a6fddc07439fb277ff.
