+++ gym_electric_motor/physical_systems/__init__.py
@@ -7,7 +7,7 @@ from .electric_motors import DcExternallyExcitedMotor, DcSeriesMotor, DcPermanen
     PermanentMagnetSynchronousMotor, ElectricMotor, SynchronousReluctanceMotor, SquirrelCageInductionMotor, \
     DoublyFedInductionMotor
 
-from .mechanical_loads import MechanicalLoad, PolynomialStaticLoad, ExternalSpeedLoad, ConstantSpeedLoad
+from .mechanical_loads import MechanicalLoad, PolynomialStaticLoad, ConstantSpeedLoad
 
 from .solvers import OdeSolver, EulerSolver, ScipyOdeIntSolver, ScipySolveIvpSolver, ScipyOdeSolver
 
@@ -46,7 +46,6 @@ register_class(NoConverter, PowerElectronicConverter, 'NoConverter')
 
 register_class(PolynomialStaticLoad, MechanicalLoad, 'PolyStaticLoad')
 register_class(ConstantSpeedLoad, MechanicalLoad, 'ConstSpeedLoad')
-register_class(ExternalSpeedLoad, MechanicalLoad, 'ExtSpeedLoad')
 
 register_class(GaussianWhiteNoiseGenerator, NoiseGenerator, 'GWN')
 
