+++ tests/testing_utils.py
@@ -387,9 +387,9 @@ class DummyElectricMotor(ElectricMotor):
         self.u_in = u_in
         return state - u_in
 
-    def reset(self, state_space, state_positions):
+    def reset(self):
         self.reset_counter += 1
-        return super().reset(state_space, state_positions)
+        return super().reset()
 
     def torque(self, currents):
         return np.prod(currents)
@@ -513,7 +513,7 @@ class DummyLoad(MechanicalLoad):
         self.reset_counter = 0
         super().__init__(tau=tau, **kwargs)
 
-    def reset(self, state_space, state_positions, nominal_state,  *_, **__):
+    def reset(self, *_, **__):
         self.reset_counter += 1
         return np.zeros(2)
 
