+++ docs/parts/physical_systems/mechanical_loads/mechanical_load.rst
@@ -7,7 +7,6 @@ Mechanical Loads
 
     polystatic
     const_speed_load
-    ext_speed_load
 
 MechanicalLoad Base Class
 *************************************

[…]
