+++ gym_electric_motor/reward_functions/__init__.py
@@ -1,5 +1,4 @@
 from .weighted_sum_of_errors import WeightedSumOfErrors, ShiftedWeightedSumOfErrors
-from .constraint_monitor import ConstraintMonitor
 from ..utils import register_class
 from .. import RewardFunction
 
