### 2020-07-17 597ff2c
episode based mean reward plot added in seperate fig. aesthetics improvement and cleanup pending
