### 2020-07-17 ff25740
legends/labels added. code commented, examples modified
