+++ gym_electric_motor/visualization/motor_dashboard.py
@@ -21,6 +21,7 @@ class MotorDashboard(ElectricMotorVisualization):
                     - {state_name}: The corresponding state is plotted
                     - reward: The reward per step is plotted
                     - action_{i}: The i-th action is plotted. 'action_0' for discrete action space
+                    - mean_reward: The mean episode reward
                     
             update_cycle(int): Number after how many steps the plot shall be updated. (default 1000)
             dark_mode(Bool):  Select a dark background for visualization by setting it to True
@@ -38,7 +39,7 @@ class MotorDashboard(ElectricMotorVisualization):
                     self._plots.append(mdp.RewardPlot())
                 elif plot.startswith('action_'):
                     self._plots.append(mdp.ActionPlot(plot))
-                elif plot.startswith('mean_'):
+                elif plot.startswith('mean_reward'):
                     self._episode_plots.append(mdp.MeanEpisodeRewardPlot())
                 else:
                     self._plots.append(mdp.StatePlot(plot))
@@ -49,11 +50,13 @@ class MotorDashboard(ElectricMotorVisualization):
     def reset(self, **__):
         """Called when the environment is reset. All subplots are reset.
         """
-        for plot in self._plots:
+
+        for plot in self._plots:  # for plot in self._plots + self._episode_plots throws warning
             plot.reset()
         for plot in self._episode_plots:
             plot.reset()
 
+        # since end of an episode can only be identified by a reset call. Episode based plot canvas updated here
         if self._figure_ep:
             self._figure_ep.canvas.draw()
             self._figure_ep.canvas.flush_events()
@@ -71,9 +74,9 @@ class MotorDashboard(ElectricMotorVisualization):
             reward(ndarray(float)): Last received reward. (None after reset)
             done(bool): Flag if the current state is terminal
         """
-        if not self._figure:
+        if not (self._figure or self._figure_ep):
             self._initialize()
-        for plot in self._plots:
+        for plot in self._plots :    # for plot in self._plots + self._episode_plots throws warning
             plot.step(k, state, reference, action, reward, done)
         for plot in self._episode_plots:
             plot.step(k, state, reference, action, reward, done)
@@ -90,7 +93,7 @@ class MotorDashboard(ElectricMotorVisualization):
             rg(ReferenceGenerator): ReferenceGenerator of the environment
             rf(RewardFunction): RewardFunction of the environment
         """
-        for plot in self._plots:
+        for plot in self._plots:  # for plot in self._plots + self._episode_plots throws warning
             plot.set_modules(ps, rg, rf)
         for plot in self._episode_plots:
             plot.set_modules(ps, rg, rf)
@@ -104,13 +107,16 @@ class MotorDashboard(ElectricMotorVisualization):
         # Use dark-mode, if selected
         if self._dark_mode:
             plt.style.use('dark_background')
+        # create seperate figures for time based and episode based plots
         self._figure, axes = plt.subplots(len(self._plots), sharex=True)
         self._figure_ep, axes_ep = plt.subplots(len(self._episode_plots))
 
         self._figure.subplots_adjust(wspace=0.0, hspace=0.2)
         self._figure_ep.subplots_adjust(wspace=0.0, hspace=0.02)
 
-        #plt.xlabel('t/s')  # adding a common x-label to all the subplots   #todo
+        # adding a common x-label to all the subplots in each figure
+        self._figure.text(0.5, 0.04, 't/s', va='center', ha='center')
+        self._figure_ep.text(0.5, 0.04, 'episode_number', va='center', ha='center')
 
         # plt.subplot() does not return an iterable var when the number of subplots==1
         if len(self._plots) == 1:
[…]
