+++ tests/test_environments/test_environments.py
@@ -277,7 +277,8 @@ def test_threephase_environments(reward_function_type, reference_generator_type,
                         converter=converter_type,
                         reward_function=reward_function_type,
                         reference_generator=reference_generator_type,
-                        visualization='MotorDashboard',
+                        #visualization is better tested manually
+                        #visualization='MotorDashboard',
                         ode_solver='euler',
                         load='PolyStaticLoad',
                         solver_kwargs=solver_kwargs,
