### 2020-07-19 #2275 @epezent
@Kevin-Mattheus-Moerman, thank you for the consideration but I don't currently have the bandwidth or experience with reinforcement learning to provide a high-quality review.

I can recommend my colleague @nathandunk
