### 2020-07-20 #2275 @dineshresearch
Hello  Kevin Mattheus Moerman,

I can do the review as I have the relevant experience in terms of technical
expertise and I will provide my review by the end of this week

On Sun, Jul 12, 2020 at 4:42 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @christiankral <https://github.com/christiankral> @dineshresearch
> <https://github.com/dineshresearch> @madcowswe
> <https://github.com/madcowswe> would you be interested in reviewing this
> submission for JOSS (The Journal of Open Source Software)
> <https://joss.theoj.org>:
> *gym-electric-motor (GEM): A Python toolbox for the simulation of electric
> drive systems*
>
> The review takes place on GitHub and focuses on this short paper:
>
> https://github.com/openjournals/joss-papers/blob/joss.02275/joss.02275/10.21105.joss.02275.pdf
> and the software:
> https://github.com/upb-lea/gym-electric-motor
>
> You can tell me if you are interested by replying here. Once I have
> sufficient reviewers I'll open a dedicated review issue.
>
