### 2020-07-21 c7a192c
remove version annotation for pytest-lambda, as that one is conflicting with the high pytest version. Confusingly, giving no version number for pytest-lambda will in fact install v 1.1.0 and there are no conflicts
