+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -1265,14 +1265,17 @@ class SynchronousReluctanceMotor(SynchronousMotor):
 
     """
     HAS_JACOBIAN = True
-    _default_motor_parameter = {'p': 2, 'l_d': 73.2e-3, 'l_q': 7.3e-3,
-                                'j_rotor': 2.45e-3, 'r_s': 0.3256}
-    _default_nominal_values = {
-        'i': 54, 'torque': 0, 'omega': 523.0, 'epsilon': np.pi, 'u': 600
-    }
 
-    _default_limits = {'i': 70, 'torque': 0, 'omega': 600.0, 'epsilon': np.pi,
-                       'u': 600}
+    #### Parameters taken from DOI: 10.1109/AMC.2008.4516099 (K. Malekian, M. R. Sharif, J. Milimonfared)
+    _default_motor_parameter = {'p': 4,
+                                'l_d': 10.1e-3,
+                                'l_q':  4.1e-3,
+                                'j_rotor': 0.8e-3,
+                                'r_s': 0.57
+                                }
+
+    _default_nominal_values = {'i': 10, 'torque': 0, 'omega': 3e3 * np.pi / 30, 'epsilon': np.pi, 'u': 100}
+    _default_limits = {'i': 13, 'torque': 0, 'omega': 4.3e3 * np.pi / 30, 'epsilon': np.pi, 'u': 100}
     _default_initializer = {'states': {'i_sq': 0.0, 'i_sd': 0.0, 'epsilon': 0.0},
                             'interval': None,
                             'random_init': None,
@@ -1401,18 +1404,20 @@ class PermanentMagnetSynchronousMotor(SynchronousMotor):
             Furthermore, if specific limits/nominal values (e.g. i_a) are not specified they are inferred from
             the general limits/nominal values (e.g. i)
     """
+
+    #### Parameters taken from DOI: 10.1109/TPEL.2020.3006779 (A. Brosch, S. Hanke, O. Wallscheid, J. Boecker)
+    #### and DOI: 10.1109/IEMDC.2019.8785122 (S. Hanke, O. Wallscheid, J. Boecker)
     _default_motor_parameter = {
-        'p': 2,
-        'l_d': 79e-3,
-        'l_q': 113e-3,
-        'j_rotor': 2.45e-3,
-        'r_s': 4.9,
-        'psi_p': 0.165,
+        'p': 3,
+        'l_d': 0.37e-3,
+        'l_q': 1.2e-3,
+        'j_rotor': 0.3883,
+        'r_s': 18e-3,
+        'psi_p': 66e-3,
     }
     HAS_JACOBIAN = True
-    _default_limits = dict(omega=80, torque=0.0, i=20, epsilon=math.pi, u=600)
-    _default_nominal_values = dict(omega=75, torque=0.0, i=12, epsilon=math.pi,
-                                   u=600)
+    _default_limits = dict(omega=12e3 * np.pi / 30, torque=0.0, i=230, epsilon=math.pi, u=300)
+    _default_nominal_values = dict(omega=3e3 * np.pi / 30, torque=0.0, i=170, epsilon=math.pi, u=300)
     _default_initializer = {'states':  {'i_sq': 0.0, 'i_sd': 0.0, 'epsilon': 0.0},
                             'interval': None,
                             'random_init': None,
@@ -1436,14 +1441,22 @@ class PermanentMagnetSynchronousMotor(SynchronousMotor):
 
     def _torque_limit(self):
         # Docstring of superclass
-        return self.torque([0, self._limits['i_sq'], 0])
+        mp = self._motor_parameter
[…]
