### 2020-07-21 76a80f8
correction of PMSM nominal current
