### 2020-07-21 530b33e
plot axes and legend modified according to review comments
