+++ gym_electric_motor/visualization/motor_dashboard.py
@@ -113,7 +113,7 @@ class MotorDashboard(ElectricMotorVisualization):
             self._figure_ep, axes_ep = plt.subplots(len(self._episode_plots))
             self._figure_ep.subplots_adjust(wspace=0.0, hspace=0.02)
             self._figure.subplots_adjust(wspace=0.0, hspace=0.2)
-            self._figure_ep.text(0.5, 0.04, 'episode_number', va='center', ha='center')
+            self._figure_ep.text(0.5, 0.04, 'episode', va='center', ha='center')
 
         # adding a common x-label to all the subplots in each figure
         self._figure.text(0.5, 0.04, 't/s', va='center', ha='center')
