### 2020-07-21 #2275 @madcowswe
Thanks for the invitation! I probably don't have time for a complete review, but here are some things I noticed while quickly looking at the paper:

1. The arrows for Usup and Uin should point from bottom to up (the other way).
2. The omega_me direction is ambigous as it's annotated to the rotation arrows in both directions
3. Figure 2: It seems that the dq frame presented for the bottom right plot (the current) is the rotor frame. However, the dq frame I am expecting for an asynchronous motor is one where the d axis is aligned to the rotor flux.
