### 2020-07-22 #2498 @whedon
[REVIEW]: gym-electric-motor (GEM): A Python toolbox for the simulation of electric drive systems
**Submitting author:** @wkirgsn (<a href="http://orcid.org/0000-0001-9490-1843">Wilhelm Kirchgässner</a>)
**Repository:** <a href="https://github.com/upb-lea/gym-electric-motor" target ="_blank">https://github.com/upb-lea/gym-electric-motor</a>
**Version:** v0.3.1
