### 2020-07-22 #2498 @wallscheid
Thanks for the comments from @madcowswe in the pre-rewie thread #2275. Please find some answers below:

>     1. Figure1: The convention I am used to (passive sign convention), the arrows for Usup and Uin should point from bottom to up (the other way).
The sign convention is an arbitrary choice. In energy conversion engineering (such as power electronic and drive systems) the so-called generator convention is mostly used in the international literature where the voltage arrow is pointing down as indicated.

>     2. Figrue1: The omega_me direction is ambigous as it's annotated to the rotation arrows in both directions
If you may have a closer look on the arrow labeling you can inspect that on the motor side we indicate both the rotation and torque orientation while on the mechanical load side we point out that the load torque has an opposite orientation compared to the motor torque (as usual in any free body diagram) and that the load torque is or might be depending on the shaft speed T(omega). Hence, on the shaft there are in total two torques indicated while the mechanical rotation orientation is only labeled once.

>     3. Figure 2: It seems that the dq frame presented for the bottom right plot (the current) is the rotor frame. However, the dq frame I am expecting for an asynchronous motor is one where the d axis is aligned to the rotor flux.
Sure, the usual field-oriented motor control definition is that the dq coordinate system is aligned towards the rotor flux (induction or synchronous machines). And this coordinate transformation can be directly used to also transform other important electrical motor quantities like the phase current or phase voltage into that very same dq coordinate framework. Based on that idea the usual field-oriented control starts with a current control loop using the dq-transformed currents. If you are interesting in more details on that I can recommend these [lecture notes](https://ei.uni-paderborn.de/fileadmin/elektrotechnik/fg/lea/Lehre/GDA/Dokumente/Geregelte_Drehstromantriebe_DE_EN.pdf) or the book by W. Leonhard 'Control of Electrical Drives' ,Springer, 2001.
