+++ examples/agents/simple_controllers.py
@@ -1,3 +1,5 @@
+import time
+
 from gym.spaces import Discrete, Box
 import sys
 import os
@@ -116,7 +118,7 @@ class PController(Controller):
     def __init__(self, environment, k_p=10, controller_no=0, state_idx=None, reference_idx=0):
         action_space = environment.action_space
         assert type(action_space) is Box and type(
-            environment.physical_system) is DcMotorSystem, 'No suitable action space for P Controller'
+            environment.physical_system), 'No suitable action space for P Controller'
         self._k_p = k_p
         self._controller_no = controller_no
         self._action_min = action_space.low[controller_no]
@@ -186,6 +188,50 @@ class IController(Controller):
         self._integrated_value = 0
 
 
+class DController(Controller):
+
+    def __init__(self, environment, k_d=1, controller_no=0, state_idx=None, reference_idx=0):
+        self._derivative_value = 0
+        self._current_time = time.time()
+        self._prev_time = self._current_time
+        self._prev_error = 0
+        action_space = environment.action_space
+        assert type(action_space) is Box and type(
+            environment.physical_system), 'No suitable action space for P Controller'
+        self._k_d = k_d
+        self._controller_no = controller_no
+        self._tau = environment.physical_system.tau
+        self._action_min = action_space.low[controller_no]
+        self._action_max = action_space.high[controller_no]
+        self._limits = environment.physical_system.limits
+        self._ref_idx = reference_idx
+        self._referenced_state = state_idx or np.argmax(environment.reference_generator.referenced_states)
+        self._referenced_state_max = self._limits[self._referenced_state] \
+                                     * environment.physical_system.state_space.high[self._referenced_state]
+        self._referenced_state_min = self._limits[self._referenced_state] \
+                                     * environment.physical_system.state_space.low[self._referenced_state]
+
+    def control(self, state, reference):
+        diff = reference[self._ref_idx] - state[self._referenced_state]
+        dt = self._current_time - self._prev_time
+        de = diff - self._prev_error
+        self._derivative_value = de / dt
+        if dt > 0:
+            return np.array([max(
+                self._action_min,
+                min(
+                    self._action_max,
+                    self._k_d * self._derivative_value
+                )
+            )
+            ])
+        self._prev_time = self._current_time
+        self._prev_error = diff
+
+    def reset(self, **__):
+        self._derivative_value = 0
+
+
 class PIController(PController):
     """
       The below Discrete PI Controller performs discrete-time PI controller computation using the error signal and proportional
@@ -195,7 +241,7 @@ class PIController(PController):
       Valid for DC Motor System
     """
 
-    def __init__(self, environment, k_p=10, k_i=0.01, controller_no=0, reference_idx=0):
+    def __init__(self, environment, k_p=10, k_i= 1, controller_no=0, reference_idx=0):
         super().__init__(environment, k_p, controller_no, reference_idx)
         action_space = environment.action_space
         assert type(action_space) is Box and type(
@@ -233,11 +279,72 @@ class PIController(PController):
                 self._action_min,
                 min(
                     self._action_max,
-                    self._k_p * (reference[0] - state[self._referenced_state])
+                    self._k_p * (reference[self._ref_idx] - state[self._referenced_state])
                     + self._tau / self._k_i * self._integrated_value
                 )
             )
         ])
+
+    def reset(self, **__):
+        self._integrated_value = 0
+
+
+class PIDController(PIController):
+    """
+      The below Discrete PI Controller performs discrete-time PI controller computation using the error signal and proportional
+      and integral gain inputs. The error signal is the difference between the reference_idx and the referenced_state.
[…]
