### 2020-07-27 38cec6b
start defining ODE of SRM
