+++ gym_electric_motor/physical_systems/electric_motors.py
@@ -878,6 +878,175 @@ class DcExternallyExcitedMotor(DcMotor):
         super()._update_limits(limit_agenda)
 
 
+class SwitchedReluctanceMotor(ElectricMotor):
+    """
+        The SwitchedReluctanceMotor implements the technical system of the switched reluctance motor.
+
+        This includes the system equations, the motor parameters, operational limits and bandwidth.
+
+        The model does not include saturation, the iron is assumed to be linear.
+
+        =====================  ==========  ============= ===========================================
+        Motor Parameter        Unit        Default Value Description
+        =====================  ==========  ============= ===========================================
+        r_s                    Ohm         0.01          Resistance of each stator phase winding
+        l_d                    H           23.60e-3      Stator inductance in aligned rotor position
+        l_q                    H           06.67e-3      Stator inductance in unaligned rotor position
+        p_s                    1           3             Number of stator pole pairs
+        alpha_s                1
+        p_r                    1           2             Number of rotor pole pairs
+        j_rotor                kg/m^2      0.0082        Moment of inertia of the rotor
+        =====================  ==========  ============= ===========================================
+
+        =============== ====== =============================================
+        Motor Currents  Unit   Description
+        =============== ====== =============================================
+
+        =============== ====== =============================================
+        =============== ====== =============================================
+        Motor Voltages  Unit   Description
+        =============== ====== =============================================
+
+        =============== ====== =============================================
+
+        ======== ===========================================================
+        Limits / Nominal Value Dictionary Entries:
+        -------- -----------------------------------------------------------
+        Entry    Description
+        ======== ===========================================================
+        i        General current limit / nominal value
+        i_a      Current in phase a
+        i_b      Current in phase b
+        i_c      Current in phase c
+        i_alpha  Current in alpha axis
+        i_beta   Current in beta axis
+        i_sd     Current in direct axis
+        i_sq     Current in quadrature axis
+        omega    Mechanical angular Velocity
+        epsilon  Electrical rotational angle
+        torque   Motor generated torque
+        u_a      Voltage in phase a
+        u_b      Voltage in phase b
+        u_c      Voltage in phase c
+        u_alpha  Voltage in alpha axis
+        u_beta   Voltage in beta axis
+        u_sd     Voltage in direct axis
+        u_sq     Voltage in quadrature axis
+        ======== ===========================================================
+    """
+    _default_motor_parameter = {'p_s': 4,
+                                'alpha_s': 20 * 2 * np.pi / 360,
+                                'p_r': 3,
+                                'alpha_r': 25 * 2 * np.pi / 360,
+                                'l_a': 72e-3,
+                                'l_u': 9e-3,
+                                'j_rotor': 0.057,
+                                'r_s': 0.9,
+                                }
+
[…]
