+++ examples/stable_baselines.py
@@ -9,11 +9,31 @@ from stable_baselines3 import DQN
 from stable_baselines3.dqn import MlpPolicy
 from gym.spaces import Discrete, Box
 from gym.wrappers import FlattenObservation
+from gym import ObservationWrapper
 
 """
 This example is based on stable baselines3 0.8.0a5. Since it is still being frequently updated 
 some parts of this code might be broken or not the latest recommended way of working with it
 """
+
+gamma = 0.99
+
+class EpsilonWrapper(ObservationWrapper):
+    """Changes Epsilon in a flattened observation to cos(epsilon) and sin(epsilon)"""
+    def __init__(self,env, epsilon_idx):
+        super(EpsilonWrapper, self).__init__(env)
+        self.EPSILON_IDX = epsilon_idx
+        new_low = np.concatenate((self.env.observation_space.low[:self.EPSILON_IDX], np.array([-1.]), self.env.observation_space.low[self.EPSILON_IDX:]))
+        new_high = np.concatenate((self.env.observation_space.high[:self.EPSILON_IDX], np.array([1.]), self.env.observation_space.high[self.EPSILON_IDX:]))
+
+        self.observation_space = Box(new_low, new_high)
+        
+    def observation(self,observation):
+        cos_eps = np.cos(observation[self.EPSILON_IDX]*np.pi)
+        sin_eps = np.sin(observation[self.EPSILON_IDX]*np.pi)
+        observation = np.concatenate((observation[:self.EPSILON_IDX], np.array([cos_eps, sin_eps]), observation[self.EPSILON_IDX+1:]))
+        return observation
+        
 
 class SqdCurrentMonitor:
     """
@@ -22,21 +42,19 @@ class SqdCurrentMonitor:
     i_sd**2 + i_sq**2 < 1.5 * nominal_limit 
     """
     def __init__(self, ):
-        self.I_SD_IDX = 0
-        self.I_SQ_IDX = 1
-        self.EPSILON_IDX = 2
-        self.CURRENTS_IDX = [0, 1]
-        self.CURRENTS = ['i_sd', 'i_sq']
-        self.VOLTAGES = ['u_sd', 'u_sq']
+        self.I_SD_IDX = 5
+        self.I_SQ_IDX = 6
     
     def __call__(self, state, observed_states, k, physical_system):
+        self.I_SD_IDX = physical_system.state_names.index('i_sd')
+        self.I_SQ_IDX = physical_system.state_names.index('i_sq')
         # normalize to limit_values, since state is normalized
         nominal_values = physical_system.nominal_state / abs(physical_system.limits)
         limits = 1.5 * nominal_values
         # calculating squared currents as observed measure 
         sqd_currents = state[self.I_SD_IDX]**2 + state[self.I_SQ_IDX]**2
         
-        return any(sqd_currents > limits)
+        return (sqd_currents > limits[self.I_SD_IDX] or sqd_currents > limits[self.I_SQ_IDX])
         
 
 motor_parameter = dict(p=3,  # [p] = 1, nb of pole pairs
@@ -60,8 +78,8 @@ rg = gem.reference_generators.MultipleReferenceGenerator([q_generator, d_generat
 env = gem.make(# define a PMSM with continuous action space
                "PMSMDisc-v1",
                
-               visualization = ConsolePrinter(verbose = 1),
-               #MotorDashboard(plots = ['i_sq', 'i_sd']),
+               #visualization = ConsolePrinter(verbose = 1),
+               visualization = MotorDashboard(plots = ['i_sq', 'i_sd', 'reward']),
     
                # parameterize the PMSM
                motor_parameter=motor_parameter,
@@ -75,46 +93,65 @@ env = gem.make(# define a PMSM with continuous action space
[…]
