### 2020-07-28 196e36e
changed tau back to 1e-5 -> Great results :)
