### 2020-07-29 e250e5b
new load - only works with hotfix of issue_#76
