### 2020-07-29 477cc41
fixed initialisation for ExternalSpeedLoad and ConstantSpeedLoad
