### 2020-07-29 b4f7e4d
added pytest for fixes and additional adaption in implementation
