### 2020-08-03 #2498 @moorepants
Here is my first pass on a review:

> Contribution and authorship: Has the submitting author (@wkirgsn) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Some of the paper authors, including the first author, do not seem to have made any commits to the software repository. Please clarify the authorship of all the authors.

These sentences "In particular, the domain of three-phase drives makes use of coordinate transformations to view these ODEs in a more interpretable frame. In GEM, both, the physical (abc-) and the simplified (dq-)coordinates are available to be used as the frame of input and output variables, allowing for easy and quick controller analysis and diagnose within the most convenient coordinate frame" are not understandable without some domain knowledge. Please define "frame" in this context as well as "physical (abc-) and the simplified (dq-)".

I don't think "generatoric" is a word in the English language.

You list 4 examples in the paper but these are not present in the documentation. I suggest adding these examples to the documentation so that users can easily find and understand the examples. The examples, as written are not easily interpretable by a user. There are very few comments in the files explaining what is going on. I think you need a tutorial style section of the documentation to teach users how to use your software. I don't think I could figure it out with the present information.

> Functionality: Have the functional claims of the software been confirmed?

This is not apparent from the documentation or the paper. You show one plot in the paper that is created from the software, but there is no associated code that I can run to reproduce that plot. I think a tutorial explaining your four examples that provide the inputs the user types and the subsequent outputs that the software produces would demonstrate the functionality (at least for those for examples).

> Installation

My attempt at installing the software failed. Here is the session:

```
moorepants@garuda:~$ conda create -n gem python=3
Collecting package metadata (current_repodata.json): done
Solving environment: done

## Package Plan ##

  environment location: /home/moorepants/miniconda3/envs/gem

  added / updated specs:
    - python=3

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    _libgcc_mutex-0.1          |      conda_forge           3 KB  conda-forge
    _openmp_mutex-4.5          |            1_gnu          22 KB  conda-forge
    certifi-2020.6.20          |   py38h32f6830_0         151 KB  conda-forge
    python-3.8.5               |h4d41432_2_cpython        71.0 MB  conda-forge
    python_abi-3.8             |           1_cp38           4 KB  conda-forge
    setuptools-49.2.1          |   py38h32f6830_0         931 KB  conda-forge
    ------------------------------------------------------------
                                           Total:        72.1 MB

The following NEW packages will be INSTALLED:

  _libgcc_mutex      conda-forge/linux-64::_libgcc_mutex-0.1-conda_forge
  _openmp_mutex      conda-forge/linux-64::_openmp_mutex-4.5-1_gnu
  ca-certificates    conda-forge/linux-64::ca-certificates-2020.6.20-hecda079_0
  certifi            conda-forge/linux-64::certifi-2020.6.20-py38h32f6830_0
  ld_impl_linux-64   conda-forge/linux-64::ld_impl_linux-64-2.34-hc38a660_9
  libffi             conda-forge/linux-64::libffi-3.2.1-he1b5a44_1007
  libgcc-ng          conda-forge/linux-64::libgcc-ng-9.3.0-h24d8f2e_14
  libgomp            conda-forge/linux-64::libgomp-9.3.0-h24d8f2e_14
  libstdcxx-ng       conda-forge/linux-64::libstdcxx-ng-9.3.0-hdf63c60_14
  ncurses            conda-forge/linux-64::ncurses-6.2-he1b5a44_1
  openssl            conda-forge/linux-64::openssl-1.1.1g-h516909a_1
  pip                conda-forge/noarch::pip-20.2-py_0
  python             conda-forge/linux-64::python-3.8.5-h4d41432_2_cpython
  python_abi         conda-forge/linux-64::python_abi-3.8-1_cp38
  readline           conda-forge/linux-64::readline-8.0-he28a2e2_2
  setuptools         conda-forge/linux-64::setuptools-49.2.1-py38h32f6830_0
  sqlite             conda-forge/linux-64::sqlite-3.32.3-hcee41ef_1
  tk                 conda-forge/linux-64::tk-8.6.10-hed695b0_0
  wheel              conda-forge/noarch::wheel-0.34.2-py_1
  xz                 conda-forge/linux-64::xz-5.2.5-h516909a_1
  zlib               conda-forge/linux-64::zlib-1.2.11-h516909a_1006

Downloading and Extracting Packages
_openmp_mutex-4.5    | 22 KB     | ############################################################## | 100%
python-3.8.5         | 71.0 MB   | ############################################################## | 100%
_libgcc_mutex-0.1    | 3 KB      | ############################################################## | 100%
setuptools-49.2.1    | 931 KB    | ############################################################## | 100%
certifi-2020.6.20    | 151 KB    | ############################################################## | 100%
python_abi-3.8       | 4 KB      | ############################################################## | 100%
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate gem
#
# To deactivate an active environment, use
#
#     $ conda deactivate

moorepants@garuda:~$ act gem
(gem) moorepants@garuda:~$ which pip
/home/moorepants/miniconda3/envs/gem/bin/pip
(gem) moorepants@garuda:~$ pip install gym-electric-motor
Collecting gym-electric-motor
  Downloading gym_electric_motor-0.2.1-py3-none-any.whl (140 kB)
     |████████████████████████████████| 140 kB 1.5 MB/s
Collecting scipy>=1.4.1
  Downloading scipy-1.5.2-cp38-cp38-manylinux1_x86_64.whl (25.7 MB)
     |████████████████████████████████| 25.7 MB 3.0 MB/s
Collecting pytest-cov
  Downloading pytest_cov-2.10.0-py2.py3-none-any.whl (19 kB)
Collecting PyQT5==5.9.2
  Downloading PyQt5-5.9.2-5.9.3-cp35.cp36.cp37-abi3-manylinux1_x86_64.whl (105.3 MB)
     |████████████████████████████████| 105.3 MB 26 kB/s
Collecting matplotlib>=3.1.0
  Downloading matplotlib-3.3.0-1-cp38-cp38-manylinux1_x86_64.whl (11.5 MB)
     |████████████████████████████████| 11.5 MB 2.5 MB/s
Collecting numpy>=1.16.4
  Downloading numpy-1.19.1-cp38-cp38-manylinux2010_x86_64.whl (14.5 MB)
     |████████████████████████████████| 14.5 MB 655 kB/s
Collecting pytest>=5.2.2
  Downloading pytest-6.0.1-py3-none-any.whl (270 kB)
     |████████████████████████████████| 270 kB 1.6 MB/s
Collecting gym>=0.15.4
  Downloading gym-0.17.2.tar.gz (1.6 MB)
     |████████████████████████████████| 1.6 MB 2.3 MB/s
Collecting coverage>=4.4
  Downloading coverage-5.2.1-cp38-cp38-manylinux1_x86_64.whl (231 kB)
     |████████████████████████████████| 231 kB 2.5 MB/s
ERROR: Could not find a version that satisfies the requirement sip<4.20,>=4.19.4 (from PyQT5==5.9.2->gym-electric-motor) (from versions: 5.0.0, 5.0.1, 5.1.0, 5.1.1, 5.1.2, 5.2.0, 5.3.0)
ERROR: No matching distribution found for sip<4.20,>=4.19.4 (from PyQT5==5.9.2->gym-electric-motor)
```

Also not clear why testing related packages are installed as dependencies, shouldn't these only be required if you are developing the package?

> Statement of need

This could be more explicit and there could be more. When I read the paper I interpreted it as Matlab does this already but doesn't connect to keras/tensorflow well, so you need this package. That is a sufficient need, but is that the main/only need?

> Installation instructions

I don't see an explicit list of dependencies. This would be helpful for setting up a conda environment, for example. There is also a dependency issue pip encounters as shown above. I don't see any info in the docs about dependencies that would help me debug this.

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There are 4 examples but they need to be improved to be understandable by a new user.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

API docs look good.

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Looks like there are tests, but haven't been able to run due to installation issue.

>  Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Needs some improvement for non-specialist audience. Define or reduce jargon, for example.

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?

Little is explained about this besides saying that Matlab does something similar. More detail is needed here.

>  Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

There are some improvements in grammar needed.

>  References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

This seems brief. Maybe you'll have more references to actual electric motor software once you improve your state of the field section.
