### 2020-08-03 #2498 @wkirgsn
Thank you for the very constructive review, @moorepants!
We appreciate every remark and are eager to improve on these discrepancies.

We are working on a major update of the repo, which becomes manifest in the `nightly` branch. There, the dependency problem preventing you from installation is already solved, along the incorporation of more features and bug fixes.
Although this development branch is already quite packed and far ahead of the `master`, we're working on completing tutorials that were also requested by your review, that will delay the eventual merge of the `nightly` into `master` until presumably mid September.

For the meantime, please use GEM `master` with python 3.6 or 3.7, or just the `nightly` branch. I am going to add a restriction on the python version for the master branch for the time being until the major update to prevent other users experiencing these issues.
