### 2020-08-03 406caa8
Prohibit python 3.8 for now. Must be reverted when nightly is merged into master! #77
