### 2020-08-07 fdc5aca
fix matplotlib requirement to v3.1.2 as unit tests break with v3.3.0
