+++ requirements.txt
@@ -1,4 +1,4 @@
-matplotlib>=3.1.0
+matplotlib==3.1.2
 numpy>=1.16.4
 scipy>=1.4.1
 gym>=0.15.4
