### 2020-08-10 a26f998
fixed error in flux initializer and limits updater
