+++ examples/stable_baselines.py
@@ -2,8 +2,7 @@ import sys, os
 sys.path.append(os.path.abspath(os.path.join('..')))
 
 import gym_electric_motor as gem
-from gym_electric_motor.physical_systems.converters import DiscMultiConverter, DiscFourQuadrantConverter, DiscTwoQuadrantConverter
-from gym_electric_motor.visualization import MotorDashboard, ConsolePrinter
+from gym_electric_motor.visualization import MotorDashboard
 import numpy as np
 from stable_baselines3 import DQN
 from stable_baselines3.dqn import MlpPolicy
@@ -18,6 +17,22 @@ some parts of this code might be broken or not the latest recommended way of wor
 """
 
 gamma = 0.99
+tau = 1e-5
+simulation_time = 5 # seconds
+buffer_size = 200000 #number of old obersation steps saved
+learning_starts = 10000 # memory warmup
+train_freq = 1 # prediction network gets an update each train_freq's step
+batch_size = 25 # mini batch size drawn at each update step
+policy_kwargs = {
+        'net_arch': [64,64] # hidden layer size of MLP
+        }
+
+exploration_fraction = 0.1 # Fraction of training steps the epsilon decays 
+target_update_interval = 1000 # Target network gets updated each target_update_interval's step
+verbose = 1 # verbosity of stable basline's prints
+
+
+
 
 class EpsilonWrapper(ObservationWrapper):
     """Changes Epsilon in a flattened observation to cos(epsilon) and sin(epsilon)"""
@@ -76,24 +91,24 @@ q_generator = gem.reference_generators.WienerProcessReferenceGenerator(reference
 d_generator = gem.reference_generators.WienerProcessReferenceGenerator(reference_state='i_sd')
 rg = gem.reference_generators.MultipleReferenceGenerator([q_generator, d_generator])
 
-env = gem.make(# define a PMSM with continuous action space
+env = gem.make(# define a PMSM with discrete action space
                "PMSMDisc-v1",
                
-               #visualization = ConsolePrinter(verbose = 1),
+               #visualize the results
                visualization = MotorDashboard(plots = ['i_sq', 'i_sd', 'reward']),
     
-               # parameterize the PMSM
+               #parameterize the PMSM
                motor_parameter=motor_parameter,
                #converter = converter,
     
-               # update the limitations of the state space
+               #update the limitations of the state space
                limit_values=limit_values,
                nominal_values=nominal_values,
                
                # define the DC link voltage
                u_sup=u_sup, 
                
-               # define the speed at which the motor is operated
+               # define the speed at which the motor is operated - should be drawn randomly at each episode
                load='ConstSpeedLoad',
                load_initializer = {'states': {'omega': 1000 * np.pi / 30,},
                                    'interval': [[-4000*2*np.pi/60, 4000*2*np.pi/60]],
@@ -102,7 +117,7 @@ env = gem.make(# define a PMSM with continuous action space
                #gem.physical_systems.ConstantSpeedLoad(omega_fixed=1000 * np.pi / 30), 
                
                # define the duration of one sampling step
-               tau=1e-5,
+               tau=tau,
                
                # turn off terminations via limit violation and parameterize the reward function
                reward_function=gem.reward_functions.WeightedSumOfErrors(observed_states=['i_sq', 'i_sd'], 
@@ -116,10 +131,7 @@ env = gem.make(# define a PMSM with continuous action space
     
                # define a numerical solver of adequate accuracy
                ode_solver='euler',
-    
[…]
