### 2020-08-17 bdc7349
random state initialization added
