### 2020-08-17 fc7b61d
fixed init for constant, external speed load
