+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -365,7 +365,7 @@ class ExternalSpeedLoad(MechanicalLoad):
         super().__init__(load_initializer=load_initializer, **kwargs)
         self.kwargs = kwargs
         self._omega_initial = omega_initial or self._initializer['states']['omega']
-        if omega_initial is not None:
+        if omega_initial:
             self._initializer['states']['omega'] = omega_initial
         self._speed_profile = speed_profile
         self._tau = tau
@@ -413,7 +413,7 @@ class ConstantSpeedLoad(MechanicalLoad):
         """
         super().__init__(load_initializer=load_initializer, **kwargs)
         self._omega = omega_fixed or self._initializer['states']['omega']
-        if omega_fixed is not None:
+        if omega_fixed:
             self._initializer['states']['omega'] = omega_fixed
 
     def mechanical_ode(self, *_, **__):
