### 2020-08-19 7790a05
Added execution timer for the DQN
