+++ examples/stable_baselines.py
@@ -9,6 +9,7 @@ from stable_baselines3.dqn import MlpPolicy
 from gym.spaces import Discrete, Box
 from gym.wrappers import FlattenObservation, TimeLimit
 from gym import ObservationWrapper
+import time
 
 """
 This example is based on stable baselines3 0.8.0a5. Since it is still being frequently updated 
@@ -142,10 +143,11 @@ env.action_space = Discrete(7)
 nb_steps = int(simulation_time // tau)
 
 
-
+start_time = time.time()
 model = DQN(MlpPolicy, env, buffer_size=buffer_size, learning_starts=learning_starts ,train_freq=train_freq, batch_size=batch_size, gamma=gamma,
             policy_kwargs=policy_kwargs, exploration_fraction=exploration_fraction, target_update_interval=target_update_interval,
             verbose=verbose).learn(total_timesteps=nb_steps)
+print(f'Execution time of stable baselines3 DQN is: {time.time()-start_time}')
 
 #in case your want to save the model for further evalutation
 model.save("dqn_PMSM")
