### 2020-08-19 bfe629a
Timer print formatting
