### 2020-08-19 d2efda8
change init from externalload to a not random kind, to avoid problems at reset. Added usage comments in examples
