+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -359,14 +359,17 @@ class ExternalSpeedLoad(MechanicalLoad):
                         amplitude(float), freq(float), time(float)
 
             omega_initial(float)): Initial value for the speed in rad/s.
+                the initial value must be the same as the speed profile
+                offset for omptimal usage
             tau(float): discrete time step of the system
             kwargs(dict): further arguments for speed_profile
         """
+        self._initializer = self._default_initializer.copy()
+        self._initializer['states']['omega'] = omega_initial
         super().__init__(load_initializer=load_initializer, **kwargs)
+
         self.kwargs = kwargs
-        self._omega_initial = omega_initial or self._initializer['states']['omega']
-        if omega_initial:
-            self._initializer['states']['omega'] = omega_initial
+        self._omega_initial = omega_initial
         self._speed_profile = speed_profile
         self._tau = tau
         #self._jacobi = jacobi
