### 2020-08-19 1d340b1
changed referencing in Joss docu from hyperlinks to reference section
