### 2020-08-19 2c4a91b
changed some jabref entries
