### 2020-08-19 63725fd
changed jabref entries a little more
