+++ docs/JOSS/Literature.bib
@@ -30,66 +30,59 @@
 }
 
 @Misc{MathWorks,
-  author       = {MATLAB},
-  howpublished = {{The MathWorks, Inc.}},
-  title        = {R2020a},
-  year         = {2020},
-  url          = {https://www.mathworks.com/products/matlab.html},
+  author = {{The MathWorks, Inc.}},
+  title  = {MATLAB - Simulink R2020a},
+  year   = {2020},
+  url    = {https://www.mathworks.com/products/matlab.html},
 }
 
 @Misc{dSPACE,
-  author       = {VEOS},
-  howpublished = {{dSPACE digital signal processing and control engineering GmbH}},
-  title        = {4.5},
-  year         = {2019},
-  url          = {https://www.dspace.com/en/inc/home/products/sw/simulation_software/veos.cfm},
+  author = {{dSPACE digital signal processing and control engineering GmbH}},
+  title  = {VEOS 4.5},
+  year   = {2019},
+  url    = {https://www.dspace.com/en/inc/home/products/sw/simulation_software/veos.cfm},
 }
 
 @Misc{OPAL-RT,
-  author       = {HYPERSIM},
-  howpublished = {{OPAL-RT TECHNOLOGIES, Inc.}},
-  title        = {2020.1},
-  year         = {2020},
-  url          = {https://www.opal-rt.com/systems-hypersim/},
+  author = {{OPAL-RT TECHNOLOGIES, Inc.}},
+  title  = {HYPERSIM 2020.1},
+  year   = {2020},
+  url    = {https://www.opal-rt.com/systems-hypersim/},
 }
 
 @Misc{Margolis,
-  author       = {SimuPy},
-  howpublished = {{Benjamin W. L. Margolis}},
-  title        = {1.0.0},
-  year         = {2017},
-  url          = {https://github.com/simupy/simupy},
+  author = {Benjamin W. L. Margolis},
+  title  = {SimuPy 1.0.0},
+  year   = {2017},
+  url    = {https://github.com/simupy/simupy},
 }
 
 @Misc{ANSYS,
-  author       = {{ANSYS Maxwell}},
-  howpublished = {{ANSYS, Inc.}},
-  title        = {2020 R1},
-  year         = {2020},
-  url          = {https://www.ansys.com/products/electronics/ansys-maxwell},
+  author = {{ANSYS, Inc.}},
+  title  = {{ANSYS Maxwell} 2020 R1},
+  year   = {2020},
+  url    = {https://www.ansys.com/products/electronics/ansys-maxwell},
 }
 
 @Misc{MotorDesignLtd,
-  author       = {{Motor-CAD}},
-  howpublished = {{Motor Design Ltd.}},
-  title        = {v12},
-  year         = {2020},
-  url          = {https://www.motor-design.com/motor-cad-software/},
+  author = {{Motor Design Ltd.}},
+  title  = {{Motor-CAD} v12},
+  year   = {2020},
+  url    = {https://www.motor-design.com/motor-cad-software/},
 }
 
 @Misc{ElectroMagneticWorks,
-  author       = {{MotorWizard}},
-  howpublished = {{ElectroMagneticWorks, Inc.}},
-  year         = {2020},
-  url          = {https://www.emworks.com/product/MotorWizard},
+  author = {{ElectroMagneticWorks, Inc.}},
+  title  = {{MotorWizard}},
+  year   = {2020},
+  url    = {https://www.emworks.com/product/MotorWizard},
 }
 
 @Misc{Meeker,
-  author       = {{FEMM (Finite Element Method Magnetics)}},
-  howpublished = {{David Meeker}},
-  title        = {4.2},
-  year         = {2020},
[…]
