### 2020-08-19 8728865
typos in joss doc
