+++ gym_electric_motor/physical_systems/mechanical_loads.py
@@ -1,18 +1,20 @@
 import numpy as np
 from scipy.stats import truncnorm
-
+import warnings
 
 # todo updateting intialization with load_initializer,
 # now can be initialized directly in load class
 class MechanicalLoad:
     """
-    The MechanicalLoad is the base class for all the mechanical systems attached to the electrical motors rotor.
+    The MechanicalLoad is the base class for all the mechanical systems attached
+    to the electrical motors rotor.
 
-    It contains an mechanical ode system as well as the state names, limits and nominal values
-    of the mechanical quantities. The only required state is 'omega' as the rotational speed of the motor shaft
-    in rad/s.
-    ExternalSpeedLoad and ConstantSpeedLoad can be initialized by giving the
-    initializer as an class parameter by instanziation
+    It contains an mechanical ode system as well as the state names, limits and
+    nominal values of the mechanical quantities. The only required state is
+    'omega' as the rotational speed of the motor shaft in rad/s.
+    ConstantSpeedLoad can be initialized with the initializer as an
+    class parameter by instantiation. ExternalSpeedLoad takes the first value
+    of the SpeedProfile as initial value.
 
     Initialization is given by initializer(dict). Can be a constant state value
     or random value in given interval.
@@ -22,6 +24,7 @@ class MechanicalLoad:
                     (only for random init), shape(num states, 2)
           'random_init'(str): 'uniform' or 'normal'
           'random_params(tuple): mue(float), sigma(int)
+
     Example initializer(dict) for constant initialization:
         { 'states': {'omega': 16.0}}
     Example  initializer(dict) for random initialization:
@@ -334,11 +337,6 @@ class ExternalSpeedLoad(MechanicalLoad):
 
     HAS_JACOBIAN = False
 
-    _default_initializer = {'states': {'omega': 0.0},
-                            'interval': None,
-                            'random_init': None,
-                            'random_params': (None, None)}
-
     @property
     def omega(self):
         """
@@ -347,8 +345,8 @@ class ExternalSpeedLoad(MechanicalLoad):
         """
         return self._omega_initial
 
-    def __init__(self, speed_profile=None, omega_initial=0,
-                 tau=1e-4, load_initializer=None, **kwargs):
+    def __init__(self, speed_profile, tau=1e-4,
+                 load_initializer=None, **kwargs):
         """
         Args:
             speed_profile(function): function or lambda expression
@@ -357,20 +355,20 @@ class ExternalSpeedLoad(MechanicalLoad):
                     (lambda t, amplitude, freq: amplitude*numpy.sin(2*pi*f)))
                     with additional parameters:
                         amplitude(float), freq(float), time(float)
-
-            omega_initial(float)): Initial value for the speed in rad/s.
-                the initial value must be the same as the speed profile
-                offset for omptimal usage
             tau(float): discrete time step of the system
             kwargs(dict): further arguments for speed_profile
         """
-        self._initializer = self._default_initializer.copy()
-        self._initializer['states']['omega'] = omega_initial
-        super().__init__(load_initializer=load_initializer, **kwargs)
+
+        super().__init__(**kwargs)
+        warnings.warn('Given initializer will be overwritten with starting value from speed-profile, '
+                      'to avoid complications at the load reset. It is recommended to choose starting value of'
+                      ' load by the defined speed-profile.',
+                      UserWarning)
 
         self.kwargs = kwargs
-        self._omega_initial = omega_initial
         self._speed_profile = speed_profile
+        # setting initial load as speed-profile at time 0
+        self._omega_initial = self._speed_profile(t=0, **self.kwargs)
         self._tau = tau
[…]
