### 2020-08-20 9ee48a2
changes to joss_docu
