+++ docs/JOSS/Literature.bib
@@ -37,7 +37,7 @@
 }
 
 @Misc{dSPACE,
-  author = {{dSPACE digital signal processing and control engineering GmbH}},
+  author = {dSPACE GmbH},
   title  = {VEOS 4.5},
   year   = {2019},
   url    = {https://www.dspace.com/en/inc/home/products/sw/simulation_software/veos.cfm},
@@ -80,7 +80,7 @@
 
 @Misc{Meeker,
   author = {David Meeker},
-  title  = {{FEMM (Finite Element Method Magnetics)} 4.2},
+  title  = {{FEMM} 4.2},
   year   = {2020},
   url    = {http://www.femm.info/wiki/HomePage},
 }
