+++ docs/JOSS/Literature.bib
@@ -37,7 +37,7 @@
 }
 
 @Misc{dSPACE,
-  author = {dSPACE GmbH},
+  author = {{dSPACE GmbH}},
   title  = {VEOS 4.5},
   year   = {2019},
   url    = {https://www.dspace.com/en/inc/home/products/sw/simulation_software/veos.cfm},
