### 2020-08-20 a27928e
add assertion whether action is in action space for all Disc converters
