### 2020-08-20 cafb2c6
add unit tests for action_space.contains(X) assertion
