### 2020-08-20 a7ba4a2
add action_space assertion also for compound converters Disc-4QC and Disc-B6C
