### 2020-08-20 f283871
add action_space assertion also for compound converters Disc-4QC and Disc-B6C
