+++ tests/test_physical_systems/test_converters.py
@@ -884,6 +884,20 @@ class TestDiscFourQuadrantConverter(TestDiscConverter):
             assert converter._subconverters[0].last_action == action // 2 + 1
             assert converter._subconverters[1].last_action == action % 2 + 1
 
+        converter = self.class_to_test()
+        time = 0
+        with pytest.raises(AssertionError) as assertText:
+            converter.set_action(-1, time)
+        assert "-1" in str(assertText.value) and "Discrete(4)" in str(assertText.value)
+
+        with pytest.raises(AssertionError) as assertText:
+            converter.set_action(int(1e9), time)
+        assert str(int(1e9)) in str(assertText.value) and "Discrete(4)" in str(assertText.value)
+
+        with pytest.raises(AssertionError) as assertText:
+            converter.set_action(np.pi, time)
+        assert str(np.pi) in str(assertText.value) and "Discrete(4)" in str(assertText.value)
+
     @pytest.mark.parametrize('i_out', [[-12], [0], [12]])
     def test_convert(self, converter, i_out):
         t = np.random.rand()
@@ -1221,6 +1235,19 @@ class TestDiscB6BridgeConverter(TestDiscConverter):
             assert converter._subconverters[2].last_t == t
             subactions = [sc.last_action % 2 for sc in converter._subconverters]
             assert action == reduce(lambda x, y: 2*x+y, subactions)
+
+        time = 0
+        with pytest.raises(AssertionError) as assertText:
+            converter.set_action(-1, time)
+        assert "-1" in str(assertText.value) and "Discrete(8)" in str(assertText.value)
+
+        with pytest.raises(AssertionError) as assertText:
+            converter.set_action(int(1e9), time)
+        assert str(int(1e9)) in str(assertText.value) and "Discrete(8)" in str(assertText.value)
+
+        with pytest.raises(AssertionError) as assertText:
+            converter.set_action(np.pi, time)
+        assert str(np.pi) in str(assertText.value) and "Discrete(8)" in str(assertText.value)
 
     @pytest.mark.parametrize('i_out', [[-1, -1, 0], [1, 1, -2], [0, 0, 1]])
     def test_convert(self, converter, i_out):
