### 2020-08-20 5cc9d57
fixed warning in external speed load
