+++ requirements.txt
@@ -5,4 +5,3 @@ gym>=0.15.4
 pytest>=5.2.2
 pytest-cov
 pytest-lambda
-warnings
