### 2020-08-20 #2498 @wkirgsn
@Kevin-Mattheus-Moerman The initial estimate of mid september for the major update still seems to hold for now.
The nightly branch is updated continuously, and the paper is also currently worked on, incorporating moorepants' remarks.
