+++ docs/JOSS/paper.md
@@ -50,59 +50,64 @@ components to be included in the simulation framework.
 
 # Field of Application
 
-Electric drive systems are an important topic both in academic and 
-industrial research due to the worldwide availability and deployment of such 
-plants. Control algorithms for these systems have usually been designed, parameterized and 
+Electric drive systems and their control are an important topic both in academic and 
+industrial research due to their worldwide usage and deployment. 
+Control algorithms for these systems have usually been designed, parameterized and 
 tested within ``MATLAB - Simulink`` [@MathWorks], 
 which is developed and promoted specifically for
 such engineering tasks. In the more recent past, however, commercial software like
 ``MATLAB`` has difficulties to stay on par with the expandability and flexibility offered 
 by open-source libraries that are available for more accessible programming languages like Python. 
+In consequence, a Python-based drive simulation framework like ``GEM`` is an evident step in order to accelerate corresponding control research and development.
 Moreover, the latest efforts concerning industrial application of reinforcement-learning control 
-algorithms heavily depend on Python packages like ``Keras`` [@Chollet2015] or ``Tensorflow`` [@tensorflow2015-whitepaper]. 
+algorithms heavily depend on Python packages like 
+``Keras`` [@Chollet2015], ``Tensorflow`` [@tensorflow2015-whitepaper] or ``PyTorch``[@NEURIPS2019_9015]. 
+Hence, the built-in OpenAI ``gym`` interface allows to easily couple ``GEM`` to other open-source reinforcement 
+learning toolboxes such as ``Stable Baselines3`` [@stable-baselines3], ``TF-Agents`` [@TFAgents] or ``keras-rl`` [@plappert2016kerasrl].
 
-In consequence, a Python-based drive simulation framework is an evident step in order to accelerate corresponding control research and development.
 Providing easy access to the non-commercial, open-source ``GEM`` library allows users from any engineering domain to 
 include accurate drive models into their simulations, also beyond the topic of control applications.
 Concerning the predominance of commercial software like ``MATLAB`` for educational purposes, a free-of-charge simulation alternative
 that does not force students or institutions to pay for licenses, has great potential to support and encourage
-training of new talents in the field of electrical drives.
+training of new talents in the field of electrical drives and neighbouring domains (e.g. power electronics or energy systems).
 
 # Related software
 
-Due to the strong dependence of industrial development chains on simulation data
-there is a comprehensive collection of commercial software that enables numerical analysis of 
-every facet of electric drives. To name just a few, ``MATLAB - Simulink``
-is probably the most popular software environment for numerical analysis in engineering.
+Due to the strong dependence of industrial development chains on simulation data there is a comprehensive collection of
+commercial software that enables numerical analysis of every facet of electric drives. To name just a few,
+``MATLAB - Simulink`` is probably the most popular software environment for numerical analysis in engineering.
 Herein, ``MATLAB`` is providing for a scientific calculation framework and ``Simulink`` for a simulation environment
-with a very large field of applications. Examples that are designed for real-time capability (e.g. for hardware-in-the-loop prototyping) can be found in 
-``VEOS`` [@dSPACE]
-or ``HYPERSIM`` [@OPAL-RT].
-Non-commercial simulation libraries exist, but they rarely come with predefined system models. 
-An exemplary package from this category is ``SimuPy`` [@Margolis], which provides lots of 
-flexibility for the synthesis of simulation models, but also requires the user to possess the necessary expert knowledge
-in order to implement a desired system model.
+with a very large field of applications. Examples that are designed for real-time capability 
+(e.g. for hardware-in-the-loop prototyping) can be found in ``VEOS`` [@dSPACE] or ``HYPERSIM`` [@OPAL-RT].
[…]
