+++ docs/JOSS/paper.md
@@ -41,25 +41,26 @@ bibliography: Literature.bib
 The ``gym-electric-motor`` (``GEM``) library provides simulation environments for 
 electrical drive systems and, therefore, allows to easily design and analyze drive control
 solutions in Python. Since ``GEM`` is strongly inspired by OpenAI's ``gym`` [@gym-whitepaper], it 
-is particulary well-equipped for (but not limited to) applications in the field of 
+is particularly well-equipped for (but not limited to) applications in the field of 
 reinforcement-learning-based control algorithms. In addition, the interface allows to plug in any expert-driven 
 control approach, such as model predictive control, to be tested  and to perform benchmark comparisons. 
 The ``GEM`` package includes a wide variety of motors, power electronic converters and mechanical load models that 
-can be flexibly selected and parameterized via the API. A modular software code also allows additional system 
+can be flexibly selected and parameterized via the API. A modular structure allows additional system 
 components to be included in the simulation framework.
 
 # Field of Application
 
-Electric drive systems and their control are an important topic both in academic and 
+Electric drive systems and their control are an important topic in both academic and 
 industrial research due to their worldwide usage and deployment. 
 Control algorithms for these systems have usually been designed, parameterized and 
 tested within ``MATLAB - Simulink`` [@MathWorks], 
 which is developed and promoted specifically for
 such engineering tasks. In the more recent past, however, commercial software like
-``MATLAB`` has difficulties to stay on par with the expandability and flexibility offered 
+``MATLAB`` has difficulties to stay on par with state-of-the-art concepts 
+for scientific modeling and the flexibility offered 
 by open-source libraries that are available for more accessible programming languages like Python. 
-In consequence, a Python-based drive simulation framework like ``GEM`` is an evident step in order to accelerate corresponding control research and development.
-Moreover, the latest efforts concerning industrial application of reinforcement-learning control 
+Consequently, a Python-based drive simulation framework like ``GEM`` is an evident step in order to accelerate corresponding control research and development.
+Specifically, the latest efforts concerning industrial application of reinforcement-learning control 
 algorithms heavily depend on Python packages like 
 ``Keras`` [@Chollet2015], ``Tensorflow`` [@tensorflow2015-whitepaper] or ``PyTorch``[@NEURIPS2019_9015]. 
 Hence, the built-in OpenAI ``gym`` interface allows to easily couple ``GEM`` to other open-source reinforcement 
@@ -67,17 +68,18 @@ learning toolboxes such as ``Stable Baselines3`` [@stable-baselines3], ``TF-Agen
 
 Providing easy access to the non-commercial, open-source ``GEM`` library allows users from any engineering domain to 
 include accurate drive models into their simulations, also beyond the topic of control applications.
-Concerning the predominance of commercial software like ``MATLAB`` for educational purposes, a free-of-charge simulation alternative
+Considering the prevalence of commercial software like ``MATLAB`` for educational purposes, a free-of-charge simulation alternative
 that does not force students or institutions to pay for licenses, has great potential to support and encourage
 training of new talents in the field of electrical drives and neighbouring domains (e.g. power electronics or energy systems).
 
 # Related software
 
-Due to the strong dependence of industrial development chains on simulation data there is a comprehensive collection of
+Due to the strong dependence of downstream industrial development on simulated environments
+there is a comprehensive variety of
 commercial software that enables numerical analysis of every facet of electric drives. To name just a few,
 ``MATLAB - Simulink`` is probably the most popular software environment for numerical analysis in engineering.
-Herein, ``MATLAB`` is providing for a scientific calculation framework and ``Simulink`` for a simulation environment
-with a very large field of applications. Examples that are designed for real-time capability 
[…]
