+++ docs/JOSS/Literature.bib
@@ -107,7 +107,7 @@
 }
 
 @Misc{TFAgents,
-  author       = {{Sergio Guadarrama and Anoop Korattikara and Oscar Ramirez and Pablo Castro and Ethan Holly and Sam Fishman and Ke Wang and Ekaterina Gonina and Neal Wu and Efi Kokiopoulou and Luciano Sbaiz and Jamie Smith and Gábor Bartók and Jesse Berent and Chris Harris and Vincent Vanhoucke and Eugene Brevdo}},
+  author       = {Sergio Guadarrama and Anoop Korattikara and Oscar Ramirez and Pablo Castro and Ethan Holly and Sam Fishman and Ke Wang and Ekaterina Gonina and Neal Wu and Efi Kokiopoulou and Luciano Sbaiz and Jamie Smith and Gábor Bartók and Jesse Berent and Chris Harris and Vincent Vanhoucke and Eugene Brevdo},
   howpublished = {\url{https://github.com/tensorflow/agents}},
   note         = {[Online; accessed 25-June-2019]},
   title        = {{TF-Agents}: A library for Reinforcement Learning in TensorFlow},
