+++ docs/JOSS/Literature.bib
@@ -29,4 +29,117 @@
   eprint = {arXiv:1606.01540},
 }
 
+@Misc{MathWorks,
+  author = {{The MathWorks, Inc.}},
+  title  = {MATLAB - Simulink R2020a},
+  year   = {2020},
+  url    = {https://www.mathworks.com/products/matlab.html},
+}
+
+@Misc{dSPACE,
+  author = {{dSPACE GmbH}},
+  title  = {VEOS 4.5},
+  year   = {2019},
+  url    = {https://www.dspace.com/en/inc/home/products/sw/simulation_software/veos.cfm},
+}
+
+@Misc{OPAL-RT,
+  author = {{OPAL-RT TECHNOLOGIES, Inc.}},
+  title  = {HYPERSIM 2020.1},
+  year   = {2020},
+  url    = {https://www.opal-rt.com/systems-hypersim/},
+}
+
+@Misc{Margolis,
+  author  = {Benjamin W. L. Margolis},
+  title   = {SimuPy 1.0.0},
+  year    = {2017},
+  journal = {GitHub repository},
+  url     = {https://github.com/simupy/simupy},
+}
+
+@Misc{ANSYS,
+  author = {{ANSYS, Inc.}},
+  title  = {{ANSYS Maxwell} 2020 R1},
+  year   = {2020},
+  url    = {https://www.ansys.com/products/electronics/ansys-maxwell},
+}
+
+@Misc{MotorDesignLtd,
+  author = {{Motor Design Ltd.}},
+  title  = {{Motor-CAD} v12},
+  year   = {2020},
+  url    = {https://www.motor-design.com/motor-cad-software/},
+}
+
+@Misc{ElectroMagneticWorks,
+  author = {{ElectroMagneticWorks, Inc.}},
+  title  = {{MotorWizard}},
+  year   = {2020},
+  url    = {https://www.emworks.com/product/MotorWizard},
+}
+
+@Misc{Meeker,
+  author = {David Meeker},
+  title  = {{FEMM} 4.2},
+  year   = {2020},
+  url    = {http://www.femm.info/wiki/HomePage},
+}
+
+@InCollection{NEURIPS2019_9015,
+  author    = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
+  booktitle = {Advances in Neural Information Processing Systems 32},
+  publisher = {Curran Associates, Inc.},
+  title     = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+  year      = {2019},
+  editor    = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
+  pages     = {8024--8035},
+  url       = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf},
+}
+
+@Misc{stable-baselines3,
+  author       = {Raffin, Antonin and Hill, Ashley and Ernestus, Maximilian and Gleave, Adam and Kanervisto, Anssi and Dormann, Noah},
+  howpublished = {\url{https://github.com/DLR-RM/stable-baselines3}},
+  title        = {Stable Baselines3},
+  year         = {2019},
+  journal      = {GitHub repository},
+  publisher    = {GitHub},
+}
+
+@Misc{TFAgents,
[…]
