### 2020-08-21 7c68a56
revert changes that are unrelated to the JOSS paper
