### 2020-08-21 806f8c3
revert changes that are unrelated to the JOSS paper
