+++ examples/dqn_series_current_control.py
@@ -55,5 +55,6 @@ if __name__ == '__main__':
     )
 
     dqn.compile(Adam(lr=1e-4), metrics=['mse'])
-    dqn.fit(env, nb_steps=200000, action_repetition=1, verbose=2, visualize=True, nb_max_episode_steps=50000, log_interval=10000)
+    dqn.fit(env, nb_steps=200000, action_repetition=1, verbose=2, visualize=True, nb_max_episode_steps=50000,
+            log_interval=10000)
     dqn.test(env, nb_episodes=3, nb_max_episode_steps=50000, visualize=True)
