### 2020-08-23 401a633
example for training and testing a dqn with tensorforce
