+++ examples/tensorforce_example.py
@@ -1,0 +1,229 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from tqdm import tqdm
+
+import gym_electric_motor as gem
+from gym_electric_motor.constraint_monitor import ConstraintMonitor
+from gym_electric_motor.reference_generators import \
+    MultipleReferenceGenerator,\
+    WienerProcessReferenceGenerator
+from gym_electric_motor.visualization import MotorDashboard
+from gym_electric_motor.physical_systems import ConstantSpeedLoad
+
+from gym.spaces import Discrete, Box
+from gym.wrappers import FlattenObservation, TimeLimit
+from gym import ObservationWrapper
+
+import tensorforce
+import tensorflow
+from tensorforce.environments import Environment
+from tensorforce.agents import Agent
+from tensorforce.execution import Runner
+
+
+class SqdCurrentMonitor:
+    """
+    monitor for squared currents:
+
+    i_sd**2 + i_sq**2 < 1.5 * nominal_limit
+    """
+
+    def __call__(self, state, observed_states, k, physical_system):
+        self.I_SD_IDX = physical_system.state_names.index('i_sd')
+        self.I_SQ_IDX = physical_system.state_names.index('i_sq')
+        # normalize to limit_values, since state is normalized
+        nominal_values = physical_system.nominal_state / abs(
+            physical_system.limits)
+        limits = 1.5 * nominal_values
+        # calculating squared currents as observed measure
+        sqd_currents = state[self.I_SD_IDX] ** 2 + state[self.I_SQ_IDX] ** 2
+
+        return (sqd_currents > limits[self.I_SD_IDX] or sqd_currents > limits[
+            self.I_SQ_IDX])
+
+
+class EpsilonWrapper(ObservationWrapper):
+    """
+    Changes Epsilon in a flattened observation to cos(epsilon)
+    and sin(epsilon)
+    """
+
+    def __init__(self, env, epsilon_idx):
+        super(EpsilonWrapper, self).__init__(env)
+        self.EPSILON_IDX = epsilon_idx
+        new_low = np.concatenate((self.env.observation_space.low[
+                                  :self.EPSILON_IDX], np.array([-1.]),
+                                  self.env.observation_space.low[
+                                  self.EPSILON_IDX:]))
+        new_high = np.concatenate((self.env.observation_space.high[
+                                   :self.EPSILON_IDX], np.array([1.]),
+                                   self.env.observation_space.high[
+                                   self.EPSILON_IDX:]))
+
+        self.observation_space = Box(new_low, new_high)
+
+    def observation(self, observation):
+        cos_eps = np.cos(observation[self.EPSILON_IDX] * np.pi)
+        sin_eps = np.sin(observation[self.EPSILON_IDX] * np.pi)
+        observation = np.concatenate((observation[:self.EPSILON_IDX],
+                                      np.array([cos_eps, sin_eps]),
+                                      observation[self.EPSILON_IDX + 1:]))
+        return observation
+
+
+# hyperparameter
+"""
+parameter for agent_config:
+
+    memory: size of replay-buffer
+    batch_size: size of mini-batch used for training
+    network: net-architect for dqn
+    update_frequency: Frequency of updates
+    start_updating: memory warm-up steps
+    learning_rate for optimizer
+    discount: gamma/ discount of future rewards
+    target_sync_frequency: Target network gets updated 'sync_freq' steps
+    target_update_weight: weight for target-network update
+
+"""
+
+
+motor_parameter = dict(p=3,  # [p] = 1, nb of pole pairs
+                       r_s=17.932e-3,  # [r_s] = Ohm, stator resistance
[…]
