+++ _sources/JOSS/paper.md.txt
@@ -41,55 +41,94 @@ bibliography: Literature.bib
 The ``gym-electric-motor`` (``GEM``) library provides simulation environments for 
 electrical drive systems and, therefore, allows to easily design and analyze drive control
 solutions in Python. Since ``GEM`` is strongly inspired by OpenAI's ``gym`` [@gym-whitepaper], it 
-is particulary well-equipped for (but not limited to) applications in the field of 
-reinforcement-learning-based control algorithms. In addition, the interface allows to plugin-in any expert-driven 
+is particularly well-equipped for (but not limited to) applications in the field of 
+reinforcement-learning-based control algorithms. In addition, the interface allows to plug in any expert-driven 
 control approach, such as model predictive control, to be tested  and to perform benchmark comparisons. 
 The ``GEM`` package includes a wide variety of motors, power electronic converters and mechanical load models that 
-can be flexibly selected and parameterized via the API. A modular software code also allows additional system 
-components to be included in the framework simulation.
+can be flexibly selected and parameterized via the API. A modular structure allows additional system 
+components to be included in the simulation framework.
 
 # Field of Application
 
-Electric drive systems are an important topic both in academic and 
-industrial research due to the worldwide availability and deployment of such 
-plants. Control algorithms for these systems have usually been designed, parameterized and 
-tested within [``MATLAB - Simulink``](https://www.mathworks.com/products/matlab.html), 
+Electric drive systems and their control are an important topic in both academic and 
+industrial research due to their worldwide usage and deployment. 
+Control algorithms for these systems have usually been designed, parameterized and 
+tested within ``MATLAB - Simulink`` [@MathWorks], 
 which is developed and promoted specifically for
 such engineering tasks. In the more recent past, however, commercial software like
-``MATLAB`` has difficulties to stay on par with the expandability and flexibility offered 
+``MATLAB`` has difficulties to stay on par with state-of-the-art concepts 
+for scientific modeling and the flexibility offered 
 by open-source libraries that are available for more accessible programming languages like Python. 
-Moreover, the latest efforts concerning industrial application of reinforcement-learning control 
-algorithms heavily depend on Python packages like ``Keras`` [@Chollet2015] or ``Tensorflow`` [@tensorflow2015-whitepaper]. 
-In order to allow easy access to an open-source drive simulation environment, the ``GEM`` library has been developed.
+Consequently, a Python-based drive simulation framework like ``GEM`` is an evident step in order to accelerate corresponding control research and development.
+Specifically, the latest efforts concerning industrial application of reinforcement-learning control 
+algorithms heavily depend on Python packages like 
+``Keras`` [@Chollet2015], ``Tensorflow`` [@tensorflow2015-whitepaper] or ``PyTorch``[@NEURIPS2019_9015]. 
+Hence, the built-in OpenAI ``gym`` interface allows to easily couple ``GEM`` to other open-source reinforcement 
+learning toolboxes such as ``Stable Baselines3`` [@stable-baselines3], ``TF-Agents`` [@TFAgents] or ``keras-rl`` [@plappert2016kerasrl].
+
+Providing easy access to the non-commercial, open-source ``GEM`` library allows users from any engineering domain to 
+include accurate drive models into their simulations, also beyond the topic of control applications.
+Considering the prevalence of commercial software like ``MATLAB`` for educational purposes, a free-of-charge simulation alternative
+that does not force students or institutions to pay for licenses, has great potential to support and encourage
+training of new talents in the field of electrical drives and neighbouring domains (e.g. power electronics or energy systems).
+
+# Related software
+
+Due to the strong dependence of downstream industrial development on simulated environments
+there is a comprehensive variety of
[…]
