### 2020-08-25 #2498 @wallscheid
As @wkirgsn announced we have revised the paper based on the reviewer remarks. Hence, we are looking forward to further comments on the paper.

In the meantime we are working on a major update of the master branch, which can already be previewed on the nightly branch.
