### 2020-08-25 3039911
changed commentary in example scripts
