+++ examples/ddpg_pmsm_dq_current_control.py
@@ -6,7 +6,7 @@ import tensorflow as tf
 from tensorflow.keras.optimizers import Adam
 from rl.agents import DDPGAgent
 from rl.memory import SequentialMemory
-from rl.random import OrnsteinUhlenbeckProcess, GaussianWhiteNoiseProcess
+from rl.random import OrnsteinUhlenbeckProcess
 from gym.wrappers import FlattenObservation
 import numpy as np
 import sys
@@ -20,83 +20,138 @@ from gym_electric_motor.physical_systems import ConstantSpeedLoad
 from gym.core import Wrapper
 from gym.spaces import Box, Tuple
 
+# This example shows how we can use GEM to train a reinforcement learning agent to control the current within
+# a permanent magnet synchronous motor three-phase drive
+# We use a deep-deterministic-policy-gradient (DDPG) agent
+# to determine which action must be taken on a continuous-control-set
+
 
 class AppendLastActionWrapper(Wrapper):
     """
     The following environment considers the dead time in the real-world motor control systems.
-    The real-world system changes its state, while the agent calculates the next action based on a previoulsly measured
-    observation. Therefore, for the agents it seems as if the applied action effects the state one step delayed.
-    (with a dead time of one time-step)
+    The real-world system changes its state, while the agent simultaneously calculates the next action based on a
+    previously measured observation.
+    Therefore, for the agents it seems as if the applied action affects the environment with one step delay
+    (with a dead time of one time step).
 
-    For complete observability of the system at each time-step we append the last played action of the agent to the
-    observation, because this action will be the one that is active in the next step.
+    For complete observability of the system at each time step we append the last selected action to the
+    observation, because this action will be the one that is active while the agent has to make the next decision.
     """
     def __init__(self, environment):
         super().__init__(environment)
+        # append the action space dimensions to the observation space dimensions
         self.observation_space = Tuple((Box(
             np.concatenate((environment.observation_space[0].low, environment.action_space.low)),
             np.concatenate((environment.observation_space[0].high, environment.action_space.high))
         ), environment.observation_space[1]))
 
     def step(self, action):
+
         (state, ref), rew, term, info = self.env.step(action)
+
+        # extend the output state by the selected action
         state = np.concatenate((state, action))
+
         return (state, ref), rew, term, info
 
     def reset(self, **kwargs):
+
         state, ref = self.env.reset()
+
+        # extend the output state by zeros after reset
+        # no action can be appended yet, but the dimension must fit
         state = np.concatenate((state, np.zeros(self.env.action_space.shape)))
+
         return state, ref
 
 
 if __name__ == '__main__':
+
     tf.compat.v1.disable_eager_execution()
 
     window_length = 1
 
-    # Changing i_q reference and constant 0 i_d reference.
-    q_generator = WienerProcessReferenceGenerator(reference_state='i_sq')
-    d_generator = ConstReferenceGenerator('i_sd', 0)
+    # Define reference generators for both currents of the flux oriented dq frame
+    d_generator = ConstReferenceGenerator('i_sd', 0)  # d current reference is chosen to be constantly at zero
+    q_generator = WienerProcessReferenceGenerator(reference_state='i_sq')  # q current changes dynamically
+    # The MultipleReferenceGenerator allows to apply these references simultaneously
     rg = MultipleReferenceGenerator([d_generator, q_generator])
 
-    # Change of the default motor parameters.
+    # Set the electric parameters of the motor
     motor_parameter = dict(
[…]

[…]
