### 2020-08-25 5ef2521
network definition fixed in examples
