+++ examples/tensorforce_for_gem/tensorforce_testing.py
@@ -1,0 +1,301 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from tqdm import tqdm
+
+import gym_electric_motor as gem
+from gym_electric_motor.constraint_monitor import ConstraintMonitor
+from gym_electric_motor.reference_generators import \
+    MultipleReferenceGenerator,\
+    WienerProcessReferenceGenerator
+from gym_electric_motor.visualization import MotorDashboard
+from gym_electric_motor.physical_systems import ConstantSpeedLoad
+
+from gym.spaces import Discrete, Box
+from gym.wrappers import FlattenObservation, TimeLimit
+from gym import ObservationWrapper
+
+from tensorforce.environments import Environment
+from tensorforce.agents import Agent
+
+
+def calc_mae(observation, reference):
+    """
+    observation: d x n
+    reference: d x n
+    """
+    mae = np.mean(np.abs(observation - reference), axis=-1)
+    return mae
+
+
+def set_plot_params(figure_title=22, figsize=(20, 14), title_size=24,
+                    label_size=20, tick_size=16, marker_size=10, line_width=3,
+                    legend_font=16, style='dark'):
+    """ setting matplotlib params """
+
+    if style == 'dark':
+        plt.style.use('dark_background')
+    else:
+        plt.style.use(style)
+
+    params = {'axes.titlesize': title_size,
+              'legend.fontsize': legend_font,
+              'figure.figsize': figsize,
+              'axes.labelsize': label_size,
+              'xtick.labelsize': label_size,
+              'ytick.labelsize': label_size,
+              'figure.titlesize': figure_title}
+    plt.rcParams.update(params)
+
+
+class SqdCurrentMonitor:
+    """
+    monitor for squared currents:
+
+    i_sd**2 + i_sq**2 < 1.5 * nominal_limit
+    """
+
+    def __call__(self, state, observed_states, k, physical_system):
+        self.I_SD_IDX = physical_system.state_names.index('i_sd')
+        self.I_SQ_IDX = physical_system.state_names.index('i_sq')
+        # normalize to limit_values, since state is normalized
+        nominal_values = physical_system.nominal_state / abs(
+            physical_system.limits)
+        limits = 1.5 * nominal_values
+        # calculating squared currents as observed measure
+        sqd_currents = state[self.I_SD_IDX] ** 2 + state[self.I_SQ_IDX] ** 2
+
+        return (sqd_currents > limits[self.I_SD_IDX] or sqd_currents > limits[
+            self.I_SQ_IDX])
+
+
+class EpsilonWrapper(ObservationWrapper):
+    """
+    Changes Epsilon in a flattened observation to cos(epsilon)
+    and sin(epsilon)
+    """
+
+    def __init__(self, env, epsilon_idx):
+        super(EpsilonWrapper, self).__init__(env)
+        self.EPSILON_IDX = epsilon_idx
+        new_low = np.concatenate((self.env.observation_space.low[
+                                  :self.EPSILON_IDX], np.array([-1.]),
+                                  self.env.observation_space.low[
+                                  self.EPSILON_IDX:]))
+        new_high = np.concatenate((self.env.observation_space.high[
+                                   :self.EPSILON_IDX], np.array([1.]),
+                                   self.env.observation_space.high[
+                                   self.EPSILON_IDX:]))
+
+        self.observation_space = Box(new_low, new_high)
+
+    def observation(self, observation):
+        cos_eps = np.cos(observation[self.EPSILON_IDX] * np.pi)
+        sin_eps = np.sin(observation[self.EPSILON_IDX] * np.pi)
[…]
