+++ examples/tensorforce_for_gem/tensorforce_training.py
@@ -23,7 +23,7 @@ import time
 print(tensorforce.__version__)
 print(tensorflow.__version__)
 
-path = input()
+#path = input()
 
 
 class SqdCurrentMonitor:
@@ -113,11 +113,11 @@ gem_env = gem.make(
                motor_parameter=motor_parameter, limit_values=limit_values,
                nominal_values=nominal_values, u_sup=350, tau=tau,
                # define the starting speed and states (randomly drawn)
-               load='ConstSpeedLoad',
-               load_initializer={'states': {'omega': 1000 * np.pi / 30,},
-                                 'interval': [[-4000*2*np.pi/60, 4000*2*np.pi/60]],
-                                 'random_init': 'uniform'},
-               motor_initializer = motor_init,
+               #load='ConstSpeedLoad',
+               #load_initializer={'states': {'omega': 1000 * np.pi / 30,},
+               #                 'interval': [[-4000*2*np.pi/60, 4000*2*np.pi/60]],
+               #                 'random_init': 'uniform'},
+               #motor_initializer = motor_init,
                # parameterize the reward function
                reward_function=gem.reward_functions.WeightedSumOfErrors(
                    observed_states=['i_sq', 'i_sd'],
@@ -165,7 +165,7 @@ epsilon_decay = {'type': 'decaying',
 net = [
     dict(type='dense', size=64, activation='relu'),
     dict(type='dense', size=64, activation='relu'),
-    dict(type='linear', size=7)
+    #dict(type='linear', size=7)
 ]
 
 agent_config = {
@@ -193,11 +193,57 @@ runner.close()
 print(f'Execution time of tensorforce dqn-training is: 'f'{time.time()-start_time:.2f} seconds')
 
 
-#path = '/home/pascal/Sciebo/Uni/Master/Semester_2/' \
-#        + 'Projektarbeit/python/Notebooks/tensorforce/saves' \
+path = '/home/pascal/Sciebo/Uni/Master/Semester_2/Projektarbeit/python/Notebooks/tensorforce/saves'
 
 dqn_agent.save(directory=path,
-               filename='dqn_tf_trained_relu',
+               filename='dqn_tf_trained_no_rand',
                format='tensorflow',
                append='timesteps')
 
+# test agent
+tau = 1e-5
+steps = 100000
+
+rewards = []
+lens = []
+obs_hist = []
+
+states = []
+references = []
+
+obs = gem_env_.reset()
+obs_hist.append(obs)
+terminal = False
+cum_rew = 0
+step_counter = 0
+eps_rew = 0
+
+for step in tqdm(range(steps)):
+    #while not terminal:
+    #gem_env_.render()
+    actions = dqn_agent.act(states=obs, independent=True)
+
+    obs, reward, terminal, _ = gem_env_.step(action=actions)
+    rewards.append(cum_rew)
+    obs_hist.append(obs)
+
+    # dqn_agent.observe(terminal, reward=reward)
+    cum_rew += reward
+    eps_rew += reward
+
+    if terminal:
+        lens.append(step_counter)
+        step_counter = 0
+        #print(f'Episode length: {episode_length} steps')
+        obs = gem_env_.reset()
+
+        obs_hist.append(obs)
+        rewards.append(eps_rew)
+
+        terminal = False
+        eps_rew = 0
+    step_counter += 1
+
+
+print(f' \n Cumulated Reward over {steps} steps is {cum_rew} \n')
+print(f' \n Longest Episode: {np.amax(lens)} steps \n')
