### 2020-08-26 7940e31
revisions for external_speed and initializer showcase, also fix some typos, TODO: constraint showcase
