+++ examples/userdefined_initialization.py
@@ -5,66 +5,85 @@ from gym_electric_motor.visualization import MotorDashboard
 from agents.simple_controllers import Controller
 import time
 
+# This code example presents how the initializer interface can be used to sample random initial states for the drive.
+# This is important when using e.g. reinforcement learning, because random initialization allows for a better
+# exploration of the state space (so called "exploring starts").
+# Initializers can be applied to electric motor state (which is also decisive for the initial torque) and
+# mechanical load (which sets initial drive speed).
+#
+# For a more general introduction to GEM, we recommend to have a look at the "_control.py" examples first.
 
-# initializations for different motor types
-# initializer for given constant initial-values
-# e.g. for series motor ('DcExtExCont-v1')
-dc_series = {'states': {'i': 12}}
-# e.g. for synchronous motor ('PMSMCont-v1')
-sync_init = {'states': {'i_sq': 5.6, 'i_sd': 6.0, 'epsilon': 3.0}}
 
-# random initial-values in nominal-value-space
+
+# Initializers use the state names that are present for the used motor:
+
+# initializer for a specific current; e.g. DC series motor ('DcSeriesCont-v1' / 'DcSeriesDisc-v1')
+dc_series_init = {'states': {'i': 12}}
+
+# initializer for a specific current and position; e.g. permanent magnet synchronous motor ('PMSMCont-v1' / 'PMSMDisc-v1')
+pmsm_init = {'states': {'i_sd': -36.0,
+                        'i_sq': 55.0,
+                        'epsilon': 3.0}}
+
+# initializer for a random initial current with gaussian distribution, parameterized with mu=25 and sigma=10
 gaussian_init = {'random_init': 'gaussian',
-                 'random_params': (None, None)}
-# random initial-values in limited-value-space (for motor with one state)
+                 'random_params': (25, 0.1),
+                 'states': {'i': 0}
+                 }
+
+# initializer for a ranom initial speed with uniform distribution within the interval omega=60 to omega=80
 uniform_init = {'random_init': 'uniform',
-                'interval': [[20, 40]]}
-# initial value for PolynomialLoad
-# (other loads can recieve an own intial value, like in external_speed_profile.py)
+                'interval': [[60, 80]],
+                'states': {'omega': 0}}
+
+# initializer for a specific speed
 load_init = {'states': {'omega': 20}}
 
-
 if __name__ == '__main__':
-
     env = gem.make(
-            # choose motor
             'DcSeriesCont-v1',
-            #'PMSMCont-v1',
-
-            motor_initializer=gaussian_init,
-            load_initializer=load_init,
-            #visualization=MotorDashboard(plots=['omega','i_sq', 'i_sd', 'reward'],
-            #                             dark_mode=True),
-            visualization=MotorDashboard(plots=['omega','i', 'reward'],
-                                         dark_mode=True),
-            motor_parameter=dict(r_a=15e-3, r_e=15e-3, l_a=1e-3, l_e=1e-3),
-            # Take standard class and pass parameters (Load)
-            load_parameter=dict(a=0.01, b=.1, c=0.1, j_load=.06),
-            # Pass a string (with extra parameters)
+            visualization=MotorDashboard(plots=['omega', 'i'],
+                                         dark_mode=False),
+            motor_parameter=dict(j_rotor=0.001),
+            load_parameter=dict(a=0, b=0.1, c=0, j_load=0.001),
             ode_solver='scipy.solve_ivp', solver_kwargs=dict(),
-            # Pass a Class with extra parameters
             reference_generator=rg.SwitchedReferenceGenerator(
                 sub_generators=[
-                    rg.SinusoidalReferenceGenerator, rg.WienerProcessReferenceGenerator(), rg.StepReferenceGenerator()
[…]
