+++ examples/external_speed_profile.py
@@ -8,59 +8,83 @@ from scipy import signal
 from gym_electric_motor.physical_systems.mechanical_loads \
     import ExternalSpeedLoad, ConstantSpeedLoad
 
-# defining different reference generators
-const_sub_gen = [rg.ConstReferenceGenerator(reference_value=0.20),
-                 rg.ConstReferenceGenerator(reference_value=0.25),
-                 rg.ConstReferenceGenerator(reference_value=0.30),
-                 rg.ConstReferenceGenerator(reference_value=0.35),
-                 rg.ConstReferenceGenerator(reference_value=0.40)]
+# This code example presents how the speed load classes can be used to
+# define a speed profile which the drive will then follow.
+# This is often useful when prototyping drive controllers that either
+# control purely electrical drive behavior (e.g. current controllers) or
+# drive torque with fixed speed requirements (e.g. traction applications, generator operation).
+#
+# For a more general introduction to GEM, we recommend to have a look at the "_control.py" examples first.
+
+
+# We will have a look at a current control scenario here
+const_sub_gen = [rg.ConstReferenceGenerator(reference_state='i', reference_value=0.1),  # operation at 10 % of the current limit
+                 rg.ConstReferenceGenerator(reference_state='i', reference_value=0.9),  # operation at 90 % of the current limit
+                 rg.ConstReferenceGenerator(reference_state='i', reference_value=1.1),  # operation at 110 % of the current limit,
+                 ]
+
+# since the case of 110 % current may lead to limit violation, we only assign a small probability of 2 % to it
 const_switch_gen = rg.SwitchedReferenceGenerator(const_sub_gen,
+                                                 p=[0.49, 0.49, 0.02],
                                                  super_episode_length=(1000, 2000))
 
-# possible speed profiles
-sinus_lambda = (lambda t,f,amp, b: amp * np.sin(2*np.pi*f*t) + b)
-constant_lambda = (lambda value: value)
-triangle_lambda = (lambda t, amp, f, b: amp * signal.sawtooth(2 * np.pi * f * t,
-                                                             width=0.5) + b)
-saw_lambda = (lambda t, amp, f, b: amp * signal.sawtooth(2 * np.pi * f * t) + b)
+# The ExternalSpeedLoad class allows to pass an arbitrary function of time which will then dictate the speed profile.
+# As shown here it can also contain more parameters. Some examples:
+# Parameterizable sine oscillation
+sinus_lambda = (lambda t, frequency, amplitude, bias: amplitude * np.sin(2 * np.pi * frequency * t) + bias)
+# Constant speed
+constant_lambda = (lambda t, value: value)
+# Parameterizable triangle oscillation
+triangle_lambda = (lambda t, amplitude, frequency, bias: amplitude * signal.sawtooth(2 * np.pi * frequency * t,
+                                                                                     width=0.5) + bias)
+# Parameterizable sawtooth oscillation
+saw_lambda = (lambda t, amplitude, frequency, bias: amplitude * signal.sawtooth(2 * np.pi * frequency * t,
+                                                                                width=0.9) + bias)
+
 
 # usage of a random load initializer is only recommended for the
 # ConstantSpeedLoad, due to the already given profile by an ExternalSpeedLoad
 load_init = {'random_init': 'uniform'},
 
-# defining a ConstandSpeedLoad
-const_load = ConstantSpeedLoad(omega_fixed=42)
 # External speed profiles can be given by an ExternalSpeedLoad,
 # inital value is given by bias of the profile
-ext_load = ExternalSpeedLoad(speed_profile=sinus_lambda, amp=10, f=2, b=42)
+sampling_time = 1e-4
[…]
