### 2020-08-26 b9d6ce5
cleaned up testing and training script
