+++ examples/tensorforce_for_gem/tensorforce_training.py
@@ -19,11 +19,22 @@ from tensorforce.environments import Environment
 from tensorforce.agents import Agent
 from tensorforce.execution import Runner
 import time
+import json
 
-print(tensorforce.__version__)
-print(tensorflow.__version__)
 
-#path = input()
+def save_to_json(data, name, indent=4, sort=False, ending='.json'):
+    """"""
+
+    with open(name + ending, 'w') as json_file:
+        json.dump(data, json_file, indent=indent, sort_keys=sort)
+    print('Done writing')
+
+
+def load_json(filepath):
+    """"""
+    data = json.loads(filepath)
+    print('Loaded')
+    return data
 
 
 class SqdCurrentMonitor:
@@ -76,60 +87,74 @@ class EpsilonWrapper(ObservationWrapper):
         return observation
 
 
-sqd_current_monitor = ConstraintMonitor(external_monitor=SqdCurrentMonitor)
+config_path = '/home/pascal/Sciebo/Uni/Master/Semester_2/Projektarbeit/' + \
+              'python/saves/configs/'
+agent_path = '/home/pascal/Sciebo/Uni/Master/Semester_2/Projektarbeit/' + \
+             'python/saves/agents/'
 
+env_name = 'env_config_zero_init'
+agent_config_name = 'default_agent'
+
+
+# define motor arguments
+sqd_current_monitor = ConstraintMonitor(external_monitor=SqdCurrentMonitor)
 motor_parameter = dict(p=3,  # [p] = 1, nb of pole pairs
                        r_s=17.932e-3,  # [r_s] = Ohm, stator resistance
                        l_d=0.37e-3,  # [l_d] = H, d-axis inductance
                        l_q=1.2e-3,  # [l_q] = H, q-axis inductance
                        psi_p=65.65e-3,  # [psi_p] = Vs, magnetic flux of the permanent magnet
                        )
-nominal_values=dict(omega=4000*2*np.pi/60, i=230, u=350)
+u_sup = 350
+nominal_values=dict(omega=4000*2*np.pi/60,
+                    i=230,
+                    u=u_sup
+                    )
 limit_values=nominal_values.copy()
-
-q_generator = WienerProcessReferenceGenerator(reference_state='i_sq')
-d_generator = WienerProcessReferenceGenerator(reference_state='i_sd')
-rg = MultipleReferenceGenerator([q_generator, d_generator])
-
-tau = 1e-5
-gamma = 0.99
-simulation_time = 5 # seconds
-max_eps_steps = 10000
-simulation_steps = 500000
-episodes = simulation_steps / max_eps_steps
 
 motor_init = {'random_init': 'uniform'}
 load_initializer = {'interval': [[-4000*2*np.pi/60, 4000*2*np.pi/60]],
                     'random_init': 'uniform'}
+
 const_random_load = ConstantSpeedLoad(load_initializer=load_initializer)
+q_generator = WienerProcessReferenceGenerator(reference_state='i_sq')
+d_generator = WienerProcessReferenceGenerator(reference_state='i_sd')
+rg = MultipleReferenceGenerator([q_generator, d_generator])
+tau = 1e-5
+gamma = 0.99
 
-# creating gem enviroment
+max_eps_steps = 10000
+simulation_steps = 500000
+episodes = simulation_steps / max_eps_steps
+
+environment_config = dict(
+    motor_parameter=motor_parameter,
+    limit_values=limit_values,
+    nominal_values=nominal_values,
+    u_sup=u_sup,
+    tau=tau,
+    #load='ConstSpeedLoad',
+    #load_initializer={'states': {'omega': 1000 * np.pi / 30,},
[…]
