+++ examples/tensorforce_for_gem/tensorforce_testing.py
@@ -16,27 +16,42 @@ from gym import ObservationWrapper
 
 from tensorforce.environments import Environment
 from tensorforce.agents import Agent
+import json
 
-#path = input()
 
-# def set_plot_params(figure_title=22, figsize=(20, 14), title_size=24,
-#                     label_size=20, tick_size=16, marker_size=10, line_width=3,
-#                     legend_font=16, style='dark'):
-#     """ setting matplotlib params """
-#
-#     if style == 'dark':
-#         plt.style.use('dark_background')
-#     else:
-#         plt.style.use(style)
-#
-#     params = {'axes.titlesize': title_size,
-#               'legend.fontsize': legend_font,
-#               'figure.figsize': figsize,
-#               'axes.labelsize': label_size,
-#               'xtick.labelsize': label_size,
-#               'ytick.labelsize': label_size,
-#               'figure.titlesize': figure_title}
-#     plt.rcParams.update(params)
+def save_to_json(data, name, indent=4, sort=False, ending='.json'):
+    """"""
+
+    with open(name + ending, 'w') as json_file:
+        json.dump(data, json_file, indent=indent, sort_keys=sort)
+    print('Done writing')
+
+
+def load_json(filepath):
+    """"""
+    data = json.loads(filepath)
+    print('Loaded')
+    return data
+
+
+def set_plot_params(figure_title=22, figsize=(20, 14), title_size=24,
+                    label_size=20, tick_size=16, marker_size=10, line_width=3,
+                    legend_font=16, style='dark'):
+    """ setting matplotlib params """
+
+    if style == 'dark':
+        plt.style.use('dark_background')
+    else:
+        plt.style.use(style)
+
+    params = {'axes.titlesize': title_size,
+              'legend.fontsize': legend_font,
+              'figure.figsize': figsize,
+              'axes.labelsize': label_size,
+              'xtick.labelsize': label_size,
+              'ytick.labelsize': label_size,
+              'figure.titlesize': figure_title}
+    plt.rcParams.update(params)
 
 
 class SqdCurrentMonitor:
@@ -89,139 +104,68 @@ class EpsilonWrapper(ObservationWrapper):
         return observation
 
 
+config_path = '/home/pascal/Sciebo/Uni/Master/Semester_2/Projektarbeit/' + \
+              'python/saves/configs'
+agent_path = '/home/pascal/Sciebo/Uni/Master/Semester_2/Projektarbeit/' + \
+             'python/saves/agents'
+
+env_name = 'env_config_zero_init'
+agent_config_name = 'default_agent'
+
+# define motor arguments
+load_initializer = {'interval': [[-4000*2*np.pi/60, 4000*2*np.pi/60]],
+                    'random_init': 'uniform'}
+
 sqd_current_monitor = ConstraintMonitor(external_monitor=SqdCurrentMonitor)
-
-motor_parameter = dict(p=3,  # [p] = 1, nb of pole pairs
-                       r_s=17.932e-3,  # [r_s] = Ohm, stator resistance
-                       l_d=0.37e-3,  # [l_d] = H, d-axis inductance
-                       l_q=1.2e-3,  # [l_q] = H, q-axis inductance
-                       psi_p=65.65e-3,  # [psi_p] = Vs, magnetic flux of the permanent magnet
-                       )
-u_sup = 350
-nominal_values=dict(omega=4000*2*np.pi/60,
-                    i=230,
-                    u=u_sup
-                    )
-limit_values=nominal_values.copy()
-
+const_random_load = ConstantSpeedLoad(load_initializer=load_initializer)
 q_generator = WienerProcessReferenceGenerator(reference_state='i_sq')
[…]
