### 2020-08-27 7b469ca
revision of the constraint monitor showcase and minor changes to the other examples
