+++ examples/userdefined_constraints.py
@@ -10,22 +10,16 @@ import numpy as np
 import gym
 from gym.spaces import Box
 from gym_electric_motor.constraint_monitor import ConstraintMonitor
-
-rg.SwitchedReferenceGenerator(
-            sub_generators=[
-                rg.SinusoidalReferenceGenerator, rg.WienerProcessReferenceGenerator(), rg.StepReferenceGenerator()
-            ], p=[0.1, 0.8, 0.1], super_episode_length=(1000, 10000)
-        )
-
-const_sub_gen = [rg.ConstReferenceGenerator(reference_value=0.20),
-                 rg.ConstReferenceGenerator(reference_value=0.25),
-                 rg.ConstReferenceGenerator(reference_value=0.30),
-                 rg.ConstReferenceGenerator(reference_value=0.35),
-                 rg.ConstReferenceGenerator(reference_value=0.40)]
+from gym_electric_motor.physical_systems.mechanical_loads import ConstantSpeedLoad
 
 
-const_switch_gen = rg.SwitchedReferenceGenerator(const_sub_gen,
-                                                 super_episode_length=(8000, 12000))
+# This code example presents how the constraint monitor can be used to define custom termination conditions.
+# This feature is interesting for e.g. three-phase drives, where the current needs to be monitored with
+# respect to the total current within all three phases , and not separately per phase.
+# In this example we will define an external constraint for the DcSeriesMotor which lies at 95 % of
+# the predefined current limit.
+#
+# For a more general introduction to GEM, we recommend to have a look at the "_control.py" examples first.
 
 
 # the following external monitors are examples for the general use-case.
@@ -37,18 +31,33 @@ class ExternalMonitorClass:
     constraint-conditions
     """
     def __init__(self):
-        self._constraints = Box(low=np.array([-70, -440, -70, -300, -420]),
-                                high=np.array([70, 440, 70, 300, 420]))
+        # Do some setup
+        self.upper_current_constraint = 95
+        self.lower_current_constraint = -95
 
     def __call__(self, state, observed_states, **kwargs):
 
-        return self.check_violation(state, observed_states)
+        # Read the current limit and the time since the last reset
+        physical_system = kwargs['physical_system']
+        current_limit = physical_system._limits[2]
+        time = kwargs['k'] * physical_system.tau
 
-    def check_violation(self, state, observed_states):
-        lower = self._constraints.low
-        upper = self._constraints.high
-        return float(
-            (abs(state) < lower).any() or (abs(state) > upper).any())
+        # the result of this check sets the "done" flag of the env.step() output
+        # done==True usually necessitates an env.reset()
+        return self.check_violation(denormalized_current=state[2] * current_limit, t=time)
+
+    def check_violation(self, denormalized_current, t):
+
+        # check whether the constraints are violated
+        check_violation = int(abs(denormalized_current) < self.lower_current_constraint or
+                              abs(denormalized_current) > self.upper_current_constraint)
+
+        if check_violation:
+            print(f"External current constraint was violated at t={t} s")
+            print(f"Measured current: {denormalized_current} A, constraint was at {self.upper_current_constraint} A")
+            print()
+
+        return check_violation
 
 
 def external_monitor_func(state, physical_system, k, **kwargs):
@@ -57,51 +66,64 @@ def external_monitor_func(state, physical_system, k, **kwargs):
     violation of the constraint conditions. Additional function parameters
     can be given to the ConstraintMonitor separately
     """
-    factor = 0.9
-    limits = physical_system.limits * factor
-    check = float((abs(state) > abs(limits)).any())
-    return check
+    # get the defined current limit
+    current_limit = physical_system._limits[2]
 
+    # denormalize the current value
[…]
