+++ examples/external_speed_profile.py
@@ -92,6 +92,8 @@ if __name__ == '__main__':
 
             if done:
                 state, _ = env.reset()
+                # Redefine the controller in order to reset it
+                controller = Controller.make('pi_controller', env)
             cum_rew += reward
 
         print(cum_rew)
