### 2020-08-27 62e34b7
use controller.reset instead of redefinition
