### 2020-08-28 1dee05e
added the plotting example to the JOSS paper
