+++ docs/JOSS/Literature.bib
@@ -60,7 +60,7 @@
 
 @Misc{ANSYS,
   author = {{ANSYS, Inc.}},
-  title  = {{ANSYS Maxwell} 2020 R1},
+  title  = {{ANSYS Maxwell 2020 R1}},
   year   = {2020},
   url    = {https://www.ansys.com/products/electronics/ansys-maxwell},
 }
@@ -107,39 +107,35 @@
 }
 
 @Misc{TFAgents,
-  author       = {Sergio Guadarrama and Anoop Korattikara and Oscar Ramirez and Pablo Castro and Ethan Holly and Sam Fishman and Ke Wang and Ekaterina Gonina and Neal Wu and Efi Kokiopoulou and Luciano Sbaiz and Jamie Smith and Gábor Bartók and Jesse Berent and Chris Harris and Vincent Vanhoucke and Eugene Brevdo},
-  howpublished = {\url{https://github.com/tensorflow/agents}},
-  note         = {[Online; accessed 25-June-2019]},
-  title        = {{TF-Agents}: A library for Reinforcement Learning in TensorFlow},
-  year         = {2018},
-  journal      = {GitHub repository},
-  url          = {https://github.com/tensorflow/agents},
+  author  = {Sergio Guadarrama and Anoop Korattikara and Oscar Ramirez and Pablo Castro and Ethan Holly and Sam Fishman and Ke Wang and Ekaterina Gonina and Neal Wu and Efi Kokiopoulou and Luciano Sbaiz and Jamie Smith and Gábor Bartók and Jesse Berent and Chris Harris and Vincent Vanhoucke and Eugene Brevdo},
+  note    = {[Online; accessed 25-June-2019]},
+  title   = {{TF-Agents}: A library for Reinforcement Learning in TensorFlow},
+  year    = {2018},
+  journal = {GitHub repository},
+  url     = {https://github.com/tensorflow/agents},
 }
 
 @Misc{plappert2016kerasrl,
-  author       = {Matthias Plappert},
-  howpublished = {\url{https://github.com/keras-rl/keras-rl}},
-  title        = {keras-rl},
-  year         = {2016},
-  journal      = {GitHub repository},
-  publisher    = {GitHub},
-  url          = {https://github.com/keras-rl/keras-rl},
+  author    = {Matthias Plappert},
+  title     = {keras-rl},
+  year      = {2016},
+  journal   = {GitHub repository},
+  publisher = {GitHub},
+  url       = {https://github.com/keras-rl/keras-rl},
 }
 
 @Misc{OSMC2020,
-  author       = {{Open Source Modelica Consortium}},
-  howpublished = {\url{https://openmodelica.org/}},
-  title        = {OpenModelica},
-  year         = {2020},
-  url          = {https://openmodelica.org/},
+  author = {{Open Source Modelica Consortium}},
+  title  = {OpenModelica},
+  year   = {2020},
+  url    = {https://openmodelica.org/},
 }
 
 @Misc{Scilab2020,
-  author       = {{Scilab}},
-  howpublished = {\url{https://www.scilab.org/software/xcos}},
-  title        = {Xcos (Scilab 6.1.0)},
-  year         = {2020},
-  url          = {https://www.scilab.org/software/xcos},
+  author = {{Scilab}},
+  title  = {{Xcos (Scilab 6.1.0)}},
+  year   = {2020},
+  url    = {https://www.scilab.org/software/xcos},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
