### 2020-08-28 8c849ad
correct some JabRef
