+++ docs/JOSS/Literature.bib
@@ -1,16 +1,16 @@
 % Encoding: UTF-8
 
 @Misc{Chollet2015,
-  author       = {Chollet, Fran\c{c}ois and others},
-  howpublished = {\url{https://keras.io}},
-  title        = {Keras},
-  year         = {2015},
+  author = {Chollet, Fran\c{c}ois and others},
+  title  = {Keras},
+  year   = {2015},
+  url    = {https://keras.io},
 }
 
 @Misc{tensorflow2015-whitepaper,
   author = {Mart\'{\i}n~Abadi and Ashish~Agarwal and Paul~Barham and Eugene~Brevdo and Zhifeng~Chen and Craig~Citro and Greg~S.~Corrado and Andy~Davis and Jeffrey~Dean and Matthieu~Devin and Sanjay~Ghemawat and Ian~Goodfellow and Andrew~Harp and Geoffrey~Irving and Michael~Isard and Yangqing Jia and Rafal~Jozefowicz and Lukasz~Kaiser and Manjunath~Kudlur and Josh~Levenberg and Dandelion~Man\'{e} and Rajat~Monga and Sherry~Moore and Derek~Murray and Chris~Olah and Mike~Schuster and Jonathon~Shlens and Benoit~Steiner and Ilya~Sutskever and Kunal~Talwar and Paul~Tucker and Vincent~Vanhoucke and Vijay~Vasudevan and Fernanda~Vi\'{e}gas and Oriol~Vinyals and Pete~Warden and Martin~Wattenberg and Martin~Wicke and Yuan~Yu and Xiaoqiang~Zheng},
   note   = {Software available from tensorflow.org},
-  title  = {{TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+  title  = {{TensorFlow: Large-Scale Machine Learning on Heterogeneous Systems}},
   year   = {2015},
   url    = {https://www.tensorflow.org/},
 }
@@ -24,14 +24,15 @@
 
 @Misc{gym-whitepaper,
   author = {Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
-  title  = {OpenAI Gym},
+  title  = {{OpenAI Gym}},
   year   = {2016},
   eprint = {arXiv:1606.01540},
+  url    = {https://github.com/openai/gym},
 }
 
 @Misc{MathWorks,
   author = {{The MathWorks, Inc.}},
-  title  = {MATLAB - Simulink R2020a},
+  title  = {{MATLAB - Simulink R2020a}},
   year   = {2020},
   url    = {https://www.mathworks.com/products/matlab.html},
 }
@@ -51,11 +52,11 @@
 }
 
 @Misc{Margolis,
-  author  = {Benjamin W. L. Margolis},
-  title   = {SimuPy 1.0.0},
-  year    = {2017},
-  journal = {GitHub repository},
-  url     = {https://github.com/simupy/simupy},
+  author    = {Benjamin W. L. Margolis},
+  title     = {SimuPy 1.0.0},
+  year      = {2017},
+  publisher = {GitHub},
+  url       = {https://github.com/simupy/simupy},
 }
 
 @Misc{ANSYS,
@@ -90,7 +91,7 @@
   author    = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
[…]
