### 2020-08-30 8680010
new folder structure modified train, test scripts
