+++ examples/rl_frameworks/tensorforce_training.py
@@ -1,0 +1,240 @@
+# import numpy as np
+# from tqdm import tqdm
+#
+# import gym_electric_motor as gem
+# from gym_electric_motor.constraint_monitor import ConstraintMonitor
+# from gym_electric_motor.reference_generators import \
+#     MultipleReferenceGenerator,\
+#     WienerProcessReferenceGenerator
+# from gym_electric_motor.visualization import MotorDashboard
+# from gym_electric_motor.physical_systems import ConstantSpeedLoad
+#
+# from gym.spaces import Discrete, Box
+# from gym.wrappers import FlattenObservation, TimeLimit
+# from gym import ObservationWrapper
+#
+# import tensorforce
+# import tensorflow
+from tensorforce.environments import Environment
+from tensorforce.agents import Agent
+from tensorforce.execution import Runner
+import time
+import json
+from setting_environment import set_env
+
+
+# def save_to_json(data, name, indent=4, sort=False, ending='.json'):
+#     """"""
+#
+#     with open(name + ending, 'w') as json_file:
+#         json.dump(data, json_file, indent=indent, sort_keys=sort)
+#     print('Done writing')
+#
+#
+# def load_json(filepath):
+#     """"""
+#     data = json.loads(filepath)
+#     print('Loaded')
+#     return data
+
+
+# class SqdCurrentMonitor:
+#     """
+#     monitor for squared currents:
+#
+#     i_sd**2 + i_sq**2 < 1.5 * nominal_limit
+#     """
+#
+#     def __call__(self, state, observed_states, k, physical_system):
+#         self.I_SD_IDX = physical_system.state_names.index('i_sd')
+#         self.I_SQ_IDX = physical_system.state_names.index('i_sq')
+#         # normalize to limit_values, since state is normalized
+#         nominal_values = physical_system.nominal_state / abs(
+#             physical_system.limits)
+#         limits = 1.5 * nominal_values
+#         # calculating squared currents as observed measure
+#         sqd_currents = state[self.I_SD_IDX] ** 2 + state[self.I_SQ_IDX] ** 2
+#
+#         return (sqd_currents > limits[self.I_SD_IDX] or sqd_currents > limits[
+#             self.I_SQ_IDX])
+#
+#
+# class EpsilonWrapper(ObservationWrapper):
+#     """
+#     Changes Epsilon in a flattened observation to cos(epsilon)
+#     and sin(epsilon)
+#     """
+#
+#     def __init__(self, env, epsilon_idx):
+#         super(EpsilonWrapper, self).__init__(env)
+#         self.EPSILON_IDX = epsilon_idx
+#         new_low = np.concatenate((self.env.observation_space.low[
+#                                   :self.EPSILON_IDX], np.array([-1.]),
+#                                   self.env.observation_space.low[
+#                                   self.EPSILON_IDX:]))
+#         new_high = np.concatenate((self.env.observation_space.high[
+#                                    :self.EPSILON_IDX], np.array([1.]),
+#                                    self.env.observation_space.high[
+#                                    self.EPSILON_IDX:]))
+#
+#         self.observation_space = Box(new_low, new_high)
+#
+#     def observation(self, observation):
+#         cos_eps = np.cos(observation[self.EPSILON_IDX] * np.pi)
+#         sin_eps = np.sin(observation[self.EPSILON_IDX] * np.pi)
+#         observation = np.concatenate((observation[:self.EPSILON_IDX],
+#                                       np.array([cos_eps, sin_eps]),
[…]
