### 2020-08-30 9a95b88
example files cleaned up
