+++ examples/rl_frameworks/tensorforce_testing.py
@@ -1,26 +1,12 @@
 import numpy as np
-import matplotlib.pyplot as plt
 from tqdm import tqdm
-
-import gym_electric_motor as gem
-from gym_electric_motor.constraint_monitor import ConstraintMonitor
-from gym_electric_motor.reference_generators import \
-    MultipleReferenceGenerator,\
-    WienerProcessReferenceGenerator
-from gym_electric_motor.visualization import MotorDashboard
-from gym_electric_motor.physical_systems import ConstantSpeedLoad
-
-from gym.spaces import Discrete, Box
-from gym.wrappers import FlattenObservation, TimeLimit
-from gym import ObservationWrapper
-
 from tensorforce.environments import Environment
 from tensorforce.agents import Agent
-import json
+import tensorflow as tf
 from setting_environment import set_env
 
 max_eps_steps = 10000
-gem_env = set_env()
+gem_env = set_env(time_limit=False)
 # creating tensorforce environment
 tensor_env = Environment.create(environment=gem_env,
                                 max_episode_timesteps=max_eps_steps)
@@ -49,15 +35,15 @@ epsilon_decay = {'type': 'decaying',
                  'power': 3.0}
 net = [
     dict(type='dense', size=64, activation='relu'),
-    dict(type='dense', size=64, activation='relu'),
-    dict(type='linear', size=7)
+    dict(type='dense', size=64, activation='none'),
+    #dict(type='linear', size=7)
 ]
 
 agent_config = {
     'agent': 'dqn',
     'memory': 200000,
     'batch_size': 25,
-    'network': net, #dict(type='auto', size=64, depth=2),
+    'network': net,
     'update_frequency': 1,
     'start_updating': 10000,
     'learning_rate': 1e-4,
@@ -72,19 +58,17 @@ agent_name = 'dqn_64_64'
 dqn_agent = Agent.load(
     directory=agent_path,
     filename=agent_name,
-    format='pb-actonly',
-    #environment=tensor_env,
-    #**agent_config
+    environment=tensor_env,
+    **agent_config
 )
+print('\n agent loaded \n')
 
 # test agent
 tau = 1e-5
-steps = 5000#1000000
+steps = 1000000
 
 rewards = []
-# lens = []
 # obs_hist = []
-
 states = []
 references = []
 
@@ -96,19 +80,15 @@ step_counter = 0
 eps_rew = 0
 
 for step in tqdm(range(steps)):
-    #while not terminal:
     gem_env.render()
-    print(type(obs))
     actions = dqn_agent.act(obs, evaluation=True)
     obs, reward, terminal, _ = gem_env.step(action=actions)
     rewards.append(cum_rew)
-    #obs_hist.append(obs)
-    # dqn_agent.observe(terminal, reward=reward)
+    # obs_hist.append(obs)
     cum_rew += reward
     eps_rew += reward
 
     if terminal:
-        #lens.append(step_counter)
         step_counter = 0
         #print(f'Episode length: {episode_length} steps')
         obs = gem_env.reset()
@@ -119,54 +99,3 @@ for step in tqdm(range(steps)):
     step_counter += 1
 
 print(f' \n Cumulated Reward per step is {cum_rew/steps} \n')
-#print(f' \n Longest Episode: {np.amax(lens)} steps \n')
-
-
-
-# test agent
-# tau = 1e-5
-# steps = 100000
-#
-# rewards = []
-# lens = []
-# obs_hist = []
-#
-# states = []
-# references = []
-#
-# obs = gem_env_.reset()
-# obs_hist.append(obs)
-# terminal = False
-# cum_rew = 0
-# step_counter = 0
-# eps_rew = 0
-#
[…]
