### 2020-09-01 16ad747
correct file and folder structure for sprint3
