### 2020-09-01 539c0c0
Adopted new folder structure. Added trained agent
