### 2020-09-02 5bba174
small tidy up
