### 2020-09-21 cafce30
updated environment for more stable trajectory
