+++ examples/rl_frameworks/setting_environment.py
@@ -7,7 +7,7 @@ from gym_electric_motor.reference_generators import \
 from gym_electric_motor.visualization import MotorDashboard
 from gym.spaces import Discrete, Box
 from gym.wrappers import FlattenObservation, TimeLimit
-from gym import ObservationWrapper
+from gym import ObservationWrapper, Wrapper
 
 
 class SqdCurrentMonitor:
@@ -20,15 +20,8 @@ class SqdCurrentMonitor:
     def __call__(self, state, observed_states, k, physical_system):
         self.I_SD_IDX = physical_system.state_names.index('i_sd')
         self.I_SQ_IDX = physical_system.state_names.index('i_sq')
-        # normalize to limit_values, since state is normalized
-        nominal_values = physical_system.nominal_state / abs(
-            physical_system.limits)
-        limits = 1.5 * nominal_values
-        # calculating squared currents as observed measure
         sqd_currents = state[self.I_SD_IDX] ** 2 + state[self.I_SQ_IDX] ** 2
-
-        return (sqd_currents > limits[self.I_SD_IDX] or sqd_currents > limits[
-            self.I_SQ_IDX])
+        return sqd_currents > 1
 
 
 class EpsilonWrapper(ObservationWrapper):
@@ -59,8 +52,41 @@ class EpsilonWrapper(ObservationWrapper):
                                       observation[self.EPSILON_IDX + 1:]))
         return observation
 
+class AppendNLastOberservationsWrapper(Wrapper):
 
-def set_env(time_limit=True, gamma = 0.99):
+    def __init__(self, env, N):
+        super().__init__(env)
+        self._N = N
+        self._current_step = 0
+        self._obs = None
+        new_low = self.env.observation_space.low
+        new_high = self.env.observation_space.high
+        for i in range(self._N):
+            new_low = np.concatenate((new_low, self.env.observation_space.low))
+            new_high = np.concatenate((new_high, self.env.observation_space.high))
+        self.observation_space = Box(new_low, new_high)
+
+
+    def step(self, action):
+        obs, rew, term, info = self.env.step(action)
+        self._current_step += 1
+        if self._current_step <= self._N:
+            self._obs[self._current_step*self.env.observation_space.shape[0]:(self._current_step + 1)*self.env.observation_space.shape[0]] = obs
+        else:
+            valid_obs = self._obs[self.env.observation_space.shape[0]:]
+            self._obs = np.concatenate((valid_obs, obs))
+        return self._obs, rew, term, info
+
+    def reset(self, **kwargs):
+        self._current_step = 0
+        obs = self.env.reset()
+        for i in range(self._N):
+            obs = np.concatenate((obs, np.zeros(self.env.observation_space.shape)))
+        self._obs = obs
+        return self._obs
+
+
+def set_env(time_limit=True, gamma = 0.99, N=0):
     # define motor arguments
     motor_parameter = dict(p=3,  # [p] = 1, nb of pole pairs
                            r_s=17.932e-3,  # [r_s] = Ohm, stator resistance
@@ -73,7 +99,10 @@ def set_env(time_limit=True, gamma = 0.99):
                         i=230,
                         u=u_sup
                         )
-    limit_values=nominal_values.copy()
+    limit_values=dict(omega=4000*2*np.pi/60,
+                        i=1.5*230,
+                        u=u_sup
+                        )
     q_generator = WienerProcessReferenceGenerator(reference_state='i_sq')
     d_generator = WienerProcessReferenceGenerator(reference_state='i_sd')
     rg = MultipleReferenceGenerator([q_generator, d_generator])
[…]
