### 2020-09-21 cda4945
First commit fpr Visualization refactoring
