+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -7,12 +7,20 @@ from gym.spaces import Box
 class MotorDashboardPlot:
     """Base Plot class that all plots in the MotorDashboard have to derive from."""
 
-    def __init__(self):
+    _default_line_cfg = {
+        'linestyle': '',
+        'linewidth': 0.75,
+        'marker': '.',
+        'markersize': .5
+    }
+
+    def __init__(self, line_config=None):
+
         self._axis = None
-        self._tau = None
-        self._x_points = None
-        self._t = 0
-        self._t_data = []
+        line_config = line_config or {}
+        assert type(line_config) is dict, f'The line_config of the plots needs to be a dict but is {type(line_config)}.'
+
+        self._line_cfg = self._default_line_cfg.update(line_config)
 
     def initialize(self, axis):
         """Initialization of the plot. Set labels, legends,... here.
@@ -22,19 +30,18 @@ class MotorDashboardPlot:
         """
         self._axis = axis
         self._axis.grid(True)
-        self._axis.set_xlim(-self.x_width, 0)
 
     def set_modules(self, ps, rg, rf):
         """Interconnection of the environments modules.
+
         Save all relevant information from other modules here (e.g. state_names, references,...)
+
         Args:
             ps(PhysicalSystem): The PhysicalSystem of the environment
             rg(ReferenceGenerator): The ReferenceGenerator of the environment.
             rf(RewardFunction): The RewardFunction of the environment
         """
-        self._tau = ps.tau
-        # Number of points on the x-axis in a plot (= x_width / tau)
-        self._x_points = int(self.x_width / self._tau)
+        pass
 
     def step(self, k, state, reference, action, reward, done):
         """Passing of current environmental information..
@@ -49,44 +56,59 @@ class MotorDashboardPlot:
         """
         raise NotImplementedError
 
-    def update(self):
-        """Called by the MotorDashboard each time before the figure is updated."""
-        # configure x-axis properties
-        if self.mode == 'continuous':
-            x_lim = self._axis.get_xlim()
-            upper_lim = max(self._t, x_lim[1])
-            lower_lim = upper_lim - self.x_width
-            self._axis.set_xlim(lower_lim, upper_lim)
-
     def reset(self):
         """Called by the MotorDashboard each time the environment is reset."""
         pass
 
 
-class StatePlot(MotorDashboardPlot):
+class StepPlot(MotorDashboardPlot):
+
+    def __init__(self, x_width=1, update_cycle=1000, **kwargs):
+        super().__init__(**kwargs)
+        self._x_points = None
+        self._t = 0
+        self._t_data = []
+        self._tau = None
+        self.x_width = x_width
+        self.update_cycle = update_cycle
+        self.y_lim = (-np.inf, np.inf)
+
+    def reset(self):
+        """Called by the MotorDashboard each time the environment is reset."""
+        pass
+
+    def step(self, k, state, reference, action, reward, done):
+        raise NotImplementedError
+
+    def set_modules(self, ps, rg, rf):
+        super().set_modules(ps, rg, rf)
+        self._axis.set_xlim(-self.x_width, 0)
+        self._tau = ps.tau
+
+        # Number of points on the x-axis in a plot (= x_width / tau)
+        self._x_points = int(self.x_width / self._tau)
+
+    def update(self):
+        """Called by the MotorDashboard each time before the figure is updated."""
+        # configure x-axis properties
+        x_lim = self._axis.get_xlim()
+        upper_lim = max(self._t, x_lim[1])
+        lower_lim = upper_lim - self.x_width
+        self._axis.set_xlim(lower_lim, upper_lim)
+
[…]
