### 2020-09-22 bd48c40
changed reward weights back to 1 - was an internal testing not meant to be pushed
