### 2020-09-24 d2b8fc8
finalize example revision
