### 2020-09-24 965b74f
add comments to one example I overlooked
