### 2020-09-24 d5bc788
corrected latex formulas in examples readme (?)
