+++ examples/README.md
@@ -22,4 +22,4 @@ or as feature showcases
 ### Feature Showcases
 - [external_speed_profile.py](external_speed_profile.py): presents a builtin feature that can be used to define arbitrary speed profiles, which is useful when e.g. investigating generator operation (where mechanical and thus electrical frequency is determined by external means) 
 - [userdefined_initialization.py](userdefined_initialization.py): presents a builtin feature that allows the user to determine the initial state of the motor, which comes in handy when e.g. using exploring starts in reinforcement learning applications
-- [userdefined_constraints.py](userdefined_constraints.py): presents a builtin feature that can be used to extend the operational constraints of the motor, because oftentimes emergency shutdowns should be performed on the basis of composite conditions (e.g. for three-phase drives it is important to monitor $i_s = \sqrt{i_d^2 + i_q^2}$ instead of monitoring $i_d$ and $i_q$ independently)
+- [userdefined_constraints.py](userdefined_constraints.py): presents a builtin feature that can be used to extend the operational constraints of the motor, because oftentimes emergency shutdowns should be performed on the basis of composite conditions (e.g. for three-phase drives it is important to monitor <img src="https://latex.codecogs.com/png.latex?\inline&space;\bg_white&space;i_s=\sqrt{i_d^2&plus;i_q^2}" title="i_s=\sqrt{i_d^2+i_q^2}" /> instead of monitoring <img src="https://latex.codecogs.com/png.latex?\inline&space;\bg_white&space;i_d" title="i_d" /> and <img src="https://latex.codecogs.com/png.latex?\inline&space;\bg_white&space;i_q" title="i_q" /> independently)
