### 2020-09-25 28cbc7e
Added training flag to have the same test reward function every time
