### 2020-09-27 8be3f22
uniform motor initialization - improves training a lot
