+++ examples/rl_frameworks/stable_baselines3_testing.py
@@ -3,16 +3,33 @@ import sys
 sys.path.append(str(Path().resolve().parent.parent))
 from stable_baselines3 import DQN
 from setting_environment import set_env
-
+import numpy as np
+import matplotlib.pyplot as plt
 """
 This example is based on stable baselines3 0.8.0a5. Since it is still being frequently updated 
 some parts of this code might be broken or not the latest recommended way of working with it
 """
 
-gamma = 0.99
 
-env = set_env(time_limit = False, gamma = gamma)
-model = DQN.load(Path(__file__).parent / "saved_agents" / "sb3_dqn_PMSM")  
+rewards = np.load(Path(__file__).parent / "saved_agents" / "PreTrainedEpisodeRewards.npy")[1:]
+plt.grid(True)
+plt.xlim(0,len(rewards))
+plt.ylim(min(rewards), 0)
+plt.yticks(np.arange(min(rewards), 1, 1.0))
+plt.tick_params(axis = 'y', left = False, labelleft = False)
+plt.xticks(np.arange(0, len(rewards), 10))
+plt.xlabel('#Episode')
+plt.ylabel('Mean Reward per Episode (Qualitativ)')
+plt.plot(rewards)
+plt.show()
+
+N = 0
+M = 0
+gamma = 0.99
+time_limit = False
+
+env = set_env(time_limit = time_limit, gamma = gamma, N = N, M = M, training = False)
+model = DQN.load(Path(__file__).parent / "saved_agents" / "TutorialPreTrainedAgent")  
 obs = env.reset()   
 
 visualization_steps = int(9e4) # currently this crashes for larger values
@@ -26,21 +43,25 @@ for i in range(visualization_steps):
     cum_rew_episode += reward
     env.render()
     if done:
-        print(f'Episode had {i-last_i} steps')
-        print(f'The reward per step was {cum_rew_episode/(i-last_i)}')
         last_i = i
         cum_rew_episode = 0
-        env.reset() 
+        obs = env.reset() 
         
-print("\nStart Evaluation, this may take some while")
 test_steps = int(1e6) #1 milion for stability reasons
-cum_rew_testing_period = 0
-for j in range(3):
+print(f"\nStart Evaluation for {test_steps} steps, this may take some while")
+for j in range(2):
+    episode_lengths = []
+    episode_step = 0
+    cum_rew_testing_period = 0
     for i in range(test_steps):
+        print(f"{i+1}", end = '\r')
+        episode_step += 1
         action, _states = model.predict(obs, deterministic=True)
         obs, reward, done, _ = env.step(action)
         cum_rew_testing_period += reward
         if done:
-            env.reset()
-    print(f"Reward per step for testing period {j+1} with {test_steps} steps: {cum_rew_testing_period/test_steps} ")
-    cum_rew_testing_period = 0
+            episode_lengths.append(episode_step)
+            episode_step = 0
+            obs = env.reset()
+    print(f"Reward per step for testing period {j+1} with {test_steps} steps: {cum_rew_testing_period/test_steps:.4f} ")
+    print(f"The average Episode length was: {round(np.mean(episode_lengths))} ")
