### 2020-09-30 18b573b
changed some formatting of the code
