### 2020-10-01 4871cd1
integrate remark by @wkirgsn concerning block comments
