+++ examples/dqn_series_current_control.py
@@ -17,9 +17,11 @@ from gym_electric_motor.reward_functions import WeightedSumOfErrors
 from gym_electric_motor.visualization import MotorDashboard
 from gym_electric_motor.reference_generators import WienerProcessReferenceGenerator
 
-# This example shows how we can use GEM to train a reinforcement learning agent to control the motor current
-# of a DC series motor. In this scenario, the state space is continuous while the action space is discrete.
-# We use a deep Q learning agent to determine which action must be taken on a finite-control-set
+'''
+This example shows how we can use GEM to train a reinforcement learning agent to control the motor current
+of a DC series motor. In this scenario, the state space is continuous while the action space is discrete.
+We use a deep Q learning agent to determine which action must be taken on a finite-control-set
+'''
 
 if __name__ == '__main__':
 
