### 2020-10-03 87e46c1
Visualization update almost finished. Documentation still required.
