+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -1,26 +1,30 @@
+import matplotlib.pyplot as plt
 import numpy as np
-import matplotlib.lines as lin
-from collections import deque
 from gym.spaces import Box
+from gym_electric_motor.core import Callback
 
 
-class MotorDashboardPlot:
+class MotorDashboardPlot(Callback):
     """Base Plot class that all plots in the MotorDashboard have to derive from."""
 
     _default_line_cfg = {
         'linestyle': '',
         'linewidth': 0.75,
         'marker': '.',
-        'markersize': .5
+        'markersize': .75
     }
 
     def __init__(self, line_config=None):
-
+        super().__init__()
         self._axis = None
         line_config = line_config or {}
         assert type(line_config) is dict, f'The line_config of the plots needs to be a dict but is {type(line_config)}.'
 
-        self._line_cfg = self._default_line_cfg.update(line_config)
+        self._line_cfg = self._default_line_cfg.copy()
+        self._line_cfg.update(line_config)
+        self._lines = None
+        self._labels = None
+        self._colors = [cycle['color'] for cycle in plt.rcParams['axes.prop_cycle']]
 
     def initialize(self, axis):
         """Initialization of the plot. Set labels, legends,... here.
@@ -31,86 +35,68 @@ class MotorDashboardPlot:
         self._axis = axis
         self._axis.grid(True)
 
-    def set_env(self, env):
-        """Interconnection of the environments modules.
-
-        Save all relevant information from the environment here (e.g. state_names, references,...)
-
-        Args:
-            env(ElectricMotorEnvironment): The environment to read information like state names from it.
-        """
-        pass
-
-    def on_step_begin(self, k, action):
-        """Passing of current environmental information..
-
-        Args:
-            k(int): Current episode step.
-            action(ndarray(float)): Last taken action. (None after reset)
-        """
-        raise NotImplementedError
-
-    def on_step_end(self, k, state, reference, reward, done):
-        """Passing of step result information.
-
-        Args:
-            k(int): Current episode step.
-            state(ndarray(float)): The next state of the environment.
-            reference(ndarray(float)): The reference for the current timestep.
-            reward(float): The reward received for the last taken action.
-            done(bool): Indicator, if the episode has terminated due to a constraint violation.
-        """
-        raise NotImplementedError
-
-    def on_reset_begin(self):
-        """Called by the MotorDashboard each time before the environment is reset."""
-        pass
-
-    def on_reset_end(self, k, state, reference):
-        raise NotImplementedError
-
     def render(self):
         raise NotImplementedError
 
 
 class StepPlot(MotorDashboardPlot):
 
-    def __init__(self, x_width=1, update_cycle=1000, **kwargs):
+    _default_violation_line_cfg = {
+        'color': 'red',
+        'linewidth': 0.5,
+        'linestyle': '-'
+    }
+
+    _default_reset_line_cfg = {
+        'color': 'blue',
+        'linewidth': 0.5,
+        'linestyle': '-'
+    }
+
+    def __init__(self, x_width=1, violation_line_cfg=None, reset_line_cfg=None, **kwargs):
         super().__init__(**kwargs)
+        reset_line_cfg = reset_line_cfg or {}
+        violation_line_cfg = violation_line_cfg or {}
+
+        assert type(reset_line_cfg) is dict
+        assert type(violation_line_cfg) is dict
+
+        self._reset_line_cfg = self._default_reset_line_cfg.copy()
+        self._reset_line_cfg.update(reset_line_cfg)
+
[…]
