### 2020-10-03 6152548
Updated ConsolePrinter to new Visualization Interface
