+++ gym_electric_motor/visualization/console_printer.py
@@ -22,54 +22,66 @@ class ConsolePrinter(ElectricMotorVisualization):
                 Unsigned Integer setting the update frequency if verbose is 2.
                 It's value n means that each nth step the corresponding output will be printed
         """
+        super().__init__()
         self._limits = None
+        self._state = None
+        self._reference = None
+        self._action = None
+        self._reward = None
+        self._k = 0
         self._print_freq = verbose
         self._update_freq = update_freq
         self._num_steps = 0
         self._cum_reward = 0
-        self._violation = False
+        self._done = False
         self._episode = 0
         np.set_printoptions(formatter={'float': '{:9.3f}'.format})
         self._reset = False
 
-    def on_reset_end(self, observation):
+    def on_reset_begin(self):
         """Gets called on environment reset. Handles internal value reset and External reset printing"""
         self._reset = True
+
+    def on_reset_end(self, state, reference):
+        self._reference = reference
+        self._state = state
 
     def set_env(self, env):
         """Gets the limits of the current physical system for accurate printing"""
         self._limits = env.physical_system.limits
 
-    def on_step_begin(self, action):
+    def on_step_begin(self, k, action):
+        self._k = k
+        self._action = action
+
+    def on_step_end(self, k, state, reference, reward, done):
         """Gets called at each step of the environment.
         Handles per step printing as well es constraint violation printing"""
-
-        if self._print_freq > 0:
-            self._num_steps += 1
-            self._cum_reward += reward
-            if done:
-                self._violation = True
-            if self._print_freq > 0 and self._reset:
-                print(
-                    f'\nEpisode {self._episode} ',
-                    f'Constraint Violation! ' if self._violation else 'External Reset. ',
-                    f'Number of steps: {self._num_steps: 8d} ',
-                    f'Cumulative Reward: {self._cum_reward:7.3f}\n')
-                self._violation = False
-                self._num_steps = 1
-                self._cum_reward = reward
-                self._reset = False
-                self._episode += 1
-                
-            if self._print_freq == 2 and (self._num_steps % self._update_freq) == 0:
-                print(f'Episode {self._episode} '
-                      f'State {state * self._limits} '
-                      f'Reference {reference * self._limits} '
-                      f'Reward {reward:7.3f} '
-                      f'Step {self._num_steps:08d} '
-                      f'Cumulative Reward {self._cum_reward:7.3f}',
-                      end='\r'
-                )
+        self._k = k
+        self._state = state
+        self._reference = reference
+        self._reward = reward
+        self._cum_reward += reward
+        self._done = done
 
     def render(self):
-        print('')
+        if self._print_freq > 0:
+            if self._reset:
+                print(
+                    f'\nEpisode {self._episode} ',
+                    f'Constraint Violation! ' if self._done else 'External Reset. ',
+                    f'Number of steps: {self._k: 8d} ',
+                    f'Cumulative Reward: {self._cum_reward:7.3f}\n')
+                self._cum_reward = 0
+                self._reset = False
+                self._episode += 1
+
+            if self._print_freq == 2 and (self._k % self._update_freq) == 0:
+                print(f'Episode {self._episode} '
+                      f'Step {self._k: 8d}'
+                      f'State {self._state * self._limits} '
+                      f'Reference {self._reference * self._limits} '
+                      f'Reward {self._reward:7.3f} '
+                      f'Cumulative Reward {self._cum_reward:7.3f}',
+                      end='\r'
+                      )
