### 2020-10-03 02a5399
bugfix in ConsolePrinter
