### 2020-10-04 e812aa2
Initial commit ConstraintMonitor Refactoring. Not completed.
