+++ gym_electric_motor/__init__.py
@@ -2,7 +2,7 @@ from .core import ReferenceGenerator
 from .core import PhysicalSystem
 from .core import RewardFunction
 from .core import ElectricMotorVisualization
-from .constraint_monitor import ConstraintMonitor
+from .constraints import Constraint, LimitConstraint
 from .utils import make, register_superclass
 
 register_superclass(RewardFunction)
