+++ gym_electric_motor/core.py
@@ -15,13 +15,13 @@ Each ElectricMotorEnvironment contains the four following modules:
 * ElectricMotorVisualization
     - Visualization of the PhysicalSystems state, reference and reward for the user.
 """
+from functools import reduce
 import gym
 import numpy as np
 from gym.spaces import Box
 
-from .utils import set_state_array
 from .utils import instantiate
-from .constraint_monitor import ConstraintMonitor as CM
+from .constraints import Constraint, LimitConstraint
 
 
 class ElectricMotorEnvironment(gym.core.Env):
@@ -147,8 +147,8 @@ class ElectricMotorEnvironment(gym.core.Env):
         """
         return self._physical_system.nominal_state[self.state_filter]
 
-    def __init__(self, physical_system, reference_generator, reward_function, visualization=None, state_filter=None, callbacks = [],
-                 **kwargs):
+    def __init__(self, physical_system, reference_generator, reward_function, visualization=None, state_filter=None,
+                 callbacks=(), **kwargs):
         """
         Setting and initialization of all environments' modules.
 
@@ -194,7 +194,7 @@ class ElectricMotorEnvironment(gym.core.Env):
         self.reward_range = self._reward_function.reward_range
         self._action = None
         
-        self._callbacks = callbacks
+        self._callbacks = list(callbacks)
         self._call_callbacks('set_env', self)
         
     def _call_callbacks(self, func_name, *args):
@@ -455,25 +455,6 @@ class RewardFunction:
     #: Tuple(int,int): Lower and upper possible reward (excluding limit violation reward)
     reward_range = (-np.inf, np.inf)
 
-    def __init__(self, observed_states='currents',
-                 constraint_monitor=None, **__):
-        """
-        Args:
-            observed_states(str/iterable(str)): Names of the observed states. 'all' for the observation of every state.
-                'currents' for all currents and 'voltages' for all voltages. Combinations of 'currents' and
-                'voltages' with other states are possible. Choose None for not observing states.
-            constraint_monitor(class instance): ConstraintMonitor for monitoring
-                states regarding defined constraints
-        """
-        self._physical_system = None
-        observed_states = observed_states or []
-        if not isinstance(observed_states, list):
-            observed_states = [observed_states]
-        self._observed_states = observed_states
-        self._reference_generator = None
-        self._monitor = CM(external_monitor=constraint_monitor) or CM()
-        self._limits = None
-
     def __call__(self, state, reference, k):
         """
         Call of the reward calculation.
@@ -489,6 +470,33 @@ class RewardFunction:
         """
         return self.reward(state, reference, k)
 
+    def _limit_violation_reward(self, state):
+        """
+        Called, when limits have been violated to return a special reward for this case.
+
+        Args:
+            state(ndarray(float)): Current state array of the environment.
+
+        Returns:
+            float: The limit violation reward.
+        """
+        raise NotImplementedError
+
+    def _standard_reward(self, state, reference, action):
+        """
+        Standard reward function. Called, when no limits have been violated to calculate the reward based on the state
+        and the reference.
+
+        Args:
+            state(ndarray(float)): State array of the environment.
+            reference(ndarray(float): Reference array of the environment.
+            action(element of action space): The previously taken action.
+
+        Returns:
+            float: Standard reward.
+        """
+        raise NotImplementedError
+
     def set_modules(self, physical_system, reference_generator):
         """
         Setting of the physical system, to set state arrays fitting to the environments states
@@ -497,40 +505,9 @@ class RewardFunction:
             physical_system(PhysicalSystem): The physical system of the environment
             reference_generator(ReferenceGenerator): The reference generator of the environment.
         """
-        observed_states = {}
-        allowed_observed_states = physical_system.state_names + \
-                                  ['all', 'currents', 'voltages']
[…]
