### 2020-10-10 #2498 @Kevin-Mattheus-Moerman
@dineshresearch thanks for your help here! I see you have left the `State of the field` box :point_up: unchecked. Can you summarize for the authors what action is needed here?
