### 2020-10-11 caee645
Bugfixes in examples
