+++ examples/ddpg_series_omega_control.py
@@ -25,7 +25,7 @@ if __name__ == '__main__':
     env = gem.make(
         'emotor-dc-series-cont-v1',
         # Pass a class with extra parameters
-        visualization=MotorDashboard(['omega', 'torque', 'i', 'u', 'u_sup', 'reward']), visu_period=1,
+        visualization=MotorDashboard(state_plots=['omega', 'torque', 'i', 'u', 'u_sup'], reward_plot=True),
         motor_parameter=dict(r_a=2.5, r_e=4.5, l_a=9.7e-3, l_e_prime=9.2e-3, l_e=9.2e-3, j_rotor=0.001),
         # Take standard class and pass parameters (Load)
         load_parameter=dict(a=0, b=.0, c=0.01, j_load=.001),
