### 2020-10-12 ab4f2bd
small clean up
