+++ examples/perm_DC_omega.py
@@ -29,7 +29,7 @@ if __name__ == '__main__':
                       switching states which can be generated by the converter.     
     """
     env = gem.make(
-        'DcPermExDisc-v1',  # replace with 'DcPermExCont-v1'  for continuous mode
+        'DcPermExDisc-v1',
         visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u', 'u_sup'], visu_period=1),
         ode_solver='scipy.solve_ivp', solver_kwargs=dict(),
         reference_generator=rg.SwitchedReferenceGenerator(
@@ -38,8 +38,7 @@ if __name__ == '__main__':
             ], p=[0.1, 0.8, 0.1], super_episode_length=(1000, 10000)
         )
     )
-    controller = Controller.make('p_controller', env)
-    # The above controller can be replaced with 'on_off'(disc),three_point(disc),'p_controller', 'cascaded pi' from simple_controllers class
+    controller = Controller.make('on_off', env)
     state, reference = env.reset()
     start = time.time()
     cum_rew = 0
