### 2020-10-13 3573f3e
Some language fixes for SB3 notebook, deleted unnecessary code
