### 2020-10-14 60a9399
Delete unnecessary code
