### 2020-10-14 04e0abf
updated requirements for tensorforce
