+++ examples/example_notebooks/requirements.txt
@@ -3,5 +3,6 @@ gym-electric-motor
 numpy
 stable-baselines3
 keras-rl2
+tensorforce==0.5.5
 
 
