### 2020-10-15 a960449
Added requirements to the setup.py
