+++ setup.py
@@ -18,8 +18,11 @@ setuptools.setup(
       description='An OpenAI gym environment for electric motor control.',
       packages=setuptools.find_packages(),
       install_requires=requirements,
-      extras_require={'examples': 
-                      ['keras-rl2 @git+https://github.com/wau/keras-rl2.git']
+      extras_require={'examples': [
+                          'keras-rl2 @git+https://github.com/wau/keras-rl2.git',
+                          'stable-baselines3',
+                          'tensorforce==0.5.5',
+                       ]
                      },
       author=', '.join(sorted(AUTHORS, key=lambda n: n.split()[-1].lower())),
       long_description=long_description,
