### 2020-10-16 0fe05ad
renamed set_env to get_env
