### 2020-10-16 287fb7a
added explanation for the removal of action 8
