### 2020-10-19 c1612b6
Inductance values for pi_series example are much too low -> controlled motor always crashes into overcurrent terminations.
