### 2020-10-19 25f0095
PI_series_omega example is only working with PI controller right now. Existing inline comment that it would also work with other controller types was deleted.
