### 2020-10-19 dd96b61
spelling and comment corrections
