### 2020-10-19 c638269
Unfortunately some of the simple control sub-classes lacked any logic and were obviously never tested within the simulation, because a reasonable control behaviour would not have been possible. Such classes were deleted. Furthermore, spelling corrections were made (not conclusive).
