+++ examples/agents/simple_controllers.py
@@ -39,11 +39,11 @@ class Controller:
 
 class OnOffController(Controller):
     """
-        The following controller is a simple OnOffController which allows the state to choose high_action when referenced_state is below
-        ref_idx and low_action otherwise.A 'Hysteresis' value should be integrated(chosen as 0.01) in this controller because of the
+        The following controller is a simple on-off controller with optional hysteresis which allows the state to choose high_action when referenced_state is below
+        ref_idx and low_action otherwise. A 'hysteresis' value should be integrated (chosen as 0.01) in this controller because of the
          constant switching and high frequency around the reference_idx.
 
-        Valid for Motors with Discrete Action Space
+        Valid for motors with discrete action space
     """
 
     def __init__(self, environment, hysteresis=0.01, state_idx=None, reference_idx=0):
@@ -74,10 +74,10 @@ class ThreePointController(Controller):
     """
     Below is an implementation of a 3 point controller: When state_idx is below the reference_idx, it choose high_action and when the
     state_idx is above the reference_idx,it choose low_action. If it is between the reference_idx values, it will choose idle_action
-    A 'Hysteresis' value should be integrated(chosen as 0.01) in this controller because of the constant switching and
+    A 'hysteresis' value should be integrated(chosen as 0.01) in this controller because of the constant switching and
     high frequency around the reference_idx.
 
-    Valid for Motors with Discrete Action Space
+    Valid for motors with discrete action space
 
     """
 
@@ -112,10 +112,10 @@ class PController(Controller):
     """
     In the below proportional control implementation, the controller output is proportional to the error signal,
      which is the difference between the reference_idx and the state_idx .i.e., the output of a proportional controller
-    is the multiplication product of the error signal and the proportional gain.Here kp =10 is assumed for
-    proportionality gain.
+    is the multiplication product of the error signal and the proportional gain. Here kp =10 is assumed for
+    proportionality gain by default, which is likely to require adaption to a given control plant.
 
-    Valid for DC Motor System
+    Valid for DC motor system
     """
 
     def __init__(self, environment, k_p=10, controller_no=0, state_idx=None, reference_idx=0):
@@ -148,12 +148,12 @@ class PController(Controller):
 
 class IController(Controller):
     """
-    In the below Integral control implementation, the controller output is proportional to the integral of error signal,
+    In the below integral control implementation, the controller output is proportional to the integral of error signal,
      which is the difference between the reference_idx and the state_idx .i.e., the output of a integral controller
     is the multiplication product of the error signal and  integral gain.
-    Here k_i =0.01 is assumed for proportionality gain.
+    Here k_i =0.01 is assumed for integral gain by default which is likely to require individual tuning based on the given control plant.
 
-    Valid for DC Motor System
+    Valid for DC motor system
     """
     def __init__(self, environment, k_i=0.01, controller_no=0, state_idx=None, reference_idx=0):
         action_space = environment.action_space
@@ -200,11 +200,11 @@ class IController(Controller):
 
 class PIController(PController):
     """
-      The below Discrete PI Controller performs discrete-time PI controller computation using the error signal and
+      This class performs discrete-time PI controller computation using the error signal and
       proportional and integral gain inputs. The error signal is the difference between the reference_idx and the
       referenced_state. It outputs a weighted sum of the input error signal and the integral of the input error signal.
 
-      Valid for DC Motor System
[…]
