+++ examples/example_notebooks/getting_environment.py
@@ -1,0 +1,181 @@
+import numpy as np
+
+import gym_electric_motor as gem
+from gym_electric_motor.reference_generators import \
+    MultipleReferenceGenerator,\
+    WienerProcessReferenceGenerator
+from gym_electric_motor.visualization import MotorDashboard
+from gym.spaces import Discrete, Box
+from gym.wrappers import FlattenObservation, TimeLimit
+from gym import ObservationWrapper
+
+
+class SqdCurrentMonitor:
+    """
+    Constraint monitor which monitors if the sum of the squared currents
+    is bigger than the nominal limit:
+
+    i_sd**2 + i_sq**2 < nominal_limit
+    
+    Ends the episode and throws a high negative reward, otherwise.
+    """
+
+    def __call__(self, state, observed_states, k, physical_system):
+        """
+        Gets called at each step to check if the agent's trajectory is violating the above constraint
+        """
+        self.I_SD_IDX = physical_system.state_names.index('i_sd')
+        self.I_SQ_IDX = physical_system.state_names.index('i_sq')
+        sqd_currents = state[self.I_SD_IDX] ** 2 + state[self.I_SQ_IDX] ** 2
+        return sqd_currents > 1
+
+
+class FeatureWrapper(ObservationWrapper):
+    """
+    Wrapper class which wraps the environment to change its observation. Serves
+    the purpose to improve the agent's learning speed.
+    
+    It changes epsilon to cos(epsilon) and sin(epsilon). This serves the purpose
+    to have the angles -pi and pi close to each other numerically without losing
+    any information on the angle.
+    
+    Additionally, this wrapper adds a new observation i_sd**2 + i_sq**2. This should
+    help the agent to easier detect incoming limit violations.
+    """
+
+    def __init__(self, env, epsilon_idx, i_sd_idx, i_sq_idx):
+        """
+        Changes the observation space to fit the new features
+        
+        Args:
+            env(GEM env): GEM environment to wrap
+            epsilon_idx(integer): Epsilon's index in the observation array
+            i_sd_idx(integer): I_sd's index in the observation array
+            i_sq_idx(integer): I_sq's index in the observation array
+        """
+        super(FeatureWrapper, self).__init__(env)
+        self.EPSILON_IDX = epsilon_idx
+        self.I_SQ_IDX = i_sq_idx
+        self.I_SD_IDX = i_sd_idx
+        new_low = np.concatenate((self.env.observation_space.low[
+                                  :self.EPSILON_IDX], np.array([-1.]),
+                                  self.env.observation_space.low[
+                                  self.EPSILON_IDX:], np.array([0.])))
+        new_high = np.concatenate((self.env.observation_space.high[
+                                   :self.EPSILON_IDX], np.array([1.]),
+                                   self.env.observation_space.high[
+                                   self.EPSILON_IDX:],np.array([1.])))
+
+        self.observation_space = Box(new_low, new_high)
+
+    def observation(self, observation):
+        """
+        Gets called at each return of an observation. Adds the new features to the
+        observation and removes original epsilon.
+        
+        """
+        cos_eps = np.cos(observation[self.EPSILON_IDX] * np.pi)
+        sin_eps = np.sin(observation[self.EPSILON_IDX] * np.pi)
+        currents_squared = observation[self.I_SQ_IDX]**2 + observation[self.I_SD_IDX]**2
+        observation = np.concatenate((observation[:self.EPSILON_IDX],
+                                      np.array([cos_eps, sin_eps]),
+                                      observation[self.EPSILON_IDX + 1:],
+                                      np.array([currents_squared])))
+        return observation
+
+
[…]
