### 2020-10-21 e15f9bc
polish up gem tutorial. Polish up README (links to colab missing yet). Other tutorials also need polishment yet
