+++ README.md
@@ -1,47 +1,45 @@
+# Gym Electric Motor
 ![](docs/plots/Motor_Logo.png)
+
+[**Paper**](https://arxiv.org/abs/1910.09434)
+| [**Quickstart**](#getting-started)
+| [**Install guide**](#installation)
+| [**Reference docs**](https://upb-lea.github.io/gym-electric-motor/)
+| [**Release notes**](https://github.com/upb-lea/gym-electric-motor/releases)
 
 [![Build Status](https://travis-ci.org/upb-lea/gym-electric-motor.svg?branch=master)](https://travis-ci.org/upb-lea/gym-electric-motor)
 [![codecov](https://codecov.io/gh/upb-lea/gym-electric-motor/branch/master/graph/badge.svg)](https://codecov.io/gh/upb-lea/gym-electric-motor)
 [![PyPI version shields.io](https://img.shields.io/pypi/v/gym-electric-motor.svg)](https://pypi.python.org/pypi/gym-electric-motor/)
 [![License](https://img.shields.io/github/license/mashape/apistatus.svg?maxAge=2592000)](https://github.com/upb-lea/gym-electric-motor/blob/master/LICENSE)
 
-__The gym-electric-motor (GEM) package is a software toolbox for the
-simulation of
-different electric motors.__ 
+## Overview
+The gym-electric-motor (GEM) package is a Python toolbox for the simulation and control of various electric motors.
 
-The toolbox is built upon the [OpenAI Gym Environments](https://gym.openai.com/) for reinforcement learning. 
-Therefore, the toolbox is specifically designed for running reinforcement 
-learning algorithms to train agents controlling electric motors.
+It is built upon [OpenAI Gym Environments](https://gym.openai.com/), and, therefore, can be used for both, classical control simulation and reinforcement learning experiments. It allows you to construct a typical drive train with the usual building blocks, i.e. supply voltages, converters, electric motors and load models, and obtain not only a closed-loop simulation of this physical structure, but also a rich interface for plugging in any decision making algorithm, from PI-controllers to [Deep Deterministic Policy Gradient](https://spinningup.openai.com/en/latest/algorithms/ddpg.html) agents.
 
-[Read the detailed docs!](https://upb-lea.github.io/gym-electric-motor/)
+## Getting Started
+An easy way to get started with GEM is by playing around with the following interactive notebooks in Google Colaboratory. Most important features of GEM as well as application demonstrations are showcased, and give a kickstart for engineers in industry and academia.
 
-So far, several DC-motor models and the three-phase motors permanent magnet synchronous motor (PMSM), 
-synchronous reluctance motor (SynRM), squirrel cage induction motor (SCIM) and doubly-fed induction 
-motor (DFIM) are available.
-Beside electrical motors, also converters and load models are implemented. The converters can be driven by means of a duty cycle (continuous mode) or 
-switching commands (discrete mode). 
-The figure shows the basic scheme of the converter, motor and load. 
-### Physical Structure of the Environment's Components
-![](docs/plots/SCML_Setting.svg)
-### Control Flow of a Step Cycle of the Environment 
-![](docs/plots/CycleScheme.svg)
 
-### Citation
-A whitepaper for this framework is available under [arxiv.org/abs/1910.09434
-](https://arxiv.org/abs/1910.09434). Please use the following BibTeX entry for citing us:
-```
-@misc{traue2019reinforcement,
-    title={Towards a Reinforcement Learning Environment Toolbox for Intelligent Electric Motor Control},
-    author={Arne Traue and Gerrit Book and Wilhelm Kirchgässner and Oliver Wallscheid},
-    year={2019},
-    eprint={1910.09434},
[…]
