### 2020-10-21 73260ef
add colab links (working only as of existence in the master branch)
