### 2020-10-21 e643258
Delete perm_DC_omega.py
