### 2020-10-21 07a59ea
add authors, increase version to 0.3.0
