### 2020-10-22 6b63030
optimize folder structure within examples
