### 2020-10-22 312c862
Incorporate reviewer feedback (OW). Resolve typographical errors, fix broken links
