### 2020-10-22 0a0707b
more typographical corrections
