### 2020-10-22 #2498 @wkirgsn
The toolbox just experienced its major update by merging nightly into master.
We now stand on v0.3.0, with plenty of new additions, bug fixes and comprehensive tutorials as well as example scripts.
These include the feedback of our reviewers, and we hope that discrepancies are decimated now.

We are looking forward to any feedback and appreciate your time.
