+++ _sources/parts/physical_systems/voltage_supplies/ac_1_phase_supply.rst.txt
@@ -1,0 +1,6 @@
+AC 1 Phase Supply
+##############################
+
+.. autoclass:: gym_electric_motor.physical_systems.voltage_supplies.AC1PhaseSupply
+    :members:
+    :inherited-members:
