### 2020-10-23 #2498 @moorepants
@wkirgsn Thank you. It would be helpful if you address my comments explicitly and point to specific new content that address each.
