### 2020-10-23 ef06323
refactoring of the links in JOSS and examples README
