+++ examples/README.md
@@ -26,6 +26,6 @@ The presented examples can be classified either as classical control examples (w
 - [external_speed_profile.py](environment_features/external_speed_profile.py): presents a builtin feature that can be used to define arbitrary speed profiles, which is useful when e.g. investigating generator operation (where mechanical and thus electrical frequency is determined by external means) 
 - [userdefined_initialization.py](environment_features/userdefined_initialization.py): presents a builtin feature that allows the user to determine the initial state of the motor, which comes in handy when e.g. using exploring starts in reinforcement learning applications
 - [userdefined_constraints.py](environment_features/userdefined_constraints.py): presents a builtin feature that can be used to extend the operational constraints of the motor, because oftentimes emergency shutdowns should be performed on the basis of composite conditions (e.g. for three-phase drives it is important to monitor <img src="https://latex.codecogs.com/png.latex?\inline&space;\bg_white&space;i_s=\sqrt{i_d^2&plus;i_q^2}" title="i_s=\sqrt{i_d^2+i_q^2}" /> instead of monitoring <img src="https://latex.codecogs.com/png.latex?\inline&space;\bg_white&space;i_d" title="i_d" /> and <img src="https://latex.codecogs.com/png.latex?\inline&space;\bg_white&space;i_q" title="i_q" /> independently)
-- [scim_ideal_grid_simulation.py](environment_features/scim_ideal_grod_simulation.py): simulates the start-up behavior of the squirrel cage induction motor connected to an ideal three-phase grid. 
+- [scim_ideal_grid_simulation.py](environment_features/scim_ideal_grid_simulation.py): simulates the start-up behavior of the squirrel cage induction motor connected to an ideal three-phase grid. 
 The state and action space is continuous.
 Running the example will create a formatted plot that show the motors angular velocity, the drive torque, the applied voltage in three-phase abc-coordinates and the measured current in field-oriented dq-coordinates.
