### 2020-10-23 0e15708
fix links in JOSS paper
