### 2020-10-23 3d4d82b
link notebooks to colab in examples README
