+++ examples/README.md
@@ -5,22 +5,22 @@ The presented examples can be classified either as classical control examples (w
 (which present a rather novel branch of control theory), or as feature showcases (where the focus lies in introducing interesting and useful builtins).
 
 ### Installation, Setup and Interface
-- [GEM_cookbook.ipynb](environment_features/GEM_cookbook.ipynb): a basic tutorial-style notebook that presents the basic interface and usage of GEM
+- [GEM_cookbook.ipynb](https://colab.research.google.com/github/upb-lea/gym-electric-motor/blob/master/examples/environment_features/GEM_cookbook.ipynb): a basic tutorial-style notebook that presents the basic interface and usage of GEM
 
 ### Classical Control
 - [pi_series_omega_control.py](classic_controllers/pi_series_omega_control.py): a conventional linear control algorithm applied to the speed control of a series DC motor with continuous-control-set
 - [perm_dc_omega_control.py](classic_controllers/perm_dc_omega_control.py): a conventional discontinuous control algorithm applied to the speed control of a permanently excited DC motor with finite-control-set
 
 ### Advanced Control
-- [gekko_mpc_cont_pmsm_example.ipynb](model_predictive_controllers/gekko_mpc_cont_pmsm_example.ipynb): a model predictive control solution for the currents of the three-phase permanent magnet synchronous motor on a continuous-control-set
+- [gekko_mpc_cont_pmsm_example.ipynb](https://colab.research.google.com/github/upb-lea/gym-electric-motor/blob/master/examples/model_predictive_controllers/gekko_mpc_cont_pmsm_example.ipynb): a model predictive control solution for the currents of the three-phase permanent magnet synchronous motor on a continuous-control-set
 
 ### Reinforcement-Learning Control
 - [dqn_series_current_control.py](reinforcement_learning_controllers/dqn_series_current_control.py): a deep Q-value network reinforcement-learning control approach for finite-control-set current control of a series DC motor
 - [ddpg_pmsm_dq_current_control.py](reinforcement_learning_controllers/ddpg_pmsm_dq_current_control.py): a deep deterministic policy gradient reinforcement-learning control approach applied to the current control of a permanent magnet synchronous motor within the $dq$-frame with continuous-control-set
 - [ddpg_series_omega_control.py](reinforcement_learning_controllers/ddpg_series_omega_control.py): a deep deterministic policy gradient reinforcement-learning control approach applied to the speed control of a series DC motor with continuous-control-set
-- [keras_rl2_dqn_disc_pmsm_example.ipynb](reinforcement_learning_controllers/keras_rl2_dqn_disc_pmsm_example.ipynb): a tutorial-style notebook that presents the usage of GEM in conjunction with [Keras_RL2](https://github.com/wau/keras-rl2) in the context of deep Q learning current control of a permanent magnet synchronous motor
-- [stable_baselines3_dqn_disc_pmsm_example.ipynb](reinforcement_learning_controllers/stable_baselines3_dqn_disc_pmsm_example.ipynb): a tutorial-style notebook that presents the usage of GEM in conjunction with [Stable Baselines3](https://github.com/DLR-RM/stable-baselines3) in the context of deep Q learning current control of a permanent magnet synchronous motor
-- [tensorforce_dqn_disc_pmsm_example.ipynb](reinforcement_learning_controllers/tensorforce_dqn_disc_pmsm_example.ipynb): a tutorial-style notebook that presents the usage of GEM in conjunction with [Tensorforce](https://github.com/tensorforce/tensorforce) in the context of deep Q learning current control of a permanent magnet synchronous motor
[…]
