### 2020-10-23 #2498 @wkirgsn
Hi @moorepants, please see the detailed response below:

__"Some of the paper authors, including the first author, do not seem to have made any commits to the software repository. Please clarify the authorship of all the authors."__
Email addresses were updated on github among authors, such that code contributions should be addressable now.
Github Insights might not show the updated data right off, but the box "contributors" on the repo frontpage includes all authors now, and are reflected in the git history in any case.

__These sentences "In particular, the domain of three-phase drives makes use of coordinate transformations to view these ODEs in a more interpretable frame. In GEM, both, the physical (abc-) and the simplified (dq-)coordinates are available to be used as the frame of input and output variables, allowing for easy and quick controller analysis and diagnose within the most convenient coordinate frame" are not understandable without some domain knowledge. Please define "frame" in this context as well as "physical (abc-) and the simplified (dq-)".__

Sentence was reworked in the [revised JOSS paper](https://github.com/upb-lea/gym-electric-motor/blob/master/docs/JOSS/paper.md).

__"I don't think "generatoric" is a word in the English language."__
Sentence was reworked in the [revised JOSS paper](https://github.com/upb-lea/gym-electric-motor/blob/master/docs/JOSS/paper.md).

__"You list 4 examples in the paper but these are not present in the documentation. I suggest adding these examples to the documentation so that users can easily find and understand the examples. The examples, as written are not easily interpretable by a user. There are very few comments in the files explaining what is going on. I think you need a tutorial style section of the documentation to teach users how to use your software. I don't think I could figure it out with the present information."__
The [example scripts](https://github.com/upb-lea/gym-electric-motor/tree/master/examples) are reworked completely with added documentation. Moreover, several tutorial notebooks were introduced (links in README) for a step-by-step guide to various aspects of the toolbox

__"Functionality: Have the functional claims of the software been confirmed?
This is not apparent from the documentation or the paper. You show one plot in the paper that is created from the software, but there is no associated code that I can run to reproduce that plot. I think a tutorial explaining your four examples that provide the inputs the user types and the subsequent outputs that the software produces would demonstrate the functionality (at least for those for examples)."__
Several tutorial notebooks were introduced (links in README) for a step-by-step guide to various aspects of the toolbox

__"Installation
My attempt at installing the software failed. Here is the session: .."__
Any installation related problems were rectified. In particular, Python 3.8 is supported now.

__"Also not clear why testing related packages are installed as dependencies, shouldn't these only be required if you are developing the package?"__
We feel that also mere users of the toolbox might want to ensure that what they use is also thoroughly tested.
The pytest packages are rather lightweight and shouldn't be perceived as a burden.

__"Statement of need
This could be more explicit and there could be more. When I read the paper I interpreted it as Matlab does this already but doesn't connect to keras/tensorflow well, so you need this package. That is a sufficient need, but is that the main/only need?"__
This was more elaborated on in the [revised JOSS paper](https://github.com/upb-lea/gym-electric-motor/blob/master/docs/JOSS/paper.md).

__"Installation instructions
I don't see an explicit list of dependencies. This would be helpful for setting up a conda environment, for example. There is also a dependency issue pip encounters as shown above. I don't see any info in the docs about dependencies that would help me debug this."__
Any required packages are compiled in the ```requirements.txt``` as is usual for python packages. They'll be automatically installed when executing setup.py, or alternatively when installing gem with pip.

__"Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
There are 4 examples but they need to be improved to be understandable by a new user."__
The [example scripts](https://github.com/upb-lea/gym-electric-motor/tree/master/examples) are reworked completely with added documentation. Moreover, several tutorial notebooks were introduced (links in README) for a step-by-step guide to various aspects of the toolbox

__"Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
Looks like there are tests, but haven't been able to run due to installation issue."__
All installation issues should be resolved now. Alternatively, [Travis-CI](https://travis-ci.org/upb-lea/gym-electric-motor) builds the repo by running all tests with [coverage analysis](https://codecov.io/gh/upb-lea/gym-electric-motor).

__"Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
Needs some improvement for non-specialist audience. Define or reduce jargon, for example.
State of the field: Do the authors describe how this software compares to other commonly-used packages?
Little is explained about this besides saying that Matlab does something similar. More detail is needed here.
Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
There are some improvements in grammar needed.
References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
This seems brief. Maybe you'll have more references to actual electric motor software once you improve your state of the field section."__

This was more elaborated on in the [revised JOSS paper](https://github.com/upb-lea/gym-electric-motor/blob/master/docs/JOSS/paper.md).
