+++ docs/parts/visualizations/motor_dashboard_plots.rst
@@ -1,0 +1,35 @@
+Motor Dashboard Plots
+=====================
+
+Motor Dashboard Plot Base Class
+-------------------------------
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.MotorDashboardPlot
+   :members:
+   :inherited-members:
+
+Step Plot Base Class
+--------------------
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.StepPlot
+   :members:
+   :inherited-members:
+
+Available Step Plots
+--------------------
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.StatePlot
+   :members:
+   :inherited-members:
+
+Episodic Plot Base Class
+........................
+
+Available  Episodic Plots
+........................
+
+Interval Plot Base Class
+........................
+
+Available Interval Plots
+........................
