+++ docs/parts/visualizations/motor_dashboard.rst
@@ -1,5 +1,12 @@
 Motor Dashboard
 ###############################
+The
+
+..  toctree::
+    :maxdepth: 2
+    :caption: Available Plots:
+
+    motor_dashboard_plots
 
 .. autoclass:: gym_electric_motor.visualization.motor_dashboard.MotorDashboard
    :members:
