### 2020-10-28 51f85c4
Removal of pytest-lambda
