### 2020-10-29 b012996
Fixed function naming
