+++ tests/test_physical_systems/test_mechanical_loads.py
@@ -24,7 +24,7 @@ test_bias = 10
 test_freq = 2
 
 
-def speed_profile(t, amp, freq, bias):
+def speed_profile_(t, amp, freq, bias):
     return amp*signal.sawtooth(2*np.pi*freq*t, width=0.5)+bias
 
 
@@ -249,21 +249,21 @@ class TestExtSpeedLoad(TestMechanicalLoad):
 
     key = 'ExtSpeedLoad'
     class_to_test = ExternalSpeedLoad
-    kwargs = dict(speed_profile=speed_profile,
+    kwargs = dict(speed_profile=speed_profile_,
                   amp=test_amp,
                   bias=test_bias,
                   freq=test_freq)
 
     @pytest.fixture
     def ext_speed_load(self):
-        return ExternalSpeedLoad(speed_profile=speed_profile,
+        return ExternalSpeedLoad(speed_profile=speed_profile_,
                                  amp=test_amp, bias=test_bias, freq=test_freq)
 
     def test_initialization(self):
-        load = ExternalSpeedLoad(speed_profile=speed_profile,
+        load = ExternalSpeedLoad(speed_profile=speed_profile_,
                                  amp=test_amp, bias=test_bias, freq=test_freq)
-        assert load._speed_profile == speed_profile
-        assert load.omega == speed_profile(t=0,
+        assert load._speed_profile == speed_profile_
+        assert load.omega == speed_profile_(t=0,
                                                 amp=test_amp,
                                                 bias=test_bias,
                                                 freq=test_freq)
@@ -299,7 +299,7 @@ class TestExtSpeedLoad(TestMechanicalLoad):
     ]
     )
     def test_jacobian(self, omega, omega_initial, expected):
-        test_object = self.class_to_test(speed_profile, amp=test_amp,
+        test_object = self.class_to_test(speed_profile_, amp=test_amp,
                                          bias=test_bias, freq=test_freq)
 
         # 2 Runs to test independence on time and torque
