### 2020-10-29 ab27e27
Moved functionality into the base plotting class
