+++ gym_electric_motor/visualization/motor_dashboard_plots.py
@@ -22,8 +22,12 @@ class MotorDashboardPlot(Callback):
 
         self._line_cfg = self._default_line_cfg.copy()
         self._line_cfg.update(line_config)
-        self._lines = None
-        self._labels = None
+        self._lines = []
+        self._label = ''
+        self._t_data = []
+        self._y_data = []
+        self._x_lim = None
+        self._y_lim = None
         self._colors = [cycle['color'] for cycle in plt.rcParams['axes.prop_cycle']]
 
     def initialize(self, axis):
@@ -34,9 +38,29 @@ class MotorDashboardPlot(Callback):
         """
         self._axis = axis
         self._axis.grid(True)
+        self._axis.set_ylabel(self._label)
+        if self._y_lim is None and self._x_lim is None:
+            self._axis.autoscale(True, axis='both')
+        elif self._x_lim is None:
+            self._axis.autoscale(True, axis='X')
+            self._axis.set_ylim(self._y_lim)
+        elif self._y_lim is None:
+            self._axis.autoscale(True, axis='Y')
+            self._axis.set_xlim(self._x_lim)
+        else:
+            self._axis.autoscale(False)
 
     def render(self):
-        raise NotImplementedError
+        for line, data in zip(self._lines, self._y_data):
+            line.set_data(self._t_data, data)
+        self._scale_x_axis()
+        self._scale_y_axis()
+
+    def _scale_x_axis(self):
+        pass
+
+    def _scale_y_axis(self):
+        pass
 
 
 class StepPlot(MotorDashboardPlot):
@@ -53,6 +77,11 @@ class StepPlot(MotorDashboardPlot):
         'linestyle': '-'
     }
 
+    @property
+    def data_idx(self):
+        """Returns the current index to access the time and data arrays."""
+        return self._k % self._x_width
+
     def __init__(self, violation_line_cfg=None, reset_line_cfg=None, **kwargs):
         super().__init__(**kwargs)
         reset_line_cfg = reset_line_cfg or {}
@@ -68,12 +97,9 @@ class StepPlot(MotorDashboardPlot):
         self._violation_line_cfg.update(violation_line_cfg)
 
         self._t = 0
-        self._t_data = []
-        self._y_data = []
         self._tau = None
         self._done = None
         self._x_width = 10000
-        self.y_lim = (-np.inf, np.inf)
         self._k = 0
 
     def set_width(self, width):
@@ -90,27 +116,24 @@ class StepPlot(MotorDashboardPlot):
         self._t_data = np.linspace(0, self._x_width * self._tau, self._x_width)
         self._y_data = np.nan * np.ones_like(self._t_data)
 
-    def render(self):
-        """Called by the MotorDashboard each time before the figure is updated."""
-        # configure x-axis properties
+    def on_reset_begin(self):
+        if self._done is not None:
+            if self._done:
+                self._axis.axvline(self._t, **self._violation_line_cfg)
+            else:
+                self._axis.axvline(self._t, **self._reset_line_cfg)
+        self._done = False
+
+    def on_step_end(self, k, state, reference, reward, done):
+        self._k += 1
+        self._t += self._tau
+        self._done = done
+
+    def _scale_x_axis(self):
         x_lim = self._axis.get_xlim()
         upper_lim = max(self._t, x_lim[1])
         lower_lim = upper_lim - self._x_width * self._tau
         self._axis.set_xlim(lower_lim, upper_lim)
-
-    def initialize(self, axis):
-        super().initialize(axis)
-        self._axis.set_xlim(0, self._x_width * self._tau)
-        if self.y_lim == (-np.inf, np.inf):
-            self._axis.autoscale(True, axis='Y')
-        else:
-            min_limit, max_limit = self.y_lim
-            spacing = 0.1 * (max_limit - min_limit)
[…]
