### 2020-11-02 #2498 @wkirgsn
Hi @Kevin-Mattheus-Moerman,
we revised the toolbox and the JOSS paper candidate according to the reviewers' remarks.
Can you please advice what the next steps are, or which obstacles need to be further overcome in order to proceed to an acceptance and publication?
We appreciate your and the reviewers' time and effort as always.
