+++ docs/parts/visualizations/visualization.rst
@@ -12,5 +12,5 @@ Visualization
 Visualization Base Class
 ''''''''''''''''''''''''''''''
 
-.. autoclass:: gym_electric_motor.core.ElectricMotorVisualization
-   :members:
+..  autoclass:: gym_electric_motor.core.ElectricMotorVisualization
+    :members:

[…]
