+++ docs/parts/visualizations/motor_dashboard_plots/action_plot.rst
@@ -1,0 +1,6 @@
+Action Plot
+###########
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.ActionPlot
+   :members:
+   :inherited-members:
