+++ docs/parts/visualizations/motor_dashboard_plots/state_plot.rst
@@ -1,0 +1,6 @@
+State Plot
+###########
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.StatePlot
+   :members:
+   :inherited-members:
