+++ docs/parts/visualizations/motor_dashboard_plots/time_plot.rst
@@ -1,0 +1,7 @@
+Time Plot (Abstract)
+#####################
+TheTimePLot is athe base class for all plots that plot data with the time on the x-axis.
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.TimePlot
+   :members:
+   :inherited-members:
