### 2020-11-07 7db7cd4
Finished Docu and fixed examples.
