+++ gym_electric_motor/visualization/motor_dashboard_plots/cumulative_constraint_violation_plot.py
@@ -2,6 +2,7 @@ from .base_plots import StepPlot
 
 
 class CumulativeConstraintViolationPlot(StepPlot):
+    """Plots the cumulative number of constraint violations during the runtime of the env. into an axis."""
 
     def __init__(self):
         super().__init__()
@@ -16,14 +17,16 @@ class CumulativeConstraintViolationPlot(StepPlot):
         self._lines.append(self._axis.plot(self._x_data, self._violations)[0])
 
     def on_step_end(self, k, state, reference, reward, done):
+        super().on_step_end(k, state, reference, reward, done)
         if done:
+            # Add another point for a step-like plot
+            self._x_data.append(self._k - 1)
+            self._violations.append(self._no_of_violations)
             self._no_of_violations += 1
-
-    def _set_y_data(self):
-        self._violations.append(self._no_of_violations)
-
-    def _scale_x_axis(self):
-        self._axis.set_xlim(-1, self._x_data[-1] + 1)
+            self._x_data.append(self._k)
+            self._violations.append(self._no_of_violations)
 
     def _scale_y_axis(self):
-        self._axis.set_ylim(-1, self._no_of_violations + 1)
+        # Read the limit before writing, because reading is fast, but writing is slow
+        if self._axis.get_xlim() != (-1, self._no_of_violations + 1):
+            self._axis.set_ylim(-1, self._no_of_violations + 1)
