+++ examples/reinforcement_learning_controllers/ddpg_pmsm_dq_current_control.py
@@ -30,6 +30,7 @@ We use a deep-deterministic-policy-gradient (DDPG) agent
 to determine which action must be taken on a continuous-control-set
 '''
 
+
 class AppendLastActionWrapper(Wrapper):
     """
     The following environment considers the dead time in the real-world motor control systems.
