+++ gym_electric_motor/visualization/motor_dashboard_plots/__init__.py
@@ -1,5 +1,5 @@
 from .action_plot import ActionPlot
-from .base_plots import MotorDashboardPlot, TimePlot, EpisodicPlot, StepPlot
+from .base_plots import MotorDashboardPlot, TimePlot, EpisodePlot, StepPlot
 from .cumulative_constraint_violation_plot import CumulativeConstraintViolationPlot
 from .episode_length_plot import EpisodeLengthPlot
 from .mean_episode_reward_plot import MeanEpisodeRewardPlot
