+++ gym_electric_motor/reward_functions/weighted_sum_of_errors.py
@@ -6,46 +6,82 @@ import warnings
 
 
 class WeightedSumOfErrors(RewardFunction):
-    """
-    Reward Function that calculates the reward as the weighted sum of errors with a certain power.
+    """A reward function that calculates the reward as the weighted sum of errors with a certain power.
 
     .. math::
-        reward = - reward\_weights * (abs(state - reference)/ state\_length)^{reward\_power}
+        r_{wse} = - \sum_i w_i ((|s_i-s^*_i|) / l_i) ^{n_i} + b
 
-    The reward can be normalized to the range [0, -1].
+    Notation:
+        - :math:`r_{wse}`: Weighted sum of error reward
+        - :math:`w_{i}`: Reward weight of state i
+        - :math:`s_{i}`: State value of state i
+        - :math:`s^*_{i}`: Reference value of state i
+        - :math:`l_{i}`: State length of state i
+        - :math:`n_{i}`: Reward power of state i
+        - :math:`b`: Bias
 
-    If states are observed for limit violation (current observation is defualt setting), an additional terminal reward
-    is added if limits are violated. This value depends on the factor gamma as follows
+    | :math:`l_i = 1` for states with positive values only.
+    | :math:`l_i = 2` for states with positive and negative values.
+
+    If environments constraints are violated to a certain degree, a special violation reward is returned as follows:
 
     .. math::
-        limit\_violation\_reward = -1 / (1 - \gamma).
+        r_{total} = (1.0 - d_{violation})  r_{wse} + d_{violation} r_{violation}
 
-    | state_length[i] = 1 for states with positive values only.
-    | state_length[i] = 2 for states with positive and negative values.
+    Notation:
+        - :math:`r_{total}`: Total reward
+        - :math:`r_{wse}`: Weighted sum of error reward
+        - :math:`r_{violation}`: Constraint violation reward
+        - :math:`d_{violation}`: Limit violation degree :math:`d_{violation} \in [0,1]`
+
+
+    The violation reward can be chosen freely by the user and shall punish the agents to comply with the constraints.
+    Per default, the violation reward is selected so that it is always the worst expected reward the agent could get.
+
+    .. math::
+        r_{violation} = r_{wse,min} / (1 - \gamma)
+
+    :math:`r_{wse,min}` is the minimal :math:`r_{wse}` (=reward_range[0]) and :math:`\gamma` the agents discount factor.
     """
 
-    def __init__(self, reward_weights=None, normed_reward_weights=False,
-                 gamma=0.9, reward_power=1, **kwargs):
+    def __init__(self, reward_weights=None, normed_reward_weights=False, violation_reward=None,
+                 gamma=0.9, reward_power=1, bias=0.0, **__):
         """
         Args:
             reward_weights(dict/list/ndarray(float)): Dict mapping state names to reward_weights, 0 otherwise.
                 Or an array with the reward_weights on the position of the state_names.
-            normed_reward_weights(bool): If True, the reward weights will be normalized to 1.
-            gamma(float): Discount factor for the reward punishment. Should equal agents' discount factor gamma.
+
+            normed_reward_weights(bool): If True, the reward weights will be normalized to sum up to 1.
+
+            violation_reward(None/float): The punishment reward if constraints have been violated.
+
+                - None(default): Per default, the violation reward is calculated as described above.
+                - float: This value is taken as limit violation reward.
+
+            gamma(float in [0.0, 1.0]): Discount factor for the reward punishment.
+                Should equal agents' discount factor gamma. Used only, if violation_reward=None.
+
             reward_power(dict/list(float)/float): Reward power for each of the systems states.
-            kwargs: Further arguments to pass to the superclass RewardFunction
+
+            bias(float/'positive'): Additional bias that is added to the reward.
+
[…]
