+++ gym_electric_motor/core.py
@@ -445,57 +445,32 @@ class RewardFunction:
     The abstract base class for reward functions in gym electric motor environments.
 
     The reward function is called once per step and returns reward for the current time step.
-    Furthermore, the reward function includes the limit observer. If limits have been violated and the episode ended
-    a different reward for violating the limits is returned.
 
-    reward_range:
-        Tuple(float, float): Defining lowest and highest possible rewards for non limit-violating steps.
+    Attributes:
+        reward_range(Tuple(float, float)):Defining lowest and highest possible rewards.
     """
 
-    #: Tuple(int,int): Lower and upper possible reward (excluding limit violation reward)
+    #: Tuple(int,int): Lower and upper possible reward
     reward_range = (-np.inf, np.inf)
 
-    def __call__(self, state, reference, k):
+    def __call__(self, state, reference, k, action, violation_degree):
         """
         Call of the reward calculation.
 
         Args:
-            state: State array of the environment.
-            reference: Reference array of the environment.
-            k: Systems momentary time-step
+            state(numpy.ndarray(float)): State array of the environment.
+            reference(numpy.ndarray(float)): Reference array of the environment.
+            k(int): Systems momentary time-step
+            action(element of action-space): The taken action :a_{k-1}: at the beginning of the step.
+            violation_degree(float in [0.0, 1.0]): Degree of violation of the constraints. 0.0 indicates that all
+                constraints are complied. 1.0 indicates that the constraints have been so much violated, that a reset is
+                necessary.
 
 
         Returns:
             float: The reward for the state, reference pair
         """
-        return self.reward(state, reference, k)
-
-    def _limit_violation_reward(self, state):
-        """
-        Called, when limits have been violated to return a special reward for this case.
-
-        Args:
-            state(ndarray(float)): Current state array of the environment.
-
-        Returns:
-            float: The limit violation reward.
-        """
-        raise NotImplementedError
-
-    def _standard_reward(self, state, reference, action):
-        """
-        Standard reward function. Called, when no limits have been violated to calculate the reward based on the state
-        and the reference.
-
-        Args:
-            state(ndarray(float)): State array of the environment.
-            reference(ndarray(float): Reference array of the environment.
-            action(element of action space): The previously taken action.
-
-        Returns:
-            float: Standard reward.
-        """
-        raise NotImplementedError
+        return self.reward(state, reference, k, action, violation_degree)
 
     def set_modules(self, physical_system, reference_generator):
         """
@@ -521,15 +496,14 @@ class RewardFunction:
                 necessary.
 
         Returns:
-            float: Reward for this state, reference pair.
+            float: Reward for this state, reference, action tuple.
         """
 
-        return (1.0 - violation_degree) * self._standard_reward(state, reference, action) \
-            + violation_degree * self._limit_violation_reward(state)
+        raise NotImplementedError
 
     def reset(self, initial_state=None, initial_reference=None):
-        """
-        This function is called by the environment when reset.
+        """This function is called by the environment when reset.
+
         Inner states of the reward function can be reset here, if necessary.
 
         Args:
@@ -539,9 +513,7 @@ class RewardFunction:
         pass
 
     def close(self):
-        """
-        Called, when the environment is closed to store logs, close files etc.
-        """
+        """Called, when the environment is closed to store logs, close files etc."""
         pass
 
 
@@ -670,6 +642,9 @@ class Callback:
         _env: The GEM environment. Use it to have full control over the environment on runtime. 
     """
 
+    def __init__(self):
+        self._env = None
+
     def set_env(self, env):
         """Sets the environment of the motor."""
         self._env = env
@@ -696,14 +671,52 @@ class Callback:
 
 
 class ConstraintMonitor:
[…]
