### 2020-11-11 3dc7ea5
Changed most examples and default environments to the new constraintmonitor
