+++ gym_electric_motor/envs/gym_dcm/dc_permex_motor_env.py
@@ -6,7 +6,7 @@ from gym_electric_motor.reward_functions import WeightedSumOfErrors
 
 class DcPermanentlyExcitedMotorEnvironment(ElectricMotorEnvironment):
 
-    def __init__(self, motor='DcPermEx', reward_function=None, reference_generator=None, **kwargs):
+    def __init__(self, motor='DcPermEx', reward_function=None, reference_generator=None, constraints=('i',), **kwargs):
         """
         Args:
             motor(ElectricMotor): Electric Motor used in the PhysicalSystem
@@ -18,7 +18,8 @@ class DcPermanentlyExcitedMotorEnvironment(ElectricMotorEnvironment):
         reference_generator = reference_generator or WienerProcessReferenceGenerator(**kwargs)
         reward_function = reward_function or WeightedSumOfErrors(**kwargs)
         super().__init__(
-            physical_system, reference_generator=reference_generator, reward_function=reward_function, **kwargs
+            physical_system, reference_generator=reference_generator, reward_function=reward_function,
+            constraints=constraints, **kwargs
         )
 
 
