### 2020-11-13 472a058
Finalized the Constraint Monitor. Still missing tests.
