+++ gym_electric_motor/constraints.py
@@ -4,17 +4,50 @@ from .utils import set_state_array
 
 
 class Constraint:
+    """Base class for all constraints in the ConstraintMonitor."""
 
     def __call__(self, state):
+        """Function that is called to check the constraint.
+
+        Args:
+            state(numpy.ndarray(float)): The current physical systems state.
+
+        Returns:
+              float in [0.0, 1.0]: Degree how much the constraint has been violated.
+                0.0: No violation
+                (0.0, 1.0): Undesired zone near to a full violation. No episode termination.
+                1.0: Full violation and episode termination.
+        """
         raise NotImplementedError
 
     def set_modules(self, ps):
+        """Called by the environment that the Constraint can read information from the PhysicalSystem.
+
+        Args:
+            ps(PhysicalSystem): PhysicalSystem of the environment.
+        """
         pass
 
 
 class LimitConstraint(Constraint):
+    """Constraint to observe the limits on one or more system states.
+
+    This constraint observes if any of the systems state values exceeds the limit specified in the PhysicalSystem.
+
+    .. math::
+        1.0 <= s_i / s_{i,max}
+
+    For all :math:`i` in the set of PhysicalSystems states :math:`S`.
+
+    """
 
     def __init__(self, observed_state_names='all'):
+        """
+        Args:
+            observed_state_names('all'/iterable(str)): The states to observe. \n
+                - 'all': Shortcut for observing all states.
+                - iterable(str): Pass an iterable containing all state names of the states to observe.
+        """
         self._observed_state_names = observed_state_names
         self._limits = None
         self._observed_states = None
@@ -34,8 +67,20 @@ class LimitConstraint(Constraint):
 
 
 class SquaredConstraint(Constraint):
+    """A squared constraint on multiple states as it is required oftentimes for the dq-currents in synchronous and
+    asynchronous electric motors.
+
+    .. math::
+        1.0 <= \sum_{i \in S} (s_i / s_{i,max})^2
+
+    :math:`S`: Set of the observed PhysicalSystems states
+    """
 
     def __init__(self, states=()):
+        """
+        Args:
+            states(iterable(str)): Names of all states to be observed within the SquaredConstraint.
+        """
         self._states = states
         self._state_indices = ()
         self._limits = ()
