+++ docs/parts/constraint_monitor.rst
@@ -1,7 +1,16 @@
 Constraint Monitor
 #####################
 
+..  toctree::
+    :maxdepth: 1
+    :caption: Available Constraints:
 
-.. autoclass:: gym_electric_motor.constraint_monitor.ConstraintMonitor
+    constraints/constraint
+    constraints/limit_constraint
+    constraints/squared_constraint
+
+
+
+.. autoclass:: gym_electric_motor.core.ConstraintMonitor
     :members:
     :inherited-members:
