+++ docs/parts/constraints/constraint.rst
@@ -1,0 +1,12 @@
+Constraint Base Class
+#####################
+
+How To: Define Your Own Constraints
+________________________________
+
+
+Constraint API Documentation
+____________________________
+
+.. autoclass:: gym_electric_motor.core.RewardFunction
+   :members:
