### 2020-11-13 1ab64cb
Finalized the Constraint Monitor. Still missing tests.
