+++ docs/parts/constraint_monitor.rst
@@ -9,8 +9,13 @@ Constraint Monitor
     constraints/limit_constraint
     constraints/squared_constraint
 
+Usage Guide
+___________
 
+ToDo
 
+Constraint Monitor API Documentation
+____________________________________
 .. autoclass:: gym_electric_motor.core.ConstraintMonitor
     :members:
     :inherited-members:
