### 2020-11-16 dfee9f3
Update of final remarks
