+++ docs/parts/visualizations/motor_dashboard_plots/episode_plots.rst
@@ -1,6 +1,6 @@
 Episode Plot (Abstract)
 #####################
-The EpisodePLot is  the base class for all plots that plot data with number of episodes on the x-axis.
+The EpisodePlot is the base class for all plots that plot data with number of episodes on the x-axis.
 
 .. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.EpisodePlot
    :members:
