### 2020-11-16 801dece
Update of latest PR comments
