### 2020-11-17 #2498 @moorepants
I have given the repo and paper a second pass. There has been significant improvements. All of my comments have been addressed and the package is looking really nice now. Congrats on the great work!

I am trouble getting the Colab notebooks to run and give output:

https://colab.research.google.com/github/upb-lea/gym-electric-motor/blob/master/examples/environment_features/GEM_cookbook.ipynb#scrollTo=vwvhU6IPsKQw

Last cell gives me:

![image](https://user-images.githubusercontent.com/276007/99377609-b2015700-28c6-11eb-9bbd-e7dc7458d64f.png)

https://colab.research.google.com/github/upb-lea/gym-electric-motor/blob/master/examples/model_predictive_controllers/gekko_mpc_cont_pmsm_example.ipynb#scrollTo=lBEjIJtSsEHf

give me:

![image](https://user-images.githubusercontent.com/276007/99377659-c0e80980-28c6-11eb-9362-e9842b262cb5.png)

https://colab.research.google.com/github/upb-lea/gym-electric-motor/blob/master/examples/reinforcement_learning_controllers/tensorforce_dqn_disc_pmsm_example.ipynb

gives:

![image](https://user-images.githubusercontent.com/276007/99377695-ccd3cb80-28c6-11eb-8080-e0db23ef4df6.png)

So there are some new errors.

I also still am having trouble installing locally (Ubuntu 20.04):
moorepants@nandi:~$ conda create -n gem python=3
Collecting package metadata (current_repodata.json): done
Solving environment: done

## Package Plan ##

  environment location: /home/moorepants/miniconda/envs/gem

  added / updated specs:
    - python=3

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    _libgcc_mutex-0.1          |      conda_forge           3 KB  conda-forge
    _openmp_mutex-4.5          |            1_gnu          22 KB  conda-forge
    certifi-2020.11.8          |   py39hf3d152e_0         150 KB  conda-forge
    python-3.9.0               |h2a148a8_4_cpython        24.0 MB  conda-forge
    python_abi-3.9             |           1_cp39           4 KB  conda-forge
    setuptools-49.6.0          |   py39h079e4ff_2         953 KB  conda-forge
    tzdata-2020d               |       h516909a_0         120 KB  conda-forge
    ------------------------------------------------------------
                                           Total:        25.3 MB

The following NEW packages will be INSTALLED:

  _libgcc_mutex      conda-forge/linux-64::_libgcc_mutex-0.1-conda_forge
  _openmp_mutex      conda-forge/linux-64::_openmp_mutex-4.5-1_gnu
  ca-certificates    conda-forge/linux-64::ca-certificates-2020.11.8-ha878542_0
  certifi            conda-forge/linux-64::certifi-2020.11.8-py39hf3d152e_0
  ld_impl_linux-64   conda-forge/linux-64::ld_impl_linux-64-2.35.1-hed1e6ac_0
  libffi             conda-forge/linux-64::libffi-3.2.1-he1b5a44_1007
  libgcc-ng          conda-forge/linux-64::libgcc-ng-9.3.0-h5dbcf3e_17
  libgomp            conda-forge/linux-64::libgomp-9.3.0-h5dbcf3e_17
  libstdcxx-ng       conda-forge/linux-64::libstdcxx-ng-9.3.0-h2ae2ef3_17
  ncurses            conda-forge/linux-64::ncurses-6.2-h58526e2_4
  openssl            conda-forge/linux-64::openssl-1.1.1h-h516909a_0
  pip                conda-forge/noarch::pip-20.2.4-py_0
  python             conda-forge/linux-64::python-3.9.0-h2a148a8_4_cpython
  python_abi         conda-forge/linux-64::python_abi-3.9-1_cp39
  readline           conda-forge/linux-64::readline-8.0-he28a2e2_2
  setuptools         conda-forge/linux-64::setuptools-49.6.0-py39h079e4ff_2
  sqlite             conda-forge/linux-64::sqlite-3.33.0-h4cf870e_1
  tk                 conda-forge/linux-64::tk-8.6.10-hed695b0_1
  tzdata             conda-forge/noarch::tzdata-2020d-h516909a_0
  wheel              conda-forge/noarch::wheel-0.35.1-pyh9f0ad1d_0
  xz                 conda-forge/linux-64::xz-5.2.5-h516909a_1
  zlib               conda-forge/linux-64::zlib-1.2.11-h516909a_1010

Downloading and Extracting Packages
setuptools-49.6.0    | 953 KB    | ################################################################################################# | 100%
tzdata-2020d         | 120 KB    | ################################################################################################# | 100%
python_abi-3.9       | 4 KB      | ################################################################################################# | 100%
certifi-2020.11.8    | 150 KB    | ################################################################################################# | 100%
_openmp_mutex-4.5    | 22 KB     | ################################################################################################# | 100%
python-3.9.0         | 24.0 MB   | ################################################################################################# | 100%
_libgcc_mutex-0.1    | 3 KB      | ################################################################################################# | 100%
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate gem
#
# To deactivate an active environment, use
#
#     $ conda deactivate

moorepants@nandi:~$ act gem
(gem) moorepants@nandi:~$ which pip
/home/moorepants/miniconda/envs/gem/bin/pip
(gem) moorepants@nandi:~$ pip install gym-electric-motor
Collecting gym-electric-motor
  Downloading gym_electric_motor-0.3.0-py3-none-any.whl (151 kB)
     |████████████████████████████████| 151 kB 2.9 MB/s
Collecting numpy>=1.16.4
  Downloading numpy-1.19.4-cp39-cp39-manylinux2010_x86_64.whl (14.5 MB)
     |████████████████████████████████| 14.5 MB 5.3 MB/s
Collecting gym>=0.15.4
  Downloading gym-0.17.3.tar.gz (1.6 MB)
     |████████████████████████████████| 1.6 MB 8.4 MB/s
Collecting scipy>=1.4.1
  Downloading scipy-1.5.4-cp39-cp39-manylinux1_x86_64.whl (25.8 MB)
     |████████████████████████████████| 25.8 MB 1.5 MB/s
Collecting pytest-cov
  Downloading pytest_cov-2.10.1-py2.py3-none-any.whl (19 kB)
Collecting pytest-lambda
  Downloading pytest_lambda-1.2.3-py3-none-any.whl (10 kB)
Collecting matplotlib==3.1.2
  Downloading matplotlib-3.1.2.tar.gz (40.9 MB)
     |████████████████████████████████| 40.9 MB 7.8 MB/s
Collecting pytest>=5.2.2
  Downloading pytest-6.1.2-py3-none-any.whl (272 kB)
     |████████████████████████████████| 272 kB 4.0 MB/s
Collecting pyglet<=1.5.0,>=1.4.0
  Downloading pyglet-1.5.0-py2.py3-none-any.whl (1.0 MB)
     |████████████████████████████████| 1.0 MB 9.4 MB/s
Collecting cloudpickle<1.7.0,>=1.2.0
  Downloading cloudpickle-1.6.0-py3-none-any.whl (23 kB)
Collecting coverage>=4.4
  Downloading coverage-5.3-cp39-cp39-manylinux1_x86_64.whl (228 kB)
     |████████████████████████████████| 228 kB 10.2 MB/s
Collecting wrapt<2.0.0,>=1.11.0
  Downloading wrapt-1.12.1.tar.gz (27 kB)
Collecting cycler>=0.10
  Downloading cycler-0.10.0-py2.py3-none-any.whl (6.5 kB)
Collecting kiwisolver>=1.0.1
  Downloading kiwisolver-1.3.1-cp39-cp39-manylinux1_x86_64.whl (1.2 MB)
     |████████████████████████████████| 1.2 MB 10.1 MB/s
Collecting pyparsing!=2.0.4,!=2.1.2,!=2.1.6,>=2.0.1
  Downloading pyparsing-2.4.7-py2.py3-none-any.whl (67 kB)
     |████████████████████████████████| 67 kB 8.9 MB/s
Collecting python-dateutil>=2.1
  Downloading python_dateutil-2.8.1-py2.py3-none-any.whl (227 kB)
     |████████████████████████████████| 227 kB 9.4 MB/s
Collecting packaging
  Downloading packaging-20.4-py2.py3-none-any.whl (37 kB)
Collecting pluggy<1.0,>=0.12
  Downloading pluggy-0.13.1-py2.py3-none-any.whl (18 kB)
Collecting attrs>=17.4.0
  Downloading attrs-20.3.0-py2.py3-none-any.whl (49 kB)
     |████████████████████████████████| 49 kB 12.4 MB/s
Collecting py>=1.8.2
  Downloading py-1.9.0-py2.py3-none-any.whl (99 kB)
     |████████████████████████████████| 99 kB 9.0 MB/s
Collecting toml
  Downloading toml-0.10.2-py2.py3-none-any.whl (16 kB)
Collecting iniconfig
  Downloading iniconfig-1.1.1-py2.py3-none-any.whl (5.0 kB)
Collecting future
  Downloading future-0.18.2.tar.gz (829 kB)
     |████████████████████████████████| 829 kB 9.2 MB/s
Collecting six
  Downloading six-1.15.0-py2.py3-none-any.whl (10 kB)
Building wheels for collected packages: gym, matplotlib, wrapt, future
  Building wheel for gym (setup.py) ... done
  Created wheel for gym: filename=gym-0.17.3-py3-none-any.whl size=1654652 sha256=a4ce11df7bbc54e5abc0b2f84bdbce29e77b4359bc288a25ad9cea1bbb0c4ee2
  Stored in directory: /home/moorepants/.cache/pip/wheels/5d/06/a4/57a926b2e87a5e1f21551577750549206dde639e19a5ac72d5
  Building wheel for matplotlib (setup.py) ... error
   command: /home/moorepants/miniconda/envs/gem/bin/python3.9 -u -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/tmp/pip-install-2zfv9tkf/matplotlib/setup.py'"'"'; __file__='"'"'/tmp/pip-install-2zfv9tkf/matplotlib/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-i_o21moi
       cwd: /tmp/pip-install-2zfv9tkf/matplotlib/
  Complete output (506 lines):
  ================================================================================
  Edit setup.cfg to change the build options

  BUILDING MATPLOTLIB
    matplotlib: yes [3.1.2]
        python: yes [3.9.0 | packaged by conda-forge | (default, Oct 14 2020,
                    22:59:50)  [GCC 7.5.0]]
      platform: yes [linux]

  OPTIONAL SUBPACKAGES
   sample_data: yes [installing]
         tests: no  [skipping due to configuration]

  OPTIONAL BACKEND EXTENSIONS
           agg: yes [installing]
         tkagg: yes [installing; run-time loading from Python Tcl/Tk]
        macosx: no  [Mac OS-X only]

  OPTIONAL PACKAGE DATA
          dlls: no  [skipping due to configuration]

  running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.9
  copying lib/pylab.py -> build/lib.linux-x86_64-3.9
  creating build/lib.linux-x86_64-3.9/mpl_toolkits
  copying lib/mpl_toolkits/__init__.py -> build/lib.linux-x86_64-3.9/mpl_toolkits
  creating build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/path.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/bezier.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_animation_data.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/textpath.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/backend_tools.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_color_data.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/patheffects.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/docstring.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/markers.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/__init__.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/legend_handler.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/artist.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/dates.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/axis.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/collections.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/figure.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/animation.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/dviread.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/text.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/category.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/image.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/gridspec.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_mathtext_data.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/units.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/container.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/afm.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/patches.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/tight_bbox.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/legend.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_layoutbox.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/lines.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/hatch.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/font_manager.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/backend_bases.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/offsetbox.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/contour.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_cm.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/quiver.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/mlab.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_version.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/texmanager.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/type1font.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/rcsetup.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/colorbar.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/backend_managers.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_pylab_helpers.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_cm_listed.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/_constrained_layout.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/mathtext.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/colors.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/tight_layout.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/fontconfig_pattern.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/cm.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/stackplot.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/sankey.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/streamplot.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/scale.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/ticker.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/pyplot.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/pylab.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/blocking_input.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/transforms.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/widgets.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/table.py -> build/lib.linux-x86_64-3.9/matplotlib
  copying lib/matplotlib/spines.py -> build/lib.linux-x86_64-3.9/matplotlib
  creating build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/grid_helper_curvelinear.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/axisline_style.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/__init__.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/clip_path.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/angle_helper.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/axislines.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/axes_grid.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/axes_divider.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/parasite_axes.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/axes_rgb.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/floating_axes.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/grid_finder.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  copying lib/mpl_toolkits/axisartist/axis_artist.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axisartist
  creating build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/anchored_artists.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/__init__.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/axes_size.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/axes_grid.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/colorbar.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/axes_divider.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/parasite_axes.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/inset_locator.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/axes_rgb.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  copying lib/mpl_toolkits/axes_grid1/mpl_axes.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid1
  creating build/lib.linux-x86_64-3.9/mpl_toolkits/mplot3d
  copying lib/mpl_toolkits/mplot3d/proj3d.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/mplot3d
  copying lib/mpl_toolkits/mplot3d/__init__.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/mplot3d
  copying lib/mpl_toolkits/mplot3d/art3d.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/mplot3d
  copying lib/mpl_toolkits/mplot3d/axis3d.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/mplot3d
  copying lib/mpl_toolkits/mplot3d/axes3d.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/mplot3d
  creating build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/grid_helper_curvelinear.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/axisline_style.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/anchored_artists.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/__init__.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/clip_path.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/angle_helper.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/axes_size.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/axislines.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/axes_grid.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/colorbar.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/axes_divider.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/parasite_axes.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/inset_locator.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/axes_rgb.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
  copying lib/mpl_toolkits/axes_grid/floating_axes.py -> build/lib.linux-x86_64-3.9/mpl_toolkits/axes_grid
[…]
