### 2020-11-17 #2498 @wkirgsn
Thank you @moorepants for your second review!
You raised very good points of which we weren't aware of.
After analyzing the problems we found the root cause of them all to be rather peripheral.

* Colab notebooks not showing plots:
  * This seems to be a [known issue in colab](https://github.com/matplotlib/matplotlib/issues/15161), and is connected to the fact that we have consecutively updating plots rather than one-shot plots. This is unfortunately not fixable short-term, but isn't an issue for locally run notebooks anyway. We understand this diminishes the tutorial experience, but should be updated on colab site in our opinion.
* gekko_mpc_tutorial not finding a soution:
  * We were aware of this one, and it is already fixed in the master (not on pip/pypi yet, will come soon).
* tensorforce tutorial showing errors during installation
  * These are minor discrepancies that display as error but can be understood as mere warnings actually. They are not preventing the toolbox/code/tutorial from working or mitigating the functionality in any sense (only minor niche third-party packages are showing up which aren't used anyway).
* Installation error with pip:
  * You tried the installation with a fresh environment that was setup on python 3.9, which is just a month or so old, and is not supported by many major packages such as matplotlib yet. Since our package relies on matplotlib as well, this leads to the shown error. Since the discrepancy is in the required third-party package and not in GEM, we again see action to be taken there rather than on our side. Of course this is also mitigating user experience on our side, such that we will mention the missing compatibility for the very new python 3.9 version in the next release notes. If you want to make a short installation test, please use python 3.6-3.8.

So all those points are either fixed or going to fix themselves in the medium run. Thank you for your time again, and we appreciate all feedback given so far.
