### 2020-11-17 #2498 @moorepants
Ok, I installed the package successfully with Python 3.8 and tested an example.

The colab explanation is satifactory.

I approve!
