### 2020-11-18 #2498 @moorepants
Yes, I recommend accepting the paper. The only thing would for them to push the latest release to PyPi so that the notebooks work, as mentioned. I assume that would be done in the finalization and zenodo archive, etc.
