### 2020-11-22 72c3a57
Revision of RewardFunction Tests
