+++ tests/test_reward_functions/test_weighted_sum_of_errors.py
@@ -2,142 +2,119 @@ import pytest
 import numpy as np
 
 from gym_electric_motor.reward_functions.weighted_sum_of_errors import WeightedSumOfErrors
-import gym_electric_motor.reward_functions.weighted_sum_of_errors as ws
-import tests.testing_utils as tu
+from .test_reward_functions import TestRewardFunction
+from ..testing_utils import DummyPhysicalSystem, DummyReferenceGenerator, DummyConstraintMonitor
 
 
-class TestWeightedSumOfErrors:
-    """
-    This test class consists of unit tests to verify modules in the weighted_sum_of_errors.py
-    Following modules are included:
-    WeightedSumOfErrors
-        __init__
-        set_modules
+class TestWeightedSumOfErrors(TestRewardFunction):
 
-    ShiftedWeightedSumOfErrors
-        set_modules
-    """
+    class_to_test = WeightedSumOfErrors
 
-    # Parameters used for testing
-    reward_dict = dict(omega=1, torque=1, i_a=1)
-    reward_list = [1, 1, 1]
-    # Test counter to count function calls
-    _mock_set_state_array_counter = 0
+    @pytest.fixture
+    def reward_function(self):
+        rf = WeightedSumOfErrors()
+        rf.set_modules(DummyPhysicalSystem(), DummyReferenceGenerator(), DummyConstraintMonitor())
+        return rf
 
-    @pytest.mark.parametrize("reward_weights", [None, reward_dict, reward_list])
-    @pytest.mark.parametrize("normed", [False, True])
-    @pytest.mark.parametrize("power", [-1, 1, 0])
-    def test_initialization(self, reward_weights, normed, power):
-        """
-        Tests the initialization of WeightedSumOfErrors object with defined parameters
-        :param reward_weights:
-        :param normed:
-        :param power:
-        :return:
-        """
-        test_wsoe_obj = WeightedSumOfErrors(reward_weights, normed, observed_states=None, gamma=0.9, reward_power=power)
-        # test if the correct attributes are set during init
-        assert test_wsoe_obj._n == power
-        assert test_wsoe_obj._reward_weights == reward_weights
-        assert not test_wsoe_obj._state_length
-        assert test_wsoe_obj._normed == normed
+    @pytest.mark.parametrize(['ps', 'reward_power', 'n'], [
+        [DummyPhysicalSystem(state_length=2), [1, 2], np.array([1, 2])],
+        [DummyPhysicalSystem(state_length=3), 2, np.array([2, 2, 2])],
+        [DummyPhysicalSystem(state_length=1), [1], np.array([1])],
+        [DummyPhysicalSystem(state_length=2), dict(dummy_state_0=2), np.array([2, 0])]
+    ])
+    @pytest.mark.parametrize('rg', [DummyReferenceGenerator()])
+    @pytest.mark.parametrize('cm', [DummyConstraintMonitor()])
+    def test_reward_powers(self, ps, rg, cm, reward_power, n):
+        rf = self.class_to_test(reward_power=reward_power)
+        rf.set_modules(
+            ps, rg, cm)
+        assert all(rf._n == n)
 
-    @pytest.mark.parametrize("reward_weights", [None, reward_list, reward_dict])
-    @pytest.mark.parametrize("normed", [True, False])
-    @pytest.mark.parametrize("power", [-1, 1, 2, 0])
-    @pytest.mark.parametrize("mock_ref_states", [[True, True], [True, False]])
-    def test_wsoe_set_modules(self, monkeypatch, reward_weights, normed, power, mock_ref_states):
-        """
-         test the set_modules function
-        :param monkeypatch:
-        :param reward_weights:
[…]
