+++ tests/test_core.py
@@ -216,101 +216,6 @@ class TestElectricMotorEnvironment:
         assert all(env.limits == expected_result)
 
 
-class TestRewardFunction:
-
-    @staticmethod
-    def mock_standard_reward(*_):
-        return 1
-
-    @staticmethod
-    def mock_limit_violation_reward(*_):
-        return -1
-
-    @staticmethod
-    def mock_reward_function(*_):
-        return 'Reward Function called'
-
-    @pytest.mark.parametrize("observed_states", (['dummy_state_0'], ['dummy_state_0', 'dummy_state_1']))
-    def test_initialization(self, observed_states):
-        RewardFunction(observed_states)
-
-    @pytest.mark.parametrize("reward_function, physical_system, reference_generator, expected_observed_states", (
-            (RewardFunction('dummy_state_0'), DummyPhysicalSystem(), DummyReferenceGenerator(), ['dummy_state_0']),
-            (RewardFunction(['dummy_state_0', 'dummy_state_2']), DummyPhysicalSystem(3), DummyReferenceGenerator(),
-             ['dummy_state_0', 'dummy_state_2']),
-            (RewardFunction('all'), DummyPhysicalSystem(3), DummyReferenceGenerator(),
-             ['dummy_state_0', 'dummy_state_1', 'dummy_state_2']),
-            (RewardFunction('currents'), DummyPhysicalSystem(3, 'i'), DummyReferenceGenerator(),
-             ['i_0', 'i_1', 'i_2']),
-            (RewardFunction(['voltages']), DummyPhysicalSystem(3, 'u'), DummyReferenceGenerator(),
-             ['u_0', 'u_1', 'u_2'])
-        )
-    )
-    def test_set_modules(self, monkeypatch, reward_function, physical_system, reference_generator,
-                         expected_observed_states):
-        reward_function.set_modules(physical_system, reference_generator)
-        assert np.all(
-            reward_function._observed_states == np.isin(physical_system.state_names, expected_observed_states))
-
-    @pytest.mark.parametrize("reward_function, physical_system, reference_generator, expected_observed_states", (
-            (RewardFunction(['currents', 'voltages']), DummyPhysicalSystem(5), DummyReferenceGenerator(),
-             ['i_a', 'i_e', 'u_a']),
-            (RewardFunction(['currents', 'omega']), DummyPhysicalSystem(5), DummyReferenceGenerator(),
-             ['omega', 'i_a', 'i_e']),
-            (RewardFunction(['omega', 'voltages']), DummyPhysicalSystem(5), DummyReferenceGenerator(),
-             ['omega', 'u_a']),
-            (RewardFunction(['currents', 'omega', 'voltages']), DummyPhysicalSystem(5), DummyReferenceGenerator(),
-             ['omega', 'i_a', 'i_e', 'u_a']),
-    ))
-    def test_set_modules_combined_observed_states(self, monkeypatch, reward_function, physical_system,
-                                                  reference_generator, expected_observed_states):
-        physical_states = ['i_a', 'i_e', 'u_a', 'omega', 'torque']
-        physical_system._state_names = physical_states
-        reward_function.set_modules(physical_system, reference_generator)
-        assert np.all(
-            reward_function._observed_states == np.isin(physical_system.state_names, expected_observed_states))
-
-    @pytest.mark.parametrize(
-        "physical_system, reference_generator", ((DummyPhysicalSystem(3), DummyReferenceGenerator()),)
-    )
-    @pytest.mark.parametrize("observed_state_idx, violated_state_idx", (
-            ([0, 1], [2]),
-            ([0, 1, 2], []),
-            ([2], [1, 2])
-    ))
-    def test_reward(self, monkeypatch, physical_system, reference_generator, observed_state_idx, violated_state_idx):
[…]
