+++ tests/test_environments/test_environments.py
@@ -12,12 +12,9 @@ import gym_electric_motor.envs.gym_synrm.syn_reluctance_motor_env as synrmenv
 import gym_electric_motor.envs.gym_im.squirrel_cage_induc_motor_env as scimenv
 import gym_electric_motor.envs.gym_im.doubly_fed_induc_motor_env as dfimenv
 
-import time
-
-
 from ..testing_utils import *
 from ..conf import *
-from numpy.random import randint, seed
+from numpy.random import seed
 import pytest
 
 # region first tests
@@ -27,21 +24,25 @@ g_disc_converter = ['Disc-1QC', 'Disc-2QC', 'Disc-4QC']
 g_cont_converter = ['Cont-1QC', 'Cont-2QC', 'Cont-4QC']
 g_converter = ['Disc-1QC', 'Disc-2QC', 'Disc-4QC', 'Cont-1QC', 'Cont-2QC', 'Cont-4QC']
 g_reference_generator = ['WienerProcessReference', 'SinusReference', 'StepReference']
-g_reward_functions = ['WSE', 'SWSE']
+g_reward_functions = ['WSE']
 
 
 @pytest.mark.parametrize("converter_type", g_cont_converter)
 @pytest.mark.parametrize("reward_function_type", g_reward_functions)
 @pytest.mark.parametrize("reference_generator_type", g_reference_generator)
 @pytest.mark.parametrize("tau_test", [1E-5, 1E-4, 1E-3])
-@pytest.mark.parametrize("motor_type, state_filter, motor_class, initializer",
-                         [('DcPermEx', ['omega', 'u', 'i'], ContDcPermanentlyExcitedMotorEnvironment, permex_initializer),
-                          ('DcSeries', ['torque'], ContDcSeriesMotorEnvironment, series_initializer),
-                          ('DcShunt', ['omega'], ContDcShuntMotorEnvironment, shunt_initializer),
-                          ('DcExtEx', ['torque'], ContDcExternallyExcitedMotorEnvironment, extex_initializer)])
-def test_cont_dc_motor_environments(motor_type, converter_type, reward_function_type,
-                                    reference_generator_type, tau_test, state_filter,
-                                    motor_class, turn_off_windows, initializer):
+@pytest.mark.parametrize(
+    "motor_type, state_filter, motor_class, initializer", [
+        ('DcPermEx', ['omega', 'u', 'i'], ContDcPermanentlyExcitedMotorEnvironment, permex_initializer),
+        ('DcSeries', ['torque'], ContDcSeriesMotorEnvironment, series_initializer),
+        ('DcShunt', ['omega'], ContDcShuntMotorEnvironment, shunt_initializer),
+        ('DcExtEx', ['torque'], ContDcExternallyExcitedMotorEnvironment, extex_initializer)
+    ]
+)
+def test_cont_dc_motor_environments(
+        motor_type, converter_type, reward_function_type, reference_generator_type, tau_test, state_filter,
+        motor_class, turn_off_windows, initializer
+):
     """
     test cont dc motor environments
 
@@ -126,7 +127,7 @@ def test_cont_dc_motor_environments(motor_type, converter_type, reward_function_
                                                                                                   reward_range[1],
                                                                                                   index])
             else:
-                assert reward == env.reward_function._limit_violation_reward(None)
+                assert reward == env.reward_function._violation_reward
                 break
             env.render()
         env.close()
@@ -227,7 +228,7 @@ def test_disc_dc_motor_environments(motor_type, converter_type, reward_function_
                                                                                                    reward_range[1],
                                                                                                    index])
             else:
-                reward == env.reward_function._limit_violation_reward(None)
+                reward == env.reward_function._violation_reward
                 break
             env.render()
         env.close()
[…]
