+++ tests/test_core.py
@@ -1,11 +1,12 @@
 import pytest
 import numpy as np
 from tests.testing_utils import DummyPhysicalSystem, DummyReferenceGenerator, DummyRewardFunction, DummyVisualization,\
-    DummyCallback, DummyConstraintMonitor, mock_instantiate, instantiate_dict
+    DummyCallback, DummyConstraintMonitor, DummyConstraint, mock_instantiate, instantiate_dict
 from gym.spaces import Tuple, Box
 import gym_electric_motor
 from gym_electric_motor.core import ElectricMotorEnvironment, RewardFunction, \
-    ReferenceGenerator, PhysicalSystem
+    ReferenceGenerator, PhysicalSystem, ConstraintMonitor, Constraint
+from gym_electric_motor.constraints import LimitConstraint
 import gym
 import gym_electric_motor as gem
 
@@ -201,10 +202,13 @@ class TestElectricMotorEnvironment:
         # No Exception raised
         env.step(env.action_space.sample())
 
-    @pytest.mark.parametrize("number_states, state_filter, expected_result",
-                             ((1, ['dummy_state_0'], [10]),
-                              (3, ['dummy_state_0', 'dummy_state_1', 'dummy_state_2'], [10, 20, 30]),
-                              (3, ['dummy_state_1'], [20])))
+    @pytest.mark.parametrize(
+        "number_states, state_filter, expected_result", (
+            (1, ['dummy_state_0'], [10]),
+            (3, ['dummy_state_0', 'dummy_state_1', 'dummy_state_2'], [10, 20, 30]),
+            (3, ['dummy_state_1'], [20])
+        )
+    )
     def test_limits(self, number_states, state_filter, expected_result):
         ps = DummyPhysicalSystem(state_length=number_states)
         rg = DummyReferenceGenerator()
@@ -269,3 +273,102 @@ class TestPhysicalSystem:
         assert ps.state_names == state_names
         assert ps.tau == tau
         assert ps.k == 0
+
+
+class TestConstraintMonitor:
+
+    @pytest.mark.parametrize(
+        ['ps', 'limit_constraints', 'expected_observed_states'], [
+            [DummyPhysicalSystem(3), ['dummy_state_0', 'dummy_state_2'], ['dummy_state_0', 'dummy_state_2']],
+            [DummyPhysicalSystem(1), ['dummy_state_0'], ['dummy_state_0']],
+            [DummyPhysicalSystem(2), ['all_states'], ['dummy_state_0', 'dummy_state_1']]
+        ]
+                             )
+    def test_limit_constraint_setting(self, ps, limit_constraints, expected_observed_states):
+        cm = ConstraintMonitor(limit_constraints=limit_constraints)
+        cm.set_modules(ps)
+        assert cm.constraints[0]._observed_state_names == expected_observed_states
+
+    @pytest.mark.parametrize('constraints', [
+        [lambda state: 0.0, DummyConstraint(), DummyConstraint()]
+    ])
+    @pytest.mark.parametrize('ps', [DummyPhysicalSystem(1)])
+    def test_additional_constraint_setting(self, ps, constraints):
+        cm = ConstraintMonitor(additional_constraints=constraints)
+        cm.set_modules(ps)
+        assert all(constraint in cm.constraints for constraint in constraints)
+
+    @pytest.mark.parametrize('additional_constraints', [
+        [lambda state: 0.0, DummyConstraint(), DummyConstraint()]
+    ])
+    @pytest.mark.parametrize('limit_constraints', [
+        ['all_states'], ['dummy_state_0'], []
+    ])
+    @pytest.mark.parametrize('ps', [DummyPhysicalSystem(1)])
[…]
