### 2020-11-28 4ce2649
Final example changes and bugfixes
