+++ gym_electric_motor/visualization/motor_dashboard.py
@@ -184,7 +184,7 @@ class MotorDashboard(ElectricMotorVisualization):
         # create separate figures for time based, step and episode based plots
         if len(self._time_plots) > 0:
             self._time_plot_figure, axes_step = plt.subplots(len(self._time_plots), sharex=True)
-            self._time_plot_figure.canvas.set_window_title('Step Plots')
+            self._time_plot_figure.canvas.set_window_title('Time Plots')
             axes_step = [axes_step] if len(self._time_plots) == 1 else axes_step
             self._time_plot_figure.subplots_adjust(wspace=0.0, hspace=0.2)
             axes_step[-1].set_xlabel('$t$/s')
