### 2020-11-29 e656d7b
Deletion to be consistent with the current nightly and master
