### 2020-12-08 3d67126
WIP: First draft of visu.create_new_figures() command
