+++ gym_electric_motor/visualization/motor_dashboard_plots/state_plot.py
@@ -81,9 +81,7 @@ class StatePlot(TimePlot):
         self._referenced = rg.referenced_states[self._state_idx]
         # Bool: if the data is already normalized to an interval of [-1, 1]
         self._normalized = self._limits != self._state_space[1]
-        # Initialize the data containers
-        self._state_data = np.ones(self._x_width) * np.nan
-        self._ref_data = np.ones(self._x_width) * np.nan
+        self.reset_data()
 
         min_limit = self._limits * self._state_space[0] if self._normalized else self._state_space[0]
         max_limit = self._limits * self._state_space[1] if self._normalized else self._state_space[1]
@@ -94,6 +92,12 @@ class StatePlot(TimePlot):
 
         # Set the y-axis label
         self._label = self.state_labels.get(self._state, self._state)
+
+    def reset_data(self):
+        super().reset_data()
+        # Initialize the data containers
+        self._state_data = np.ones(self._x_width) * np.nan
+        self._ref_data = np.ones(self._x_width) * np.nan
 
     def initialize(self, axis):
         # Docstring of superclass
