### 2020-12-09 4219fc3
Enabled Step and Episode Plots in Notebook
