### 2020-12-12 34c9ab0
Finished figure reset.
