### 2020-12-12 a92706a
Finished figure reset.
