### 2020-12-12 3ef0ffa
Finished figure reset.
