+++ gym_electric_motor/envs/gym_srm/srm_finite_control_env.py
@@ -1,20 +1,0 @@
-import gym
-
-
-class SRMFiniteControlEnv(gym.Env):
-    metadata = {'render.modes': ['human']}
-
-    def __init__(self):
-        pass
-
-    def step(self, action):
-        raise NotImplementedError()
-
-    def reset(self):
-        raise NotImplementedError()
-
-    def render(self, mode='human'):
-        raise NotImplementedError()
-
-    def close(self):
-        raise NotImplementedError()
