+++ gym_electric_motor/envs/gym_srm/__init__.py
@@ -1,0 +1,2 @@
+from .switched_rel_motor_env import ContSwitchedReluctanceMotorEnvironment
+from .switched_rel_motor_env import DiscSwitchedReluctanceMotorEnvironment

[…]
