### 2020-12-16 #2498 @Kevin-Mattheus-Moerman
Unfortunately we have been unable to reach @dineshresearch in relation to the last unticked box. However, they did previously state:

> Hello I have completed the review checklist..

Furthermore @moorepants has also completed the review and recommends acceptance. Finally I have also reviewed the paper and sections relating to "State of the field", and feel they are sufficient. I will there now process this work for acceptance in JOSS to avoid further delay.
