### 2020-12-16 #2498 @Kevin-Mattheus-Moerman
@wkirgsn I will be processing this work for acceptance. I might have some final points on the paper later on but for the moment can you:

- [ ] Archive a copy of this work on [ZENODO](https://zenodo.org/) and report back here with the archive DOI? As you create this archive please **ensure that the metadata matches that of the paper**, this includes the title, author names, number of authors, and author order.
- [ ] Can you verify that the version label is accurate? Are you still at `v0.2.1` or has it moved on?
- [ ] This paper is about to be accepted in JOSS. I recommend you read through the paper yourself once more and check if the author names and acknowledgements for instance are accurate.
