### 2020-12-17 1020a1c
use of np.full instead of np.zeros_like
