### 2020-12-18 ddd387b
remove urls from setup.py to be pypi conform
