### 2020-12-18 #2498 @wkirgsn
Here's the badge:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4355691.svg)](https://doi.org/10.5281/zenodo.4355691)

I am afraid the author names, their number and their order do not match that of the JOSS paper, as in the mean time there were several other contributors (which is natural for open-source SW), that were extracted automatically from github for the zenodo doi generation.
However, the original authors of the JOSS paper are still those with the most significant contribution. Is it mandatory to include the new contributors as well? I feel this can get cluttered and bloated with github nicknames.

The current version number is v0.3.1.

I've read the paper again and all information in the JOSS paper are still valid.
Thank you
