+++ _sources/parts/visualizations/motor_dashboard_plots/step_plot.rst.txt
@@ -1,0 +1,7 @@
+Step Plot (Abstract)
+#####################
+The StepPlot is the base class for all plots that plot data with the cumulative number of steps on the x-axis.
+
+.. autoclass:: gym_electric_motor.visualization.motor_dashboard_plots.TimePlot
+   :members:
+   :inherited-members:
