+++ _sources/parts/constraint_monitor.rst.txt
@@ -1,7 +1,21 @@
 Constraint Monitor
 #####################
 
+..  toctree::
+    :maxdepth: 1
+    :caption: Available Constraints:
 
-.. autoclass:: gym_electric_motor.constraint_monitor.ConstraintMonitor
+    constraints/constraint
+    constraints/limit_constraint
+    constraints/squared_constraint
+
+Usage Guide
+___________
+
+ToDo
+
+Constraint Monitor API Documentation
+____________________________________
+.. autoclass:: gym_electric_motor.core.ConstraintMonitor
     :members:
     :inherited-members:
