+++ examples/classic_controllers/classic_controllers.py
@@ -1,0 +1,322 @@
+from gym.spaces import Discrete, Box, MultiDiscrete
+import sys
+import os
+from gym_electric_motor.physical_systems import SynchronousMotorSystem, DcMotorSystem
+import numpy as np
+
+sys.path.append(os.path.abspath(os.path.join('..')))
+
+
+class Controller:
+
+    @classmethod
+    def make(cls, controller_type, environment, **controller_kwargs):
+        assert controller_type in _controllers.keys(), f'Controller {controller_type} unknown'
+        controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
+        return controller
+
+    def control(self, state, reference):
+        pass
+
+    def reset(self):
+        pass
+        
+class ContinuousActionController(Controller):
+    def __init__(self, controller_type, environment, **kwargs):
+        self.action_space = environment.action_space
+        assert type(self.action_space) is Box and type(
+            environment.physical_system) is DcMotorSystem, 'No suitable action space for Continuous Action Controller'
+        self.ref_state_idx = np.where(environment.reference_generator.referenced_states == True)[0]
+        self.controller = _controllers[controller_type][1].make(controller_type, environment, **kwargs)
+        self.mp = environment.physical_system.electrical_motor.motor_parameter
+        if 'psi_e' in self.mp.keys():
+            self.psi_e = self.mp['psi_e']
+        else:
+            self.l_e = self.mp['l_e_prime']
+            self.psi_e = None
+        self.i_idx = environment.physical_system.CURRENTS_IDX[-1]
+        self.u_idx = environment.physical_system.VOLTAGES_IDX[-1]
+        self.limit = environment.physical_system.limits[environment.state_filter]
+        self.omega_idx = environment.state_names.index('omega')
+        self.action = np.zeros(self.action_space.shape[0])
+        if self.action_space.shape[0] > 1:
+            self.action[1] = self.limit[self.i_idx] * self.mp['r_e'] / self.limit[self.u_idx]
+
+    def control(self, state, reference):
+        self.action[0] = self.controller.control(state[self.ref_state_idx], reference[0]) + self.decoupling(state)
+        if self.action_space.low[0] <= self.action.all() <= self.action_space.high[0]:
+            self.controller.integrate(state[self.ref_state_idx], reference[0])
+        return np.clip(self.action, self.action_space.low, self.action_space.high)
+
+    def reset(self):
+        self.controller.reset()
+
+    def decoupling(self, state):
+        psi_e = self.psi_e or self.l_e * state[self.i_idx] * self.limit[self.i_idx]
+        return (state[self.omega_idx] * self.limit[self.omega_idx] * psi_e) / self.limit[self.u_idx]
+
+class DiscreteActionController(Controller):
+    def __init__(self, controller_type, environment, **kwargs):
+        assert type(environment.action_space) in [Discrete, MultiDiscrete] and type(
+            environment.physical_system) is DcMotorSystem, 'No suitable action space for Discrete Action Controller'
+        self.action_space = environment.action_space
+        self.limit = environment.physical_system.limits
+        self.u_idx = environment.physical_system.VOLTAGES_IDX[-1]
+        self.i_idx = environment.physical_system.CURRENTS_IDX[-1]
+        self.mp = environment.physical_system.electrical_motor.motor_parameter
+        if self.action_space.shape != ():
+            self.action = np.zeros(self.action_space.nvec.shape, dtype='int')
+            self.k = 0
+            self.u_e_limit = int(self.limit[self.u_idx] / (self.limit[self.i_idx] * self.mp['r_e'])) + 1
+        self.controller = _controllers[controller_type][1].make(controller_type, environment, **kwargs)
[…]
