### 2020-12-21 f530699
Update test_classic_controllers.py
