### 2020-12-21 #2498 @Kevin-Mattheus-Moerman
@wkirgsn  thanks for those updates. I set the archive and update the version tag. You do need to amend the archive metadata for the title and authors to match that of the paper.

- [ ] You need to either add the additional contributors as co-authors on the paper (you'll ideally need their full names/affiliations), or you need to remove them from the archive listing on Zenodo.
- [ ] Change the title for the Zenodo archive from `upb-lea/gym-electric-motor: ConstraintMonitor and Visualization` to the paper title `gym-electric-motor (GEM): A Python toolbox for the simulation of electric drive systems`.
