### 2021-01-03 8ccae43
First Dummy for the new Environment interface
