+++ gym_electric_motor/envs/gym_dcm/dc_permex_motor_env.py
@@ -1,25 +1,98 @@
-from gym_electric_motor.core import ElectricMotorEnvironment
+from gym_electric_motor.core import ElectricMotorEnvironment, ReferenceGenerator, RewardFunction, \
+    ElectricMotorVisualization
 from gym_electric_motor.physical_systems.physical_systems import DcMotorSystem
+from gym_electric_motor.visualization import MotorDashboard
 from gym_electric_motor.reference_generators.wiener_process_reference_generator import WienerProcessReferenceGenerator
+from gym_electric_motor import physical_systems as ps
 from gym_electric_motor.reward_functions import WeightedSumOfErrors
+from gym_electric_motor.utils import initialize
 
 
-class DcPermanentlyExcitedMotorEnvironment(ElectricMotorEnvironment):
+class DiscSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
+    """
+        Description:
+            Environment to simulate a discretely speed controlled permanently excited DC Motor
 
-    def __init__(self, motor='DcPermEx', reward_function=None, reference_generator=None, constraints=('i',), **kwargs):
+        Key:
+            `Disc-SC-DcPermEx-v0`
+
+        Default Modules:
+
+            Physical System:
+                SCMLSystem/DcMotorSystem with:
+                    | IdealVoltageSupply
+                    | DiscFourQuadrantConverter
+                    | DcPermanentlyExcitedMotor
+                    | PolynomialStaticLoad
+                    | GaussianWhiteNoiseGenerator
+                    | EulerSolver
+                    | tau=1e-5
+
+            Reference Generator:
+                WienerProcessReferenceGenerator
+                    Reference Quantity. 'omega'
+
+            Reward Function:
+                WeightedSumOfErrors: reward_weights 'omega' = 1
+
+            Visualization:
+                *MotorDashboard* - Plots: omega, action
+
+        State Variables:
+            ``['omega' , 'torque', 'i', 'u', 'u_sup']``
+
+        Observation Space:
+            Type: Tuple(State_Space, Reference_Space)
+
+        State Space:
+            Box(low=[-1, -1, -1, -1, 0], high=[1, 1, 1, 1, 1])
+
+        Reference Space:
+            Box(low=[-1], high=[1])
+
+        Action Space:
+            Type: Box(low=[-1], high=[1])
+
+        Starting State:
+            Zeros on all state variables.
+
+        Episode Termination:
+            Termination if current limits are violated. The terminal reward -10 is used as reward.
+            (Have a look at the reward functions.)
+        """
+    def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None, noise_generator=None,
+                 reward_function=None, reference_generator=None, visualization=None,
+                 constraints=('i',), calc_jacobian=True, tau=1e-5):
         """
         Args:
             motor(ElectricMotor): Electric Motor used in the PhysicalSystem
             reward_function(RewardFunction): Reward Function for the environment
             reference_generator(ReferenceGenerator): Reference Generator for the environment
-            kwargs(dict): Further kwargs to pass to the superclass and the submodules
         """
-        physical_system = DcMotorSystem(motor=motor, **kwargs)
-        reference_generator = reference_generator or WienerProcessReferenceGenerator(**kwargs)
-        reward_function = reward_function or WeightedSumOfErrors(**kwargs)
+
+        physical_system = DcMotorSystem(
+            supply=initialize(ps.VoltageSupply, supply, ps.IdealVoltageSupply, dict(u_nominal=420.0)),
+            converter=initialize(ps.PowerElectronicConverter, converter, ps.DiscFourQuadrantConverter, dict()),
[…]
