### 2021-01-03 5eb9e13
First Dummy for the new Environment interface
