+++ gym_electric_motor/envs/gym_dcm/dc_permex_motor_env.py
@@ -96,116 +96,433 @@ class DiscSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
         )
 
 
-class DiscDcPermanentlyExcitedMotorEnvironment(DcPermanentlyExcitedMotorEnvironment):
+class ContSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
     """
-    Description:
-        Environment to simulate a discretely controlled permanently excited DC Motor
+        Description:
+            Environment to simulate a discretely speed controlled permanently excited DC Motor
 
-    Key:
-        `DcPermExDisc-v1`
+        Key:
+            `Disc-SC-DcPermEx-v0`
 
-    Default Modules:
+        Default Modules:
 
-        Physical System:
-            SCMLSystem/DcMotorSystem with:
-                | IdealVoltageSupply
-                | DiscFourQuadrantConverter
-                | DcPermanentlyExcitedMotor
-                | PolynomialStaticLoad
-                | GaussianWhiteNoiseGenerator
-                | EulerSolver
-                | tau=1e-5
+            Physical System:
+                SCMLSystem/DcMotorSystem with:
+                    | IdealVoltageSupply
+                    | ContFourQuadrantConverter
+                    | DcPermanentlyExcitedMotor
+                    | PolynomialStaticLoad
+                    | GaussianWhiteNoiseGenerator
+                    | EulerSolver
+                    | tau=1e-4
 
-        Reference Generator:
-            WienerProcessReferenceGenerator
-                Reference Quantity. 'omega'
+            Reference Generator:
+                WienerProcessReferenceGenerator
+                    Reference Quantity. 'omega'
 
-        Reward Function:
-            WeightedSumOfErrors(reward_weights= {'omega': 1 })
+            Reward Function:
+                WeightedSumOfErrors: reward_weights 'omega' = 1
 
-        Visualization:
-            ElectricMotorVisualization (Dummy for no Visualization)
+            Visualization:
+                *MotorDashboard* - Plots: omega, action
 
-    State Variables:
-        ``['omega' , 'torque', 'i', 'u', 'u_sup']``
+        State Variables:
+            ``['omega' , 'torque', 'i', 'u', 'u_sup']``
 
-    Observation Space:
-        Type: Tuple(State_Space, Reference_Space)
+        Observation Space:
+            Type: Tuple(State_Space, Reference_Space)
 
-    State Space:
-        Box(low=[-1, -1, -1, -1, 0], high=[1, 1, 1, 1, 1])
+        State Space:
+            Box(low=[-1, -1, -1, -1, 0], high=[1, 1, 1, 1, 1])
 
-    Reference Space:
-        Box(low=[-1], high=[1])
+        Reference Space:
+            Box(low=[-1], high=[1])
 
-    Action Space:
-        Type: Box(low=[-1], high=[1])
+        Action Space:
+            Type: Box(low=[-1], high=[1])
 
-    Starting State:
-        Zeros on all state variables.
+        Starting State:
+            Zeros on all state variables.
 
-    Episode Termination:
-        Termination if current limits are violated. The terminal reward -10 is used as reward.
-        (Have a look at the reward functions.)
+        Episode Termination:
+            Termination if current limits are violated. The terminal reward -10 is used as reward.
+            (Have a look at the reward functions.)
+        """
+    def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None, noise_generator=None,
+                 reward_function=None, reference_generator=None, visualization=None,
+                 constraints=('i',), calc_jacobian=True, tau=1e-4):
+        """
+        Args:
+            motor(ElectricMotor): Electric Motor used in the PhysicalSystem
[…]
