+++ examples/classic_controllers/classic_controllers.py
@@ -1,18 +1,31 @@
 from gym.spaces import Discrete, Box, MultiDiscrete
-import sys
-import os
 from gym_electric_motor.physical_systems import SynchronousMotorSystem, DcMotorSystem
 import numpy as np
-
-sys.path.append(os.path.abspath(os.path.join('..')))
 
 
 class Controller:
 
     @classmethod
-    def make(cls, controller_type, environment, **controller_kwargs):
-        assert controller_type in _controllers.keys(), f'Controller {controller_type} unknown'
-        controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
+    def make(cls, environment, controller_type=None, **controller_kwargs):
+
+        ref_states = []
+        for i in range(np.size(np.where(environment.reference_generator.referenced_states == True)[0])):
+            ref_states.append(
+                environment.state_names[np.where(environment.reference_generator.referenced_states == True)[0][i]])
+
+        if controller_type != None:
+
+            assert controller_type in _controllers.keys(), f'Controller {controller_type} unknown'
+            controller_kwargs = cls.automated_gain(cls, controller_type, environment, ref_states, **controller_kwargs)
+            controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
+
+        else:
+
+            controller_type, controller_kwargs = cls.automated_controller_design(cls, environment, controller_type,
+                                                                                 ref_states, **controller_kwargs)
+            controller_kwargs = cls.automated_gain(cls, controller_type, environment, ref_states, **controller_kwargs)
+            controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
+
         return controller
 
     def control(self, state, reference):
@@ -20,13 +33,130 @@ class Controller:
 
     def reset(self):
         pass
-        
+
+    def automated_controller_design(self, environment, controller_type, ref_states, **controller_kwargs):
+        action_space = type(environment.action_space)
+
+        if type(environment.physical_system) == DcMotorSystem:
+
+            if 'omega' in ref_states:
+                controller_type = 'cascaded_controller'
+                stages = controller_kwargs['stages'] if 'stages' in controller_kwargs.keys() else []
+                for i in range(len(stages), 2):
+                    if i == 0:
+                        if action_space == Box:
+                            stages.append({'controller_type': 'pi_controller'})
+                        else:
+                            stages.append({'controller_type': 'three_point'})
+                    else:
+                        stages.append({'controller_type': 'pi_controller'})
+
+                controller_kwargs['stages'] = stages
+
+            elif 'i' in ref_states or 'i_a' in ref_states or 'torque' in ref_states:
+
+                if action_space == Discrete or action_space == MultiDiscrete:
+                    controller_type = 'three_point'
+                elif action_space == Box:
+                    controller_type = 'pi_controller'
+
+
+        elif type(environment.physical_system) == SynchronousMotorSystem:
+            controller_type = 'foc_controller'
+            if 'q_controller' not in controller_kwargs.keys():
+                controller_kwargs['q_controller'] = {'controller_type': 'pi_controller'}
+            if 'd_controller' not in controller_kwargs.keys():
+                controller_kwargs['d_controller'] = {'controller_type': 'pi_controller'}
+
+        return controller_type, controller_kwargs
+
+    def automated_gain(self, controller_type, environment, ref_states, **controller_kwargs):
+        mp = environment.physical_system.electrical_motor.motor_parameter
[…]
