### 2021-01-06 #2498 @wkirgsn
Hi @Kevin-Mattheus-Moerman, happy new year!
Due to your help I figured how to change the zenodo details. However, currently I experience the following error message issued by zenodo when I want to "publish" the changes:
```Internal Server Error Error ID: 30a8cc6068734a1eaf80eaf96cc3299e```
I will let you know when this is fixed and I was able to push the details to zenodo.
