### 2021-01-06 f4e5418
Update test_classic_controllers.py
