+++ examples/classic_controllers/test_classic_controllers.py
@@ -1,8 +1,7 @@
 from classic_controllers import Controller
-import sys, os
 
-sys.path.append(os.path.abspath(os.path.join('..')))
 import gym_electric_motor as gem
+from gym.spaces import Box
 from gym_electric_motor.reference_generators import StepReferenceGenerator, SinusoidalReferenceGenerator, SwitchedReferenceGenerator, MultipleReferenceGenerator, SubepisodedReferenceGenerator, WienerProcessReferenceGenerator, TriangularReferenceGenerator, ConstReferenceGenerator
 from gym_electric_motor.visualization import MotorDashboard
 from gym_electric_motor import reference_generators as rg
@@ -13,17 +12,18 @@ import numpy as np
 
 if __name__ == '__main__':
 
-ref = 'omega'
+
+    ref = 'omega'
     env = gem.make(
-        #'DcExtExDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u_a', 'u_e'], visu_period=1), motor_parameter=dict(j_rotor=0.00005), load_parameter={'a': 0, 'b': 0, 'c': 0, 'j_load': 0},
+        #'DcExtExDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u_a', 'u_e'], visu_period=1),
         #'DcPermExDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u'], visu_period=1),
         #'DcSeriesDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u'], visu_period=1),
-        #'DcShuntDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u'], visu_period=1), motor_parameter=dict(j_rotor=0.001), load_parameter={'a': 0, 'b': 0, 'c': 0, 'j_load': 0.001},
+        #'DcShuntDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u'], visu_period=1),
 
-        #'DcExtExCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u_a', 'u_e'], visu_period=1), motor_parameter=dict(j_rotor=0.00005), load_parameter={'a': 0, 'b': 0, 'c': 0, 'j_load': 0},
+        #'DcExtExCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u_a', 'u_e'], visu_period=1),
         'DcPermExCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u'], visu_period=1),
         #'DcSeriesCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u'], visu_period=1),
-        #'DcShuntCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u'], visu_period=1), motor_parameter=dict(j_rotor=0.001), load_parameter={'a': 0, 'b': 0, 'c': 0.5, 'j_load': 0},
+        #'DcShuntCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u'], visu_period=1),
 
         ode_solver='scipy.solve_ivp', solver_kwargs=dict(),
         #load=ConstantSpeedLoad(omega_fixed=50 * np.pi / 30),
@@ -36,35 +36,9 @@ ref = 'omega'
 
     )
 
+    controller = Controller.make(env)
 
-    try:
[…]
