### 2021-01-11 #2498 @wkirgsn
Hi @Kevin-Mattheus-Moerman,
since Zenodo still gives me an internal server error upon clicking on "publish", whereas "saving" the changes went fine, I come to think that the required changes just happened already and there is no need to re-publish the changes (albeit showing the internal server error and a big blue publish button a somewhat confusing user experience is).

Can you confirm that the zenodo article has all required details now? If that is the case, I think we can continue the process.
