+++ gym_electric_motor/__init__.py
@@ -24,10 +24,19 @@ from gym.envs.registration import register
 envs_path = 'gym_electric_motor.envs:'
 
 # Version 1
-register(id='DcPermExCont-v1',
-         entry_point=envs_path+'ContDcPermanentlyExcitedMotorEnvironment')
-register(id='DcPermExDisc-v1',
-         entry_point=envs_path+'DiscDcPermanentlyExcitedMotorEnvironment')
+register(id='Disc-SC-PermExDc-v0',
+         entry_point=envs_path+'DiscSpeedControlDcPermanentlyExcitedMotorEnv')
+register(id='Cont-SC-PermExDc-v0',
+         entry_point=envs_path+'ContSpeedControlDcPermanentlyExcitedMotorEnv')
+register(id='Disc-TC-PermExDc-v0',
+         entry_point=envs_path+'DiscTorqueControlDcPermanentlyExcitedMotorEnv')
+register(id='Cont-TC-PermExDc-v0',
+         entry_point=envs_path+'ContTorqueControlDcPermanentlyExcitedMotorEnv')
+register(id='Disc-CC-PermExDc-v0',
+         entry_point=envs_path+'DiscCurrentControlDcPermanentlyExcitedMotorEnv')
+register(id='Cont-CC-PermExDc-v0',
+         entry_point=envs_path+'ContCurrentControlDcPermanentlyExcitedMotorEnv')
+
 register(id='DcExtExCont-v1',
          entry_point=envs_path+'ContDcExternallyExcitedMotorEnvironment')
 register(id='DcExtExDisc-v1',
