+++ gym_electric_motor/envs/gym_dcm/dc_permex_motor_env.py
@@ -14,19 +14,15 @@ class DiscSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
             Environment to simulate a discretely speed controlled permanently excited DC Motor
 
         Key:
-            `Disc-SC-DcPermEx-v0`
+            `Disc-SC-PermExDc-v0`
 
-        Default Modules:
-
-            Physical System:
-                SCMLSystem/DcMotorSystem with:
-                    | IdealVoltageSupply
-                    | DiscFourQuadrantConverter
-                    | DcPermanentlyExcitedMotor
-                    | PolynomialStaticLoad
-                    | GaussianWhiteNoiseGenerator
-                    | EulerSolver
-                    | tau=1e-5
+        Default Components:
+            Supply: IdealVoltageSupply
+            Converter: DiscFourQuadrantConverter
+            Motor: DcPermanentlyExcitedMotor
+            Load: PolynomialStaticLoad
+            Ode-Solver: EulerSolver
+            Noise: None
 
             Reference Generator:
                 WienerProcessReferenceGenerator
@@ -36,7 +32,13 @@ class DiscSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
                 WeightedSumOfErrors: reward_weights 'omega' = 1
 
             Visualization:
-                *MotorDashboard* - Plots: omega, action
+                MotorDashboard: omega and action plots
+
+            Constraints:
+                Current Limit on 'i'
+
+        Control Cycle Time:
+            tau = 1e-5 seconds
 
         State Variables:
             ``['omega' , 'torque', 'i', 'u', 'u_sup']``
@@ -51,23 +53,48 @@ class DiscSpeedControlDcPermanentlyExcitedMotorEnv(ElectricMotorEnvironment):
             Box(low=[-1], high=[1])
 
         Action Space:
-            Type: Box(low=[-1], high=[1])
+            Type: Discrete(4)
 
         Starting State:
             Zeros on all state variables.
 
         Episode Termination:
-            Termination if current limits are violated. The terminal reward -10 is used as reward.
-            (Have a look at the reward functions.)
+            Termination if current limits are violated.
         """
     def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None, noise_generator=None,
-                 reward_function=None, reference_generator=None, visualization=None,
+                 reward_function=None, reference_generator=None, visualization=None, state_filter=None, callbacks=(),
                  constraints=('i',), calc_jacobian=True, tau=1e-5):
         """
         Args:
-            motor(ElectricMotor): Electric Motor used in the PhysicalSystem
-            reward_function(RewardFunction): Reward Function for the environment
-            reference_generator(ReferenceGenerator): Reference Generator for the environment
+            supply(env-arg): Specification of the supply to be used in the environment
+            converter(env-arg): Specification of the converter to be used in the environment
+            motor(env-arg): Specification of the motor to be used in the environment
+            load(env-arg): Specification of the load to be used in the environment
+            ode_solver(env-arg): Specification of the ode_solver to be used in the environment
+            noise_generator(env-arg): Specification of the noise_generator to be used in the environment
+            reward_function(env-arg: Specification of the reward_function to be used in the environment
+            reference_generator(env-arg): Specification of the reference generator to be used in the environment
+            visualization(env-arg): Specification of the visualization to be used in the environment
+            constraints(iterable(str/Constraint)): All Constraints of the environment.
+                - str: A LimitConstraints for states (episode terminates, if the quantity exceeds the limit)
+                 can be directly specified by passing the state name here (e.g. 'i', 'omega')
+                 - instance of Constraint: More complex constraints (e.g. the SquaredConstraint can be initialized and
+                 passed to the environment.
[…]
