### 2021-01-13 d55ba97
Documentation added for the new environments
