+++ docs/parts/environments/environment.rst
@@ -2,45 +2,88 @@ Environments
 ############
 
 On this page, all environments with their environment-id are listed.
-Furthermore, a short introduction into the Interface of the environments is given.
+In general, all environment-ids are structured as follows:
 
-========================================================= ==============================
-Environment                                                   environment-id
-========================================================= ==============================
-Continuous Dc Series Motor Environment                    ``'DcSeriesCont-v1'``
-Discrete Dc Series Motor Environment                      ``'DcSeriesDisc-v1'``
+``ControlType-ControlTask-MotorType-v0``
 
-Continuous Dc Shunt Motor Environment                     ``'DcShuntCont-v1'``
-Discrete Dc Shunt Motor Environment                       ``'DcShuntDisc-v1'``
+- The ``ControlType`` is in ``\{Disc / Cont\}``
+- The ``ControlTask`` is in ``\{TC / SC / CC\}`` (Torque / Speed Current Control)
+- The ``MotorType`` is in ``\{PermExDc / ExtExDc / SeriesDc / ShuntDc / PMSM / SynRM / DFIM / SCIM \}``
 
-Continuous Dc Permanently Excited Motor Environment       ``'DcPermExCont-v1'``
-Discrete Dc Permanently Excited Motor Environment         ``'DcPermExDisc-v1'``
 
-Continuous Dc Externally Excited Motor Environment        ``'DcExtExCont-v1'``
-Discrete Dc Externally Excited Motor Environment          ``'DcExtExDisc-v1'``
 
-Continuous Permanent Magnet Synchronous Motor Environment ``'PMSMCont-v1'``
-Discrete Permanent Magnet Synchronous Motor Environment   ``'PMSMDisc-v1'``
+=================================================================== ==============================
+Environment                                                          environment-id
+=================================================================== ==============================
+**Permanently Excited DC Motor Environments**
 
-Continuous Synchronous Reluctance Motor Environment       ``'SynRMCont-v1'``
-Discrete Synchronous Reluctance Motor Environment         ``'SynRMDisc-v1'``
+Discrete Torque Control Permanently Excited DC Motor Environment     ``'Disc-TC-PermExDc-v0'``
+Continuous Torque Control Permanently Excited DC Motor Environment   ``'Cont-TC-PermExDc-v0'``
+Discrete Speed Control Permanently Excited DC Motor Environment      ``'Disc-SC-PermExDc-v0'``
+Continuous Speed Control Permanently Excited DC Motor Environment    ``'Cont-SC-PermExDc-v0'``
+Discrete Current Control Permanently Excited DC Motor Environment    ``'Disc-CC-PermExDc-v0'``
+Continuous Current Control Permanently Excited DC Motor Environment  ``'Cont-CC-PermExDc-v0'``
 
-Continuous Squirrel Cage Induction Motor Environment      ``'SCIMCont-v1'``
-Discrete Squirrel Cage Induction Motor Environment        ``'SCIMDisc-v1'``
+**Externally Excited DC Motor Environments**
 
-Continuous Doubly Fed Induction Motor Environment         ``'DFIMCont-v1'``
-Discrete Doubly Fed Induction Motor Environment           ``'DFIMDisc-v1'``
+Discrete Torque Control Externally Excited DC Motor Environment      ``'Disc-TC-ExtExDc-v0'``
+Continuous Torque Control Externally Excited DC Motor Environment    ``'Cont-TC-ExtExDc-v0'``
+Discrete Speed Control Externally Excited DC Motor Environment       ``'Disc-SC-ExtExDc-v0'``
[…]
