+++ docs/parts/environments/disc_sc_permex.rst
@@ -1,0 +1,4 @@
+Discrete Speed Control DC Permanently Excited Motor Environment
+*******************************************************
+.. autoclass:: gym_electric_motor.envs.DiscSpeedControlDcPermanentlyExcitedMotorEnv
+   :members:
