+++ docs/parts/environments/disc_tc_permex.rst
@@ -1,4 +1,4 @@
-Discrete TorqueControl DC Permanently Excited Motor Environment
+Discrete Torque Control DC Permanently Excited Motor Environment
 *******************************************************
 .. autoclass:: gym_electric_motor.envs.DiscTorqueControlDcPermanentlyExcitedMotorEnv
    :members:
