+++ docs/parts/environments/disc_cc_permex.rst
@@ -1,0 +1,4 @@
+Discrete Current Control DC Permanently Excited Motor Environment
+*******************************************************
+.. autoclass:: gym_electric_motor.envs.DiscCurrentControlDcPermanentlyExcitedMotorEnv
+   :members:
