+++ examples/classic_controllers/classic_controllers.py
@@ -1,5 +1,7 @@
 from gym.spaces import Discrete, Box, MultiDiscrete
-from gym_electric_motor.physical_systems import SynchronousMotorSystem, DcMotorSystem
+from gym_electric_motor.physical_systems import SynchronousMotorSystem, DcMotorSystem, DcSeriesMotor, DcExternallyExcitedMotor
+from gym_electric_motor.reference_generators import MultipleReferenceGenerator, SwitchedReferenceGenerator
+from gym_electric_motor.envs import ContDcExternallyExcitedMotorEnvironment
 import numpy as np
 
 
@@ -8,23 +10,18 @@ class Controller:
     @classmethod
     def make(cls, environment, controller_type=None, **controller_kwargs):
 
-        ref_states = []
-        for i in range(np.size(np.where(environment.reference_generator.referenced_states == True)[0])):
-            ref_states.append(
-                environment.state_names[np.where(environment.reference_generator.referenced_states == True)[0][i]])
+        controller_kwargs = cls.reference_states(cls, environment, **controller_kwargs)
 
         if controller_type != None:
-
             assert controller_type in _controllers.keys(), f'Controller {controller_type} unknown'
-            controller_kwargs = cls.automated_gain(cls, controller_type, environment, ref_states, **controller_kwargs)
+            controller_kwargs = cls.automated_gain(cls, controller_type, environment, **controller_kwargs)
             controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
 
         else:
-
             controller_type, controller_kwargs = cls.automated_controller_design(cls, environment, controller_type,
-                                                                                 ref_states, **controller_kwargs)
-            controller_kwargs = cls.automated_gain(cls, controller_type, environment, ref_states, **controller_kwargs)
-            controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
+                                                                                 **controller_kwargs)
+            controller_kwargs = cls.automated_gain(cls, controller_type, environment, **controller_kwargs)
+            controller = _controllers[controller_type][0](controller_type, environment,  **controller_kwargs)
 
         return controller
 
@@ -34,12 +31,26 @@ class Controller:
     def reset(self):
         pass
 
-    def automated_controller_design(self, environment, controller_type, ref_states, **controller_kwargs):
-        action_space = type(environment.action_space)
+    def reference_states(self, environment, **controller_kwargs):
+        ref_states = []
+        if type(environment.reference_generator) == MultipleReferenceGenerator:
+            for rg in environment.reference_generator._sub_generators:
 
+                ref_states.append(rg._sub_generators[0]._reference_state)
+        elif type(environment.reference_generator) == SwitchedReferenceGenerator:
+            ref_states.append(environment.reference_generator._sub_generators[0]._reference_state)
+        else:
+            ref_states.append(environment.reference_generator._reference_state)
+        controller_kwargs['ref_states'] = np.array(ref_states)
+        return controller_kwargs
+
+
+    def automated_controller_design(self, environment, controller_type, **controller_kwargs):
+        action_space = type(environment.action_space)
+        ref_states = controller_kwargs['ref_states']
         if type(environment.physical_system) == DcMotorSystem:
 
-            if 'omega' in ref_states:
+            if 'omega' in ref_states or 'torque' in ref_states:
                 controller_type = 'cascaded_controller'
                 stages = controller_kwargs['stages'] if 'stages' in controller_kwargs.keys() else []
                 for i in range(len(stages), 2):
[…]
