### 2021-01-20 ceaed15
Update test_classic_controllers.py
