+++ examples/classic_controllers/test_classic_controllers.py
@@ -1,19 +1,44 @@
 from classic_controllers import Controller
 
 import gym_electric_motor as gem
+from gym_electric_motor.physical_systems.mechanical_loads import PolynomialStaticLoad, ConstantSpeedLoad
 from gym.spaces import Box
 from gym_electric_motor.reference_generators import StepReferenceGenerator, SinusoidalReferenceGenerator, SwitchedReferenceGenerator, MultipleReferenceGenerator, SubepisodedReferenceGenerator, WienerProcessReferenceGenerator, TriangularReferenceGenerator, ConstReferenceGenerator
 from gym_electric_motor.visualization import MotorDashboard
 from gym_electric_motor import reference_generators as rg
-from gym_electric_motor.physical_systems.mechanical_loads import ConstantSpeedLoad
 from matplotlib import pyplot as plt
 import numpy as np
 
 
 if __name__ == '__main__':
 
+    ref = ['omega']
 
-    ref = 'omega'
+    if len(ref) == 1:
+        reference_generator = rg.SwitchedReferenceGenerator(
+            sub_generators=[
+                rg.SinusoidalReferenceGenerator(reference_state=ref[0], amplitude_range=(0, 0.3), offset_range=(0, 0.2)),
+                rg.WienerProcessReferenceGenerator(reference_state=ref[0], amplitude_range=(0, 0.3),
+                                                   offset_range=(0, 0.2)),
+                rg.StepReferenceGenerator(reference_state=ref[0], amplitude_range=(0, 0.3), offset_range=(0, 0.2))
+            ], p=[0.5, 0.25, 0.25], super_episode_length=(10000, 100000)
+        )
+    else:
+        ref_gen = []
+        for i in ref:
+            ref_gen.append(rg.SwitchedReferenceGenerator(
+                    sub_generators=[
+                    rg.SinusoidalReferenceGenerator(reference_state=i, amplitude_range=(0, 0.3), offset_range=(0, 0.2)),
+                    rg.WienerProcessReferenceGenerator(reference_state=i, amplitude_range=(0, 0.3),
+                                                       offset_range=(0, 0.2)),
+                    rg.StepReferenceGenerator(reference_state=i, amplitude_range=(0, 0.3), offset_range=(0, 0.2))
+                    ], p=[0.5, 0.25, 0.25], super_episode_length=(10000, 100000)
+                )
+            )
+        reference_generator = MultipleReferenceGenerator(ref_gen)
+
+    limit_values = dict(omega=100*np.pi/30)
+
     env = gem.make(
         #'DcExtExDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u_a', 'u_e'], visu_period=1),
         #'DcPermExDisc-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i', 'u'], visu_period=1),
@@ -26,20 +51,14 @@ if __name__ == '__main__':
         #'DcShuntCont-v1', visualization=MotorDashboard(plots=['omega', 'torque', 'i_a', 'i_e', 'u'], visu_period=1),
 
         ode_solver='scipy.solve_ivp', solver_kwargs=dict(),
+        #limit_values=limit_values,
         #load=ConstantSpeedLoad(omega_fixed=50 * np.pi / 30),
-        reference_generator=rg.SwitchedReferenceGenerator(
-            sub_generators=[
[…]
