### 2021-01-25 acb2297
update readme with new bibtex citation
