### 2021-01-25 #2498 @wkirgsn
The error still persisted so I discarded the draft changes and retyped everything again (copy&pasted).
Had to save as another draft, and then the publishing worked like a charm.

@Kevin-Mattheus-Moerman it is finally done - sorry for the delay.
