### 2021-01-27 #2498 @wkirgsn
@Kevin-Mattheus-Moerman This article can be processed now.
Zenodo badge:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4355691.svg)](https://doi.org/10.5281/zenodo.4355691)
with version v3.1.0
