+++ gym_electric_motor/envs/gym_dcm/__init__.py
@@ -4,9 +4,13 @@ from .dc_permex_motor_env import ContTorqueControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import DiscTorqueControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import ContCurrentControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import DiscCurrentControlDcPermanentlyExcitedMotorEnv
+from .dc_extex_motor_env import ContSpeedControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import DiscSpeedControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import ContTorqueControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import DiscTorqueControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import ContCurrentControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import DiscCurrentControlDcExternallyExcitedMotorEnv
 from .dc_shunt_motor_env import ContDcShuntMotorEnvironment
 from .dc_shunt_motor_env import DiscDcShuntMotorEnvironment
 from .dc_series_motor_env import ContDcSeriesMotorEnvironment
 from .dc_series_motor_env import DiscDcSeriesMotorEnvironment
-from .dc_extex_motor_env import ContDcExternallyExcitedMotorEnvironment
-from .dc_extex_motor_env import DiscDcExternallyExcitedMotorEnvironment
