+++ gym_electric_motor/envs/__init__.py
@@ -6,9 +6,12 @@ from .gym_dcm.dc_permex_motor_env import DiscTorqueControlDcPermanentlyExcitedMo
 from .gym_dcm.dc_permex_motor_env import ContCurrentControlDcPermanentlyExcitedMotorEnv
 from .gym_dcm.dc_permex_motor_env import DiscCurrentControlDcPermanentlyExcitedMotorEnv
 
-from .gym_dcm.dc_extex_motor_env import ContDcExternallyExcitedMotorEnvironment
-from .gym_dcm.dc_extex_motor_env import DiscDcExternallyExcitedMotorEnvironment
-
+from .gym_dcm.dc_extex_motor_env import ContSpeedControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import DiscSpeedControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import ContTorqueControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import DiscTorqueControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import ContCurrentControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import DiscCurrentControlDcExternallyExcitedMotorEnv
 from .gym_dcm.dc_series_motor_env import ContDcSeriesMotorEnvironment
 from .gym_dcm.dc_series_motor_env import DiscDcSeriesMotorEnvironment
 
