+++ gym_electric_motor/__init__.py
@@ -37,10 +37,19 @@ register(id='Disc-CC-PermExDc-v0',
 register(id='Cont-CC-PermExDc-v0',
          entry_point=envs_path+'ContCurrentControlDcPermanentlyExcitedMotorEnv')
 
-register(id='DcExtExCont-v1',
-         entry_point=envs_path+'ContDcExternallyExcitedMotorEnvironment')
-register(id='DcExtExDisc-v1',
-         entry_point=envs_path+'DiscDcExternallyExcitedMotorEnvironment')
+register(id='Disc-SC-ExtExDc-v0',
+         entry_point=envs_path+'DiscSpeedControlDcExternallyExcitedMotorEnv')
+register(id='Cont-SC-ExtExDc-v0',
+         entry_point=envs_path+'ContSpeedControlDcExternallyExcitedMotorEnv')
+register(id='Disc-TC-ExtExDc-v0',
+         entry_point=envs_path+'DiscTorqueControlDcExternallyExcitedMotorEnv')
+register(id='Cont-TC-ExtExDc-v0',
+         entry_point=envs_path+'ContTorqueControlDcExternallyExcitedMotorEnv')
+register(id='Disc-CC-ExtExDc-v0',
+         entry_point=envs_path+'DiscCurrentControlDcExternallyExcitedMotorEnv')
+register(id='Cont-CC-ExtExDc-v0',
+         entry_point=envs_path+'ContCurrentControlDcExternallyExcitedMotorEnv')
+
 register(id='DcSeriesCont-v1',
          entry_point=envs_path+'ContDcSeriesMotorEnvironment')
 register(id='DcSeriesDisc-v1',
