+++ examples/classic_controllers/classic_controllers.py
@@ -1,27 +1,28 @@
 from gym.spaces import Discrete, Box, MultiDiscrete
 from gym_electric_motor.physical_systems import SynchronousMotorSystem, DcMotorSystem, DcSeriesMotor, DcExternallyExcitedMotor
 from gym_electric_motor.reference_generators import MultipleReferenceGenerator, SwitchedReferenceGenerator
-from gym_electric_motor.envs import ContDcExternallyExcitedMotorEnvironment
+from gym_electric_motor.envs import ContDcExternallyExcitedMotorEnvironment, DiscDcExternallyExcitedMotorEnvironment
 import numpy as np
 
 
 class Controller:
 
     @classmethod
-    def make(cls, environment, controller_type=None, **controller_kwargs):
+    def make(cls, environment, stages=None, **controller_kwargs):
 
         controller_kwargs = cls.reference_states(cls, environment, **controller_kwargs)
 
-        if controller_type != None:
+        if stages != None:
+            controller_type, stages = cls.find_controller_type(cls, environment, stages)
             assert controller_type in _controllers.keys(), f'Controller {controller_type} unknown'
-            controller_kwargs = cls.automated_gain(cls, controller_type, environment, **controller_kwargs)
-            controller = _controllers[controller_type][0](controller_type, environment, **controller_kwargs)
+            stages = cls.automated_gain(cls, environment, stages, controller_type, **controller_kwargs)
+            controller = _controllers[controller_type][0](environment, stages, **controller_kwargs)
 
         else:
-            controller_type, controller_kwargs = cls.automated_controller_design(cls, environment, controller_type,
-                                                                                 **controller_kwargs)
-            controller_kwargs = cls.automated_gain(cls, controller_type, environment, **controller_kwargs)
-            controller = _controllers[controller_type][0](controller_type, environment,  **controller_kwargs)
+            controller_type, stages = cls.automated_controller_design(cls, environment, **controller_kwargs)
+            stages = cls.automated_gain(cls, environment, stages, controller_type, **controller_kwargs)
+
+            controller = _controllers[controller_type][0](environment, stages, **controller_kwargs)
 
         return controller
 
@@ -44,15 +45,41 @@ class Controller:
         controller_kwargs['ref_states'] = np.array(ref_states)
         return controller_kwargs
 
+    def find_controller_type(self, environment, stages):
+        _stages = stages
+        print(stages)
+        if type(environment.physical_system) == DcMotorSystem:
+            if type(stages) == list:
+                if len(stages) > 1:
+                    if type(stages[0]) == list:
+                        stages = stages[0]
+                    if len(stages) > 1:
+                        controller_type = 'cascaded_controller'
+                    else:
+                        controller_type = stages['controller_type']
+                else:
+                    controller_type = stages[0]['controller_type']
+            else:
+                if type(stages) == dict:
+                    controller_type = stages['controller_type']
+                    _stages = [stages]
+                else:
+                    controller_type = stages
+                    _stages = [{'controller_type': stages}]
+        elif type(environment.physical_system) == SynchronousMotorSystem:
+            controller_type = 'foc_controller'
 
-    def automated_controller_design(self, environment, controller_type, **controller_kwargs):
+        return controller_type, _stages
+
+    def automated_controller_design(self, environment, **controller_kwargs):
         action_space = type(environment.action_space)
         ref_states = controller_kwargs['ref_states']
+        stages = []
         if type(environment.physical_system) == DcMotorSystem:
 
[…]
