+++ gym_electric_motor/envs/gym_dcm/dc_series_motor_env.py
@@ -1,138 +1,690 @@
-from ...core import ElectricMotorEnvironment
-from ...physical_systems.physical_systems import DcMotorSystem
-from ...reference_generators.wiener_process_reference_generator import WienerProcessReferenceGenerator
-from ...reward_functions import WeightedSumOfErrors
+from gym_electric_motor.core import ElectricMotorEnvironment, ReferenceGenerator, RewardFunction, \
+    ElectricMotorVisualization
+from gym_electric_motor.physical_systems.physical_systems import DcMotorSystem
+from gym_electric_motor.visualization import MotorDashboard
+from gym_electric_motor.reference_generators.wiener_process_reference_generator import WienerProcessReferenceGenerator
+from gym_electric_motor import physical_systems as ps
+from gym_electric_motor.reward_functions import WeightedSumOfErrors
+from gym_electric_motor.utils import initialize
 
 
-class DcSeriesMotorEnvironment(ElectricMotorEnvironment):
+class DiscSpeedControlDcSeriesMotorEnv(ElectricMotorEnvironment):
+    """
+        Description:
+            Environment to simulate a discretely speed controlled series DC Motor
 
-    def __init__(self, motor='DcSeries', reward_function=None, reference_generator=None, constraints=('i',), **kwargs):
+        Key:
+            `Disc-SC-SeriesDc-v0`
+
+        Default Components:
+            Supply: IdealVoltageSupply
+            Converter: DiscOneQuadrantConverter
+            Motor: DcSeriesMotor
+            Load: PolynomialStaticLoad
+            Ode-Solver: EulerSolver
+            Noise: None
+
+            Reference Generator:
+                WienerProcessReferenceGenerator
+                    Reference Quantity. 'omega'
+
+            Reward Function:
+                WeightedSumOfErrors: reward_weights 'omega' = 1
+
+            Visualization:
+                MotorDashboard: omega and action plots
+
+            Constraints:
+                Current Limit on 'i'
+
+        Control Cycle Time:
+            tau = 1e-5 seconds
+
+        State Variables:
+            ``['omega' , 'torque', 'i', 'u', 'u_sup']``
+
+        Observation Space:
+            Type: Tuple(State_Space, Reference_Space)
+
+        State Space:
+            Box(low=[-1, -1, -1, -1, 0], high=[1, 1, 1, 1, 1])
+
+        Reference Space:
+            Box(low=[-1], high=[1])
+
+        Action Space:
+            Type: Discrete(2)
+
+        Starting State:
+            Zeros on all state variables.
+
+        Episode Termination:
+            Termination if current limits are violated.
+        """
+    def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None, noise_generator=None,
+                 reward_function=None, reference_generator=None, visualization=None, state_filter=None, callbacks=(),
+                 constraints=('i',), calc_jacobian=True, tau=1e-5):
         """
         Args:
-            motor(ElectricMotor): Electric Motor used in the PhysicalSystem
-            reward_function(RewardFunction): Reward Function for the environment
-            reference_generator(ReferenceGenerator): Reference Generator for the environment
-            kwargs(dict): Further kwargs to pass to the superclass and the submodules
+            supply(env-arg): Specification of the supply to be used in the environment
+            converter(env-arg): Specification of the converter to be used in the environment
+            motor(env-arg): Specification of the motor to be used in the environment
+            load(env-arg): Specification of the load to be used in the environment
+            ode_solver(env-arg): Specification of the ode_solver to be used in the environment
+            noise_generator(env-arg): Specification of the noise_generator to be used in the environment
+            reward_function(env-arg: Specification of the reward_function to be used in the environment
[…]
