+++ gym_electric_motor/envs/__init__.py
@@ -12,11 +12,20 @@ from .gym_dcm.dc_extex_motor_env import ContTorqueControlDcExternallyExcitedMoto
 from .gym_dcm.dc_extex_motor_env import DiscTorqueControlDcExternallyExcitedMotorEnv
 from .gym_dcm.dc_extex_motor_env import ContCurrentControlDcExternallyExcitedMotorEnv
 from .gym_dcm.dc_extex_motor_env import DiscCurrentControlDcExternallyExcitedMotorEnv
-from .gym_dcm.dc_series_motor_env import ContDcSeriesMotorEnvironment
-from .gym_dcm.dc_series_motor_env import DiscDcSeriesMotorEnvironment
 
-from .gym_dcm.dc_shunt_motor_env import ContDcShuntMotorEnvironment
-from .gym_dcm.dc_shunt_motor_env import DiscDcShuntMotorEnvironment
+from .gym_dcm.dc_series_motor_env import ContSpeedControlDcSeriesMotorEnv
+from .gym_dcm.dc_series_motor_env import DiscSpeedControlDcSeriesMotorEnv
+from .gym_dcm.dc_series_motor_env import ContTorqueControlDcSeriesMotorEnv
+from .gym_dcm.dc_series_motor_env import DiscTorqueControlDcSeriesMotorEnv
+from .gym_dcm.dc_series_motor_env import ContCurrentControlDcSeriesMotorEnv
+from .gym_dcm.dc_series_motor_env import DiscCurrentControlDcSeriesMotorEnv
+
+from .gym_dcm.dc_shunt_motor_env import ContSpeedControlDcShuntMotorEnv
+from .gym_dcm.dc_shunt_motor_env import DiscSpeedControlDcShuntMotorEnv
+from .gym_dcm.dc_shunt_motor_env import ContTorqueControlDcShuntMotorEnv
+from .gym_dcm.dc_shunt_motor_env import DiscTorqueControlDcShuntMotorEnv
+from .gym_dcm.dc_shunt_motor_env import ContCurrentControlDcShuntMotorEnv
+from .gym_dcm.dc_shunt_motor_env import DiscCurrentControlDcShuntMotorEnv
 
 from .gym_pmsm.perm_mag_syn_motor_env import ContPermanentMagnetSynchronousMotorEnvironment
 from .gym_pmsm.perm_mag_syn_motor_env import DiscPermanentMagnetSynchronousMotorEnvironment
