+++ gym_electric_motor/envs/gym_dcm/__init__.py
@@ -4,13 +4,24 @@ from .dc_permex_motor_env import ContTorqueControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import DiscTorqueControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import ContCurrentControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import DiscCurrentControlDcPermanentlyExcitedMotorEnv
+
 from .dc_extex_motor_env import ContSpeedControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import DiscSpeedControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import ContTorqueControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import DiscTorqueControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import ContCurrentControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import DiscCurrentControlDcExternallyExcitedMotorEnv
-from .dc_shunt_motor_env import ContDcShuntMotorEnvironment
-from .dc_shunt_motor_env import DiscDcShuntMotorEnvironment
-from .dc_series_motor_env import ContDcSeriesMotorEnvironment
-from .dc_series_motor_env import DiscDcSeriesMotorEnvironment
+
+from .dc_series_motor_env import ContSpeedControlDcSeriesMotorEnv
+from .dc_series_motor_env import DiscSpeedControlDcSeriesMotorEnv
+from .dc_series_motor_env import ContTorqueControlDcSeriesMotorEnv
+from .dc_series_motor_env import DiscTorqueControlDcSeriesMotorEnv
+from .dc_series_motor_env import ContCurrentControlDcSeriesMotorEnv
+from .dc_series_motor_env import DiscCurrentControlDcSeriesMotorEnv
+
+from .dc_shunt_motor_env import ContSpeedControlDcShuntMotorEnv
+from .dc_shunt_motor_env import DiscSpeedControlDcShuntMotorEnv
+from .dc_shunt_motor_env import ContTorqueControlDcShuntMotorEnv
+from .dc_shunt_motor_env import DiscTorqueControlDcShuntMotorEnv
+from .dc_shunt_motor_env import ContCurrentControlDcShuntMotorEnv
+from .dc_shunt_motor_env import DiscCurrentControlDcShuntMotorEnv

[…]
