### 2021-02-03 6b5ca4f
Changed 'Disc' to 'Finite' in PermEx Envs and Converters
