+++ gym_electric_motor/envs/gym_im/doubly_fed_induc_motor_env.py
@@ -35,7 +35,7 @@ class DiscDoublyFedInductionMotorEnvironment(DoublyFedInductionMotorEnvironment)
         Physical System:
             SCMLSystem/DcMotorSystem with:
                 | IdealVoltageSupply
-                | DiscB6BridgeConverter
+                | FiniteB6BridgeConverter
                 | PermanentMagnetSynchronousMotor
                 | PolynomialStaticLoad
                 | GaussianWhiteNoiseGenerator
