+++ gym_electric_motor/envs/gym_dcm/__init__.py
@@ -1,9 +1,9 @@
 from .dc_permex_motor_env import ContSpeedControlDcPermanentlyExcitedMotorEnv
-from .dc_permex_motor_env import DiscSpeedControlDcPermanentlyExcitedMotorEnv
+from .dc_permex_motor_env import FiniteSpeedControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import ContTorqueControlDcPermanentlyExcitedMotorEnv
-from .dc_permex_motor_env import DiscTorqueControlDcPermanentlyExcitedMotorEnv
+from .dc_permex_motor_env import FiniteTorqueControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import ContCurrentControlDcPermanentlyExcitedMotorEnv
-from .dc_permex_motor_env import DiscCurrentControlDcPermanentlyExcitedMotorEnv
+from .dc_permex_motor_env import FiniteCurrentControlDcPermanentlyExcitedMotorEnv
 from .dc_shunt_motor_env import ContDcShuntMotorEnvironment
 from .dc_shunt_motor_env import DiscDcShuntMotorEnvironment
 from .dc_series_motor_env import ContDcSeriesMotorEnvironment
