+++ gym_electric_motor/envs/__init__.py
@@ -7,11 +7,11 @@ from .gym_dcm.dc_permex_motor_env import ContCurrentControlDcPermanentlyExcitedM
 from .gym_dcm.dc_permex_motor_env import FiniteCurrentControlDcPermanentlyExcitedMotorEnv
 
 from .gym_dcm.dc_extex_motor_env import ContSpeedControlDcExternallyExcitedMotorEnv
-from .gym_dcm.dc_extex_motor_env import DiscSpeedControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import FiniteSpeedControlDcExternallyExcitedMotorEnv
 from .gym_dcm.dc_extex_motor_env import ContTorqueControlDcExternallyExcitedMotorEnv
-from .gym_dcm.dc_extex_motor_env import DiscTorqueControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import FiniteTorqueControlDcExternallyExcitedMotorEnv
 from .gym_dcm.dc_extex_motor_env import ContCurrentControlDcExternallyExcitedMotorEnv
-from .gym_dcm.dc_extex_motor_env import DiscCurrentControlDcExternallyExcitedMotorEnv
+from .gym_dcm.dc_extex_motor_env import FiniteCurrentControlDcExternallyExcitedMotorEnv
 
 from .gym_dcm.dc_series_motor_env import ContSpeedControlDcSeriesMotorEnv
 from .gym_dcm.dc_series_motor_env import DiscSpeedControlDcSeriesMotorEnv
