+++ gym_electric_motor/envs/gym_dcm/__init__.py
@@ -1,16 +1,16 @@
 from .dc_permex_motor_env import ContSpeedControlDcPermanentlyExcitedMotorEnv
-from .dc_permex_motor_env import DiscSpeedControlDcPermanentlyExcitedMotorEnv
+from .dc_permex_motor_env import FiniteSpeedControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import ContTorqueControlDcPermanentlyExcitedMotorEnv
-from .dc_permex_motor_env import DiscTorqueControlDcPermanentlyExcitedMotorEnv
+from .dc_permex_motor_env import FiniteTorqueControlDcPermanentlyExcitedMotorEnv
 from .dc_permex_motor_env import ContCurrentControlDcPermanentlyExcitedMotorEnv
-from .dc_permex_motor_env import DiscCurrentControlDcPermanentlyExcitedMotorEnv
+from .dc_permex_motor_env import FiniteCurrentControlDcPermanentlyExcitedMotorEnv
 
 from .dc_extex_motor_env import ContSpeedControlDcExternallyExcitedMotorEnv
-from .dc_extex_motor_env import DiscSpeedControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import FiniteSpeedControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import ContTorqueControlDcExternallyExcitedMotorEnv
-from .dc_extex_motor_env import DiscTorqueControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import FiniteTorqueControlDcExternallyExcitedMotorEnv
 from .dc_extex_motor_env import ContCurrentControlDcExternallyExcitedMotorEnv
-from .dc_extex_motor_env import DiscCurrentControlDcExternallyExcitedMotorEnv
+from .dc_extex_motor_env import FiniteCurrentControlDcExternallyExcitedMotorEnv
 
 from .dc_series_motor_env import ContSpeedControlDcSeriesMotorEnv
 from .dc_series_motor_env import DiscSpeedControlDcSeriesMotorEnv
