### 2021-02-05 #2498 @kyleniemeyer
Hello @wkirgsn, I am the EIC on duty this week doing final checks before publishing—I can see that this submission is at that stage.

I just have some minor fixes for you:
- Could you ensure that "e.g." is followed by a comma ("e.g.,") throughout?
- It looks like many of your citations are duplicated. My guess is that your `.bib` file has duplicates with different cite keys, and both are being used. Could you consolidate all of those into a single citation each?

Also, after your article was originally submitted, JOSS made [explicit](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) our requirement for a Statement of Need section that clearly illustrates the research purpose of the software.

Since your article came in before that requirement was explicit, I'm going to leave it up to you if you would like to add one—but, you could likely just repurpose some text that you already have into this short section.

Thanks!
