### 2021-02-06 52ba435
final updates in JOSS paper before publishing
