### 2021-02-06 8ddfad1
add reference to rl lecture material in JOSS paper
