+++ docs/JOSS/Literature.bib
@@ -22,6 +22,13 @@
   url    = {https://upb-lea.github.io/gym-electric-motor/},
 }
 
+@Misc{rl-lecture,
+  author = {Wilhelm Kirchg\"assner and Maximilian Schenke and Oliver Wallscheid and Daniel Weber},
+  title = {Paderborn University: Reinforcement learning Course Materials},
+  year = {2020},
+  url = {https://github.com/upb-lea/reinforcement_learning_course_materials},
+}
+
 @Misc{gym-whitepaper,
   author = {Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
   title  = {{OpenAI Gym}},
