### 2021-02-06 #2498 @wkirgsn
Hi @kyleniemeyer, thank you for picking us up!
I just added the missing commas, removed the duplicate .bib entries, and renamed a subsection to be the statement of need, as it has fit perfectly for that already.
The JOSS paper should be in perfect shape now.
