### 2021-02-07 #2498 @arfon
@moorepants, @dineshresearch - many thanks for your reviews here and to @Kevin-Mattheus-Moerman for editing this submission. JOSS relies upon the volunteer efforts of folks like yourselves and we simply wouldn't be able to do this without you! ✨
