### 2021-02-17 #2498 @lnielsen
> / cc @lnielsen any chance you could point us in the right direction? ☝️

Sorry, just seeing this now :-( The fastest way to get a reply is likely https://zenodo.org/support or alternatively ping @slint who's in charge of service operation.
